@@ -1,345 +1,400 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: DeltaNeutralContract.proto
+// Protobuf C++ Version: 5.29.3
 
-#ifndef GOOGLE_PROTOBUF_INCLUDED_DeltaNeutralContract_2eproto
-#define GOOGLE_PROTOBUF_INCLUDED_DeltaNeutralContract_2eproto
+#ifndef DeltaNeutralContract_2eproto_2epb_2eh
+#define DeltaNeutralContract_2eproto_2epb_2eh
 
 #include <limits>
 #include <string>
+#include <type_traits>
+#include <utility>
 
-#include <google/protobuf/port_def.inc>
-#if PROTOBUF_VERSION < 3012000
-#error This file was generated by a newer version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please update
-#error your headers.
+#include "google/protobuf/runtime_version.h"
+#if PROTOBUF_VERSION != 5029003
+#error "Protobuf C++ gencode is built with an incompatible version of"
+#error "Protobuf C++ headers/runtime. See"
+#error "https://protobuf.dev/support/cross-version-runtime-guarantee/#cpp"
 #endif
-#if 3012004 < PROTOBUF_MIN_PROTOC_VERSION
-#error This file was generated by an older version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please
-#error regenerate this file with a newer version of protoc.
-#endif
-
-#include <google/protobuf/port_undef.inc>
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/arena.h>
-#include <google/protobuf/arenastring.h>
-#include <google/protobuf/generated_message_table_driven.h>
-#include <google/protobuf/generated_message_util.h>
-#include <google/protobuf/inlined_string_field.h>
-#include <google/protobuf/metadata_lite.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/message.h>
-#include <google/protobuf/repeated_field.h>  // IWYU pragma: export
-#include <google/protobuf/extension_set.h>  // IWYU pragma: export
-#include <google/protobuf/unknown_field_set.h>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/arena.h"
+#include "google/protobuf/arenastring.h"
+#include "google/protobuf/generated_message_tctable_decl.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/metadata_lite.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/message.h"
+#include "google/protobuf/message_lite.h"
+#include "google/protobuf/repeated_field.h"  // IWYU pragma: export
+#include "google/protobuf/extension_set.h"  // IWYU pragma: export
+#include "google/protobuf/unknown_field_set.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
+
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+
 #define PROTOBUF_INTERNAL_EXPORT_DeltaNeutralContract_2eproto
-PROTOBUF_NAMESPACE_OPEN
+
+namespace google {
+namespace protobuf {
 namespace internal {
-class AnyMetadata;
+template <typename T>
+::absl::string_view GetAnyMessageName();
 }  // namespace internal
-PROTOBUF_NAMESPACE_CLOSE
+}  // namespace protobuf
+}  // namespace google
 
 // Internal implementation detail -- do not use these members.
 struct TableStruct_DeltaNeutralContract_2eproto {
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTableField entries[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxillaryParseTableField aux[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTable schema[1]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::FieldMetadata field_metadata[];
-  static const ::PROTOBUF_NAMESPACE_ID::internal::SerializationTable serialization_table[];
-  static const ::PROTOBUF_NAMESPACE_ID::uint32 offsets[];
+  static const ::uint32_t offsets[];
 };
-extern const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_DeltaNeutralContract_2eproto;
+extern const ::google::protobuf::internal::DescriptorTable
+    descriptor_table_DeltaNeutralContract_2eproto;
 namespace protobuf {
 class DeltaNeutralContract;
-class DeltaNeutralContractDefaultTypeInternal;
+struct DeltaNeutralContractDefaultTypeInternal;
 extern DeltaNeutralContractDefaultTypeInternal _DeltaNeutralContract_default_instance_;
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> ::protobuf::DeltaNeutralContract* Arena::CreateMaybeMessage<::protobuf::DeltaNeutralContract>(Arena*);
-PROTOBUF_NAMESPACE_CLOSE
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
+
 namespace protobuf {
 
 // ===================================================================
 
-class DeltaNeutralContract PROTOBUF_FINAL :
-    public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:protobuf.DeltaNeutralContract) */ {
+
+// -------------------------------------------------------------------
+
+class DeltaNeutralContract final : public ::google::protobuf::Message
+/* @@protoc_insertion_point(class_definition:protobuf.DeltaNeutralContract) */ {
  public:
-  inline DeltaNeutralContract() : DeltaNeutralContract(nullptr) {};
-  virtual ~DeltaNeutralContract();
+  inline DeltaNeutralContract() : DeltaNeutralContract(nullptr) {}
+  ~DeltaNeutralContract() PROTOBUF_FINAL;
 
-  DeltaNeutralContract(const DeltaNeutralContract& from);
-  DeltaNeutralContract(DeltaNeutralContract&& from) noexcept
-    : DeltaNeutralContract() {
-    *this = ::std::move(from);
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+  void operator delete(DeltaNeutralContract* msg, std::destroying_delete_t) {
+    SharedDtor(*msg);
+    ::google::protobuf::internal::SizedDelete(msg, sizeof(DeltaNeutralContract));
   }
+#endif
 
+  template <typename = void>
+  explicit PROTOBUF_CONSTEXPR DeltaNeutralContract(
+      ::google::protobuf::internal::ConstantInitialized);
+
+  inline DeltaNeutralContract(const DeltaNeutralContract& from) : DeltaNeutralContract(nullptr, from) {}
+  inline DeltaNeutralContract(DeltaNeutralContract&& from) noexcept
+      : DeltaNeutralContract(nullptr, std::move(from)) {}
   inline DeltaNeutralContract& operator=(const DeltaNeutralContract& from) {
     CopyFrom(from);
     return *this;
   }
   inline DeltaNeutralContract& operator=(DeltaNeutralContract&& from) noexcept {
-    if (GetArena() == from.GetArena()) {
-      if (this != &from) InternalSwap(&from);
+    if (this == &from) return *this;
+    if (::google::protobuf::internal::CanMoveWithInternalSwap(GetArena(), from.GetArena())) {
+      InternalSwap(&from);
     } else {
       CopyFrom(from);
     }
     return *this;
   }
 
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* descriptor() {
+  inline const ::google::protobuf::UnknownFieldSet& unknown_fields() const
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance);
+  }
+  inline ::google::protobuf::UnknownFieldSet* mutable_unknown_fields()
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.mutable_unknown_fields<::google::protobuf::UnknownFieldSet>();
+  }
+
+  static const ::google::protobuf::Descriptor* descriptor() {
     return GetDescriptor();
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* GetDescriptor() {
-    return GetMetadataStatic().descriptor;
+  static const ::google::protobuf::Descriptor* GetDescriptor() {
+    return default_instance().GetMetadata().descriptor;
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Reflection* GetReflection() {
-    return GetMetadataStatic().reflection;
+  static const ::google::protobuf::Reflection* GetReflection() {
+    return default_instance().GetMetadata().reflection;
   }
-  static const DeltaNeutralContract& default_instance();
-
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
+  static const DeltaNeutralContract& default_instance() {
+    return *internal_default_instance();
+  }
   static inline const DeltaNeutralContract* internal_default_instance() {
     return reinterpret_cast<const DeltaNeutralContract*>(
-               &_DeltaNeutralContract_default_instance_);
+        &_DeltaNeutralContract_default_instance_);
   }
-  static constexpr int kIndexInFileMessages =
-    0;
-
-  friend void swap(DeltaNeutralContract& a, DeltaNeutralContract& b) {
-    a.Swap(&b);
-  }
+  static constexpr int kIndexInFileMessages = 0;
+  friend void swap(DeltaNeutralContract& a, DeltaNeutralContract& b) { a.Swap(&b); }
   inline void Swap(DeltaNeutralContract* other) {
     if (other == this) return;
-    if (GetArena() == other->GetArena()) {
+    if (::google::protobuf::internal::CanUseInternalSwap(GetArena(), other->GetArena())) {
       InternalSwap(other);
     } else {
-      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+      ::google::protobuf::internal::GenericSwap(this, other);
     }
   }
   void UnsafeArenaSwap(DeltaNeutralContract* other) {
     if (other == this) return;
-    GOOGLE_DCHECK(GetArena() == other->GetArena());
+    ABSL_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
   // implements Message ----------------------------------------------
 
-  inline DeltaNeutralContract* New() const final {
-    return CreateMaybeMessage<DeltaNeutralContract>(nullptr);
+  DeltaNeutralContract* New(::google::protobuf::Arena* arena = nullptr) const {
+    return ::google::protobuf::Message::DefaultConstruct<DeltaNeutralContract>(arena);
   }
-
-  DeltaNeutralContract* New(::PROTOBUF_NAMESPACE_ID::Arena* arena) const final {
-    return CreateMaybeMessage<DeltaNeutralContract>(arena);
-  }
-  void CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
-  void MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
+  using ::google::protobuf::Message::CopyFrom;
   void CopyFrom(const DeltaNeutralContract& from);
-  void MergeFrom(const DeltaNeutralContract& from);
-  PROTOBUF_ATTRIBUTE_REINITIALIZES void Clear() final;
-  bool IsInitialized() const final;
-
-  size_t ByteSizeLong() const final;
-  const char* _InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) final;
-  ::PROTOBUF_NAMESPACE_ID::uint8* _InternalSerialize(
-      ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const final;
-  int GetCachedSize() const final { return _cached_size_.Get(); }
+  using ::google::protobuf::Message::MergeFrom;
+  void MergeFrom(const DeltaNeutralContract& from) { DeltaNeutralContract::MergeImpl(*this, from); }
 
   private:
-  inline void SharedCtor();
-  inline void SharedDtor();
-  void SetCachedSize(int size) const final;
+  static void MergeImpl(
+      ::google::protobuf::MessageLite& to_msg,
+      const ::google::protobuf::MessageLite& from_msg);
+
+  public:
+  bool IsInitialized() const {
+    return true;
+  }
+  ABSL_ATTRIBUTE_REINITIALIZES void Clear() PROTOBUF_FINAL;
+  #if defined(PROTOBUF_CUSTOM_VTABLE)
+  private:
+  static ::size_t ByteSizeLong(const ::google::protobuf::MessageLite& msg);
+  static ::uint8_t* _InternalSerialize(
+      const MessageLite& msg, ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream);
+
+  public:
+  ::size_t ByteSizeLong() const { return ByteSizeLong(*this); }
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+    return _InternalSerialize(*this, target, stream);
+  }
+  #else   // PROTOBUF_CUSTOM_VTABLE
+  ::size_t ByteSizeLong() const final;
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const final;
+  #endif  // PROTOBUF_CUSTOM_VTABLE
+  int GetCachedSize() const { return _impl_._cached_size_.Get(); }
+
+  private:
+  void SharedCtor(::google::protobuf::Arena* arena);
+  static void SharedDtor(MessageLite& self);
   void InternalSwap(DeltaNeutralContract* other);
-  friend class ::PROTOBUF_NAMESPACE_ID::internal::AnyMetadata;
-  static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
-    return "protobuf.DeltaNeutralContract";
+ private:
+  template <typename T>
+  friend ::absl::string_view(
+      ::google::protobuf::internal::GetAnyMessageName)();
+  static ::absl::string_view FullMessageName() { return "protobuf.DeltaNeutralContract"; }
+
+ protected:
+  explicit DeltaNeutralContract(::google::protobuf::Arena* arena);
+  DeltaNeutralContract(::google::protobuf::Arena* arena, const DeltaNeutralContract& from);
+  DeltaNeutralContract(::google::protobuf::Arena* arena, DeltaNeutralContract&& from) noexcept
+      : DeltaNeutralContract(arena) {
+    *this = ::std::move(from);
   }
-  protected:
-  explicit DeltaNeutralContract(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  private:
-  static void ArenaDtor(void* object);
-  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  public:
+  const ::google::protobuf::internal::ClassData* GetClassData() const PROTOBUF_FINAL;
+  static void* PlacementNew_(const void*, void* mem,
+                             ::google::protobuf::Arena* arena);
+  static constexpr auto InternalNewImpl_();
+  static const ::google::protobuf::internal::ClassDataFull _class_data_;
 
-  ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
-  private:
-  static ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadataStatic() {
-    ::PROTOBUF_NAMESPACE_ID::internal::AssignDescriptors(&::descriptor_table_DeltaNeutralContract_2eproto);
-    return ::descriptor_table_DeltaNeutralContract_2eproto.file_level_metadata[kIndexInFileMessages];
-  }
-
-  public:
-
+ public:
+  ::google::protobuf::Metadata GetMetadata() const;
   // nested types ----------------------------------------------------
 
   // accessors -------------------------------------------------------
-
   enum : int {
     kDeltaFieldNumber = 2,
     kPriceFieldNumber = 3,
     kConIdFieldNumber = 1,
   };
-  // double delta = 2;
+  // optional double delta = 2;
   bool has_delta() const;
-  private:
-  bool _internal_has_delta() const;
-  public:
-  void clear_delta();
+  void clear_delta() ;
   double delta() const;
   void set_delta(double value);
+
   private:
   double _internal_delta() const;
   void _internal_set_delta(double value);
-  public:
 
-  // double price = 3;
-  bool has_price() const;
-  private:
-  bool _internal_has_price() const;
   public:
-  void clear_price();
+  // optional double price = 3;
+  bool has_price() const;
+  void clear_price() ;
   double price() const;
   void set_price(double value);
+
   private:
   double _internal_price() const;
   void _internal_set_price(double value);
-  public:
 
-  // int32 conId = 1;
+  public:
+  // optional int32 conId = 1;
   bool has_conid() const;
-  private:
-  bool _internal_has_conid() const;
-  public:
-  void clear_conid();
-  ::PROTOBUF_NAMESPACE_ID::int32 conid() const;
-  void set_conid(::PROTOBUF_NAMESPACE_ID::int32 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int32 _internal_conid() const;
-  void _internal_set_conid(::PROTOBUF_NAMESPACE_ID::int32 value);
-  public:
+  void clear_conid() ;
+  ::int32_t conid() const;
+  void set_conid(::int32_t value);
 
+  private:
+  ::int32_t _internal_conid() const;
+  void _internal_set_conid(::int32_t value);
+
+  public:
   // @@protoc_insertion_point(class_scope:protobuf.DeltaNeutralContract)
  private:
   class _Internal;
+  friend class ::google::protobuf::internal::TcParser;
+  static const ::google::protobuf::internal::TcParseTable<
+      2, 3, 0,
+      0, 2>
+      _table_;
 
-  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
-  typedef void InternalArenaConstructable_;
-  typedef void DestructorSkippable_;
-  ::PROTOBUF_NAMESPACE_ID::internal::HasBits<1> _has_bits_;
-  mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
-  double delta_;
-  double price_;
-  ::PROTOBUF_NAMESPACE_ID::int32 conid_;
+  friend class ::google::protobuf::MessageLite;
+  friend class ::google::protobuf::Arena;
+  template <typename T>
+  friend class ::google::protobuf::Arena::InternalHelper;
+  using InternalArenaConstructable_ = void;
+  using DestructorSkippable_ = void;
+  struct Impl_ {
+    inline explicit constexpr Impl_(
+        ::google::protobuf::internal::ConstantInitialized) noexcept;
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena);
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena, const Impl_& from,
+                          const DeltaNeutralContract& from_msg);
+    ::google::protobuf::internal::HasBits<1> _has_bits_;
+    ::google::protobuf::internal::CachedSize _cached_size_;
+    double delta_;
+    double price_;
+    ::int32_t conid_;
+    PROTOBUF_TSAN_DECLARE_MEMBER
+  };
+  union { Impl_ _impl_; };
   friend struct ::TableStruct_DeltaNeutralContract_2eproto;
 };
+
 // ===================================================================
 
 
+
+
 // ===================================================================
+
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic push
-  #pragma GCC diagnostic ignored "-Wstrict-aliasing"
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wstrict-aliasing"
 #endif  // __GNUC__
+// -------------------------------------------------------------------
+
 // DeltaNeutralContract
 
-// int32 conId = 1;
-inline bool DeltaNeutralContract::_internal_has_conid() const {
-  bool value = (_has_bits_[0] & 0x00000004u) != 0;
+// optional int32 conId = 1;
+inline bool DeltaNeutralContract::has_conid() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000004u) != 0;
   return value;
 }
-inline bool DeltaNeutralContract::has_conid() const {
-  return _internal_has_conid();
-}
 inline void DeltaNeutralContract::clear_conid() {
-  conid_ = 0;
-  _has_bits_[0] &= ~0x00000004u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.conid_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000004u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int32 DeltaNeutralContract::_internal_conid() const {
-  return conid_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int32 DeltaNeutralContract::conid() const {
+inline ::int32_t DeltaNeutralContract::conid() const {
   // @@protoc_insertion_point(field_get:protobuf.DeltaNeutralContract.conId)
   return _internal_conid();
 }
-inline void DeltaNeutralContract::_internal_set_conid(::PROTOBUF_NAMESPACE_ID::int32 value) {
-  _has_bits_[0] |= 0x00000004u;
-  conid_ = value;
-}
-inline void DeltaNeutralContract::set_conid(::PROTOBUF_NAMESPACE_ID::int32 value) {
+inline void DeltaNeutralContract::set_conid(::int32_t value) {
   _internal_set_conid(value);
+  _impl_._has_bits_[0] |= 0x00000004u;
   // @@protoc_insertion_point(field_set:protobuf.DeltaNeutralContract.conId)
 }
+inline ::int32_t DeltaNeutralContract::_internal_conid() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.conid_;
+}
+inline void DeltaNeutralContract::_internal_set_conid(::int32_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.conid_ = value;
+}
 
-// double delta = 2;
-inline bool DeltaNeutralContract::_internal_has_delta() const {
-  bool value = (_has_bits_[0] & 0x00000001u) != 0;
+// optional double delta = 2;
+inline bool DeltaNeutralContract::has_delta() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000001u) != 0;
   return value;
 }
-inline bool DeltaNeutralContract::has_delta() const {
-  return _internal_has_delta();
-}
 inline void DeltaNeutralContract::clear_delta() {
-  delta_ = 0;
-  _has_bits_[0] &= ~0x00000001u;
-}
-inline double DeltaNeutralContract::_internal_delta() const {
-  return delta_;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.delta_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000001u;
 }
 inline double DeltaNeutralContract::delta() const {
   // @@protoc_insertion_point(field_get:protobuf.DeltaNeutralContract.delta)
   return _internal_delta();
 }
-inline void DeltaNeutralContract::_internal_set_delta(double value) {
-  _has_bits_[0] |= 0x00000001u;
-  delta_ = value;
-}
 inline void DeltaNeutralContract::set_delta(double value) {
   _internal_set_delta(value);
+  _impl_._has_bits_[0] |= 0x00000001u;
   // @@protoc_insertion_point(field_set:protobuf.DeltaNeutralContract.delta)
 }
+inline double DeltaNeutralContract::_internal_delta() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.delta_;
+}
+inline void DeltaNeutralContract::_internal_set_delta(double value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.delta_ = value;
+}
 
-// double price = 3;
-inline bool DeltaNeutralContract::_internal_has_price() const {
-  bool value = (_has_bits_[0] & 0x00000002u) != 0;
+// optional double price = 3;
+inline bool DeltaNeutralContract::has_price() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000002u) != 0;
   return value;
 }
-inline bool DeltaNeutralContract::has_price() const {
-  return _internal_has_price();
-}
 inline void DeltaNeutralContract::clear_price() {
-  price_ = 0;
-  _has_bits_[0] &= ~0x00000002u;
-}
-inline double DeltaNeutralContract::_internal_price() const {
-  return price_;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.price_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000002u;
 }
 inline double DeltaNeutralContract::price() const {
   // @@protoc_insertion_point(field_get:protobuf.DeltaNeutralContract.price)
   return _internal_price();
 }
-inline void DeltaNeutralContract::_internal_set_price(double value) {
-  _has_bits_[0] |= 0x00000002u;
-  price_ = value;
-}
 inline void DeltaNeutralContract::set_price(double value) {
   _internal_set_price(value);
+  _impl_._has_bits_[0] |= 0x00000002u;
   // @@protoc_insertion_point(field_set:protobuf.DeltaNeutralContract.price)
+}
+inline double DeltaNeutralContract::_internal_price() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.price_;
+}
+inline void DeltaNeutralContract::_internal_set_price(double value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.price_ = value;
 }
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic pop
+#pragma GCC diagnostic pop
 #endif  // __GNUC__
 
 // @@protoc_insertion_point(namespace_scope)
-
 }  // namespace protobuf
+
 
 // @@protoc_insertion_point(global_scope)
 
-#include <google/protobuf/port_undef.inc>
-#endif  // GOOGLE_PROTOBUF_INCLUDED_GOOGLE_PROTOBUF_INCLUDED_DeltaNeutralContract_2eproto
+#include "google/protobuf/port_undef.inc"
+
+#endif  // DeltaNeutralContract_2eproto_2epb_2eh
