@@ -1,183 +1,218 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: ErrorMessage.proto
+// Protobuf C++ Version: 5.29.3
 
-#ifndef GOOGLE_PROTOBUF_INCLUDED_ErrorMessage_2eproto
-#define GOOGLE_PROTOBUF_INCLUDED_ErrorMessage_2eproto
+#ifndef ErrorMessage_2eproto_2epb_2eh
+#define ErrorMessage_2eproto_2epb_2eh
 
 #include <limits>
 #include <string>
+#include <type_traits>
+#include <utility>
 
-#include <google/protobuf/port_def.inc>
-#if PROTOBUF_VERSION < 3012000
-#error This file was generated by a newer version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please update
-#error your headers.
+#include "google/protobuf/runtime_version.h"
+#if PROTOBUF_VERSION != 5029003
+#error "Protobuf C++ gencode is built with an incompatible version of"
+#error "Protobuf C++ headers/runtime. See"
+#error "https://protobuf.dev/support/cross-version-runtime-guarantee/#cpp"
 #endif
-#if 3012004 < PROTOBUF_MIN_PROTOC_VERSION
-#error This file was generated by an older version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please
-#error regenerate this file with a newer version of protoc.
-#endif
-
-#include <google/protobuf/port_undef.inc>
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/arena.h>
-#include <google/protobuf/arenastring.h>
-#include <google/protobuf/generated_message_table_driven.h>
-#include <google/protobuf/generated_message_util.h>
-#include <google/protobuf/inlined_string_field.h>
-#include <google/protobuf/metadata_lite.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/message.h>
-#include <google/protobuf/repeated_field.h>  // IWYU pragma: export
-#include <google/protobuf/extension_set.h>  // IWYU pragma: export
-#include <google/protobuf/unknown_field_set.h>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/arena.h"
+#include "google/protobuf/arenastring.h"
+#include "google/protobuf/generated_message_tctable_decl.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/metadata_lite.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/message.h"
+#include "google/protobuf/message_lite.h"
+#include "google/protobuf/repeated_field.h"  // IWYU pragma: export
+#include "google/protobuf/extension_set.h"  // IWYU pragma: export
+#include "google/protobuf/unknown_field_set.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
+
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+
 #define PROTOBUF_INTERNAL_EXPORT_ErrorMessage_2eproto
-PROTOBUF_NAMESPACE_OPEN
+
+namespace google {
+namespace protobuf {
 namespace internal {
-class AnyMetadata;
+template <typename T>
+::absl::string_view GetAnyMessageName();
 }  // namespace internal
-PROTOBUF_NAMESPACE_CLOSE
+}  // namespace protobuf
+}  // namespace google
 
 // Internal implementation detail -- do not use these members.
 struct TableStruct_ErrorMessage_2eproto {
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTableField entries[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxillaryParseTableField aux[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTable schema[1]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::FieldMetadata field_metadata[];
-  static const ::PROTOBUF_NAMESPACE_ID::internal::SerializationTable serialization_table[];
-  static const ::PROTOBUF_NAMESPACE_ID::uint32 offsets[];
+  static const ::uint32_t offsets[];
 };
-extern const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_ErrorMessage_2eproto;
+extern const ::google::protobuf::internal::DescriptorTable
+    descriptor_table_ErrorMessage_2eproto;
 namespace protobuf {
 class ErrorMessage;
-class ErrorMessageDefaultTypeInternal;
+struct ErrorMessageDefaultTypeInternal;
 extern ErrorMessageDefaultTypeInternal _ErrorMessage_default_instance_;
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> ::protobuf::ErrorMessage* Arena::CreateMaybeMessage<::protobuf::ErrorMessage>(Arena*);
-PROTOBUF_NAMESPACE_CLOSE
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
+
 namespace protobuf {
 
 // ===================================================================
 
-class ErrorMessage PROTOBUF_FINAL :
-    public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:protobuf.ErrorMessage) */ {
+
+// -------------------------------------------------------------------
+
+class ErrorMessage final : public ::google::protobuf::Message
+/* @@protoc_insertion_point(class_definition:protobuf.ErrorMessage) */ {
  public:
-  inline ErrorMessage() : ErrorMessage(nullptr) {};
-  virtual ~ErrorMessage();
+  inline ErrorMessage() : ErrorMessage(nullptr) {}
+  ~ErrorMessage() PROTOBUF_FINAL;
 
-  ErrorMessage(const ErrorMessage& from);
-  ErrorMessage(ErrorMessage&& from) noexcept
-    : ErrorMessage() {
-    *this = ::std::move(from);
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+  void operator delete(ErrorMessage* msg, std::destroying_delete_t) {
+    SharedDtor(*msg);
+    ::google::protobuf::internal::SizedDelete(msg, sizeof(ErrorMessage));
   }
+#endif
 
+  template <typename = void>
+  explicit PROTOBUF_CONSTEXPR ErrorMessage(
+      ::google::protobuf::internal::ConstantInitialized);
+
+  inline ErrorMessage(const ErrorMessage& from) : ErrorMessage(nullptr, from) {}
+  inline ErrorMessage(ErrorMessage&& from) noexcept
+      : ErrorMessage(nullptr, std::move(from)) {}
   inline ErrorMessage& operator=(const ErrorMessage& from) {
     CopyFrom(from);
     return *this;
   }
   inline ErrorMessage& operator=(ErrorMessage&& from) noexcept {
-    if (GetArena() == from.GetArena()) {
-      if (this != &from) InternalSwap(&from);
+    if (this == &from) return *this;
+    if (::google::protobuf::internal::CanMoveWithInternalSwap(GetArena(), from.GetArena())) {
+      InternalSwap(&from);
     } else {
       CopyFrom(from);
     }
     return *this;
   }
 
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* descriptor() {
+  inline const ::google::protobuf::UnknownFieldSet& unknown_fields() const
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance);
+  }
+  inline ::google::protobuf::UnknownFieldSet* mutable_unknown_fields()
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.mutable_unknown_fields<::google::protobuf::UnknownFieldSet>();
+  }
+
+  static const ::google::protobuf::Descriptor* descriptor() {
     return GetDescriptor();
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* GetDescriptor() {
-    return GetMetadataStatic().descriptor;
+  static const ::google::protobuf::Descriptor* GetDescriptor() {
+    return default_instance().GetMetadata().descriptor;
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Reflection* GetReflection() {
-    return GetMetadataStatic().reflection;
+  static const ::google::protobuf::Reflection* GetReflection() {
+    return default_instance().GetMetadata().reflection;
   }
-  static const ErrorMessage& default_instance();
-
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
+  static const ErrorMessage& default_instance() {
+    return *internal_default_instance();
+  }
   static inline const ErrorMessage* internal_default_instance() {
     return reinterpret_cast<const ErrorMessage*>(
-               &_ErrorMessage_default_instance_);
+        &_ErrorMessage_default_instance_);
   }
-  static constexpr int kIndexInFileMessages =
-    0;
-
-  friend void swap(ErrorMessage& a, ErrorMessage& b) {
-    a.Swap(&b);
-  }
+  static constexpr int kIndexInFileMessages = 0;
+  friend void swap(ErrorMessage& a, ErrorMessage& b) { a.Swap(&b); }
   inline void Swap(ErrorMessage* other) {
     if (other == this) return;
-    if (GetArena() == other->GetArena()) {
+    if (::google::protobuf::internal::CanUseInternalSwap(GetArena(), other->GetArena())) {
       InternalSwap(other);
     } else {
-      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+      ::google::protobuf::internal::GenericSwap(this, other);
     }
   }
   void UnsafeArenaSwap(ErrorMessage* other) {
     if (other == this) return;
-    GOOGLE_DCHECK(GetArena() == other->GetArena());
+    ABSL_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
   // implements Message ----------------------------------------------
 
-  inline ErrorMessage* New() const final {
-    return CreateMaybeMessage<ErrorMessage>(nullptr);
+  ErrorMessage* New(::google::protobuf::Arena* arena = nullptr) const {
+    return ::google::protobuf::Message::DefaultConstruct<ErrorMessage>(arena);
   }
-
-  ErrorMessage* New(::PROTOBUF_NAMESPACE_ID::Arena* arena) const final {
-    return CreateMaybeMessage<ErrorMessage>(arena);
-  }
-  void CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
-  void MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
+  using ::google::protobuf::Message::CopyFrom;
   void CopyFrom(const ErrorMessage& from);
-  void MergeFrom(const ErrorMessage& from);
-  PROTOBUF_ATTRIBUTE_REINITIALIZES void Clear() final;
-  bool IsInitialized() const final;
-
-  size_t ByteSizeLong() const final;
-  const char* _InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) final;
-  ::PROTOBUF_NAMESPACE_ID::uint8* _InternalSerialize(
-      ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const final;
-  int GetCachedSize() const final { return _cached_size_.Get(); }
+  using ::google::protobuf::Message::MergeFrom;
+  void MergeFrom(const ErrorMessage& from) { ErrorMessage::MergeImpl(*this, from); }
 
   private:
-  inline void SharedCtor();
-  inline void SharedDtor();
-  void SetCachedSize(int size) const final;
+  static void MergeImpl(
+      ::google::protobuf::MessageLite& to_msg,
+      const ::google::protobuf::MessageLite& from_msg);
+
+  public:
+  bool IsInitialized() const {
+    return true;
+  }
+  ABSL_ATTRIBUTE_REINITIALIZES void Clear() PROTOBUF_FINAL;
+  #if defined(PROTOBUF_CUSTOM_VTABLE)
+  private:
+  static ::size_t ByteSizeLong(const ::google::protobuf::MessageLite& msg);
+  static ::uint8_t* _InternalSerialize(
+      const MessageLite& msg, ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream);
+
+  public:
+  ::size_t ByteSizeLong() const { return ByteSizeLong(*this); }
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+    return _InternalSerialize(*this, target, stream);
+  }
+  #else   // PROTOBUF_CUSTOM_VTABLE
+  ::size_t ByteSizeLong() const final;
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const final;
+  #endif  // PROTOBUF_CUSTOM_VTABLE
+  int GetCachedSize() const { return _impl_._cached_size_.Get(); }
+
+  private:
+  void SharedCtor(::google::protobuf::Arena* arena);
+  static void SharedDtor(MessageLite& self);
   void InternalSwap(ErrorMessage* other);
-  friend class ::PROTOBUF_NAMESPACE_ID::internal::AnyMetadata;
-  static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
-    return "protobuf.ErrorMessage";
+ private:
+  template <typename T>
+  friend ::absl::string_view(
+      ::google::protobuf::internal::GetAnyMessageName)();
+  static ::absl::string_view FullMessageName() { return "protobuf.ErrorMessage"; }
+
+ protected:
+  explicit ErrorMessage(::google::protobuf::Arena* arena);
+  ErrorMessage(::google::protobuf::Arena* arena, const ErrorMessage& from);
+  ErrorMessage(::google::protobuf::Arena* arena, ErrorMessage&& from) noexcept
+      : ErrorMessage(arena) {
+    *this = ::std::move(from);
   }
-  protected:
-  explicit ErrorMessage(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  private:
-  static void ArenaDtor(void* object);
-  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  public:
+  const ::google::protobuf::internal::ClassData* GetClassData() const PROTOBUF_FINAL;
+  static void* PlacementNew_(const void*, void* mem,
+                             ::google::protobuf::Arena* arena);
+  static constexpr auto InternalNewImpl_();
+  static const ::google::protobuf::internal::ClassDataFull _class_data_;
 
-  ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
-  private:
-  static ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadataStatic() {
-    ::PROTOBUF_NAMESPACE_ID::internal::AssignDescriptors(&::descriptor_table_ErrorMessage_2eproto);
-    return ::descriptor_table_ErrorMessage_2eproto.file_level_metadata[kIndexInFileMessages];
-  }
-
-  public:
-
+ public:
+  ::google::protobuf::Metadata GetMetadata() const;
   // nested types ----------------------------------------------------
 
   // accessors -------------------------------------------------------
-
   enum : int {
     kErrorMsgFieldNumber = 4,
     kAdvancedOrderRejectJsonFieldNumber = 5,
@@ -185,409 +220,357 @@ class ErrorMessage PROTOBUF_FINAL :
     kIdFieldNumber = 1,
     kErrorCodeFieldNumber = 3,
   };
-  // string errorMsg = 4;
+  // optional string errorMsg = 4;
   bool has_errormsg() const;
-  private:
-  bool _internal_has_errormsg() const;
-  public:
-  void clear_errormsg();
+  void clear_errormsg() ;
   const std::string& errormsg() const;
-  void set_errormsg(const std::string& value);
-  void set_errormsg(std::string&& value);
-  void set_errormsg(const char* value);
-  void set_errormsg(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_errormsg(Arg_&& arg, Args_... args);
   std::string* mutable_errormsg();
-  std::string* release_errormsg();
-  void set_allocated_errormsg(std::string* errormsg);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_errormsg();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_errormsg(
-      std::string* errormsg);
+  PROTOBUF_NODISCARD std::string* release_errormsg();
+  void set_allocated_errormsg(std::string* value);
+
   private:
   const std::string& _internal_errormsg() const;
-  void _internal_set_errormsg(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_errormsg(
+      const std::string& value);
   std::string* _internal_mutable_errormsg();
-  public:
 
-  // string advancedOrderRejectJson = 5;
-  bool has_advancedorderrejectjson() const;
-  private:
-  bool _internal_has_advancedorderrejectjson() const;
   public:
-  void clear_advancedorderrejectjson();
+  // optional string advancedOrderRejectJson = 5;
+  bool has_advancedorderrejectjson() const;
+  void clear_advancedorderrejectjson() ;
   const std::string& advancedorderrejectjson() const;
-  void set_advancedorderrejectjson(const std::string& value);
-  void set_advancedorderrejectjson(std::string&& value);
-  void set_advancedorderrejectjson(const char* value);
-  void set_advancedorderrejectjson(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_advancedorderrejectjson(Arg_&& arg, Args_... args);
   std::string* mutable_advancedorderrejectjson();
-  std::string* release_advancedorderrejectjson();
-  void set_allocated_advancedorderrejectjson(std::string* advancedorderrejectjson);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_advancedorderrejectjson();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_advancedorderrejectjson(
-      std::string* advancedorderrejectjson);
+  PROTOBUF_NODISCARD std::string* release_advancedorderrejectjson();
+  void set_allocated_advancedorderrejectjson(std::string* value);
+
   private:
   const std::string& _internal_advancedorderrejectjson() const;
-  void _internal_set_advancedorderrejectjson(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_advancedorderrejectjson(
+      const std::string& value);
   std::string* _internal_mutable_advancedorderrejectjson();
-  public:
 
-  // int64 errorTime = 2;
+  public:
+  // optional int64 errorTime = 2;
   bool has_errortime() const;
-  private:
-  bool _internal_has_errortime() const;
-  public:
-  void clear_errortime();
-  ::PROTOBUF_NAMESPACE_ID::int64 errortime() const;
-  void set_errortime(::PROTOBUF_NAMESPACE_ID::int64 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int64 _internal_errortime() const;
-  void _internal_set_errortime(::PROTOBUF_NAMESPACE_ID::int64 value);
-  public:
+  void clear_errortime() ;
+  ::int64_t errortime() const;
+  void set_errortime(::int64_t value);
 
-  // int32 id = 1;
+  private:
+  ::int64_t _internal_errortime() const;
+  void _internal_set_errortime(::int64_t value);
+
+  public:
+  // optional int32 id = 1;
   bool has_id() const;
-  private:
-  bool _internal_has_id() const;
-  public:
-  void clear_id();
-  ::PROTOBUF_NAMESPACE_ID::int32 id() const;
-  void set_id(::PROTOBUF_NAMESPACE_ID::int32 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int32 _internal_id() const;
-  void _internal_set_id(::PROTOBUF_NAMESPACE_ID::int32 value);
-  public:
+  void clear_id() ;
+  ::int32_t id() const;
+  void set_id(::int32_t value);
 
-  // int32 errorCode = 3;
+  private:
+  ::int32_t _internal_id() const;
+  void _internal_set_id(::int32_t value);
+
+  public:
+  // optional int32 errorCode = 3;
   bool has_errorcode() const;
-  private:
-  bool _internal_has_errorcode() const;
-  public:
-  void clear_errorcode();
-  ::PROTOBUF_NAMESPACE_ID::int32 errorcode() const;
-  void set_errorcode(::PROTOBUF_NAMESPACE_ID::int32 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int32 _internal_errorcode() const;
-  void _internal_set_errorcode(::PROTOBUF_NAMESPACE_ID::int32 value);
-  public:
+  void clear_errorcode() ;
+  ::int32_t errorcode() const;
+  void set_errorcode(::int32_t value);
 
+  private:
+  ::int32_t _internal_errorcode() const;
+  void _internal_set_errorcode(::int32_t value);
+
+  public:
   // @@protoc_insertion_point(class_scope:protobuf.ErrorMessage)
  private:
   class _Internal;
+  friend class ::google::protobuf::internal::TcParser;
+  static const ::google::protobuf::internal::TcParseTable<
+      3, 5, 0,
+      61, 2>
+      _table_;
 
-  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
-  typedef void InternalArenaConstructable_;
-  typedef void DestructorSkippable_;
-  ::PROTOBUF_NAMESPACE_ID::internal::HasBits<1> _has_bits_;
-  mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr errormsg_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr advancedorderrejectjson_;
-  ::PROTOBUF_NAMESPACE_ID::int64 errortime_;
-  ::PROTOBUF_NAMESPACE_ID::int32 id_;
-  ::PROTOBUF_NAMESPACE_ID::int32 errorcode_;
+  friend class ::google::protobuf::MessageLite;
+  friend class ::google::protobuf::Arena;
+  template <typename T>
+  friend class ::google::protobuf::Arena::InternalHelper;
+  using InternalArenaConstructable_ = void;
+  using DestructorSkippable_ = void;
+  struct Impl_ {
+    inline explicit constexpr Impl_(
+        ::google::protobuf::internal::ConstantInitialized) noexcept;
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena);
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena, const Impl_& from,
+                          const ErrorMessage& from_msg);
+    ::google::protobuf::internal::HasBits<1> _has_bits_;
+    ::google::protobuf::internal::CachedSize _cached_size_;
+    ::google::protobuf::internal::ArenaStringPtr errormsg_;
+    ::google::protobuf::internal::ArenaStringPtr advancedorderrejectjson_;
+    ::int64_t errortime_;
+    ::int32_t id_;
+    ::int32_t errorcode_;
+    PROTOBUF_TSAN_DECLARE_MEMBER
+  };
+  union { Impl_ _impl_; };
   friend struct ::TableStruct_ErrorMessage_2eproto;
 };
+
 // ===================================================================
 
 
+
+
 // ===================================================================
+
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic push
-  #pragma GCC diagnostic ignored "-Wstrict-aliasing"
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wstrict-aliasing"
 #endif  // __GNUC__
+// -------------------------------------------------------------------
+
 // ErrorMessage
 
-// int32 id = 1;
-inline bool ErrorMessage::_internal_has_id() const {
-  bool value = (_has_bits_[0] & 0x00000008u) != 0;
+// optional int32 id = 1;
+inline bool ErrorMessage::has_id() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000008u) != 0;
   return value;
 }
-inline bool ErrorMessage::has_id() const {
-  return _internal_has_id();
-}
 inline void ErrorMessage::clear_id() {
-  id_ = 0;
-  _has_bits_[0] &= ~0x00000008u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.id_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000008u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int32 ErrorMessage::_internal_id() const {
-  return id_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int32 ErrorMessage::id() const {
+inline ::int32_t ErrorMessage::id() const {
   // @@protoc_insertion_point(field_get:protobuf.ErrorMessage.id)
   return _internal_id();
 }
-inline void ErrorMessage::_internal_set_id(::PROTOBUF_NAMESPACE_ID::int32 value) {
-  _has_bits_[0] |= 0x00000008u;
-  id_ = value;
-}
-inline void ErrorMessage::set_id(::PROTOBUF_NAMESPACE_ID::int32 value) {
+inline void ErrorMessage::set_id(::int32_t value) {
   _internal_set_id(value);
+  _impl_._has_bits_[0] |= 0x00000008u;
   // @@protoc_insertion_point(field_set:protobuf.ErrorMessage.id)
 }
+inline ::int32_t ErrorMessage::_internal_id() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.id_;
+}
+inline void ErrorMessage::_internal_set_id(::int32_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.id_ = value;
+}
 
-// int64 errorTime = 2;
-inline bool ErrorMessage::_internal_has_errortime() const {
-  bool value = (_has_bits_[0] & 0x00000004u) != 0;
+// optional int64 errorTime = 2;
+inline bool ErrorMessage::has_errortime() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000004u) != 0;
   return value;
 }
-inline bool ErrorMessage::has_errortime() const {
-  return _internal_has_errortime();
-}
 inline void ErrorMessage::clear_errortime() {
-  errortime_ = PROTOBUF_LONGLONG(0);
-  _has_bits_[0] &= ~0x00000004u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.errortime_ = ::int64_t{0};
+  _impl_._has_bits_[0] &= ~0x00000004u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int64 ErrorMessage::_internal_errortime() const {
-  return errortime_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int64 ErrorMessage::errortime() const {
+inline ::int64_t ErrorMessage::errortime() const {
   // @@protoc_insertion_point(field_get:protobuf.ErrorMessage.errorTime)
   return _internal_errortime();
 }
-inline void ErrorMessage::_internal_set_errortime(::PROTOBUF_NAMESPACE_ID::int64 value) {
-  _has_bits_[0] |= 0x00000004u;
-  errortime_ = value;
-}
-inline void ErrorMessage::set_errortime(::PROTOBUF_NAMESPACE_ID::int64 value) {
+inline void ErrorMessage::set_errortime(::int64_t value) {
   _internal_set_errortime(value);
+  _impl_._has_bits_[0] |= 0x00000004u;
   // @@protoc_insertion_point(field_set:protobuf.ErrorMessage.errorTime)
 }
+inline ::int64_t ErrorMessage::_internal_errortime() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.errortime_;
+}
+inline void ErrorMessage::_internal_set_errortime(::int64_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.errortime_ = value;
+}
 
-// int32 errorCode = 3;
-inline bool ErrorMessage::_internal_has_errorcode() const {
-  bool value = (_has_bits_[0] & 0x00000010u) != 0;
+// optional int32 errorCode = 3;
+inline bool ErrorMessage::has_errorcode() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000010u) != 0;
   return value;
 }
-inline bool ErrorMessage::has_errorcode() const {
-  return _internal_has_errorcode();
-}
 inline void ErrorMessage::clear_errorcode() {
-  errorcode_ = 0;
-  _has_bits_[0] &= ~0x00000010u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.errorcode_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000010u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int32 ErrorMessage::_internal_errorcode() const {
-  return errorcode_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int32 ErrorMessage::errorcode() const {
+inline ::int32_t ErrorMessage::errorcode() const {
   // @@protoc_insertion_point(field_get:protobuf.ErrorMessage.errorCode)
   return _internal_errorcode();
 }
-inline void ErrorMessage::_internal_set_errorcode(::PROTOBUF_NAMESPACE_ID::int32 value) {
-  _has_bits_[0] |= 0x00000010u;
-  errorcode_ = value;
-}
-inline void ErrorMessage::set_errorcode(::PROTOBUF_NAMESPACE_ID::int32 value) {
+inline void ErrorMessage::set_errorcode(::int32_t value) {
   _internal_set_errorcode(value);
+  _impl_._has_bits_[0] |= 0x00000010u;
   // @@protoc_insertion_point(field_set:protobuf.ErrorMessage.errorCode)
 }
+inline ::int32_t ErrorMessage::_internal_errorcode() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.errorcode_;
+}
+inline void ErrorMessage::_internal_set_errorcode(::int32_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.errorcode_ = value;
+}
 
-// string errorMsg = 4;
-inline bool ErrorMessage::_internal_has_errormsg() const {
-  bool value = (_has_bits_[0] & 0x00000001u) != 0;
+// optional string errorMsg = 4;
+inline bool ErrorMessage::has_errormsg() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000001u) != 0;
   return value;
 }
-inline bool ErrorMessage::has_errormsg() const {
-  return _internal_has_errormsg();
-}
 inline void ErrorMessage::clear_errormsg() {
-  errormsg_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000001u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.errormsg_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000001u;
 }
-inline const std::string& ErrorMessage::errormsg() const {
+inline const std::string& ErrorMessage::errormsg() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.ErrorMessage.errorMsg)
   return _internal_errormsg();
 }
-inline void ErrorMessage::set_errormsg(const std::string& value) {
-  _internal_set_errormsg(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void ErrorMessage::set_errormsg(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  _impl_.errormsg_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.ErrorMessage.errorMsg)
 }
-inline std::string* ErrorMessage::mutable_errormsg() {
+inline std::string* ErrorMessage::mutable_errormsg() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_errormsg();
   // @@protoc_insertion_point(field_mutable:protobuf.ErrorMessage.errorMsg)
-  return _internal_mutable_errormsg();
+  return _s;
 }
 inline const std::string& ErrorMessage::_internal_errormsg() const {
-  return errormsg_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.errormsg_.Get();
 }
 inline void ErrorMessage::_internal_set_errormsg(const std::string& value) {
-  _has_bits_[0] |= 0x00000001u;
-  errormsg_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void ErrorMessage::set_errormsg(std::string&& value) {
-  _has_bits_[0] |= 0x00000001u;
-  errormsg_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.ErrorMessage.errorMsg)
-}
-inline void ErrorMessage::set_errormsg(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000001u;
-  errormsg_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.ErrorMessage.errorMsg)
-}
-inline void ErrorMessage::set_errormsg(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000001u;
-  errormsg_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.ErrorMessage.errorMsg)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  _impl_.errormsg_.Set(value, GetArena());
 }
 inline std::string* ErrorMessage::_internal_mutable_errormsg() {
-  _has_bits_[0] |= 0x00000001u;
-  return errormsg_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  return _impl_.errormsg_.Mutable( GetArena());
 }
 inline std::string* ErrorMessage::release_errormsg() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.ErrorMessage.errorMsg)
-  if (!_internal_has_errormsg()) {
+  if ((_impl_._has_bits_[0] & 0x00000001u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000001u;
-  return errormsg_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void ErrorMessage::set_allocated_errormsg(std::string* errormsg) {
-  if (errormsg != nullptr) {
-    _has_bits_[0] |= 0x00000001u;
-  } else {
-    _has_bits_[0] &= ~0x00000001u;
+  _impl_._has_bits_[0] &= ~0x00000001u;
+  auto* released = _impl_.errormsg_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.errormsg_.Set("", GetArena());
   }
-  errormsg_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), errormsg,
-      GetArena());
+  return released;
+}
+inline void ErrorMessage::set_allocated_errormsg(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000001u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000001u;
+  }
+  _impl_.errormsg_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.errormsg_.IsDefault()) {
+    _impl_.errormsg_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.ErrorMessage.errorMsg)
 }
-inline std::string* ErrorMessage::unsafe_arena_release_errormsg() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.ErrorMessage.errorMsg)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000001u;
-  return errormsg_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void ErrorMessage::unsafe_arena_set_allocated_errormsg(
-    std::string* errormsg) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (errormsg != nullptr) {
-    _has_bits_[0] |= 0x00000001u;
-  } else {
-    _has_bits_[0] &= ~0x00000001u;
-  }
-  errormsg_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      errormsg, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.ErrorMessage.errorMsg)
-}
 
-// string advancedOrderRejectJson = 5;
-inline bool ErrorMessage::_internal_has_advancedorderrejectjson() const {
-  bool value = (_has_bits_[0] & 0x00000002u) != 0;
+// optional string advancedOrderRejectJson = 5;
+inline bool ErrorMessage::has_advancedorderrejectjson() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000002u) != 0;
   return value;
 }
-inline bool ErrorMessage::has_advancedorderrejectjson() const {
-  return _internal_has_advancedorderrejectjson();
-}
 inline void ErrorMessage::clear_advancedorderrejectjson() {
-  advancedorderrejectjson_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000002u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.advancedorderrejectjson_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000002u;
 }
-inline const std::string& ErrorMessage::advancedorderrejectjson() const {
+inline const std::string& ErrorMessage::advancedorderrejectjson() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.ErrorMessage.advancedOrderRejectJson)
   return _internal_advancedorderrejectjson();
 }
-inline void ErrorMessage::set_advancedorderrejectjson(const std::string& value) {
-  _internal_set_advancedorderrejectjson(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void ErrorMessage::set_advancedorderrejectjson(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  _impl_.advancedorderrejectjson_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.ErrorMessage.advancedOrderRejectJson)
 }
-inline std::string* ErrorMessage::mutable_advancedorderrejectjson() {
+inline std::string* ErrorMessage::mutable_advancedorderrejectjson() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_advancedorderrejectjson();
   // @@protoc_insertion_point(field_mutable:protobuf.ErrorMessage.advancedOrderRejectJson)
-  return _internal_mutable_advancedorderrejectjson();
+  return _s;
 }
 inline const std::string& ErrorMessage::_internal_advancedorderrejectjson() const {
-  return advancedorderrejectjson_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.advancedorderrejectjson_.Get();
 }
 inline void ErrorMessage::_internal_set_advancedorderrejectjson(const std::string& value) {
-  _has_bits_[0] |= 0x00000002u;
-  advancedorderrejectjson_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void ErrorMessage::set_advancedorderrejectjson(std::string&& value) {
-  _has_bits_[0] |= 0x00000002u;
-  advancedorderrejectjson_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.ErrorMessage.advancedOrderRejectJson)
-}
-inline void ErrorMessage::set_advancedorderrejectjson(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000002u;
-  advancedorderrejectjson_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.ErrorMessage.advancedOrderRejectJson)
-}
-inline void ErrorMessage::set_advancedorderrejectjson(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000002u;
-  advancedorderrejectjson_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.ErrorMessage.advancedOrderRejectJson)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  _impl_.advancedorderrejectjson_.Set(value, GetArena());
 }
 inline std::string* ErrorMessage::_internal_mutable_advancedorderrejectjson() {
-  _has_bits_[0] |= 0x00000002u;
-  return advancedorderrejectjson_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  return _impl_.advancedorderrejectjson_.Mutable( GetArena());
 }
 inline std::string* ErrorMessage::release_advancedorderrejectjson() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.ErrorMessage.advancedOrderRejectJson)
-  if (!_internal_has_advancedorderrejectjson()) {
+  if ((_impl_._has_bits_[0] & 0x00000002u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000002u;
-  return advancedorderrejectjson_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void ErrorMessage::set_allocated_advancedorderrejectjson(std::string* advancedorderrejectjson) {
-  if (advancedorderrejectjson != nullptr) {
-    _has_bits_[0] |= 0x00000002u;
-  } else {
-    _has_bits_[0] &= ~0x00000002u;
+  _impl_._has_bits_[0] &= ~0x00000002u;
+  auto* released = _impl_.advancedorderrejectjson_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.advancedorderrejectjson_.Set("", GetArena());
   }
-  advancedorderrejectjson_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), advancedorderrejectjson,
-      GetArena());
+  return released;
+}
+inline void ErrorMessage::set_allocated_advancedorderrejectjson(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000002u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000002u;
+  }
+  _impl_.advancedorderrejectjson_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.advancedorderrejectjson_.IsDefault()) {
+    _impl_.advancedorderrejectjson_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.ErrorMessage.advancedOrderRejectJson)
-}
-inline std::string* ErrorMessage::unsafe_arena_release_advancedorderrejectjson() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.ErrorMessage.advancedOrderRejectJson)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000002u;
-  return advancedorderrejectjson_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void ErrorMessage::unsafe_arena_set_allocated_advancedorderrejectjson(
-    std::string* advancedorderrejectjson) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (advancedorderrejectjson != nullptr) {
-    _has_bits_[0] |= 0x00000002u;
-  } else {
-    _has_bits_[0] &= ~0x00000002u;
-  }
-  advancedorderrejectjson_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      advancedorderrejectjson, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.ErrorMessage.advancedOrderRejectJson)
 }
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic pop
+#pragma GCC diagnostic pop
 #endif  // __GNUC__
 
 // @@protoc_insertion_point(namespace_scope)
-
 }  // namespace protobuf
+
 
 // @@protoc_insertion_point(global_scope)
 
-#include <google/protobuf/port_undef.inc>
-#endif  // GOOGLE_PROTOBUF_INCLUDED_GOOGLE_PROTOBUF_INCLUDED_ErrorMessage_2eproto
+#include "google/protobuf/port_undef.inc"
+
+#endif  // ErrorMessage_2eproto_2epb_2eh
