@@ -1,455 +1,536 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: ExecutionDetails.proto
+// Protobuf C++ Version: 5.29.3
 
-#ifndef GOOGLE_PROTOBUF_INCLUDED_ExecutionDetails_2eproto
-#define GOOGLE_PROTOBUF_INCLUDED_ExecutionDetails_2eproto
+#ifndef ExecutionDetails_2eproto_2epb_2eh
+#define ExecutionDetails_2eproto_2epb_2eh
 
 #include <limits>
 #include <string>
+#include <type_traits>
+#include <utility>
 
-#include <google/protobuf/port_def.inc>
-#if PROTOBUF_VERSION < 3012000
-#error This file was generated by a newer version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please update
-#error your headers.
+#include "google/protobuf/runtime_version.h"
+#if PROTOBUF_VERSION != 5029003
+#error "Protobuf C++ gencode is built with an incompatible version of"
+#error "Protobuf C++ headers/runtime. See"
+#error "https://protobuf.dev/support/cross-version-runtime-guarantee/#cpp"
 #endif
-#if 3012004 < PROTOBUF_MIN_PROTOC_VERSION
-#error This file was generated by an older version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please
-#error regenerate this file with a newer version of protoc.
-#endif
-
-#include <google/protobuf/port_undef.inc>
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/arena.h>
-#include <google/protobuf/arenastring.h>
-#include <google/protobuf/generated_message_table_driven.h>
-#include <google/protobuf/generated_message_util.h>
-#include <google/protobuf/inlined_string_field.h>
-#include <google/protobuf/metadata_lite.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/message.h>
-#include <google/protobuf/repeated_field.h>  // IWYU pragma: export
-#include <google/protobuf/extension_set.h>  // IWYU pragma: export
-#include <google/protobuf/unknown_field_set.h>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/arena.h"
+#include "google/protobuf/arenastring.h"
+#include "google/protobuf/generated_message_tctable_decl.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/metadata_lite.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/message.h"
+#include "google/protobuf/message_lite.h"
+#include "google/protobuf/repeated_field.h"  // IWYU pragma: export
+#include "google/protobuf/extension_set.h"  // IWYU pragma: export
+#include "google/protobuf/unknown_field_set.h"
 #include "Contract.pb.h"
 #include "Execution.pb.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
+
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+
 #define PROTOBUF_INTERNAL_EXPORT_ExecutionDetails_2eproto
-PROTOBUF_NAMESPACE_OPEN
+
+namespace google {
+namespace protobuf {
 namespace internal {
-class AnyMetadata;
+template <typename T>
+::absl::string_view GetAnyMessageName();
 }  // namespace internal
-PROTOBUF_NAMESPACE_CLOSE
+}  // namespace protobuf
+}  // namespace google
 
 // Internal implementation detail -- do not use these members.
 struct TableStruct_ExecutionDetails_2eproto {
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTableField entries[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxillaryParseTableField aux[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTable schema[1]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::FieldMetadata field_metadata[];
-  static const ::PROTOBUF_NAMESPACE_ID::internal::SerializationTable serialization_table[];
-  static const ::PROTOBUF_NAMESPACE_ID::uint32 offsets[];
+  static const ::uint32_t offsets[];
 };
-extern const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_ExecutionDetails_2eproto;
+extern const ::google::protobuf::internal::DescriptorTable
+    descriptor_table_ExecutionDetails_2eproto;
 namespace protobuf {
 class ExecutionDetails;
-class ExecutionDetailsDefaultTypeInternal;
+struct ExecutionDetailsDefaultTypeInternal;
 extern ExecutionDetailsDefaultTypeInternal _ExecutionDetails_default_instance_;
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> ::protobuf::ExecutionDetails* Arena::CreateMaybeMessage<::protobuf::ExecutionDetails>(Arena*);
-PROTOBUF_NAMESPACE_CLOSE
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
+
 namespace protobuf {
 
 // ===================================================================
 
-class ExecutionDetails PROTOBUF_FINAL :
-    public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:protobuf.ExecutionDetails) */ {
+
+// -------------------------------------------------------------------
+
+class ExecutionDetails final : public ::google::protobuf::Message
+/* @@protoc_insertion_point(class_definition:protobuf.ExecutionDetails) */ {
  public:
-  inline ExecutionDetails() : ExecutionDetails(nullptr) {};
-  virtual ~ExecutionDetails();
+  inline ExecutionDetails() : ExecutionDetails(nullptr) {}
+  ~ExecutionDetails() PROTOBUF_FINAL;
 
-  ExecutionDetails(const ExecutionDetails& from);
-  ExecutionDetails(ExecutionDetails&& from) noexcept
-    : ExecutionDetails() {
-    *this = ::std::move(from);
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+  void operator delete(ExecutionDetails* msg, std::destroying_delete_t) {
+    SharedDtor(*msg);
+    ::google::protobuf::internal::SizedDelete(msg, sizeof(ExecutionDetails));
   }
+#endif
 
+  template <typename = void>
+  explicit PROTOBUF_CONSTEXPR ExecutionDetails(
+      ::google::protobuf::internal::ConstantInitialized);
+
+  inline ExecutionDetails(const ExecutionDetails& from) : ExecutionDetails(nullptr, from) {}
+  inline ExecutionDetails(ExecutionDetails&& from) noexcept
+      : ExecutionDetails(nullptr, std::move(from)) {}
   inline ExecutionDetails& operator=(const ExecutionDetails& from) {
     CopyFrom(from);
     return *this;
   }
   inline ExecutionDetails& operator=(ExecutionDetails&& from) noexcept {
-    if (GetArena() == from.GetArena()) {
-      if (this != &from) InternalSwap(&from);
+    if (this == &from) return *this;
+    if (::google::protobuf::internal::CanMoveWithInternalSwap(GetArena(), from.GetArena())) {
+      InternalSwap(&from);
     } else {
       CopyFrom(from);
     }
     return *this;
   }
 
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* descriptor() {
+  inline const ::google::protobuf::UnknownFieldSet& unknown_fields() const
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance);
+  }
+  inline ::google::protobuf::UnknownFieldSet* mutable_unknown_fields()
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.mutable_unknown_fields<::google::protobuf::UnknownFieldSet>();
+  }
+
+  static const ::google::protobuf::Descriptor* descriptor() {
     return GetDescriptor();
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* GetDescriptor() {
-    return GetMetadataStatic().descriptor;
+  static const ::google::protobuf::Descriptor* GetDescriptor() {
+    return default_instance().GetMetadata().descriptor;
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Reflection* GetReflection() {
-    return GetMetadataStatic().reflection;
+  static const ::google::protobuf::Reflection* GetReflection() {
+    return default_instance().GetMetadata().reflection;
   }
-  static const ExecutionDetails& default_instance();
-
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
+  static const ExecutionDetails& default_instance() {
+    return *internal_default_instance();
+  }
   static inline const ExecutionDetails* internal_default_instance() {
     return reinterpret_cast<const ExecutionDetails*>(
-               &_ExecutionDetails_default_instance_);
+        &_ExecutionDetails_default_instance_);
   }
-  static constexpr int kIndexInFileMessages =
-    0;
-
-  friend void swap(ExecutionDetails& a, ExecutionDetails& b) {
-    a.Swap(&b);
-  }
+  static constexpr int kIndexInFileMessages = 0;
+  friend void swap(ExecutionDetails& a, ExecutionDetails& b) { a.Swap(&b); }
   inline void Swap(ExecutionDetails* other) {
     if (other == this) return;
-    if (GetArena() == other->GetArena()) {
+    if (::google::protobuf::internal::CanUseInternalSwap(GetArena(), other->GetArena())) {
       InternalSwap(other);
     } else {
-      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+      ::google::protobuf::internal::GenericSwap(this, other);
     }
   }
   void UnsafeArenaSwap(ExecutionDetails* other) {
     if (other == this) return;
-    GOOGLE_DCHECK(GetArena() == other->GetArena());
+    ABSL_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
   // implements Message ----------------------------------------------
 
-  inline ExecutionDetails* New() const final {
-    return CreateMaybeMessage<ExecutionDetails>(nullptr);
+  ExecutionDetails* New(::google::protobuf::Arena* arena = nullptr) const {
+    return ::google::protobuf::Message::DefaultConstruct<ExecutionDetails>(arena);
   }
-
-  ExecutionDetails* New(::PROTOBUF_NAMESPACE_ID::Arena* arena) const final {
-    return CreateMaybeMessage<ExecutionDetails>(arena);
-  }
-  void CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
-  void MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
+  using ::google::protobuf::Message::CopyFrom;
   void CopyFrom(const ExecutionDetails& from);
-  void MergeFrom(const ExecutionDetails& from);
-  PROTOBUF_ATTRIBUTE_REINITIALIZES void Clear() final;
-  bool IsInitialized() const final;
-
-  size_t ByteSizeLong() const final;
-  const char* _InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) final;
-  ::PROTOBUF_NAMESPACE_ID::uint8* _InternalSerialize(
-      ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const final;
-  int GetCachedSize() const final { return _cached_size_.Get(); }
+  using ::google::protobuf::Message::MergeFrom;
+  void MergeFrom(const ExecutionDetails& from) { ExecutionDetails::MergeImpl(*this, from); }
 
   private:
-  inline void SharedCtor();
-  inline void SharedDtor();
-  void SetCachedSize(int size) const final;
+  static void MergeImpl(
+      ::google::protobuf::MessageLite& to_msg,
+      const ::google::protobuf::MessageLite& from_msg);
+
+  public:
+  bool IsInitialized() const {
+    return true;
+  }
+  ABSL_ATTRIBUTE_REINITIALIZES void Clear() PROTOBUF_FINAL;
+  #if defined(PROTOBUF_CUSTOM_VTABLE)
+  private:
+  static ::size_t ByteSizeLong(const ::google::protobuf::MessageLite& msg);
+  static ::uint8_t* _InternalSerialize(
+      const MessageLite& msg, ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream);
+
+  public:
+  ::size_t ByteSizeLong() const { return ByteSizeLong(*this); }
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+    return _InternalSerialize(*this, target, stream);
+  }
+  #else   // PROTOBUF_CUSTOM_VTABLE
+  ::size_t ByteSizeLong() const final;
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const final;
+  #endif  // PROTOBUF_CUSTOM_VTABLE
+  int GetCachedSize() const { return _impl_._cached_size_.Get(); }
+
+  private:
+  void SharedCtor(::google::protobuf::Arena* arena);
+  static void SharedDtor(MessageLite& self);
   void InternalSwap(ExecutionDetails* other);
-  friend class ::PROTOBUF_NAMESPACE_ID::internal::AnyMetadata;
-  static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
-    return "protobuf.ExecutionDetails";
+ private:
+  template <typename T>
+  friend ::absl::string_view(
+      ::google::protobuf::internal::GetAnyMessageName)();
+  static ::absl::string_view FullMessageName() { return "protobuf.ExecutionDetails"; }
+
+ protected:
+  explicit ExecutionDetails(::google::protobuf::Arena* arena);
+  ExecutionDetails(::google::protobuf::Arena* arena, const ExecutionDetails& from);
+  ExecutionDetails(::google::protobuf::Arena* arena, ExecutionDetails&& from) noexcept
+      : ExecutionDetails(arena) {
+    *this = ::std::move(from);
   }
-  protected:
-  explicit ExecutionDetails(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  private:
-  static void ArenaDtor(void* object);
-  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  public:
+  const ::google::protobuf::internal::ClassData* GetClassData() const PROTOBUF_FINAL;
+  static void* PlacementNew_(const void*, void* mem,
+                             ::google::protobuf::Arena* arena);
+  static constexpr auto InternalNewImpl_();
+  static const ::google::protobuf::internal::ClassDataFull _class_data_;
 
-  ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
-  private:
-  static ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadataStatic() {
-    ::PROTOBUF_NAMESPACE_ID::internal::AssignDescriptors(&::descriptor_table_ExecutionDetails_2eproto);
-    return ::descriptor_table_ExecutionDetails_2eproto.file_level_metadata[kIndexInFileMessages];
-  }
-
-  public:
-
+ public:
+  ::google::protobuf::Metadata GetMetadata() const;
   // nested types ----------------------------------------------------
 
   // accessors -------------------------------------------------------
-
   enum : int {
     kContractFieldNumber = 2,
     kExecutionFieldNumber = 3,
     kReqIdFieldNumber = 1,
   };
-  // .protobuf.Contract contract = 2;
+  // optional .protobuf.Contract contract = 2;
   bool has_contract() const;
-  private:
-  bool _internal_has_contract() const;
-  public:
-  void clear_contract();
+  void clear_contract() ;
   const ::protobuf::Contract& contract() const;
-  ::protobuf::Contract* release_contract();
+  PROTOBUF_NODISCARD ::protobuf::Contract* release_contract();
   ::protobuf::Contract* mutable_contract();
-  void set_allocated_contract(::protobuf::Contract* contract);
+  void set_allocated_contract(::protobuf::Contract* value);
+  void unsafe_arena_set_allocated_contract(::protobuf::Contract* value);
+  ::protobuf::Contract* unsafe_arena_release_contract();
+
   private:
   const ::protobuf::Contract& _internal_contract() const;
   ::protobuf::Contract* _internal_mutable_contract();
-  public:
-  void unsafe_arena_set_allocated_contract(
-      ::protobuf::Contract* contract);
-  ::protobuf::Contract* unsafe_arena_release_contract();
 
-  // .protobuf.Execution execution = 3;
-  bool has_execution() const;
-  private:
-  bool _internal_has_execution() const;
   public:
-  void clear_execution();
+  // optional .protobuf.Execution execution = 3;
+  bool has_execution() const;
+  void clear_execution() ;
   const ::protobuf::Execution& execution() const;
-  ::protobuf::Execution* release_execution();
+  PROTOBUF_NODISCARD ::protobuf::Execution* release_execution();
   ::protobuf::Execution* mutable_execution();
-  void set_allocated_execution(::protobuf::Execution* execution);
+  void set_allocated_execution(::protobuf::Execution* value);
+  void unsafe_arena_set_allocated_execution(::protobuf::Execution* value);
+  ::protobuf::Execution* unsafe_arena_release_execution();
+
   private:
   const ::protobuf::Execution& _internal_execution() const;
   ::protobuf::Execution* _internal_mutable_execution();
-  public:
-  void unsafe_arena_set_allocated_execution(
-      ::protobuf::Execution* execution);
-  ::protobuf::Execution* unsafe_arena_release_execution();
 
-  // int32 reqId = 1;
+  public:
+  // optional int32 reqId = 1;
   bool has_reqid() const;
-  private:
-  bool _internal_has_reqid() const;
-  public:
-  void clear_reqid();
-  ::PROTOBUF_NAMESPACE_ID::int32 reqid() const;
-  void set_reqid(::PROTOBUF_NAMESPACE_ID::int32 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int32 _internal_reqid() const;
-  void _internal_set_reqid(::PROTOBUF_NAMESPACE_ID::int32 value);
-  public:
+  void clear_reqid() ;
+  ::int32_t reqid() const;
+  void set_reqid(::int32_t value);
 
+  private:
+  ::int32_t _internal_reqid() const;
+  void _internal_set_reqid(::int32_t value);
+
+  public:
   // @@protoc_insertion_point(class_scope:protobuf.ExecutionDetails)
  private:
   class _Internal;
+  friend class ::google::protobuf::internal::TcParser;
+  static const ::google::protobuf::internal::TcParseTable<
+      2, 3, 2,
+      0, 2>
+      _table_;
 
-  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
-  typedef void InternalArenaConstructable_;
-  typedef void DestructorSkippable_;
-  ::PROTOBUF_NAMESPACE_ID::internal::HasBits<1> _has_bits_;
-  mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
-  ::protobuf::Contract* contract_;
-  ::protobuf::Execution* execution_;
-  ::PROTOBUF_NAMESPACE_ID::int32 reqid_;
+  friend class ::google::protobuf::MessageLite;
+  friend class ::google::protobuf::Arena;
+  template <typename T>
+  friend class ::google::protobuf::Arena::InternalHelper;
+  using InternalArenaConstructable_ = void;
+  using DestructorSkippable_ = void;
+  struct Impl_ {
+    inline explicit constexpr Impl_(
+        ::google::protobuf::internal::ConstantInitialized) noexcept;
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena);
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena, const Impl_& from,
+                          const ExecutionDetails& from_msg);
+    ::google::protobuf::internal::HasBits<1> _has_bits_;
+    ::google::protobuf::internal::CachedSize _cached_size_;
+    ::protobuf::Contract* contract_;
+    ::protobuf::Execution* execution_;
+    ::int32_t reqid_;
+    PROTOBUF_TSAN_DECLARE_MEMBER
+  };
+  union { Impl_ _impl_; };
   friend struct ::TableStruct_ExecutionDetails_2eproto;
 };
+
 // ===================================================================
 
 
+
+
 // ===================================================================
+
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic push
-  #pragma GCC diagnostic ignored "-Wstrict-aliasing"
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wstrict-aliasing"
 #endif  // __GNUC__
+// -------------------------------------------------------------------
+
 // ExecutionDetails
 
-// int32 reqId = 1;
-inline bool ExecutionDetails::_internal_has_reqid() const {
-  bool value = (_has_bits_[0] & 0x00000004u) != 0;
+// optional int32 reqId = 1;
+inline bool ExecutionDetails::has_reqid() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000004u) != 0;
   return value;
 }
-inline bool ExecutionDetails::has_reqid() const {
-  return _internal_has_reqid();
-}
 inline void ExecutionDetails::clear_reqid() {
-  reqid_ = 0;
-  _has_bits_[0] &= ~0x00000004u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.reqid_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000004u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int32 ExecutionDetails::_internal_reqid() const {
-  return reqid_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int32 ExecutionDetails::reqid() const {
+inline ::int32_t ExecutionDetails::reqid() const {
   // @@protoc_insertion_point(field_get:protobuf.ExecutionDetails.reqId)
   return _internal_reqid();
 }
-inline void ExecutionDetails::_internal_set_reqid(::PROTOBUF_NAMESPACE_ID::int32 value) {
-  _has_bits_[0] |= 0x00000004u;
-  reqid_ = value;
-}
-inline void ExecutionDetails::set_reqid(::PROTOBUF_NAMESPACE_ID::int32 value) {
+inline void ExecutionDetails::set_reqid(::int32_t value) {
   _internal_set_reqid(value);
+  _impl_._has_bits_[0] |= 0x00000004u;
   // @@protoc_insertion_point(field_set:protobuf.ExecutionDetails.reqId)
 }
+inline ::int32_t ExecutionDetails::_internal_reqid() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.reqid_;
+}
+inline void ExecutionDetails::_internal_set_reqid(::int32_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.reqid_ = value;
+}
 
-// .protobuf.Contract contract = 2;
-inline bool ExecutionDetails::_internal_has_contract() const {
-  bool value = (_has_bits_[0] & 0x00000001u) != 0;
-  PROTOBUF_ASSUME(!value || contract_ != nullptr);
+// optional .protobuf.Contract contract = 2;
+inline bool ExecutionDetails::has_contract() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000001u) != 0;
+  PROTOBUF_ASSUME(!value || _impl_.contract_ != nullptr);
   return value;
 }
-inline bool ExecutionDetails::has_contract() const {
-  return _internal_has_contract();
-}
 inline const ::protobuf::Contract& ExecutionDetails::_internal_contract() const {
-  const ::protobuf::Contract* p = contract_;
-  return p != nullptr ? *p : *reinterpret_cast<const ::protobuf::Contract*>(
-      &::protobuf::_Contract_default_instance_);
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  const ::protobuf::Contract* p = _impl_.contract_;
+  return p != nullptr ? *p : reinterpret_cast<const ::protobuf::Contract&>(::protobuf::_Contract_default_instance_);
 }
-inline const ::protobuf::Contract& ExecutionDetails::contract() const {
+inline const ::protobuf::Contract& ExecutionDetails::contract() const ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.ExecutionDetails.contract)
   return _internal_contract();
 }
-inline void ExecutionDetails::unsafe_arena_set_allocated_contract(
-    ::protobuf::Contract* contract) {
+inline void ExecutionDetails::unsafe_arena_set_allocated_contract(::protobuf::Contract* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (GetArena() == nullptr) {
-    delete reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(contract_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.contract_);
   }
-  contract_ = contract;
-  if (contract) {
-    _has_bits_[0] |= 0x00000001u;
+  _impl_.contract_ = reinterpret_cast<::protobuf::Contract*>(value);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000001u;
   } else {
-    _has_bits_[0] &= ~0x00000001u;
+    _impl_._has_bits_[0] &= ~0x00000001u;
   }
   // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.ExecutionDetails.contract)
 }
 inline ::protobuf::Contract* ExecutionDetails::release_contract() {
-  auto temp = unsafe_arena_release_contract();
-  if (GetArena() != nullptr) {
-    temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+
+  _impl_._has_bits_[0] &= ~0x00000001u;
+  ::protobuf::Contract* released = _impl_.contract_;
+  _impl_.contract_ = nullptr;
+  if (::google::protobuf::internal::DebugHardenForceCopyInRelease()) {
+    auto* old = reinterpret_cast<::google::protobuf::MessageLite*>(released);
+    released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    if (GetArena() == nullptr) {
+      delete old;
+    }
+  } else {
+    if (GetArena() != nullptr) {
+      released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    }
   }
-  return temp;
+  return released;
 }
 inline ::protobuf::Contract* ExecutionDetails::unsafe_arena_release_contract() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.ExecutionDetails.contract)
-  _has_bits_[0] &= ~0x00000001u;
-  ::protobuf::Contract* temp = contract_;
-  contract_ = nullptr;
+
+  _impl_._has_bits_[0] &= ~0x00000001u;
+  ::protobuf::Contract* temp = _impl_.contract_;
+  _impl_.contract_ = nullptr;
   return temp;
 }
 inline ::protobuf::Contract* ExecutionDetails::_internal_mutable_contract() {
-  _has_bits_[0] |= 0x00000001u;
-  if (contract_ == nullptr) {
-    auto* p = CreateMaybeMessage<::protobuf::Contract>(GetArena());
-    contract_ = p;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (_impl_.contract_ == nullptr) {
+    auto* p = ::google::protobuf::Message::DefaultConstruct<::protobuf::Contract>(GetArena());
+    _impl_.contract_ = reinterpret_cast<::protobuf::Contract*>(p);
   }
-  return contract_;
+  return _impl_.contract_;
 }
-inline ::protobuf::Contract* ExecutionDetails::mutable_contract() {
+inline ::protobuf::Contract* ExecutionDetails::mutable_contract() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  _impl_._has_bits_[0] |= 0x00000001u;
+  ::protobuf::Contract* _msg = _internal_mutable_contract();
   // @@protoc_insertion_point(field_mutable:protobuf.ExecutionDetails.contract)
-  return _internal_mutable_contract();
+  return _msg;
 }
-inline void ExecutionDetails::set_allocated_contract(::protobuf::Contract* contract) {
-  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArena();
+inline void ExecutionDetails::set_allocated_contract(::protobuf::Contract* value) {
+  ::google::protobuf::Arena* message_arena = GetArena();
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (message_arena == nullptr) {
-    delete reinterpret_cast< ::PROTOBUF_NAMESPACE_ID::MessageLite*>(contract_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.contract_);
   }
-  if (contract) {
-    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena =
-      reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(contract)->GetArena();
+
+  if (value != nullptr) {
+    ::google::protobuf::Arena* submessage_arena = reinterpret_cast<::google::protobuf::MessageLite*>(value)->GetArena();
     if (message_arena != submessage_arena) {
-      contract = ::PROTOBUF_NAMESPACE_ID::internal::GetOwnedMessage(
-          message_arena, contract, submessage_arena);
+      value = ::google::protobuf::internal::GetOwnedMessage(message_arena, value, submessage_arena);
     }
-    _has_bits_[0] |= 0x00000001u;
+    _impl_._has_bits_[0] |= 0x00000001u;
   } else {
-    _has_bits_[0] &= ~0x00000001u;
+    _impl_._has_bits_[0] &= ~0x00000001u;
   }
-  contract_ = contract;
+
+  _impl_.contract_ = reinterpret_cast<::protobuf::Contract*>(value);
   // @@protoc_insertion_point(field_set_allocated:protobuf.ExecutionDetails.contract)
 }
 
-// .protobuf.Execution execution = 3;
-inline bool ExecutionDetails::_internal_has_execution() const {
-  bool value = (_has_bits_[0] & 0x00000002u) != 0;
-  PROTOBUF_ASSUME(!value || execution_ != nullptr);
+// optional .protobuf.Execution execution = 3;
+inline bool ExecutionDetails::has_execution() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000002u) != 0;
+  PROTOBUF_ASSUME(!value || _impl_.execution_ != nullptr);
   return value;
 }
-inline bool ExecutionDetails::has_execution() const {
-  return _internal_has_execution();
-}
 inline const ::protobuf::Execution& ExecutionDetails::_internal_execution() const {
-  const ::protobuf::Execution* p = execution_;
-  return p != nullptr ? *p : *reinterpret_cast<const ::protobuf::Execution*>(
-      &::protobuf::_Execution_default_instance_);
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  const ::protobuf::Execution* p = _impl_.execution_;
+  return p != nullptr ? *p : reinterpret_cast<const ::protobuf::Execution&>(::protobuf::_Execution_default_instance_);
 }
-inline const ::protobuf::Execution& ExecutionDetails::execution() const {
+inline const ::protobuf::Execution& ExecutionDetails::execution() const ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.ExecutionDetails.execution)
   return _internal_execution();
 }
-inline void ExecutionDetails::unsafe_arena_set_allocated_execution(
-    ::protobuf::Execution* execution) {
+inline void ExecutionDetails::unsafe_arena_set_allocated_execution(::protobuf::Execution* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (GetArena() == nullptr) {
-    delete reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(execution_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.execution_);
   }
-  execution_ = execution;
-  if (execution) {
-    _has_bits_[0] |= 0x00000002u;
+  _impl_.execution_ = reinterpret_cast<::protobuf::Execution*>(value);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000002u;
   } else {
-    _has_bits_[0] &= ~0x00000002u;
+    _impl_._has_bits_[0] &= ~0x00000002u;
   }
   // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.ExecutionDetails.execution)
 }
 inline ::protobuf::Execution* ExecutionDetails::release_execution() {
-  auto temp = unsafe_arena_release_execution();
-  if (GetArena() != nullptr) {
-    temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+
+  _impl_._has_bits_[0] &= ~0x00000002u;
+  ::protobuf::Execution* released = _impl_.execution_;
+  _impl_.execution_ = nullptr;
+  if (::google::protobuf::internal::DebugHardenForceCopyInRelease()) {
+    auto* old = reinterpret_cast<::google::protobuf::MessageLite*>(released);
+    released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    if (GetArena() == nullptr) {
+      delete old;
+    }
+  } else {
+    if (GetArena() != nullptr) {
+      released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    }
   }
-  return temp;
+  return released;
 }
 inline ::protobuf::Execution* ExecutionDetails::unsafe_arena_release_execution() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.ExecutionDetails.execution)
-  _has_bits_[0] &= ~0x00000002u;
-  ::protobuf::Execution* temp = execution_;
-  execution_ = nullptr;
+
+  _impl_._has_bits_[0] &= ~0x00000002u;
+  ::protobuf::Execution* temp = _impl_.execution_;
+  _impl_.execution_ = nullptr;
   return temp;
 }
 inline ::protobuf::Execution* ExecutionDetails::_internal_mutable_execution() {
-  _has_bits_[0] |= 0x00000002u;
-  if (execution_ == nullptr) {
-    auto* p = CreateMaybeMessage<::protobuf::Execution>(GetArena());
-    execution_ = p;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (_impl_.execution_ == nullptr) {
+    auto* p = ::google::protobuf::Message::DefaultConstruct<::protobuf::Execution>(GetArena());
+    _impl_.execution_ = reinterpret_cast<::protobuf::Execution*>(p);
   }
-  return execution_;
+  return _impl_.execution_;
 }
-inline ::protobuf::Execution* ExecutionDetails::mutable_execution() {
+inline ::protobuf::Execution* ExecutionDetails::mutable_execution() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  _impl_._has_bits_[0] |= 0x00000002u;
+  ::protobuf::Execution* _msg = _internal_mutable_execution();
   // @@protoc_insertion_point(field_mutable:protobuf.ExecutionDetails.execution)
-  return _internal_mutable_execution();
+  return _msg;
 }
-inline void ExecutionDetails::set_allocated_execution(::protobuf::Execution* execution) {
-  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArena();
+inline void ExecutionDetails::set_allocated_execution(::protobuf::Execution* value) {
+  ::google::protobuf::Arena* message_arena = GetArena();
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (message_arena == nullptr) {
-    delete reinterpret_cast< ::PROTOBUF_NAMESPACE_ID::MessageLite*>(execution_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.execution_);
   }
-  if (execution) {
-    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena =
-      reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(execution)->GetArena();
+
+  if (value != nullptr) {
+    ::google::protobuf::Arena* submessage_arena = reinterpret_cast<::google::protobuf::MessageLite*>(value)->GetArena();
     if (message_arena != submessage_arena) {
-      execution = ::PROTOBUF_NAMESPACE_ID::internal::GetOwnedMessage(
-          message_arena, execution, submessage_arena);
+      value = ::google::protobuf::internal::GetOwnedMessage(message_arena, value, submessage_arena);
     }
-    _has_bits_[0] |= 0x00000002u;
+    _impl_._has_bits_[0] |= 0x00000002u;
   } else {
-    _has_bits_[0] &= ~0x00000002u;
+    _impl_._has_bits_[0] &= ~0x00000002u;
   }
-  execution_ = execution;
+
+  _impl_.execution_ = reinterpret_cast<::protobuf::Execution*>(value);
   // @@protoc_insertion_point(field_set_allocated:protobuf.ExecutionDetails.execution)
 }
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic pop
+#pragma GCC diagnostic pop
 #endif  // __GNUC__
 
 // @@protoc_insertion_point(namespace_scope)
-
 }  // namespace protobuf
+
 
 // @@protoc_insertion_point(global_scope)
 
-#include <google/protobuf/port_undef.inc>
-#endif  // GOOGLE_PROTOBUF_INCLUDED_GOOGLE_PROTOBUF_INCLUDED_ExecutionDetails_2eproto
+#include "google/protobuf/port_undef.inc"
+
+#endif  // ExecutionDetails_2eproto_2epb_2eh
