@@ -1,314 +1,386 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: GlobalCancelRequest.proto
+// Protobuf C++ Version: 5.29.3
 
-#ifndef GOOGLE_PROTOBUF_INCLUDED_GlobalCancelRequest_2eproto
-#define GOOGLE_PROTOBUF_INCLUDED_GlobalCancelRequest_2eproto
+#ifndef GlobalCancelRequest_2eproto_2epb_2eh
+#define GlobalCancelRequest_2eproto_2epb_2eh
 
 #include <limits>
 #include <string>
+#include <type_traits>
+#include <utility>
 
-#include <google/protobuf/port_def.inc>
-#if PROTOBUF_VERSION < 3012000
-#error This file was generated by a newer version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please update
-#error your headers.
+#include "google/protobuf/runtime_version.h"
+#if PROTOBUF_VERSION != 5029003
+#error "Protobuf C++ gencode is built with an incompatible version of"
+#error "Protobuf C++ headers/runtime. See"
+#error "https://protobuf.dev/support/cross-version-runtime-guarantee/#cpp"
 #endif
-#if 3012004 < PROTOBUF_MIN_PROTOC_VERSION
-#error This file was generated by an older version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please
-#error regenerate this file with a newer version of protoc.
-#endif
-
-#include <google/protobuf/port_undef.inc>
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/arena.h>
-#include <google/protobuf/arenastring.h>
-#include <google/protobuf/generated_message_table_driven.h>
-#include <google/protobuf/generated_message_util.h>
-#include <google/protobuf/inlined_string_field.h>
-#include <google/protobuf/metadata_lite.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/message.h>
-#include <google/protobuf/repeated_field.h>  // IWYU pragma: export
-#include <google/protobuf/extension_set.h>  // IWYU pragma: export
-#include <google/protobuf/unknown_field_set.h>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/arena.h"
+#include "google/protobuf/arenastring.h"
+#include "google/protobuf/generated_message_tctable_decl.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/metadata_lite.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/message.h"
+#include "google/protobuf/message_lite.h"
+#include "google/protobuf/repeated_field.h"  // IWYU pragma: export
+#include "google/protobuf/extension_set.h"  // IWYU pragma: export
+#include "google/protobuf/unknown_field_set.h"
 #include "OrderCancel.pb.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
+
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+
 #define PROTOBUF_INTERNAL_EXPORT_GlobalCancelRequest_2eproto
-PROTOBUF_NAMESPACE_OPEN
+
+namespace google {
+namespace protobuf {
 namespace internal {
-class AnyMetadata;
+template <typename T>
+::absl::string_view GetAnyMessageName();
 }  // namespace internal
-PROTOBUF_NAMESPACE_CLOSE
+}  // namespace protobuf
+}  // namespace google
 
 // Internal implementation detail -- do not use these members.
 struct TableStruct_GlobalCancelRequest_2eproto {
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTableField entries[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxillaryParseTableField aux[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTable schema[1]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::FieldMetadata field_metadata[];
-  static const ::PROTOBUF_NAMESPACE_ID::internal::SerializationTable serialization_table[];
-  static const ::PROTOBUF_NAMESPACE_ID::uint32 offsets[];
+  static const ::uint32_t offsets[];
 };
-extern const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_GlobalCancelRequest_2eproto;
+extern const ::google::protobuf::internal::DescriptorTable
+    descriptor_table_GlobalCancelRequest_2eproto;
 namespace protobuf {
 class GlobalCancelRequest;
-class GlobalCancelRequestDefaultTypeInternal;
+struct GlobalCancelRequestDefaultTypeInternal;
 extern GlobalCancelRequestDefaultTypeInternal _GlobalCancelRequest_default_instance_;
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> ::protobuf::GlobalCancelRequest* Arena::CreateMaybeMessage<::protobuf::GlobalCancelRequest>(Arena*);
-PROTOBUF_NAMESPACE_CLOSE
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
+
 namespace protobuf {
 
 // ===================================================================
 
-class GlobalCancelRequest PROTOBUF_FINAL :
-    public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:protobuf.GlobalCancelRequest) */ {
+
+// -------------------------------------------------------------------
+
+class GlobalCancelRequest final : public ::google::protobuf::Message
+/* @@protoc_insertion_point(class_definition:protobuf.GlobalCancelRequest) */ {
  public:
-  inline GlobalCancelRequest() : GlobalCancelRequest(nullptr) {};
-  virtual ~GlobalCancelRequest();
+  inline GlobalCancelRequest() : GlobalCancelRequest(nullptr) {}
+  ~GlobalCancelRequest() PROTOBUF_FINAL;
 
-  GlobalCancelRequest(const GlobalCancelRequest& from);
-  GlobalCancelRequest(GlobalCancelRequest&& from) noexcept
-    : GlobalCancelRequest() {
-    *this = ::std::move(from);
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+  void operator delete(GlobalCancelRequest* msg, std::destroying_delete_t) {
+    SharedDtor(*msg);
+    ::google::protobuf::internal::SizedDelete(msg, sizeof(GlobalCancelRequest));
   }
+#endif
 
+  template <typename = void>
+  explicit PROTOBUF_CONSTEXPR GlobalCancelRequest(
+      ::google::protobuf::internal::ConstantInitialized);
+
+  inline GlobalCancelRequest(const GlobalCancelRequest& from) : GlobalCancelRequest(nullptr, from) {}
+  inline GlobalCancelRequest(GlobalCancelRequest&& from) noexcept
+      : GlobalCancelRequest(nullptr, std::move(from)) {}
   inline GlobalCancelRequest& operator=(const GlobalCancelRequest& from) {
     CopyFrom(from);
     return *this;
   }
   inline GlobalCancelRequest& operator=(GlobalCancelRequest&& from) noexcept {
-    if (GetArena() == from.GetArena()) {
-      if (this != &from) InternalSwap(&from);
+    if (this == &from) return *this;
+    if (::google::protobuf::internal::CanMoveWithInternalSwap(GetArena(), from.GetArena())) {
+      InternalSwap(&from);
     } else {
       CopyFrom(from);
     }
     return *this;
   }
 
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* descriptor() {
+  inline const ::google::protobuf::UnknownFieldSet& unknown_fields() const
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance);
+  }
+  inline ::google::protobuf::UnknownFieldSet* mutable_unknown_fields()
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.mutable_unknown_fields<::google::protobuf::UnknownFieldSet>();
+  }
+
+  static const ::google::protobuf::Descriptor* descriptor() {
     return GetDescriptor();
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* GetDescriptor() {
-    return GetMetadataStatic().descriptor;
+  static const ::google::protobuf::Descriptor* GetDescriptor() {
+    return default_instance().GetMetadata().descriptor;
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Reflection* GetReflection() {
-    return GetMetadataStatic().reflection;
+  static const ::google::protobuf::Reflection* GetReflection() {
+    return default_instance().GetMetadata().reflection;
   }
-  static const GlobalCancelRequest& default_instance();
-
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
+  static const GlobalCancelRequest& default_instance() {
+    return *internal_default_instance();
+  }
   static inline const GlobalCancelRequest* internal_default_instance() {
     return reinterpret_cast<const GlobalCancelRequest*>(
-               &_GlobalCancelRequest_default_instance_);
+        &_GlobalCancelRequest_default_instance_);
   }
-  static constexpr int kIndexInFileMessages =
-    0;
-
-  friend void swap(GlobalCancelRequest& a, GlobalCancelRequest& b) {
-    a.Swap(&b);
-  }
+  static constexpr int kIndexInFileMessages = 0;
+  friend void swap(GlobalCancelRequest& a, GlobalCancelRequest& b) { a.Swap(&b); }
   inline void Swap(GlobalCancelRequest* other) {
     if (other == this) return;
-    if (GetArena() == other->GetArena()) {
+    if (::google::protobuf::internal::CanUseInternalSwap(GetArena(), other->GetArena())) {
       InternalSwap(other);
     } else {
-      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+      ::google::protobuf::internal::GenericSwap(this, other);
     }
   }
   void UnsafeArenaSwap(GlobalCancelRequest* other) {
     if (other == this) return;
-    GOOGLE_DCHECK(GetArena() == other->GetArena());
+    ABSL_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
   // implements Message ----------------------------------------------
 
-  inline GlobalCancelRequest* New() const final {
-    return CreateMaybeMessage<GlobalCancelRequest>(nullptr);
+  GlobalCancelRequest* New(::google::protobuf::Arena* arena = nullptr) const {
+    return ::google::protobuf::Message::DefaultConstruct<GlobalCancelRequest>(arena);
   }
-
-  GlobalCancelRequest* New(::PROTOBUF_NAMESPACE_ID::Arena* arena) const final {
-    return CreateMaybeMessage<GlobalCancelRequest>(arena);
-  }
-  void CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
-  void MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
+  using ::google::protobuf::Message::CopyFrom;
   void CopyFrom(const GlobalCancelRequest& from);
-  void MergeFrom(const GlobalCancelRequest& from);
-  PROTOBUF_ATTRIBUTE_REINITIALIZES void Clear() final;
-  bool IsInitialized() const final;
-
-  size_t ByteSizeLong() const final;
-  const char* _InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) final;
-  ::PROTOBUF_NAMESPACE_ID::uint8* _InternalSerialize(
-      ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const final;
-  int GetCachedSize() const final { return _cached_size_.Get(); }
+  using ::google::protobuf::Message::MergeFrom;
+  void MergeFrom(const GlobalCancelRequest& from) { GlobalCancelRequest::MergeImpl(*this, from); }
 
   private:
-  inline void SharedCtor();
-  inline void SharedDtor();
-  void SetCachedSize(int size) const final;
+  static void MergeImpl(
+      ::google::protobuf::MessageLite& to_msg,
+      const ::google::protobuf::MessageLite& from_msg);
+
+  public:
+  bool IsInitialized() const {
+    return true;
+  }
+  ABSL_ATTRIBUTE_REINITIALIZES void Clear() PROTOBUF_FINAL;
+  #if defined(PROTOBUF_CUSTOM_VTABLE)
+  private:
+  static ::size_t ByteSizeLong(const ::google::protobuf::MessageLite& msg);
+  static ::uint8_t* _InternalSerialize(
+      const MessageLite& msg, ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream);
+
+  public:
+  ::size_t ByteSizeLong() const { return ByteSizeLong(*this); }
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+    return _InternalSerialize(*this, target, stream);
+  }
+  #else   // PROTOBUF_CUSTOM_VTABLE
+  ::size_t ByteSizeLong() const final;
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const final;
+  #endif  // PROTOBUF_CUSTOM_VTABLE
+  int GetCachedSize() const { return _impl_._cached_size_.Get(); }
+
+  private:
+  void SharedCtor(::google::protobuf::Arena* arena);
+  static void SharedDtor(MessageLite& self);
   void InternalSwap(GlobalCancelRequest* other);
-  friend class ::PROTOBUF_NAMESPACE_ID::internal::AnyMetadata;
-  static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
-    return "protobuf.GlobalCancelRequest";
+ private:
+  template <typename T>
+  friend ::absl::string_view(
+      ::google::protobuf::internal::GetAnyMessageName)();
+  static ::absl::string_view FullMessageName() { return "protobuf.GlobalCancelRequest"; }
+
+ protected:
+  explicit GlobalCancelRequest(::google::protobuf::Arena* arena);
+  GlobalCancelRequest(::google::protobuf::Arena* arena, const GlobalCancelRequest& from);
+  GlobalCancelRequest(::google::protobuf::Arena* arena, GlobalCancelRequest&& from) noexcept
+      : GlobalCancelRequest(arena) {
+    *this = ::std::move(from);
   }
-  protected:
-  explicit GlobalCancelRequest(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  private:
-  static void ArenaDtor(void* object);
-  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  public:
+  const ::google::protobuf::internal::ClassData* GetClassData() const PROTOBUF_FINAL;
+  static void* PlacementNew_(const void*, void* mem,
+                             ::google::protobuf::Arena* arena);
+  static constexpr auto InternalNewImpl_();
+  static const ::google::protobuf::internal::ClassDataFull _class_data_;
 
-  ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
-  private:
-  static ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadataStatic() {
-    ::PROTOBUF_NAMESPACE_ID::internal::AssignDescriptors(&::descriptor_table_GlobalCancelRequest_2eproto);
-    return ::descriptor_table_GlobalCancelRequest_2eproto.file_level_metadata[kIndexInFileMessages];
-  }
-
-  public:
-
+ public:
+  ::google::protobuf::Metadata GetMetadata() const;
   // nested types ----------------------------------------------------
 
   // accessors -------------------------------------------------------
-
   enum : int {
     kOrderCancelFieldNumber = 1,
   };
-  // .protobuf.OrderCancel orderCancel = 1;
+  // optional .protobuf.OrderCancel orderCancel = 1;
   bool has_ordercancel() const;
-  private:
-  bool _internal_has_ordercancel() const;
-  public:
-  void clear_ordercancel();
+  void clear_ordercancel() ;
   const ::protobuf::OrderCancel& ordercancel() const;
-  ::protobuf::OrderCancel* release_ordercancel();
+  PROTOBUF_NODISCARD ::protobuf::OrderCancel* release_ordercancel();
   ::protobuf::OrderCancel* mutable_ordercancel();
-  void set_allocated_ordercancel(::protobuf::OrderCancel* ordercancel);
+  void set_allocated_ordercancel(::protobuf::OrderCancel* value);
+  void unsafe_arena_set_allocated_ordercancel(::protobuf::OrderCancel* value);
+  ::protobuf::OrderCancel* unsafe_arena_release_ordercancel();
+
   private:
   const ::protobuf::OrderCancel& _internal_ordercancel() const;
   ::protobuf::OrderCancel* _internal_mutable_ordercancel();
-  public:
-  void unsafe_arena_set_allocated_ordercancel(
-      ::protobuf::OrderCancel* ordercancel);
-  ::protobuf::OrderCancel* unsafe_arena_release_ordercancel();
 
+  public:
   // @@protoc_insertion_point(class_scope:protobuf.GlobalCancelRequest)
  private:
   class _Internal;
+  friend class ::google::protobuf::internal::TcParser;
+  static const ::google::protobuf::internal::TcParseTable<
+      0, 1, 1,
+      0, 2>
+      _table_;
 
-  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
-  typedef void InternalArenaConstructable_;
-  typedef void DestructorSkippable_;
-  ::PROTOBUF_NAMESPACE_ID::internal::HasBits<1> _has_bits_;
-  mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
-  ::protobuf::OrderCancel* ordercancel_;
+  friend class ::google::protobuf::MessageLite;
+  friend class ::google::protobuf::Arena;
+  template <typename T>
+  friend class ::google::protobuf::Arena::InternalHelper;
+  using InternalArenaConstructable_ = void;
+  using DestructorSkippable_ = void;
+  struct Impl_ {
+    inline explicit constexpr Impl_(
+        ::google::protobuf::internal::ConstantInitialized) noexcept;
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena);
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena, const Impl_& from,
+                          const GlobalCancelRequest& from_msg);
+    ::google::protobuf::internal::HasBits<1> _has_bits_;
+    ::google::protobuf::internal::CachedSize _cached_size_;
+    ::protobuf::OrderCancel* ordercancel_;
+    PROTOBUF_TSAN_DECLARE_MEMBER
+  };
+  union { Impl_ _impl_; };
   friend struct ::TableStruct_GlobalCancelRequest_2eproto;
 };
+
 // ===================================================================
 
 
+
+
 // ===================================================================
+
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic push
-  #pragma GCC diagnostic ignored "-Wstrict-aliasing"
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wstrict-aliasing"
 #endif  // __GNUC__
+// -------------------------------------------------------------------
+
 // GlobalCancelRequest
 
-// .protobuf.OrderCancel orderCancel = 1;
-inline bool GlobalCancelRequest::_internal_has_ordercancel() const {
-  bool value = (_has_bits_[0] & 0x00000001u) != 0;
-  PROTOBUF_ASSUME(!value || ordercancel_ != nullptr);
+// optional .protobuf.OrderCancel orderCancel = 1;
+inline bool GlobalCancelRequest::has_ordercancel() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000001u) != 0;
+  PROTOBUF_ASSUME(!value || _impl_.ordercancel_ != nullptr);
   return value;
 }
-inline bool GlobalCancelRequest::has_ordercancel() const {
-  return _internal_has_ordercancel();
-}
 inline const ::protobuf::OrderCancel& GlobalCancelRequest::_internal_ordercancel() const {
-  const ::protobuf::OrderCancel* p = ordercancel_;
-  return p != nullptr ? *p : *reinterpret_cast<const ::protobuf::OrderCancel*>(
-      &::protobuf::_OrderCancel_default_instance_);
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  const ::protobuf::OrderCancel* p = _impl_.ordercancel_;
+  return p != nullptr ? *p : reinterpret_cast<const ::protobuf::OrderCancel&>(::protobuf::_OrderCancel_default_instance_);
 }
-inline const ::protobuf::OrderCancel& GlobalCancelRequest::ordercancel() const {
+inline const ::protobuf::OrderCancel& GlobalCancelRequest::ordercancel() const ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.GlobalCancelRequest.orderCancel)
   return _internal_ordercancel();
 }
-inline void GlobalCancelRequest::unsafe_arena_set_allocated_ordercancel(
-    ::protobuf::OrderCancel* ordercancel) {
+inline void GlobalCancelRequest::unsafe_arena_set_allocated_ordercancel(::protobuf::OrderCancel* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (GetArena() == nullptr) {
-    delete reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(ordercancel_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.ordercancel_);
   }
-  ordercancel_ = ordercancel;
-  if (ordercancel) {
-    _has_bits_[0] |= 0x00000001u;
+  _impl_.ordercancel_ = reinterpret_cast<::protobuf::OrderCancel*>(value);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000001u;
   } else {
-    _has_bits_[0] &= ~0x00000001u;
+    _impl_._has_bits_[0] &= ~0x00000001u;
   }
   // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.GlobalCancelRequest.orderCancel)
 }
 inline ::protobuf::OrderCancel* GlobalCancelRequest::release_ordercancel() {
-  auto temp = unsafe_arena_release_ordercancel();
-  if (GetArena() != nullptr) {
-    temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+
+  _impl_._has_bits_[0] &= ~0x00000001u;
+  ::protobuf::OrderCancel* released = _impl_.ordercancel_;
+  _impl_.ordercancel_ = nullptr;
+  if (::google::protobuf::internal::DebugHardenForceCopyInRelease()) {
+    auto* old = reinterpret_cast<::google::protobuf::MessageLite*>(released);
+    released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    if (GetArena() == nullptr) {
+      delete old;
+    }
+  } else {
+    if (GetArena() != nullptr) {
+      released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    }
   }
-  return temp;
+  return released;
 }
 inline ::protobuf::OrderCancel* GlobalCancelRequest::unsafe_arena_release_ordercancel() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.GlobalCancelRequest.orderCancel)
-  _has_bits_[0] &= ~0x00000001u;
-  ::protobuf::OrderCancel* temp = ordercancel_;
-  ordercancel_ = nullptr;
+
+  _impl_._has_bits_[0] &= ~0x00000001u;
+  ::protobuf::OrderCancel* temp = _impl_.ordercancel_;
+  _impl_.ordercancel_ = nullptr;
   return temp;
 }
 inline ::protobuf::OrderCancel* GlobalCancelRequest::_internal_mutable_ordercancel() {
-  _has_bits_[0] |= 0x00000001u;
-  if (ordercancel_ == nullptr) {
-    auto* p = CreateMaybeMessage<::protobuf::OrderCancel>(GetArena());
-    ordercancel_ = p;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (_impl_.ordercancel_ == nullptr) {
+    auto* p = ::google::protobuf::Message::DefaultConstruct<::protobuf::OrderCancel>(GetArena());
+    _impl_.ordercancel_ = reinterpret_cast<::protobuf::OrderCancel*>(p);
   }
-  return ordercancel_;
+  return _impl_.ordercancel_;
 }
-inline ::protobuf::OrderCancel* GlobalCancelRequest::mutable_ordercancel() {
+inline ::protobuf::OrderCancel* GlobalCancelRequest::mutable_ordercancel() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  _impl_._has_bits_[0] |= 0x00000001u;
+  ::protobuf::OrderCancel* _msg = _internal_mutable_ordercancel();
   // @@protoc_insertion_point(field_mutable:protobuf.GlobalCancelRequest.orderCancel)
-  return _internal_mutable_ordercancel();
+  return _msg;
 }
-inline void GlobalCancelRequest::set_allocated_ordercancel(::protobuf::OrderCancel* ordercancel) {
-  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArena();
+inline void GlobalCancelRequest::set_allocated_ordercancel(::protobuf::OrderCancel* value) {
+  ::google::protobuf::Arena* message_arena = GetArena();
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (message_arena == nullptr) {
-    delete reinterpret_cast< ::PROTOBUF_NAMESPACE_ID::MessageLite*>(ordercancel_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.ordercancel_);
   }
-  if (ordercancel) {
-    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena =
-      reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(ordercancel)->GetArena();
+
+  if (value != nullptr) {
+    ::google::protobuf::Arena* submessage_arena = reinterpret_cast<::google::protobuf::MessageLite*>(value)->GetArena();
     if (message_arena != submessage_arena) {
-      ordercancel = ::PROTOBUF_NAMESPACE_ID::internal::GetOwnedMessage(
-          message_arena, ordercancel, submessage_arena);
+      value = ::google::protobuf::internal::GetOwnedMessage(message_arena, value, submessage_arena);
     }
-    _has_bits_[0] |= 0x00000001u;
+    _impl_._has_bits_[0] |= 0x00000001u;
   } else {
-    _has_bits_[0] &= ~0x00000001u;
+    _impl_._has_bits_[0] &= ~0x00000001u;
   }
-  ordercancel_ = ordercancel;
+
+  _impl_.ordercancel_ = reinterpret_cast<::protobuf::OrderCancel*>(value);
   // @@protoc_insertion_point(field_set_allocated:protobuf.GlobalCancelRequest.orderCancel)
 }
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic pop
+#pragma GCC diagnostic pop
 #endif  // __GNUC__
 
 // @@protoc_insertion_point(namespace_scope)
-
 }  // namespace protobuf
+
 
 // @@protoc_insertion_point(global_scope)
 
-#include <google/protobuf/port_undef.inc>
-#endif  // GOOGLE_PROTOBUF_INCLUDED_GOOGLE_PROTOBUF_INCLUDED_GlobalCancelRequest_2eproto
+#include "google/protobuf/port_undef.inc"
+
+#endif  // GlobalCancelRequest_2eproto_2epb_2eh
