@@ -1,553 +1,645 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: OpenOrder.proto
+// Protobuf C++ Version: 5.29.3
 
-#ifndef GOOGLE_PROTOBUF_INCLUDED_OpenOrder_2eproto
-#define GOOGLE_PROTOBUF_INCLUDED_OpenOrder_2eproto
+#ifndef OpenOrder_2eproto_2epb_2eh
+#define OpenOrder_2eproto_2epb_2eh
 
 #include <limits>
 #include <string>
+#include <type_traits>
+#include <utility>
 
-#include <google/protobuf/port_def.inc>
-#if PROTOBUF_VERSION < 3012000
-#error This file was generated by a newer version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please update
-#error your headers.
+#include "google/protobuf/runtime_version.h"
+#if PROTOBUF_VERSION != 5029003
+#error "Protobuf C++ gencode is built with an incompatible version of"
+#error "Protobuf C++ headers/runtime. See"
+#error "https://protobuf.dev/support/cross-version-runtime-guarantee/#cpp"
 #endif
-#if 3012004 < PROTOBUF_MIN_PROTOC_VERSION
-#error This file was generated by an older version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please
-#error regenerate this file with a newer version of protoc.
-#endif
-
-#include <google/protobuf/port_undef.inc>
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/arena.h>
-#include <google/protobuf/arenastring.h>
-#include <google/protobuf/generated_message_table_driven.h>
-#include <google/protobuf/generated_message_util.h>
-#include <google/protobuf/inlined_string_field.h>
-#include <google/protobuf/metadata_lite.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/message.h>
-#include <google/protobuf/repeated_field.h>  // IWYU pragma: export
-#include <google/protobuf/extension_set.h>  // IWYU pragma: export
-#include <google/protobuf/unknown_field_set.h>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/arena.h"
+#include "google/protobuf/arenastring.h"
+#include "google/protobuf/generated_message_tctable_decl.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/metadata_lite.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/message.h"
+#include "google/protobuf/message_lite.h"
+#include "google/protobuf/repeated_field.h"  // IWYU pragma: export
+#include "google/protobuf/extension_set.h"  // IWYU pragma: export
+#include "google/protobuf/unknown_field_set.h"
 #include "Contract.pb.h"
 #include "Order.pb.h"
 #include "OrderState.pb.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
+
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+
 #define PROTOBUF_INTERNAL_EXPORT_OpenOrder_2eproto
-PROTOBUF_NAMESPACE_OPEN
+
+namespace google {
+namespace protobuf {
 namespace internal {
-class AnyMetadata;
+template <typename T>
+::absl::string_view GetAnyMessageName();
 }  // namespace internal
-PROTOBUF_NAMESPACE_CLOSE
+}  // namespace protobuf
+}  // namespace google
 
 // Internal implementation detail -- do not use these members.
 struct TableStruct_OpenOrder_2eproto {
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTableField entries[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxillaryParseTableField aux[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTable schema[1]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::FieldMetadata field_metadata[];
-  static const ::PROTOBUF_NAMESPACE_ID::internal::SerializationTable serialization_table[];
-  static const ::PROTOBUF_NAMESPACE_ID::uint32 offsets[];
+  static const ::uint32_t offsets[];
 };
-extern const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_OpenOrder_2eproto;
+extern const ::google::protobuf::internal::DescriptorTable
+    descriptor_table_OpenOrder_2eproto;
 namespace protobuf {
 class OpenOrder;
-class OpenOrderDefaultTypeInternal;
+struct OpenOrderDefaultTypeInternal;
 extern OpenOrderDefaultTypeInternal _OpenOrder_default_instance_;
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> ::protobuf::OpenOrder* Arena::CreateMaybeMessage<::protobuf::OpenOrder>(Arena*);
-PROTOBUF_NAMESPACE_CLOSE
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
+
 namespace protobuf {
 
 // ===================================================================
 
-class OpenOrder PROTOBUF_FINAL :
-    public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:protobuf.OpenOrder) */ {
+
+// -------------------------------------------------------------------
+
+class OpenOrder final : public ::google::protobuf::Message
+/* @@protoc_insertion_point(class_definition:protobuf.OpenOrder) */ {
  public:
-  inline OpenOrder() : OpenOrder(nullptr) {};
-  virtual ~OpenOrder();
+  inline OpenOrder() : OpenOrder(nullptr) {}
+  ~OpenOrder() PROTOBUF_FINAL;
 
-  OpenOrder(const OpenOrder& from);
-  OpenOrder(OpenOrder&& from) noexcept
-    : OpenOrder() {
-    *this = ::std::move(from);
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+  void operator delete(OpenOrder* msg, std::destroying_delete_t) {
+    SharedDtor(*msg);
+    ::google::protobuf::internal::SizedDelete(msg, sizeof(OpenOrder));
   }
+#endif
 
+  template <typename = void>
+  explicit PROTOBUF_CONSTEXPR OpenOrder(
+      ::google::protobuf::internal::ConstantInitialized);
+
+  inline OpenOrder(const OpenOrder& from) : OpenOrder(nullptr, from) {}
+  inline OpenOrder(OpenOrder&& from) noexcept
+      : OpenOrder(nullptr, std::move(from)) {}
   inline OpenOrder& operator=(const OpenOrder& from) {
     CopyFrom(from);
     return *this;
   }
   inline OpenOrder& operator=(OpenOrder&& from) noexcept {
-    if (GetArena() == from.GetArena()) {
-      if (this != &from) InternalSwap(&from);
+    if (this == &from) return *this;
+    if (::google::protobuf::internal::CanMoveWithInternalSwap(GetArena(), from.GetArena())) {
+      InternalSwap(&from);
     } else {
       CopyFrom(from);
     }
     return *this;
   }
 
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* descriptor() {
+  inline const ::google::protobuf::UnknownFieldSet& unknown_fields() const
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance);
+  }
+  inline ::google::protobuf::UnknownFieldSet* mutable_unknown_fields()
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.mutable_unknown_fields<::google::protobuf::UnknownFieldSet>();
+  }
+
+  static const ::google::protobuf::Descriptor* descriptor() {
     return GetDescriptor();
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* GetDescriptor() {
-    return GetMetadataStatic().descriptor;
+  static const ::google::protobuf::Descriptor* GetDescriptor() {
+    return default_instance().GetMetadata().descriptor;
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Reflection* GetReflection() {
-    return GetMetadataStatic().reflection;
+  static const ::google::protobuf::Reflection* GetReflection() {
+    return default_instance().GetMetadata().reflection;
   }
-  static const OpenOrder& default_instance();
-
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
+  static const OpenOrder& default_instance() {
+    return *internal_default_instance();
+  }
   static inline const OpenOrder* internal_default_instance() {
     return reinterpret_cast<const OpenOrder*>(
-               &_OpenOrder_default_instance_);
+        &_OpenOrder_default_instance_);
   }
-  static constexpr int kIndexInFileMessages =
-    0;
-
-  friend void swap(OpenOrder& a, OpenOrder& b) {
-    a.Swap(&b);
-  }
+  static constexpr int kIndexInFileMessages = 0;
+  friend void swap(OpenOrder& a, OpenOrder& b) { a.Swap(&b); }
   inline void Swap(OpenOrder* other) {
     if (other == this) return;
-    if (GetArena() == other->GetArena()) {
+    if (::google::protobuf::internal::CanUseInternalSwap(GetArena(), other->GetArena())) {
       InternalSwap(other);
     } else {
-      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+      ::google::protobuf::internal::GenericSwap(this, other);
     }
   }
   void UnsafeArenaSwap(OpenOrder* other) {
     if (other == this) return;
-    GOOGLE_DCHECK(GetArena() == other->GetArena());
+    ABSL_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
   // implements Message ----------------------------------------------
 
-  inline OpenOrder* New() const final {
-    return CreateMaybeMessage<OpenOrder>(nullptr);
+  OpenOrder* New(::google::protobuf::Arena* arena = nullptr) const {
+    return ::google::protobuf::Message::DefaultConstruct<OpenOrder>(arena);
   }
-
-  OpenOrder* New(::PROTOBUF_NAMESPACE_ID::Arena* arena) const final {
-    return CreateMaybeMessage<OpenOrder>(arena);
-  }
-  void CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
-  void MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
+  using ::google::protobuf::Message::CopyFrom;
   void CopyFrom(const OpenOrder& from);
-  void MergeFrom(const OpenOrder& from);
-  PROTOBUF_ATTRIBUTE_REINITIALIZES void Clear() final;
-  bool IsInitialized() const final;
-
-  size_t ByteSizeLong() const final;
-  const char* _InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) final;
-  ::PROTOBUF_NAMESPACE_ID::uint8* _InternalSerialize(
-      ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const final;
-  int GetCachedSize() const final { return _cached_size_.Get(); }
+  using ::google::protobuf::Message::MergeFrom;
+  void MergeFrom(const OpenOrder& from) { OpenOrder::MergeImpl(*this, from); }
 
   private:
-  inline void SharedCtor();
-  inline void SharedDtor();
-  void SetCachedSize(int size) const final;
+  static void MergeImpl(
+      ::google::protobuf::MessageLite& to_msg,
+      const ::google::protobuf::MessageLite& from_msg);
+
+  public:
+  bool IsInitialized() const {
+    return true;
+  }
+  ABSL_ATTRIBUTE_REINITIALIZES void Clear() PROTOBUF_FINAL;
+  #if defined(PROTOBUF_CUSTOM_VTABLE)
+  private:
+  static ::size_t ByteSizeLong(const ::google::protobuf::MessageLite& msg);
+  static ::uint8_t* _InternalSerialize(
+      const MessageLite& msg, ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream);
+
+  public:
+  ::size_t ByteSizeLong() const { return ByteSizeLong(*this); }
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+    return _InternalSerialize(*this, target, stream);
+  }
+  #else   // PROTOBUF_CUSTOM_VTABLE
+  ::size_t ByteSizeLong() const final;
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const final;
+  #endif  // PROTOBUF_CUSTOM_VTABLE
+  int GetCachedSize() const { return _impl_._cached_size_.Get(); }
+
+  private:
+  void SharedCtor(::google::protobuf::Arena* arena);
+  static void SharedDtor(MessageLite& self);
   void InternalSwap(OpenOrder* other);
-  friend class ::PROTOBUF_NAMESPACE_ID::internal::AnyMetadata;
-  static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
-    return "protobuf.OpenOrder";
+ private:
+  template <typename T>
+  friend ::absl::string_view(
+      ::google::protobuf::internal::GetAnyMessageName)();
+  static ::absl::string_view FullMessageName() { return "protobuf.OpenOrder"; }
+
+ protected:
+  explicit OpenOrder(::google::protobuf::Arena* arena);
+  OpenOrder(::google::protobuf::Arena* arena, const OpenOrder& from);
+  OpenOrder(::google::protobuf::Arena* arena, OpenOrder&& from) noexcept
+      : OpenOrder(arena) {
+    *this = ::std::move(from);
   }
-  protected:
-  explicit OpenOrder(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  private:
-  static void ArenaDtor(void* object);
-  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  public:
+  const ::google::protobuf::internal::ClassData* GetClassData() const PROTOBUF_FINAL;
+  static void* PlacementNew_(const void*, void* mem,
+                             ::google::protobuf::Arena* arena);
+  static constexpr auto InternalNewImpl_();
+  static const ::google::protobuf::internal::ClassDataFull _class_data_;
 
-  ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
-  private:
-  static ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadataStatic() {
-    ::PROTOBUF_NAMESPACE_ID::internal::AssignDescriptors(&::descriptor_table_OpenOrder_2eproto);
-    return ::descriptor_table_OpenOrder_2eproto.file_level_metadata[kIndexInFileMessages];
-  }
-
-  public:
-
+ public:
+  ::google::protobuf::Metadata GetMetadata() const;
   // nested types ----------------------------------------------------
 
   // accessors -------------------------------------------------------
-
   enum : int {
     kContractFieldNumber = 2,
     kOrderFieldNumber = 3,
     kOrderStateFieldNumber = 4,
     kOrderIdFieldNumber = 1,
   };
-  // .protobuf.Contract contract = 2;
+  // optional .protobuf.Contract contract = 2;
   bool has_contract() const;
-  private:
-  bool _internal_has_contract() const;
-  public:
-  void clear_contract();
+  void clear_contract() ;
   const ::protobuf::Contract& contract() const;
-  ::protobuf::Contract* release_contract();
+  PROTOBUF_NODISCARD ::protobuf::Contract* release_contract();
   ::protobuf::Contract* mutable_contract();
-  void set_allocated_contract(::protobuf::Contract* contract);
+  void set_allocated_contract(::protobuf::Contract* value);
+  void unsafe_arena_set_allocated_contract(::protobuf::Contract* value);
+  ::protobuf::Contract* unsafe_arena_release_contract();
+
   private:
   const ::protobuf::Contract& _internal_contract() const;
   ::protobuf::Contract* _internal_mutable_contract();
-  public:
-  void unsafe_arena_set_allocated_contract(
-      ::protobuf::Contract* contract);
-  ::protobuf::Contract* unsafe_arena_release_contract();
 
-  // .protobuf.Order order = 3;
-  bool has_order() const;
-  private:
-  bool _internal_has_order() const;
   public:
-  void clear_order();
+  // optional .protobuf.Order order = 3;
+  bool has_order() const;
+  void clear_order() ;
   const ::protobuf::Order& order() const;
-  ::protobuf::Order* release_order();
+  PROTOBUF_NODISCARD ::protobuf::Order* release_order();
   ::protobuf::Order* mutable_order();
-  void set_allocated_order(::protobuf::Order* order);
+  void set_allocated_order(::protobuf::Order* value);
+  void unsafe_arena_set_allocated_order(::protobuf::Order* value);
+  ::protobuf::Order* unsafe_arena_release_order();
+
   private:
   const ::protobuf::Order& _internal_order() const;
   ::protobuf::Order* _internal_mutable_order();
-  public:
-  void unsafe_arena_set_allocated_order(
-      ::protobuf::Order* order);
-  ::protobuf::Order* unsafe_arena_release_order();
 
-  // .protobuf.OrderState orderState = 4;
-  bool has_orderstate() const;
-  private:
-  bool _internal_has_orderstate() const;
   public:
-  void clear_orderstate();
+  // optional .protobuf.OrderState orderState = 4;
+  bool has_orderstate() const;
+  void clear_orderstate() ;
   const ::protobuf::OrderState& orderstate() const;
-  ::protobuf::OrderState* release_orderstate();
+  PROTOBUF_NODISCARD ::protobuf::OrderState* release_orderstate();
   ::protobuf::OrderState* mutable_orderstate();
-  void set_allocated_orderstate(::protobuf::OrderState* orderstate);
+  void set_allocated_orderstate(::protobuf::OrderState* value);
+  void unsafe_arena_set_allocated_orderstate(::protobuf::OrderState* value);
+  ::protobuf::OrderState* unsafe_arena_release_orderstate();
+
   private:
   const ::protobuf::OrderState& _internal_orderstate() const;
   ::protobuf::OrderState* _internal_mutable_orderstate();
-  public:
-  void unsafe_arena_set_allocated_orderstate(
-      ::protobuf::OrderState* orderstate);
-  ::protobuf::OrderState* unsafe_arena_release_orderstate();
 
-  // int32 orderId = 1;
+  public:
+  // optional int32 orderId = 1;
   bool has_orderid() const;
-  private:
-  bool _internal_has_orderid() const;
-  public:
-  void clear_orderid();
-  ::PROTOBUF_NAMESPACE_ID::int32 orderid() const;
-  void set_orderid(::PROTOBUF_NAMESPACE_ID::int32 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int32 _internal_orderid() const;
-  void _internal_set_orderid(::PROTOBUF_NAMESPACE_ID::int32 value);
-  public:
+  void clear_orderid() ;
+  ::int32_t orderid() const;
+  void set_orderid(::int32_t value);
 
+  private:
+  ::int32_t _internal_orderid() const;
+  void _internal_set_orderid(::int32_t value);
+
+  public:
   // @@protoc_insertion_point(class_scope:protobuf.OpenOrder)
  private:
   class _Internal;
+  friend class ::google::protobuf::internal::TcParser;
+  static const ::google::protobuf::internal::TcParseTable<
+      2, 4, 3,
+      0, 2>
+      _table_;
 
-  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
-  typedef void InternalArenaConstructable_;
-  typedef void DestructorSkippable_;
-  ::PROTOBUF_NAMESPACE_ID::internal::HasBits<1> _has_bits_;
-  mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
-  ::protobuf::Contract* contract_;
-  ::protobuf::Order* order_;
-  ::protobuf::OrderState* orderstate_;
-  ::PROTOBUF_NAMESPACE_ID::int32 orderid_;
+  friend class ::google::protobuf::MessageLite;
+  friend class ::google::protobuf::Arena;
+  template <typename T>
+  friend class ::google::protobuf::Arena::InternalHelper;
+  using InternalArenaConstructable_ = void;
+  using DestructorSkippable_ = void;
+  struct Impl_ {
+    inline explicit constexpr Impl_(
+        ::google::protobuf::internal::ConstantInitialized) noexcept;
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena);
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena, const Impl_& from,
+                          const OpenOrder& from_msg);
+    ::google::protobuf::internal::HasBits<1> _has_bits_;
+    ::google::protobuf::internal::CachedSize _cached_size_;
+    ::protobuf::Contract* contract_;
+    ::protobuf::Order* order_;
+    ::protobuf::OrderState* orderstate_;
+    ::int32_t orderid_;
+    PROTOBUF_TSAN_DECLARE_MEMBER
+  };
+  union { Impl_ _impl_; };
   friend struct ::TableStruct_OpenOrder_2eproto;
 };
+
 // ===================================================================
 
 
+
+
 // ===================================================================
+
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic push
-  #pragma GCC diagnostic ignored "-Wstrict-aliasing"
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wstrict-aliasing"
 #endif  // __GNUC__
+// -------------------------------------------------------------------
+
 // OpenOrder
 
-// int32 orderId = 1;
-inline bool OpenOrder::_internal_has_orderid() const {
-  bool value = (_has_bits_[0] & 0x00000008u) != 0;
+// optional int32 orderId = 1;
+inline bool OpenOrder::has_orderid() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000008u) != 0;
   return value;
 }
-inline bool OpenOrder::has_orderid() const {
-  return _internal_has_orderid();
-}
 inline void OpenOrder::clear_orderid() {
-  orderid_ = 0;
-  _has_bits_[0] &= ~0x00000008u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.orderid_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000008u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int32 OpenOrder::_internal_orderid() const {
-  return orderid_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int32 OpenOrder::orderid() const {
+inline ::int32_t OpenOrder::orderid() const {
   // @@protoc_insertion_point(field_get:protobuf.OpenOrder.orderId)
   return _internal_orderid();
 }
-inline void OpenOrder::_internal_set_orderid(::PROTOBUF_NAMESPACE_ID::int32 value) {
-  _has_bits_[0] |= 0x00000008u;
-  orderid_ = value;
-}
-inline void OpenOrder::set_orderid(::PROTOBUF_NAMESPACE_ID::int32 value) {
+inline void OpenOrder::set_orderid(::int32_t value) {
   _internal_set_orderid(value);
+  _impl_._has_bits_[0] |= 0x00000008u;
   // @@protoc_insertion_point(field_set:protobuf.OpenOrder.orderId)
 }
+inline ::int32_t OpenOrder::_internal_orderid() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.orderid_;
+}
+inline void OpenOrder::_internal_set_orderid(::int32_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.orderid_ = value;
+}
 
-// .protobuf.Contract contract = 2;
-inline bool OpenOrder::_internal_has_contract() const {
-  bool value = (_has_bits_[0] & 0x00000001u) != 0;
-  PROTOBUF_ASSUME(!value || contract_ != nullptr);
+// optional .protobuf.Contract contract = 2;
+inline bool OpenOrder::has_contract() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000001u) != 0;
+  PROTOBUF_ASSUME(!value || _impl_.contract_ != nullptr);
   return value;
 }
-inline bool OpenOrder::has_contract() const {
-  return _internal_has_contract();
-}
 inline const ::protobuf::Contract& OpenOrder::_internal_contract() const {
-  const ::protobuf::Contract* p = contract_;
-  return p != nullptr ? *p : *reinterpret_cast<const ::protobuf::Contract*>(
-      &::protobuf::_Contract_default_instance_);
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  const ::protobuf::Contract* p = _impl_.contract_;
+  return p != nullptr ? *p : reinterpret_cast<const ::protobuf::Contract&>(::protobuf::_Contract_default_instance_);
 }
-inline const ::protobuf::Contract& OpenOrder::contract() const {
+inline const ::protobuf::Contract& OpenOrder::contract() const ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.OpenOrder.contract)
   return _internal_contract();
 }
-inline void OpenOrder::unsafe_arena_set_allocated_contract(
-    ::protobuf::Contract* contract) {
+inline void OpenOrder::unsafe_arena_set_allocated_contract(::protobuf::Contract* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (GetArena() == nullptr) {
-    delete reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(contract_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.contract_);
   }
-  contract_ = contract;
-  if (contract) {
-    _has_bits_[0] |= 0x00000001u;
+  _impl_.contract_ = reinterpret_cast<::protobuf::Contract*>(value);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000001u;
   } else {
-    _has_bits_[0] &= ~0x00000001u;
+    _impl_._has_bits_[0] &= ~0x00000001u;
   }
   // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.OpenOrder.contract)
 }
 inline ::protobuf::Contract* OpenOrder::release_contract() {
-  auto temp = unsafe_arena_release_contract();
-  if (GetArena() != nullptr) {
-    temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+
+  _impl_._has_bits_[0] &= ~0x00000001u;
+  ::protobuf::Contract* released = _impl_.contract_;
+  _impl_.contract_ = nullptr;
+  if (::google::protobuf::internal::DebugHardenForceCopyInRelease()) {
+    auto* old = reinterpret_cast<::google::protobuf::MessageLite*>(released);
+    released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    if (GetArena() == nullptr) {
+      delete old;
+    }
+  } else {
+    if (GetArena() != nullptr) {
+      released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    }
   }
-  return temp;
+  return released;
 }
 inline ::protobuf::Contract* OpenOrder::unsafe_arena_release_contract() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.OpenOrder.contract)
-  _has_bits_[0] &= ~0x00000001u;
-  ::protobuf::Contract* temp = contract_;
-  contract_ = nullptr;
+
+  _impl_._has_bits_[0] &= ~0x00000001u;
+  ::protobuf::Contract* temp = _impl_.contract_;
+  _impl_.contract_ = nullptr;
   return temp;
 }
 inline ::protobuf::Contract* OpenOrder::_internal_mutable_contract() {
-  _has_bits_[0] |= 0x00000001u;
-  if (contract_ == nullptr) {
-    auto* p = CreateMaybeMessage<::protobuf::Contract>(GetArena());
-    contract_ = p;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (_impl_.contract_ == nullptr) {
+    auto* p = ::google::protobuf::Message::DefaultConstruct<::protobuf::Contract>(GetArena());
+    _impl_.contract_ = reinterpret_cast<::protobuf::Contract*>(p);
   }
-  return contract_;
+  return _impl_.contract_;
 }
-inline ::protobuf::Contract* OpenOrder::mutable_contract() {
+inline ::protobuf::Contract* OpenOrder::mutable_contract() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  _impl_._has_bits_[0] |= 0x00000001u;
+  ::protobuf::Contract* _msg = _internal_mutable_contract();
   // @@protoc_insertion_point(field_mutable:protobuf.OpenOrder.contract)
-  return _internal_mutable_contract();
+  return _msg;
 }
-inline void OpenOrder::set_allocated_contract(::protobuf::Contract* contract) {
-  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArena();
+inline void OpenOrder::set_allocated_contract(::protobuf::Contract* value) {
+  ::google::protobuf::Arena* message_arena = GetArena();
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (message_arena == nullptr) {
-    delete reinterpret_cast< ::PROTOBUF_NAMESPACE_ID::MessageLite*>(contract_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.contract_);
   }
-  if (contract) {
-    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena =
-      reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(contract)->GetArena();
+
+  if (value != nullptr) {
+    ::google::protobuf::Arena* submessage_arena = reinterpret_cast<::google::protobuf::MessageLite*>(value)->GetArena();
     if (message_arena != submessage_arena) {
-      contract = ::PROTOBUF_NAMESPACE_ID::internal::GetOwnedMessage(
-          message_arena, contract, submessage_arena);
+      value = ::google::protobuf::internal::GetOwnedMessage(message_arena, value, submessage_arena);
     }
-    _has_bits_[0] |= 0x00000001u;
+    _impl_._has_bits_[0] |= 0x00000001u;
   } else {
-    _has_bits_[0] &= ~0x00000001u;
+    _impl_._has_bits_[0] &= ~0x00000001u;
   }
-  contract_ = contract;
+
+  _impl_.contract_ = reinterpret_cast<::protobuf::Contract*>(value);
   // @@protoc_insertion_point(field_set_allocated:protobuf.OpenOrder.contract)
 }
 
-// .protobuf.Order order = 3;
-inline bool OpenOrder::_internal_has_order() const {
-  bool value = (_has_bits_[0] & 0x00000002u) != 0;
-  PROTOBUF_ASSUME(!value || order_ != nullptr);
+// optional .protobuf.Order order = 3;
+inline bool OpenOrder::has_order() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000002u) != 0;
+  PROTOBUF_ASSUME(!value || _impl_.order_ != nullptr);
   return value;
 }
-inline bool OpenOrder::has_order() const {
-  return _internal_has_order();
-}
 inline const ::protobuf::Order& OpenOrder::_internal_order() const {
-  const ::protobuf::Order* p = order_;
-  return p != nullptr ? *p : *reinterpret_cast<const ::protobuf::Order*>(
-      &::protobuf::_Order_default_instance_);
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  const ::protobuf::Order* p = _impl_.order_;
+  return p != nullptr ? *p : reinterpret_cast<const ::protobuf::Order&>(::protobuf::_Order_default_instance_);
 }
-inline const ::protobuf::Order& OpenOrder::order() const {
+inline const ::protobuf::Order& OpenOrder::order() const ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.OpenOrder.order)
   return _internal_order();
 }
-inline void OpenOrder::unsafe_arena_set_allocated_order(
-    ::protobuf::Order* order) {
+inline void OpenOrder::unsafe_arena_set_allocated_order(::protobuf::Order* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (GetArena() == nullptr) {
-    delete reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(order_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.order_);
   }
-  order_ = order;
-  if (order) {
-    _has_bits_[0] |= 0x00000002u;
+  _impl_.order_ = reinterpret_cast<::protobuf::Order*>(value);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000002u;
   } else {
-    _has_bits_[0] &= ~0x00000002u;
+    _impl_._has_bits_[0] &= ~0x00000002u;
   }
   // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.OpenOrder.order)
 }
 inline ::protobuf::Order* OpenOrder::release_order() {
-  auto temp = unsafe_arena_release_order();
-  if (GetArena() != nullptr) {
-    temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+
+  _impl_._has_bits_[0] &= ~0x00000002u;
+  ::protobuf::Order* released = _impl_.order_;
+  _impl_.order_ = nullptr;
+  if (::google::protobuf::internal::DebugHardenForceCopyInRelease()) {
+    auto* old = reinterpret_cast<::google::protobuf::MessageLite*>(released);
+    released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    if (GetArena() == nullptr) {
+      delete old;
+    }
+  } else {
+    if (GetArena() != nullptr) {
+      released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    }
   }
-  return temp;
+  return released;
 }
 inline ::protobuf::Order* OpenOrder::unsafe_arena_release_order() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.OpenOrder.order)
-  _has_bits_[0] &= ~0x00000002u;
-  ::protobuf::Order* temp = order_;
-  order_ = nullptr;
+
+  _impl_._has_bits_[0] &= ~0x00000002u;
+  ::protobuf::Order* temp = _impl_.order_;
+  _impl_.order_ = nullptr;
   return temp;
 }
 inline ::protobuf::Order* OpenOrder::_internal_mutable_order() {
-  _has_bits_[0] |= 0x00000002u;
-  if (order_ == nullptr) {
-    auto* p = CreateMaybeMessage<::protobuf::Order>(GetArena());
-    order_ = p;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (_impl_.order_ == nullptr) {
+    auto* p = ::google::protobuf::Message::DefaultConstruct<::protobuf::Order>(GetArena());
+    _impl_.order_ = reinterpret_cast<::protobuf::Order*>(p);
   }
-  return order_;
+  return _impl_.order_;
 }
-inline ::protobuf::Order* OpenOrder::mutable_order() {
+inline ::protobuf::Order* OpenOrder::mutable_order() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  _impl_._has_bits_[0] |= 0x00000002u;
+  ::protobuf::Order* _msg = _internal_mutable_order();
   // @@protoc_insertion_point(field_mutable:protobuf.OpenOrder.order)
-  return _internal_mutable_order();
+  return _msg;
 }
-inline void OpenOrder::set_allocated_order(::protobuf::Order* order) {
-  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArena();
+inline void OpenOrder::set_allocated_order(::protobuf::Order* value) {
+  ::google::protobuf::Arena* message_arena = GetArena();
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (message_arena == nullptr) {
-    delete reinterpret_cast< ::PROTOBUF_NAMESPACE_ID::MessageLite*>(order_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.order_);
   }
-  if (order) {
-    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena =
-      reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(order)->GetArena();
+
+  if (value != nullptr) {
+    ::google::protobuf::Arena* submessage_arena = reinterpret_cast<::google::protobuf::MessageLite*>(value)->GetArena();
     if (message_arena != submessage_arena) {
-      order = ::PROTOBUF_NAMESPACE_ID::internal::GetOwnedMessage(
-          message_arena, order, submessage_arena);
+      value = ::google::protobuf::internal::GetOwnedMessage(message_arena, value, submessage_arena);
     }
-    _has_bits_[0] |= 0x00000002u;
+    _impl_._has_bits_[0] |= 0x00000002u;
   } else {
-    _has_bits_[0] &= ~0x00000002u;
+    _impl_._has_bits_[0] &= ~0x00000002u;
   }
-  order_ = order;
+
+  _impl_.order_ = reinterpret_cast<::protobuf::Order*>(value);
   // @@protoc_insertion_point(field_set_allocated:protobuf.OpenOrder.order)
 }
 
-// .protobuf.OrderState orderState = 4;
-inline bool OpenOrder::_internal_has_orderstate() const {
-  bool value = (_has_bits_[0] & 0x00000004u) != 0;
-  PROTOBUF_ASSUME(!value || orderstate_ != nullptr);
+// optional .protobuf.OrderState orderState = 4;
+inline bool OpenOrder::has_orderstate() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000004u) != 0;
+  PROTOBUF_ASSUME(!value || _impl_.orderstate_ != nullptr);
   return value;
 }
-inline bool OpenOrder::has_orderstate() const {
-  return _internal_has_orderstate();
-}
 inline const ::protobuf::OrderState& OpenOrder::_internal_orderstate() const {
-  const ::protobuf::OrderState* p = orderstate_;
-  return p != nullptr ? *p : *reinterpret_cast<const ::protobuf::OrderState*>(
-      &::protobuf::_OrderState_default_instance_);
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  const ::protobuf::OrderState* p = _impl_.orderstate_;
+  return p != nullptr ? *p : reinterpret_cast<const ::protobuf::OrderState&>(::protobuf::_OrderState_default_instance_);
 }
-inline const ::protobuf::OrderState& OpenOrder::orderstate() const {
+inline const ::protobuf::OrderState& OpenOrder::orderstate() const ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.OpenOrder.orderState)
   return _internal_orderstate();
 }
-inline void OpenOrder::unsafe_arena_set_allocated_orderstate(
-    ::protobuf::OrderState* orderstate) {
+inline void OpenOrder::unsafe_arena_set_allocated_orderstate(::protobuf::OrderState* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (GetArena() == nullptr) {
-    delete reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(orderstate_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.orderstate_);
   }
-  orderstate_ = orderstate;
-  if (orderstate) {
-    _has_bits_[0] |= 0x00000004u;
+  _impl_.orderstate_ = reinterpret_cast<::protobuf::OrderState*>(value);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000004u;
   } else {
-    _has_bits_[0] &= ~0x00000004u;
+    _impl_._has_bits_[0] &= ~0x00000004u;
   }
   // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.OpenOrder.orderState)
 }
 inline ::protobuf::OrderState* OpenOrder::release_orderstate() {
-  auto temp = unsafe_arena_release_orderstate();
-  if (GetArena() != nullptr) {
-    temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+
+  _impl_._has_bits_[0] &= ~0x00000004u;
+  ::protobuf::OrderState* released = _impl_.orderstate_;
+  _impl_.orderstate_ = nullptr;
+  if (::google::protobuf::internal::DebugHardenForceCopyInRelease()) {
+    auto* old = reinterpret_cast<::google::protobuf::MessageLite*>(released);
+    released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    if (GetArena() == nullptr) {
+      delete old;
+    }
+  } else {
+    if (GetArena() != nullptr) {
+      released = ::google::protobuf::internal::DuplicateIfNonNull(released);
+    }
   }
-  return temp;
+  return released;
 }
 inline ::protobuf::OrderState* OpenOrder::unsafe_arena_release_orderstate() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.OpenOrder.orderState)
-  _has_bits_[0] &= ~0x00000004u;
-  ::protobuf::OrderState* temp = orderstate_;
-  orderstate_ = nullptr;
+
+  _impl_._has_bits_[0] &= ~0x00000004u;
+  ::protobuf::OrderState* temp = _impl_.orderstate_;
+  _impl_.orderstate_ = nullptr;
   return temp;
 }
 inline ::protobuf::OrderState* OpenOrder::_internal_mutable_orderstate() {
-  _has_bits_[0] |= 0x00000004u;
-  if (orderstate_ == nullptr) {
-    auto* p = CreateMaybeMessage<::protobuf::OrderState>(GetArena());
-    orderstate_ = p;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (_impl_.orderstate_ == nullptr) {
+    auto* p = ::google::protobuf::Message::DefaultConstruct<::protobuf::OrderState>(GetArena());
+    _impl_.orderstate_ = reinterpret_cast<::protobuf::OrderState*>(p);
   }
-  return orderstate_;
+  return _impl_.orderstate_;
 }
-inline ::protobuf::OrderState* OpenOrder::mutable_orderstate() {
+inline ::protobuf::OrderState* OpenOrder::mutable_orderstate() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  _impl_._has_bits_[0] |= 0x00000004u;
+  ::protobuf::OrderState* _msg = _internal_mutable_orderstate();
   // @@protoc_insertion_point(field_mutable:protobuf.OpenOrder.orderState)
-  return _internal_mutable_orderstate();
+  return _msg;
 }
-inline void OpenOrder::set_allocated_orderstate(::protobuf::OrderState* orderstate) {
-  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArena();
+inline void OpenOrder::set_allocated_orderstate(::protobuf::OrderState* value) {
+  ::google::protobuf::Arena* message_arena = GetArena();
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   if (message_arena == nullptr) {
-    delete reinterpret_cast< ::PROTOBUF_NAMESPACE_ID::MessageLite*>(orderstate_);
+    delete reinterpret_cast<::google::protobuf::MessageLite*>(_impl_.orderstate_);
   }
-  if (orderstate) {
-    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena =
-      reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(orderstate)->GetArena();
+
+  if (value != nullptr) {
+    ::google::protobuf::Arena* submessage_arena = reinterpret_cast<::google::protobuf::MessageLite*>(value)->GetArena();
     if (message_arena != submessage_arena) {
-      orderstate = ::PROTOBUF_NAMESPACE_ID::internal::GetOwnedMessage(
-          message_arena, orderstate, submessage_arena);
+      value = ::google::protobuf::internal::GetOwnedMessage(message_arena, value, submessage_arena);
     }
-    _has_bits_[0] |= 0x00000004u;
+    _impl_._has_bits_[0] |= 0x00000004u;
   } else {
-    _has_bits_[0] &= ~0x00000004u;
+    _impl_._has_bits_[0] &= ~0x00000004u;
   }
-  orderstate_ = orderstate;
+
+  _impl_.orderstate_ = reinterpret_cast<::protobuf::OrderState*>(value);
   // @@protoc_insertion_point(field_set_allocated:protobuf.OpenOrder.orderState)
 }
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic pop
+#pragma GCC diagnostic pop
 #endif  // __GNUC__
 
 // @@protoc_insertion_point(namespace_scope)
-
 }  // namespace protobuf
+
 
 // @@protoc_insertion_point(global_scope)
 
-#include <google/protobuf/port_undef.inc>
-#endif  // GOOGLE_PROTOBUF_INCLUDED_GOOGLE_PROTOBUF_INCLUDED_OpenOrder_2eproto
+#include "google/protobuf/port_undef.inc"
+
+#endif  // OpenOrder_2eproto_2epb_2eh
