@@ -1,507 +1,494 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: OrderCancel.proto
+// Protobuf C++ Version: 5.29.3
 
-#ifndef GOOGLE_PROTOBUF_INCLUDED_OrderCancel_2eproto
-#define GOOGLE_PROTOBUF_INCLUDED_OrderCancel_2eproto
+#ifndef OrderCancel_2eproto_2epb_2eh
+#define OrderCancel_2eproto_2epb_2eh
 
 #include <limits>
 #include <string>
+#include <type_traits>
+#include <utility>
 
-#include <google/protobuf/port_def.inc>
-#if PROTOBUF_VERSION < 3012000
-#error This file was generated by a newer version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please update
-#error your headers.
+#include "google/protobuf/runtime_version.h"
+#if PROTOBUF_VERSION != 5029003
+#error "Protobuf C++ gencode is built with an incompatible version of"
+#error "Protobuf C++ headers/runtime. See"
+#error "https://protobuf.dev/support/cross-version-runtime-guarantee/#cpp"
 #endif
-#if 3012004 < PROTOBUF_MIN_PROTOC_VERSION
-#error This file was generated by an older version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please
-#error regenerate this file with a newer version of protoc.
-#endif
-
-#include <google/protobuf/port_undef.inc>
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/arena.h>
-#include <google/protobuf/arenastring.h>
-#include <google/protobuf/generated_message_table_driven.h>
-#include <google/protobuf/generated_message_util.h>
-#include <google/protobuf/inlined_string_field.h>
-#include <google/protobuf/metadata_lite.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/message.h>
-#include <google/protobuf/repeated_field.h>  // IWYU pragma: export
-#include <google/protobuf/extension_set.h>  // IWYU pragma: export
-#include <google/protobuf/unknown_field_set.h>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/arena.h"
+#include "google/protobuf/arenastring.h"
+#include "google/protobuf/generated_message_tctable_decl.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/metadata_lite.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/message.h"
+#include "google/protobuf/message_lite.h"
+#include "google/protobuf/repeated_field.h"  // IWYU pragma: export
+#include "google/protobuf/extension_set.h"  // IWYU pragma: export
+#include "google/protobuf/unknown_field_set.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
+
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+
 #define PROTOBUF_INTERNAL_EXPORT_OrderCancel_2eproto
-PROTOBUF_NAMESPACE_OPEN
+
+namespace google {
+namespace protobuf {
 namespace internal {
-class AnyMetadata;
+template <typename T>
+::absl::string_view GetAnyMessageName();
 }  // namespace internal
-PROTOBUF_NAMESPACE_CLOSE
+}  // namespace protobuf
+}  // namespace google
 
 // Internal implementation detail -- do not use these members.
 struct TableStruct_OrderCancel_2eproto {
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTableField entries[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxillaryParseTableField aux[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTable schema[1]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::FieldMetadata field_metadata[];
-  static const ::PROTOBUF_NAMESPACE_ID::internal::SerializationTable serialization_table[];
-  static const ::PROTOBUF_NAMESPACE_ID::uint32 offsets[];
+  static const ::uint32_t offsets[];
 };
-extern const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_OrderCancel_2eproto;
+extern const ::google::protobuf::internal::DescriptorTable
+    descriptor_table_OrderCancel_2eproto;
 namespace protobuf {
 class OrderCancel;
-class OrderCancelDefaultTypeInternal;
+struct OrderCancelDefaultTypeInternal;
 extern OrderCancelDefaultTypeInternal _OrderCancel_default_instance_;
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> ::protobuf::OrderCancel* Arena::CreateMaybeMessage<::protobuf::OrderCancel>(Arena*);
-PROTOBUF_NAMESPACE_CLOSE
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
+
 namespace protobuf {
 
 // ===================================================================
 
-class OrderCancel PROTOBUF_FINAL :
-    public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:protobuf.OrderCancel) */ {
+
+// -------------------------------------------------------------------
+
+class OrderCancel final : public ::google::protobuf::Message
+/* @@protoc_insertion_point(class_definition:protobuf.OrderCancel) */ {
  public:
-  inline OrderCancel() : OrderCancel(nullptr) {};
-  virtual ~OrderCancel();
+  inline OrderCancel() : OrderCancel(nullptr) {}
+  ~OrderCancel() PROTOBUF_FINAL;
 
-  OrderCancel(const OrderCancel& from);
-  OrderCancel(OrderCancel&& from) noexcept
-    : OrderCancel() {
-    *this = ::std::move(from);
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+  void operator delete(OrderCancel* msg, std::destroying_delete_t) {
+    SharedDtor(*msg);
+    ::google::protobuf::internal::SizedDelete(msg, sizeof(OrderCancel));
   }
+#endif
 
+  template <typename = void>
+  explicit PROTOBUF_CONSTEXPR OrderCancel(
+      ::google::protobuf::internal::ConstantInitialized);
+
+  inline OrderCancel(const OrderCancel& from) : OrderCancel(nullptr, from) {}
+  inline OrderCancel(OrderCancel&& from) noexcept
+      : OrderCancel(nullptr, std::move(from)) {}
   inline OrderCancel& operator=(const OrderCancel& from) {
     CopyFrom(from);
     return *this;
   }
   inline OrderCancel& operator=(OrderCancel&& from) noexcept {
-    if (GetArena() == from.GetArena()) {
-      if (this != &from) InternalSwap(&from);
+    if (this == &from) return *this;
+    if (::google::protobuf::internal::CanMoveWithInternalSwap(GetArena(), from.GetArena())) {
+      InternalSwap(&from);
     } else {
       CopyFrom(from);
     }
     return *this;
   }
 
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* descriptor() {
+  inline const ::google::protobuf::UnknownFieldSet& unknown_fields() const
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance);
+  }
+  inline ::google::protobuf::UnknownFieldSet* mutable_unknown_fields()
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.mutable_unknown_fields<::google::protobuf::UnknownFieldSet>();
+  }
+
+  static const ::google::protobuf::Descriptor* descriptor() {
     return GetDescriptor();
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* GetDescriptor() {
-    return GetMetadataStatic().descriptor;
+  static const ::google::protobuf::Descriptor* GetDescriptor() {
+    return default_instance().GetMetadata().descriptor;
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Reflection* GetReflection() {
-    return GetMetadataStatic().reflection;
+  static const ::google::protobuf::Reflection* GetReflection() {
+    return default_instance().GetMetadata().reflection;
   }
-  static const OrderCancel& default_instance();
-
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
+  static const OrderCancel& default_instance() {
+    return *internal_default_instance();
+  }
   static inline const OrderCancel* internal_default_instance() {
     return reinterpret_cast<const OrderCancel*>(
-               &_OrderCancel_default_instance_);
+        &_OrderCancel_default_instance_);
   }
-  static constexpr int kIndexInFileMessages =
-    0;
-
-  friend void swap(OrderCancel& a, OrderCancel& b) {
-    a.Swap(&b);
-  }
+  static constexpr int kIndexInFileMessages = 0;
+  friend void swap(OrderCancel& a, OrderCancel& b) { a.Swap(&b); }
   inline void Swap(OrderCancel* other) {
     if (other == this) return;
-    if (GetArena() == other->GetArena()) {
+    if (::google::protobuf::internal::CanUseInternalSwap(GetArena(), other->GetArena())) {
       InternalSwap(other);
     } else {
-      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+      ::google::protobuf::internal::GenericSwap(this, other);
     }
   }
   void UnsafeArenaSwap(OrderCancel* other) {
     if (other == this) return;
-    GOOGLE_DCHECK(GetArena() == other->GetArena());
+    ABSL_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
   // implements Message ----------------------------------------------
 
-  inline OrderCancel* New() const final {
-    return CreateMaybeMessage<OrderCancel>(nullptr);
+  OrderCancel* New(::google::protobuf::Arena* arena = nullptr) const {
+    return ::google::protobuf::Message::DefaultConstruct<OrderCancel>(arena);
   }
-
-  OrderCancel* New(::PROTOBUF_NAMESPACE_ID::Arena* arena) const final {
-    return CreateMaybeMessage<OrderCancel>(arena);
-  }
-  void CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
-  void MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
+  using ::google::protobuf::Message::CopyFrom;
   void CopyFrom(const OrderCancel& from);
-  void MergeFrom(const OrderCancel& from);
-  PROTOBUF_ATTRIBUTE_REINITIALIZES void Clear() final;
-  bool IsInitialized() const final;
-
-  size_t ByteSizeLong() const final;
-  const char* _InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) final;
-  ::PROTOBUF_NAMESPACE_ID::uint8* _InternalSerialize(
-      ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const final;
-  int GetCachedSize() const final { return _cached_size_.Get(); }
+  using ::google::protobuf::Message::MergeFrom;
+  void MergeFrom(const OrderCancel& from) { OrderCancel::MergeImpl(*this, from); }
 
   private:
-  inline void SharedCtor();
-  inline void SharedDtor();
-  void SetCachedSize(int size) const final;
+  static void MergeImpl(
+      ::google::protobuf::MessageLite& to_msg,
+      const ::google::protobuf::MessageLite& from_msg);
+
+  public:
+  bool IsInitialized() const {
+    return true;
+  }
+  ABSL_ATTRIBUTE_REINITIALIZES void Clear() PROTOBUF_FINAL;
+  #if defined(PROTOBUF_CUSTOM_VTABLE)
+  private:
+  static ::size_t ByteSizeLong(const ::google::protobuf::MessageLite& msg);
+  static ::uint8_t* _InternalSerialize(
+      const MessageLite& msg, ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream);
+
+  public:
+  ::size_t ByteSizeLong() const { return ByteSizeLong(*this); }
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+    return _InternalSerialize(*this, target, stream);
+  }
+  #else   // PROTOBUF_CUSTOM_VTABLE
+  ::size_t ByteSizeLong() const final;
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const final;
+  #endif  // PROTOBUF_CUSTOM_VTABLE
+  int GetCachedSize() const { return _impl_._cached_size_.Get(); }
+
+  private:
+  void SharedCtor(::google::protobuf::Arena* arena);
+  static void SharedDtor(MessageLite& self);
   void InternalSwap(OrderCancel* other);
-  friend class ::PROTOBUF_NAMESPACE_ID::internal::AnyMetadata;
-  static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
-    return "protobuf.OrderCancel";
+ private:
+  template <typename T>
+  friend ::absl::string_view(
+      ::google::protobuf::internal::GetAnyMessageName)();
+  static ::absl::string_view FullMessageName() { return "protobuf.OrderCancel"; }
+
+ protected:
+  explicit OrderCancel(::google::protobuf::Arena* arena);
+  OrderCancel(::google::protobuf::Arena* arena, const OrderCancel& from);
+  OrderCancel(::google::protobuf::Arena* arena, OrderCancel&& from) noexcept
+      : OrderCancel(arena) {
+    *this = ::std::move(from);
   }
-  protected:
-  explicit OrderCancel(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  private:
-  static void ArenaDtor(void* object);
-  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  public:
+  const ::google::protobuf::internal::ClassData* GetClassData() const PROTOBUF_FINAL;
+  static void* PlacementNew_(const void*, void* mem,
+                             ::google::protobuf::Arena* arena);
+  static constexpr auto InternalNewImpl_();
+  static const ::google::protobuf::internal::ClassDataFull _class_data_;
 
-  ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
-  private:
-  static ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadataStatic() {
-    ::PROTOBUF_NAMESPACE_ID::internal::AssignDescriptors(&::descriptor_table_OrderCancel_2eproto);
-    return ::descriptor_table_OrderCancel_2eproto.file_level_metadata[kIndexInFileMessages];
-  }
-
-  public:
-
+ public:
+  ::google::protobuf::Metadata GetMetadata() const;
   // nested types ----------------------------------------------------
 
   // accessors -------------------------------------------------------
-
   enum : int {
     kManualOrderCancelTimeFieldNumber = 1,
     kExtOperatorFieldNumber = 2,
     kManualOrderIndicatorFieldNumber = 3,
   };
-  // string manualOrderCancelTime = 1;
+  // optional string manualOrderCancelTime = 1;
   bool has_manualordercanceltime() const;
-  private:
-  bool _internal_has_manualordercanceltime() const;
-  public:
-  void clear_manualordercanceltime();
+  void clear_manualordercanceltime() ;
   const std::string& manualordercanceltime() const;
-  void set_manualordercanceltime(const std::string& value);
-  void set_manualordercanceltime(std::string&& value);
-  void set_manualordercanceltime(const char* value);
-  void set_manualordercanceltime(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_manualordercanceltime(Arg_&& arg, Args_... args);
   std::string* mutable_manualordercanceltime();
-  std::string* release_manualordercanceltime();
-  void set_allocated_manualordercanceltime(std::string* manualordercanceltime);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_manualordercanceltime();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_manualordercanceltime(
-      std::string* manualordercanceltime);
+  PROTOBUF_NODISCARD std::string* release_manualordercanceltime();
+  void set_allocated_manualordercanceltime(std::string* value);
+
   private:
   const std::string& _internal_manualordercanceltime() const;
-  void _internal_set_manualordercanceltime(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_manualordercanceltime(
+      const std::string& value);
   std::string* _internal_mutable_manualordercanceltime();
-  public:
 
-  // string extOperator = 2;
-  bool has_extoperator() const;
-  private:
-  bool _internal_has_extoperator() const;
   public:
-  void clear_extoperator();
+  // optional string extOperator = 2;
+  bool has_extoperator() const;
+  void clear_extoperator() ;
   const std::string& extoperator() const;
-  void set_extoperator(const std::string& value);
-  void set_extoperator(std::string&& value);
-  void set_extoperator(const char* value);
-  void set_extoperator(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_extoperator(Arg_&& arg, Args_... args);
   std::string* mutable_extoperator();
-  std::string* release_extoperator();
-  void set_allocated_extoperator(std::string* extoperator);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_extoperator();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_extoperator(
-      std::string* extoperator);
+  PROTOBUF_NODISCARD std::string* release_extoperator();
+  void set_allocated_extoperator(std::string* value);
+
   private:
   const std::string& _internal_extoperator() const;
-  void _internal_set_extoperator(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_extoperator(
+      const std::string& value);
   std::string* _internal_mutable_extoperator();
-  public:
 
-  // int32 manualOrderIndicator = 3;
+  public:
+  // optional int32 manualOrderIndicator = 3;
   bool has_manualorderindicator() const;
-  private:
-  bool _internal_has_manualorderindicator() const;
-  public:
-  void clear_manualorderindicator();
-  ::PROTOBUF_NAMESPACE_ID::int32 manualorderindicator() const;
-  void set_manualorderindicator(::PROTOBUF_NAMESPACE_ID::int32 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int32 _internal_manualorderindicator() const;
-  void _internal_set_manualorderindicator(::PROTOBUF_NAMESPACE_ID::int32 value);
-  public:
+  void clear_manualorderindicator() ;
+  ::int32_t manualorderindicator() const;
+  void set_manualorderindicator(::int32_t value);
 
+  private:
+  ::int32_t _internal_manualorderindicator() const;
+  void _internal_set_manualorderindicator(::int32_t value);
+
+  public:
   // @@protoc_insertion_point(class_scope:protobuf.OrderCancel)
  private:
   class _Internal;
+  friend class ::google::protobuf::internal::TcParser;
+  static const ::google::protobuf::internal::TcParseTable<
+      2, 3, 0,
+      61, 2>
+      _table_;
 
-  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
-  typedef void InternalArenaConstructable_;
-  typedef void DestructorSkippable_;
-  ::PROTOBUF_NAMESPACE_ID::internal::HasBits<1> _has_bits_;
-  mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr manualordercanceltime_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr extoperator_;
-  ::PROTOBUF_NAMESPACE_ID::int32 manualorderindicator_;
+  friend class ::google::protobuf::MessageLite;
+  friend class ::google::protobuf::Arena;
+  template <typename T>
+  friend class ::google::protobuf::Arena::InternalHelper;
+  using InternalArenaConstructable_ = void;
+  using DestructorSkippable_ = void;
+  struct Impl_ {
+    inline explicit constexpr Impl_(
+        ::google::protobuf::internal::ConstantInitialized) noexcept;
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena);
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena, const Impl_& from,
+                          const OrderCancel& from_msg);
+    ::google::protobuf::internal::HasBits<1> _has_bits_;
+    ::google::protobuf::internal::CachedSize _cached_size_;
+    ::google::protobuf::internal::ArenaStringPtr manualordercanceltime_;
+    ::google::protobuf::internal::ArenaStringPtr extoperator_;
+    ::int32_t manualorderindicator_;
+    PROTOBUF_TSAN_DECLARE_MEMBER
+  };
+  union { Impl_ _impl_; };
   friend struct ::TableStruct_OrderCancel_2eproto;
 };
+
 // ===================================================================
 
 
+
+
 // ===================================================================
+
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic push
-  #pragma GCC diagnostic ignored "-Wstrict-aliasing"
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wstrict-aliasing"
 #endif  // __GNUC__
+// -------------------------------------------------------------------
+
 // OrderCancel
 
-// string manualOrderCancelTime = 1;
-inline bool OrderCancel::_internal_has_manualordercanceltime() const {
-  bool value = (_has_bits_[0] & 0x00000001u) != 0;
+// optional string manualOrderCancelTime = 1;
+inline bool OrderCancel::has_manualordercanceltime() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000001u) != 0;
   return value;
 }
-inline bool OrderCancel::has_manualordercanceltime() const {
-  return _internal_has_manualordercanceltime();
-}
 inline void OrderCancel::clear_manualordercanceltime() {
-  manualordercanceltime_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000001u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.manualordercanceltime_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000001u;
 }
-inline const std::string& OrderCancel::manualordercanceltime() const {
+inline const std::string& OrderCancel::manualordercanceltime() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.OrderCancel.manualOrderCancelTime)
   return _internal_manualordercanceltime();
 }
-inline void OrderCancel::set_manualordercanceltime(const std::string& value) {
-  _internal_set_manualordercanceltime(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void OrderCancel::set_manualordercanceltime(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  _impl_.manualordercanceltime_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.OrderCancel.manualOrderCancelTime)
 }
-inline std::string* OrderCancel::mutable_manualordercanceltime() {
+inline std::string* OrderCancel::mutable_manualordercanceltime() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_manualordercanceltime();
   // @@protoc_insertion_point(field_mutable:protobuf.OrderCancel.manualOrderCancelTime)
-  return _internal_mutable_manualordercanceltime();
+  return _s;
 }
 inline const std::string& OrderCancel::_internal_manualordercanceltime() const {
-  return manualordercanceltime_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.manualordercanceltime_.Get();
 }
 inline void OrderCancel::_internal_set_manualordercanceltime(const std::string& value) {
-  _has_bits_[0] |= 0x00000001u;
-  manualordercanceltime_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void OrderCancel::set_manualordercanceltime(std::string&& value) {
-  _has_bits_[0] |= 0x00000001u;
-  manualordercanceltime_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.OrderCancel.manualOrderCancelTime)
-}
-inline void OrderCancel::set_manualordercanceltime(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000001u;
-  manualordercanceltime_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.OrderCancel.manualOrderCancelTime)
-}
-inline void OrderCancel::set_manualordercanceltime(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000001u;
-  manualordercanceltime_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.OrderCancel.manualOrderCancelTime)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  _impl_.manualordercanceltime_.Set(value, GetArena());
 }
 inline std::string* OrderCancel::_internal_mutable_manualordercanceltime() {
-  _has_bits_[0] |= 0x00000001u;
-  return manualordercanceltime_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  return _impl_.manualordercanceltime_.Mutable( GetArena());
 }
 inline std::string* OrderCancel::release_manualordercanceltime() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.OrderCancel.manualOrderCancelTime)
-  if (!_internal_has_manualordercanceltime()) {
+  if ((_impl_._has_bits_[0] & 0x00000001u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000001u;
-  return manualordercanceltime_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void OrderCancel::set_allocated_manualordercanceltime(std::string* manualordercanceltime) {
-  if (manualordercanceltime != nullptr) {
-    _has_bits_[0] |= 0x00000001u;
-  } else {
-    _has_bits_[0] &= ~0x00000001u;
+  _impl_._has_bits_[0] &= ~0x00000001u;
+  auto* released = _impl_.manualordercanceltime_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.manualordercanceltime_.Set("", GetArena());
   }
-  manualordercanceltime_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), manualordercanceltime,
-      GetArena());
+  return released;
+}
+inline void OrderCancel::set_allocated_manualordercanceltime(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000001u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000001u;
+  }
+  _impl_.manualordercanceltime_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.manualordercanceltime_.IsDefault()) {
+    _impl_.manualordercanceltime_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.OrderCancel.manualOrderCancelTime)
 }
-inline std::string* OrderCancel::unsafe_arena_release_manualordercanceltime() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.OrderCancel.manualOrderCancelTime)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000001u;
-  return manualordercanceltime_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void OrderCancel::unsafe_arena_set_allocated_manualordercanceltime(
-    std::string* manualordercanceltime) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (manualordercanceltime != nullptr) {
-    _has_bits_[0] |= 0x00000001u;
-  } else {
-    _has_bits_[0] &= ~0x00000001u;
-  }
-  manualordercanceltime_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      manualordercanceltime, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.OrderCancel.manualOrderCancelTime)
-}
 
-// string extOperator = 2;
-inline bool OrderCancel::_internal_has_extoperator() const {
-  bool value = (_has_bits_[0] & 0x00000002u) != 0;
+// optional string extOperator = 2;
+inline bool OrderCancel::has_extoperator() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000002u) != 0;
   return value;
 }
-inline bool OrderCancel::has_extoperator() const {
-  return _internal_has_extoperator();
-}
 inline void OrderCancel::clear_extoperator() {
-  extoperator_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000002u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.extoperator_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000002u;
 }
-inline const std::string& OrderCancel::extoperator() const {
+inline const std::string& OrderCancel::extoperator() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.OrderCancel.extOperator)
   return _internal_extoperator();
 }
-inline void OrderCancel::set_extoperator(const std::string& value) {
-  _internal_set_extoperator(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void OrderCancel::set_extoperator(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  _impl_.extoperator_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.OrderCancel.extOperator)
 }
-inline std::string* OrderCancel::mutable_extoperator() {
+inline std::string* OrderCancel::mutable_extoperator() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_extoperator();
   // @@protoc_insertion_point(field_mutable:protobuf.OrderCancel.extOperator)
-  return _internal_mutable_extoperator();
+  return _s;
 }
 inline const std::string& OrderCancel::_internal_extoperator() const {
-  return extoperator_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.extoperator_.Get();
 }
 inline void OrderCancel::_internal_set_extoperator(const std::string& value) {
-  _has_bits_[0] |= 0x00000002u;
-  extoperator_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void OrderCancel::set_extoperator(std::string&& value) {
-  _has_bits_[0] |= 0x00000002u;
-  extoperator_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.OrderCancel.extOperator)
-}
-inline void OrderCancel::set_extoperator(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000002u;
-  extoperator_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.OrderCancel.extOperator)
-}
-inline void OrderCancel::set_extoperator(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000002u;
-  extoperator_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.OrderCancel.extOperator)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  _impl_.extoperator_.Set(value, GetArena());
 }
 inline std::string* OrderCancel::_internal_mutable_extoperator() {
-  _has_bits_[0] |= 0x00000002u;
-  return extoperator_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  return _impl_.extoperator_.Mutable( GetArena());
 }
 inline std::string* OrderCancel::release_extoperator() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.OrderCancel.extOperator)
-  if (!_internal_has_extoperator()) {
+  if ((_impl_._has_bits_[0] & 0x00000002u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000002u;
-  return extoperator_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void OrderCancel::set_allocated_extoperator(std::string* extoperator) {
-  if (extoperator != nullptr) {
-    _has_bits_[0] |= 0x00000002u;
-  } else {
-    _has_bits_[0] &= ~0x00000002u;
+  _impl_._has_bits_[0] &= ~0x00000002u;
+  auto* released = _impl_.extoperator_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.extoperator_.Set("", GetArena());
   }
-  extoperator_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), extoperator,
-      GetArena());
+  return released;
+}
+inline void OrderCancel::set_allocated_extoperator(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000002u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000002u;
+  }
+  _impl_.extoperator_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.extoperator_.IsDefault()) {
+    _impl_.extoperator_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.OrderCancel.extOperator)
 }
-inline std::string* OrderCancel::unsafe_arena_release_extoperator() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.OrderCancel.extOperator)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000002u;
-  return extoperator_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void OrderCancel::unsafe_arena_set_allocated_extoperator(
-    std::string* extoperator) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (extoperator != nullptr) {
-    _has_bits_[0] |= 0x00000002u;
-  } else {
-    _has_bits_[0] &= ~0x00000002u;
-  }
-  extoperator_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      extoperator, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.OrderCancel.extOperator)
-}
 
-// int32 manualOrderIndicator = 3;
-inline bool OrderCancel::_internal_has_manualorderindicator() const {
-  bool value = (_has_bits_[0] & 0x00000004u) != 0;
+// optional int32 manualOrderIndicator = 3;
+inline bool OrderCancel::has_manualorderindicator() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000004u) != 0;
   return value;
 }
-inline bool OrderCancel::has_manualorderindicator() const {
-  return _internal_has_manualorderindicator();
-}
 inline void OrderCancel::clear_manualorderindicator() {
-  manualorderindicator_ = 0;
-  _has_bits_[0] &= ~0x00000004u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.manualorderindicator_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000004u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int32 OrderCancel::_internal_manualorderindicator() const {
-  return manualorderindicator_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int32 OrderCancel::manualorderindicator() const {
+inline ::int32_t OrderCancel::manualorderindicator() const {
   // @@protoc_insertion_point(field_get:protobuf.OrderCancel.manualOrderIndicator)
   return _internal_manualorderindicator();
 }
-inline void OrderCancel::_internal_set_manualorderindicator(::PROTOBUF_NAMESPACE_ID::int32 value) {
-  _has_bits_[0] |= 0x00000004u;
-  manualorderindicator_ = value;
-}
-inline void OrderCancel::set_manualorderindicator(::PROTOBUF_NAMESPACE_ID::int32 value) {
+inline void OrderCancel::set_manualorderindicator(::int32_t value) {
   _internal_set_manualorderindicator(value);
+  _impl_._has_bits_[0] |= 0x00000004u;
   // @@protoc_insertion_point(field_set:protobuf.OrderCancel.manualOrderIndicator)
+}
+inline ::int32_t OrderCancel::_internal_manualorderindicator() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.manualorderindicator_;
+}
+inline void OrderCancel::_internal_set_manualorderindicator(::int32_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.manualorderindicator_ = value;
 }
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic pop
+#pragma GCC diagnostic pop
 #endif  // __GNUC__
 
 // @@protoc_insertion_point(namespace_scope)
-
 }  // namespace protobuf
+
 
 // @@protoc_insertion_point(global_scope)
 
-#include <google/protobuf/port_undef.inc>
-#endif  // GOOGLE_PROTOBUF_INCLUDED_GOOGLE_PROTOBUF_INCLUDED_OrderCancel_2eproto
+#include "google/protobuf/port_undef.inc"
+
+#endif  // OrderCancel_2eproto_2epb_2eh
