@@ -1,338 +1,383 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: CancelOrderRequest.proto
+// Protobuf C++ Version: 5.29.3
 
 #include "CancelOrderRequest.pb.h"
 
 #include <algorithm>
-
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/extension_set.h>
-#include <google/protobuf/wire_format_lite.h>
-#include <google/protobuf/descriptor.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/reflection_ops.h>
-#include <google/protobuf/wire_format.h>
+#include <type_traits>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/generated_message_tctable_impl.h"
+#include "google/protobuf/extension_set.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/wire_format_lite.h"
+#include "google/protobuf/descriptor.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/reflection_ops.h"
+#include "google/protobuf/wire_format.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
-extern PROTOBUF_INTERNAL_EXPORT_OrderCancel_2eproto ::PROTOBUF_NAMESPACE_ID::internal::SCCInfo<0> scc_info_OrderCancel_OrderCancel_2eproto;
-namespace protobuf {
-class CancelOrderRequestDefaultTypeInternal {
- public:
-  ::PROTOBUF_NAMESPACE_ID::internal::ExplicitlyConstructed<CancelOrderRequest> _instance;
-} _CancelOrderRequest_default_instance_;
-}  // namespace protobuf
-static void InitDefaultsscc_info_CancelOrderRequest_CancelOrderRequest_2eproto() {
-  GOOGLE_PROTOBUF_VERIFY_VERSION;
 
-  {
-    void* ptr = &::protobuf::_CancelOrderRequest_default_instance_;
-    new (ptr) ::protobuf::CancelOrderRequest();
-    ::PROTOBUF_NAMESPACE_ID::internal::OnShutdownDestroyMessage(ptr);
-  }
-  ::protobuf::CancelOrderRequest::InitAsDefaultInstance();
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+PROTOBUF_PRAGMA_INIT_SEG
+namespace _pb = ::google::protobuf;
+namespace _pbi = ::google::protobuf::internal;
+namespace _fl = ::google::protobuf::internal::field_layout;
+namespace protobuf {
+
+inline constexpr CancelOrderRequest::Impl_::Impl_(
+    ::_pbi::ConstantInitialized) noexcept
+      : _cached_size_{0},
+        ordercancel_{nullptr},
+        orderid_{0} {}
+
+template <typename>
+PROTOBUF_CONSTEXPR CancelOrderRequest::CancelOrderRequest(::_pbi::ConstantInitialized)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(_class_data_.base()),
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(),
+#endif  // PROTOBUF_CUSTOM_VTABLE
+      _impl_(::_pbi::ConstantInitialized()) {
 }
-
-::PROTOBUF_NAMESPACE_ID::internal::SCCInfo<1> scc_info_CancelOrderRequest_CancelOrderRequest_2eproto =
-    {{ATOMIC_VAR_INIT(::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase::kUninitialized), 1, 0, InitDefaultsscc_info_CancelOrderRequest_CancelOrderRequest_2eproto}, {
-      &scc_info_OrderCancel_OrderCancel_2eproto.base,}};
-
-static ::PROTOBUF_NAMESPACE_ID::Metadata file_level_metadata_CancelOrderRequest_2eproto[1];
-static constexpr ::PROTOBUF_NAMESPACE_ID::EnumDescriptor const** file_level_enum_descriptors_CancelOrderRequest_2eproto = nullptr;
-static constexpr ::PROTOBUF_NAMESPACE_ID::ServiceDescriptor const** file_level_service_descriptors_CancelOrderRequest_2eproto = nullptr;
-
-const ::PROTOBUF_NAMESPACE_ID::uint32 TableStruct_CancelOrderRequest_2eproto::offsets[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  PROTOBUF_FIELD_OFFSET(::protobuf::CancelOrderRequest, _has_bits_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::CancelOrderRequest, _internal_metadata_),
-  ~0u,  // no _extensions_
-  ~0u,  // no _oneof_case_
-  ~0u,  // no _weak_field_map_
-  PROTOBUF_FIELD_OFFSET(::protobuf::CancelOrderRequest, orderid_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::CancelOrderRequest, ordercancel_),
-  1,
-  0,
-};
-static const ::PROTOBUF_NAMESPACE_ID::internal::MigrationSchema schemas[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  { 0, 7, sizeof(::protobuf::CancelOrderRequest)},
+struct CancelOrderRequestDefaultTypeInternal {
+  PROTOBUF_CONSTEXPR CancelOrderRequestDefaultTypeInternal() : _instance(::_pbi::ConstantInitialized{}) {}
+  ~CancelOrderRequestDefaultTypeInternal() {}
+  union {
+    CancelOrderRequest _instance;
+  };
 };
 
-static ::PROTOBUF_NAMESPACE_ID::Message const * const file_default_instances[] = {
-  reinterpret_cast<const ::PROTOBUF_NAMESPACE_ID::Message*>(&::protobuf::_CancelOrderRequest_default_instance_),
+PROTOBUF_ATTRIBUTE_NO_DESTROY PROTOBUF_CONSTINIT
+    PROTOBUF_ATTRIBUTE_INIT_PRIORITY1 CancelOrderRequestDefaultTypeInternal _CancelOrderRequest_default_instance_;
+}  // namespace protobuf
+static constexpr const ::_pb::EnumDescriptor**
+    file_level_enum_descriptors_CancelOrderRequest_2eproto = nullptr;
+static constexpr const ::_pb::ServiceDescriptor**
+    file_level_service_descriptors_CancelOrderRequest_2eproto = nullptr;
+const ::uint32_t
+    TableStruct_CancelOrderRequest_2eproto::offsets[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+        protodesc_cold) = {
+        PROTOBUF_FIELD_OFFSET(::protobuf::CancelOrderRequest, _impl_._has_bits_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::CancelOrderRequest, _internal_metadata_),
+        ~0u,  // no _extensions_
+        ~0u,  // no _oneof_case_
+        ~0u,  // no _weak_field_map_
+        ~0u,  // no _inlined_string_donated_
+        ~0u,  // no _split_
+        ~0u,  // no sizeof(Split)
+        PROTOBUF_FIELD_OFFSET(::protobuf::CancelOrderRequest, _impl_.orderid_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::CancelOrderRequest, _impl_.ordercancel_),
+        1,
+        0,
 };
 
-const char descriptor_table_protodef_CancelOrderRequest_2eproto[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) =
-  "\n\030CancelOrderRequest.proto\022\010protobuf\032\021Or"
-  "derCancel.proto\"w\n\022CancelOrderRequest\022\024\n"
-  "\007orderId\030\001 \001(\005H\000\210\001\001\022/\n\013orderCancel\030\002 \001(\013"
-  "2\025.protobuf.OrderCancelH\001\210\001\001B\n\n\010_orderId"
-  "B\016\n\014_orderCancelBB\n\026com.ib.client.protob"
-  "ufB\027CancelOrderRequestProto\252\002\016IBApi.prot"
-  "obufb\006proto3"
-  ;
-static const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable*const descriptor_table_CancelOrderRequest_2eproto_deps[1] = {
-  &::descriptor_table_OrderCancel_2eproto,
+static const ::_pbi::MigrationSchema
+    schemas[] ABSL_ATTRIBUTE_SECTION_VARIABLE(protodesc_cold) = {
+        {0, 10, -1, sizeof(::protobuf::CancelOrderRequest)},
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase*const descriptor_table_CancelOrderRequest_2eproto_sccs[1] = {
-  &scc_info_CancelOrderRequest_CancelOrderRequest_2eproto.base,
+static const ::_pb::Message* const file_default_instances[] = {
+    &::protobuf::_CancelOrderRequest_default_instance_._instance,
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::once_flag descriptor_table_CancelOrderRequest_2eproto_once;
-const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_CancelOrderRequest_2eproto = {
-  false, false, descriptor_table_protodef_CancelOrderRequest_2eproto, "CancelOrderRequest.proto", 252,
-  &descriptor_table_CancelOrderRequest_2eproto_once, descriptor_table_CancelOrderRequest_2eproto_sccs, descriptor_table_CancelOrderRequest_2eproto_deps, 1, 1,
-  schemas, file_default_instances, TableStruct_CancelOrderRequest_2eproto::offsets,
-  file_level_metadata_CancelOrderRequest_2eproto, 1, file_level_enum_descriptors_CancelOrderRequest_2eproto, file_level_service_descriptors_CancelOrderRequest_2eproto,
+const char descriptor_table_protodef_CancelOrderRequest_2eproto[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+    protodesc_cold) = {
+    "\n\030CancelOrderRequest.proto\022\010protobuf\032\021Or"
+    "derCancel.proto\"w\n\022CancelOrderRequest\022\024\n"
+    "\007orderId\030\001 \001(\005H\000\210\001\001\022/\n\013orderCancel\030\002 \001(\013"
+    "2\025.protobuf.OrderCancelH\001\210\001\001B\n\n\010_orderId"
+    "B\016\n\014_orderCancelBB\n\026com.ib.client.protob"
+    "ufB\027CancelOrderRequestProto\252\002\016IBApi.prot"
+    "obufb\006proto3"
 };
-
-// Force running AddDescriptors() at dynamic initialization time.
-static bool dynamic_init_dummy_CancelOrderRequest_2eproto = (static_cast<void>(::PROTOBUF_NAMESPACE_ID::internal::AddDescriptors(&descriptor_table_CancelOrderRequest_2eproto)), true);
+static const ::_pbi::DescriptorTable* const descriptor_table_CancelOrderRequest_2eproto_deps[1] =
+    {
+        &::descriptor_table_OrderCancel_2eproto,
+};
+static ::absl::once_flag descriptor_table_CancelOrderRequest_2eproto_once;
+PROTOBUF_CONSTINIT const ::_pbi::DescriptorTable descriptor_table_CancelOrderRequest_2eproto = {
+    false,
+    false,
+    252,
+    descriptor_table_protodef_CancelOrderRequest_2eproto,
+    "CancelOrderRequest.proto",
+    &descriptor_table_CancelOrderRequest_2eproto_once,
+    descriptor_table_CancelOrderRequest_2eproto_deps,
+    1,
+    1,
+    schemas,
+    file_default_instances,
+    TableStruct_CancelOrderRequest_2eproto::offsets,
+    file_level_enum_descriptors_CancelOrderRequest_2eproto,
+    file_level_service_descriptors_CancelOrderRequest_2eproto,
+};
 namespace protobuf {
-
 // ===================================================================
 
-void CancelOrderRequest::InitAsDefaultInstance() {
-  ::protobuf::_CancelOrderRequest_default_instance_._instance.get_mutable()->ordercancel_ = const_cast< ::protobuf::OrderCancel*>(
-      ::protobuf::OrderCancel::internal_default_instance());
-}
 class CancelOrderRequest::_Internal {
  public:
-  using HasBits = decltype(std::declval<CancelOrderRequest>()._has_bits_);
-  static void set_has_orderid(HasBits* has_bits) {
-    (*has_bits)[0] |= 2u;
-  }
-  static const ::protobuf::OrderCancel& ordercancel(const CancelOrderRequest* msg);
-  static void set_has_ordercancel(HasBits* has_bits) {
-    (*has_bits)[0] |= 1u;
-  }
+  using HasBits =
+      decltype(std::declval<CancelOrderRequest>()._impl_._has_bits_);
+  static constexpr ::int32_t kHasBitsOffset =
+      8 * PROTOBUF_FIELD_OFFSET(CancelOrderRequest, _impl_._has_bits_);
 };
 
-const ::protobuf::OrderCancel&
-CancelOrderRequest::_Internal::ordercancel(const CancelOrderRequest* msg) {
-  return *msg->ordercancel_;
-}
 void CancelOrderRequest::clear_ordercancel() {
-  if (GetArena() == nullptr && ordercancel_ != nullptr) {
-    delete ordercancel_;
-  }
-  ordercancel_ = nullptr;
-  _has_bits_[0] &= ~0x00000001u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (_impl_.ordercancel_ != nullptr) _impl_.ordercancel_->Clear();
+  _impl_._has_bits_[0] &= ~0x00000001u;
 }
-CancelOrderRequest::CancelOrderRequest(::PROTOBUF_NAMESPACE_ID::Arena* arena)
-  : ::PROTOBUF_NAMESPACE_ID::Message(arena) {
-  SharedCtor();
-  RegisterArenaDtor(arena);
+CancelOrderRequest::CancelOrderRequest(::google::protobuf::Arena* arena)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  SharedCtor(arena);
   // @@protoc_insertion_point(arena_constructor:protobuf.CancelOrderRequest)
 }
-CancelOrderRequest::CancelOrderRequest(const CancelOrderRequest& from)
-  : ::PROTOBUF_NAMESPACE_ID::Message(),
-      _has_bits_(from._has_bits_) {
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  if (from._internal_has_ordercancel()) {
-    ordercancel_ = new ::protobuf::OrderCancel(*from.ordercancel_);
-  } else {
-    ordercancel_ = nullptr;
-  }
-  orderid_ = from.orderid_;
+inline PROTOBUF_NDEBUG_INLINE CancelOrderRequest::Impl_::Impl_(
+    ::google::protobuf::internal::InternalVisibility visibility, ::google::protobuf::Arena* arena,
+    const Impl_& from, const ::protobuf::CancelOrderRequest& from_msg)
+      : _has_bits_{from._has_bits_},
+        _cached_size_{0} {}
+
+CancelOrderRequest::CancelOrderRequest(
+    ::google::protobuf::Arena* arena,
+    const CancelOrderRequest& from)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  CancelOrderRequest* const _this = this;
+  (void)_this;
+  _internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(
+      from._internal_metadata_);
+  new (&_impl_) Impl_(internal_visibility(), arena, from._impl_, from);
+  ::uint32_t cached_has_bits = _impl_._has_bits_[0];
+  _impl_.ordercancel_ = (cached_has_bits & 0x00000001u) ? ::google::protobuf::Message::CopyConstruct<::protobuf::OrderCancel>(
+                              arena, *from._impl_.ordercancel_)
+                        : nullptr;
+  _impl_.orderid_ = from._impl_.orderid_;
+
   // @@protoc_insertion_point(copy_constructor:protobuf.CancelOrderRequest)
 }
+inline PROTOBUF_NDEBUG_INLINE CancelOrderRequest::Impl_::Impl_(
+    ::google::protobuf::internal::InternalVisibility visibility,
+    ::google::protobuf::Arena* arena)
+      : _cached_size_{0} {}
 
-void CancelOrderRequest::SharedCtor() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&scc_info_CancelOrderRequest_CancelOrderRequest_2eproto.base);
-  ::memset(&ordercancel_, 0, static_cast<size_t>(
-      reinterpret_cast<char*>(&orderid_) -
-      reinterpret_cast<char*>(&ordercancel_)) + sizeof(orderid_));
+inline void CancelOrderRequest::SharedCtor(::_pb::Arena* arena) {
+  new (&_impl_) Impl_(internal_visibility(), arena);
+  ::memset(reinterpret_cast<char *>(&_impl_) +
+               offsetof(Impl_, ordercancel_),
+           0,
+           offsetof(Impl_, orderid_) -
+               offsetof(Impl_, ordercancel_) +
+               sizeof(Impl_::orderid_));
 }
-
 CancelOrderRequest::~CancelOrderRequest() {
   // @@protoc_insertion_point(destructor:protobuf.CancelOrderRequest)
-  SharedDtor();
-  _internal_metadata_.Delete<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  SharedDtor(*this);
+}
+inline void CancelOrderRequest::SharedDtor(MessageLite& self) {
+  CancelOrderRequest& this_ = static_cast<CancelOrderRequest&>(self);
+  this_._internal_metadata_.Delete<::google::protobuf::UnknownFieldSet>();
+  ABSL_DCHECK(this_.GetArena() == nullptr);
+  delete this_._impl_.ordercancel_;
+  this_._impl_.~Impl_();
 }
 
-void CancelOrderRequest::SharedDtor() {
-  GOOGLE_DCHECK(GetArena() == nullptr);
-  if (this != internal_default_instance()) delete ordercancel_;
+inline void* CancelOrderRequest::PlacementNew_(const void*, void* mem,
+                                        ::google::protobuf::Arena* arena) {
+  return ::new (mem) CancelOrderRequest(arena);
 }
+constexpr auto CancelOrderRequest::InternalNewImpl_() {
+  return ::google::protobuf::internal::MessageCreator::ZeroInit(sizeof(CancelOrderRequest),
+                                            alignof(CancelOrderRequest));
+}
+PROTOBUF_CONSTINIT
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::google::protobuf::internal::ClassDataFull CancelOrderRequest::_class_data_ = {
+    ::google::protobuf::internal::ClassData{
+        &_CancelOrderRequest_default_instance_._instance,
+        &_table_.header,
+        nullptr,  // OnDemandRegisterArenaDtor
+        nullptr,  // IsInitialized
+        &CancelOrderRequest::MergeImpl,
+        ::google::protobuf::Message::GetNewImpl<CancelOrderRequest>(),
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        &CancelOrderRequest::SharedDtor,
+        ::google::protobuf::Message::GetClearImpl<CancelOrderRequest>(), &CancelOrderRequest::ByteSizeLong,
+            &CancelOrderRequest::_InternalSerialize,
+#endif  // PROTOBUF_CUSTOM_VTABLE
+        PROTOBUF_FIELD_OFFSET(CancelOrderRequest, _impl_._cached_size_),
+        false,
+    },
+    &CancelOrderRequest::kDescriptorMethods,
+    &descriptor_table_CancelOrderRequest_2eproto,
+    nullptr,  // tracker
+};
+const ::google::protobuf::internal::ClassData* CancelOrderRequest::GetClassData() const {
+  ::google::protobuf::internal::PrefetchToLocalCache(&_class_data_);
+  ::google::protobuf::internal::PrefetchToLocalCache(_class_data_.tc_table);
+  return _class_data_.base();
+}
+PROTOBUF_CONSTINIT PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::_pbi::TcParseTable<1, 2, 1, 0, 2> CancelOrderRequest::_table_ = {
+  {
+    PROTOBUF_FIELD_OFFSET(CancelOrderRequest, _impl_._has_bits_),
+    0, // no _extensions_
+    2, 8,  // max_field_number, fast_idx_mask
+    offsetof(decltype(_table_), field_lookup_table),
+    4294967292,  // skipmap
+    offsetof(decltype(_table_), field_entries),
+    2,  // num_field_entries
+    1,  // num_aux_entries
+    offsetof(decltype(_table_), aux_entries),
+    _class_data_.base(),
+    nullptr,  // post_loop_handler
+    ::_pbi::TcParser::GenericFallback,  // fallback
+    #ifdef PROTOBUF_PREFETCH_PARSE_TABLE
+    ::_pbi::TcParser::GetTable<::protobuf::CancelOrderRequest>(),  // to_prefetch
+    #endif  // PROTOBUF_PREFETCH_PARSE_TABLE
+  }, {{
+    // optional .protobuf.OrderCancel orderCancel = 2;
+    {::_pbi::TcParser::FastMtS1,
+     {18, 0, 0, PROTOBUF_FIELD_OFFSET(CancelOrderRequest, _impl_.ordercancel_)}},
+    // optional int32 orderId = 1;
+    {::_pbi::TcParser::SingularVarintNoZag1<::uint32_t, offsetof(CancelOrderRequest, _impl_.orderid_), 1>(),
+     {8, 1, 0, PROTOBUF_FIELD_OFFSET(CancelOrderRequest, _impl_.orderid_)}},
+  }}, {{
+    65535, 65535
+  }}, {{
+    // optional int32 orderId = 1;
+    {PROTOBUF_FIELD_OFFSET(CancelOrderRequest, _impl_.orderid_), _Internal::kHasBitsOffset + 1, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kInt32)},
+    // optional .protobuf.OrderCancel orderCancel = 2;
+    {PROTOBUF_FIELD_OFFSET(CancelOrderRequest, _impl_.ordercancel_), _Internal::kHasBitsOffset + 0, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kMessage | ::_fl::kTvTable)},
+  }}, {{
+    {::_pbi::TcParser::GetTable<::protobuf::OrderCancel>()},
+  }}, {{
+  }},
+};
 
-void CancelOrderRequest::ArenaDtor(void* object) {
-  CancelOrderRequest* _this = reinterpret_cast< CancelOrderRequest* >(object);
-  (void)_this;
-}
-void CancelOrderRequest::RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena*) {
-}
-void CancelOrderRequest::SetCachedSize(int size) const {
-  _cached_size_.Set(size);
-}
-const CancelOrderRequest& CancelOrderRequest::default_instance() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&::scc_info_CancelOrderRequest_CancelOrderRequest_2eproto.base);
-  return *internal_default_instance();
-}
-
-
-void CancelOrderRequest::Clear() {
+PROTOBUF_NOINLINE void CancelOrderRequest::Clear() {
 // @@protoc_insertion_point(message_clear_start:protobuf.CancelOrderRequest)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  ::uint32_t cached_has_bits = 0;
   // Prevent compiler warnings about cached_has_bits being unused
   (void) cached_has_bits;
 
-  cached_has_bits = _has_bits_[0];
+  cached_has_bits = _impl_._has_bits_[0];
   if (cached_has_bits & 0x00000001u) {
-    if (GetArena() == nullptr && ordercancel_ != nullptr) {
-      delete ordercancel_;
-    }
-    ordercancel_ = nullptr;
+    ABSL_DCHECK(_impl_.ordercancel_ != nullptr);
+    _impl_.ordercancel_->Clear();
   }
-  orderid_ = 0;
-  _has_bits_.Clear();
-  _internal_metadata_.Clear<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  _impl_.orderid_ = 0;
+  _impl_._has_bits_.Clear();
+  _internal_metadata_.Clear<::google::protobuf::UnknownFieldSet>();
 }
 
-const char* CancelOrderRequest::_InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) {
-#define CHK_(x) if (PROTOBUF_PREDICT_FALSE(!(x))) goto failure
-  _Internal::HasBits has_bits{};
-  ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena(); (void)arena;
-  while (!ctx->Done(&ptr)) {
-    ::PROTOBUF_NAMESPACE_ID::uint32 tag;
-    ptr = ::PROTOBUF_NAMESPACE_ID::internal::ReadTag(ptr, &tag);
-    CHK_(ptr);
-    switch (tag >> 3) {
-      // int32 orderId = 1;
-      case 1:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 8)) {
-          _Internal::set_has_orderid(&has_bits);
-          orderid_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // .protobuf.OrderCancel orderCancel = 2;
-      case 2:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 18)) {
-          ptr = ctx->ParseMessage(_internal_mutable_ordercancel(), ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      default: {
-      handle_unusual:
-        if ((tag & 7) == 4 || tag == 0) {
-          ctx->SetLastTag(tag);
-          goto success;
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::uint8_t* CancelOrderRequest::_InternalSerialize(
+            const MessageLite& base, ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) {
+          const CancelOrderRequest& this_ = static_cast<const CancelOrderRequest&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::uint8_t* CancelOrderRequest::_InternalSerialize(
+            ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+          const CancelOrderRequest& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(serialize_to_array_start:protobuf.CancelOrderRequest)
+          ::uint32_t cached_has_bits = 0;
+          (void)cached_has_bits;
+
+          cached_has_bits = this_._impl_._has_bits_[0];
+          // optional int32 orderId = 1;
+          if (cached_has_bits & 0x00000002u) {
+            target = ::google::protobuf::internal::WireFormatLite::
+                WriteInt32ToArrayWithField<1>(
+                    stream, this_._internal_orderid(), target);
+          }
+
+          // optional .protobuf.OrderCancel orderCancel = 2;
+          if (cached_has_bits & 0x00000001u) {
+            target = ::google::protobuf::internal::WireFormatLite::InternalWriteMessage(
+                2, *this_._impl_.ordercancel_, this_._impl_.ordercancel_->GetCachedSize(), target,
+                stream);
+          }
+
+          if (PROTOBUF_PREDICT_FALSE(this_._internal_metadata_.have_unknown_fields())) {
+            target =
+                ::_pbi::WireFormat::InternalSerializeUnknownFieldsToArray(
+                    this_._internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance), target, stream);
+          }
+          // @@protoc_insertion_point(serialize_to_array_end:protobuf.CancelOrderRequest)
+          return target;
         }
-        ptr = UnknownFieldParse(tag,
-            _internal_metadata_.mutable_unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(),
-            ptr, ctx);
-        CHK_(ptr != nullptr);
-        continue;
+
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::size_t CancelOrderRequest::ByteSizeLong(const MessageLite& base) {
+          const CancelOrderRequest& this_ = static_cast<const CancelOrderRequest&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::size_t CancelOrderRequest::ByteSizeLong() const {
+          const CancelOrderRequest& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(message_byte_size_start:protobuf.CancelOrderRequest)
+          ::size_t total_size = 0;
+
+          ::uint32_t cached_has_bits = 0;
+          // Prevent compiler warnings about cached_has_bits being unused
+          (void)cached_has_bits;
+
+          ::_pbi::Prefetch5LinesFrom7Lines(&this_);
+          cached_has_bits = this_._impl_._has_bits_[0];
+          if (cached_has_bits & 0x00000003u) {
+            // optional .protobuf.OrderCancel orderCancel = 2;
+            if (cached_has_bits & 0x00000001u) {
+              total_size += 1 +
+                            ::google::protobuf::internal::WireFormatLite::MessageSize(*this_._impl_.ordercancel_);
+            }
+            // optional int32 orderId = 1;
+            if (cached_has_bits & 0x00000002u) {
+              total_size += ::_pbi::WireFormatLite::Int32SizePlusOne(
+                  this_._internal_orderid());
+            }
+          }
+          return this_.MaybeComputeUnknownFieldsSize(total_size,
+                                                     &this_._impl_._cached_size_);
+        }
+
+void CancelOrderRequest::MergeImpl(::google::protobuf::MessageLite& to_msg, const ::google::protobuf::MessageLite& from_msg) {
+  auto* const _this = static_cast<CancelOrderRequest*>(&to_msg);
+  auto& from = static_cast<const CancelOrderRequest&>(from_msg);
+  ::google::protobuf::Arena* arena = _this->GetArena();
+  // @@protoc_insertion_point(class_specific_merge_from_start:protobuf.CancelOrderRequest)
+  ABSL_DCHECK_NE(&from, _this);
+  ::uint32_t cached_has_bits = 0;
+  (void) cached_has_bits;
+
+  cached_has_bits = from._impl_._has_bits_[0];
+  if (cached_has_bits & 0x00000003u) {
+    if (cached_has_bits & 0x00000001u) {
+      ABSL_DCHECK(from._impl_.ordercancel_ != nullptr);
+      if (_this->_impl_.ordercancel_ == nullptr) {
+        _this->_impl_.ordercancel_ =
+            ::google::protobuf::Message::CopyConstruct<::protobuf::OrderCancel>(arena, *from._impl_.ordercancel_);
+      } else {
+        _this->_impl_.ordercancel_->MergeFrom(*from._impl_.ordercancel_);
       }
-    }  // switch
-  }  // while
-success:
-  _has_bits_.Or(has_bits);
-  return ptr;
-failure:
-  ptr = nullptr;
-  goto success;
-#undef CHK_
-}
-
-::PROTOBUF_NAMESPACE_ID::uint8* CancelOrderRequest::_InternalSerialize(
-    ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const {
-  // @@protoc_insertion_point(serialize_to_array_start:protobuf.CancelOrderRequest)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  (void) cached_has_bits;
-
-  // int32 orderId = 1;
-  if (_internal_has_orderid()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteInt32ToArray(1, this->_internal_orderid(), target);
-  }
-
-  // .protobuf.OrderCancel orderCancel = 2;
-  if (_internal_has_ordercancel()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::
-      InternalWriteMessage(
-        2, _Internal::ordercancel(this), target, stream);
-  }
-
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormat::InternalSerializeUnknownFieldsToArray(
-        _internal_metadata_.unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(::PROTOBUF_NAMESPACE_ID::UnknownFieldSet::default_instance), target, stream);
-  }
-  // @@protoc_insertion_point(serialize_to_array_end:protobuf.CancelOrderRequest)
-  return target;
-}
-
-size_t CancelOrderRequest::ByteSizeLong() const {
-// @@protoc_insertion_point(message_byte_size_start:protobuf.CancelOrderRequest)
-  size_t total_size = 0;
-
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  // Prevent compiler warnings about cached_has_bits being unused
-  (void) cached_has_bits;
-
-  cached_has_bits = _has_bits_[0];
-  if (cached_has_bits & 0x00000003u) {
-    // .protobuf.OrderCancel orderCancel = 2;
-    if (cached_has_bits & 0x00000001u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::MessageSize(
-          *ordercancel_);
-    }
-
-    // int32 orderId = 1;
-    if (cached_has_bits & 0x00000002u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::Int32Size(
-          this->_internal_orderid());
-    }
-
-  }
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    return ::PROTOBUF_NAMESPACE_ID::internal::ComputeUnknownFieldsSize(
-        _internal_metadata_, total_size, &_cached_size_);
-  }
-  int cached_size = ::PROTOBUF_NAMESPACE_ID::internal::ToCachedSize(total_size);
-  SetCachedSize(cached_size);
-  return total_size;
-}
-
-void CancelOrderRequest::MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_merge_from_start:protobuf.CancelOrderRequest)
-  GOOGLE_DCHECK_NE(&from, this);
-  const CancelOrderRequest* source =
-      ::PROTOBUF_NAMESPACE_ID::DynamicCastToGenerated<CancelOrderRequest>(
-          &from);
-  if (source == nullptr) {
-  // @@protoc_insertion_point(generalized_merge_from_cast_fail:protobuf.CancelOrderRequest)
-    ::PROTOBUF_NAMESPACE_ID::internal::ReflectionOps::Merge(from, this);
-  } else {
-  // @@protoc_insertion_point(generalized_merge_from_cast_success:protobuf.CancelOrderRequest)
-    MergeFrom(*source);
-  }
-}
-
-void CancelOrderRequest::MergeFrom(const CancelOrderRequest& from) {
-// @@protoc_insertion_point(class_specific_merge_from_start:protobuf.CancelOrderRequest)
-  GOOGLE_DCHECK_NE(&from, this);
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  (void) cached_has_bits;
-
-  cached_has_bits = from._has_bits_[0];
-  if (cached_has_bits & 0x00000003u) {
-    if (cached_has_bits & 0x00000001u) {
-      _internal_mutable_ordercancel()->::protobuf::OrderCancel::MergeFrom(from._internal_ordercancel());
     }
     if (cached_has_bits & 0x00000002u) {
-      orderid_ = from.orderid_;
+      _this->_impl_.orderid_ = from._impl_.orderid_;
     }
-    _has_bits_[0] |= cached_has_bits;
   }
-}
-
-void CancelOrderRequest::CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_copy_from_start:protobuf.CancelOrderRequest)
-  if (&from == this) return;
-  Clear();
-  MergeFrom(from);
+  _this->_impl_._has_bits_[0] |= cached_has_bits;
+  _this->_internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(from._internal_metadata_);
 }
 
 void CancelOrderRequest::CopyFrom(const CancelOrderRequest& from) {
@@ -342,34 +387,31 @@ void CancelOrderRequest::CopyFrom(const CancelOrderRequest& from) {
   MergeFrom(from);
 }
 
-bool CancelOrderRequest::IsInitialized() const {
-  return true;
-}
 
-void CancelOrderRequest::InternalSwap(CancelOrderRequest* other) {
+void CancelOrderRequest::InternalSwap(CancelOrderRequest* PROTOBUF_RESTRICT other) {
   using std::swap;
-  _internal_metadata_.Swap<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(&other->_internal_metadata_);
-  swap(_has_bits_[0], other->_has_bits_[0]);
-  ::PROTOBUF_NAMESPACE_ID::internal::memswap<
-      PROTOBUF_FIELD_OFFSET(CancelOrderRequest, orderid_)
-      + sizeof(CancelOrderRequest::orderid_)
-      - PROTOBUF_FIELD_OFFSET(CancelOrderRequest, ordercancel_)>(
-          reinterpret_cast<char*>(&ordercancel_),
-          reinterpret_cast<char*>(&other->ordercancel_));
+  _internal_metadata_.InternalSwap(&other->_internal_metadata_);
+  swap(_impl_._has_bits_[0], other->_impl_._has_bits_[0]);
+  ::google::protobuf::internal::memswap<
+      PROTOBUF_FIELD_OFFSET(CancelOrderRequest, _impl_.orderid_)
+      + sizeof(CancelOrderRequest::_impl_.orderid_)
+      - PROTOBUF_FIELD_OFFSET(CancelOrderRequest, _impl_.ordercancel_)>(
+          reinterpret_cast<char*>(&_impl_.ordercancel_),
+          reinterpret_cast<char*>(&other->_impl_.ordercancel_));
 }
 
-::PROTOBUF_NAMESPACE_ID::Metadata CancelOrderRequest::GetMetadata() const {
-  return GetMetadataStatic();
+::google::protobuf::Metadata CancelOrderRequest::GetMetadata() const {
+  return ::google::protobuf::Message::GetMetadataImpl(GetClassData()->full());
 }
-
-
 // @@protoc_insertion_point(namespace_scope)
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> PROTOBUF_NOINLINE ::protobuf::CancelOrderRequest* Arena::CreateMaybeMessage< ::protobuf::CancelOrderRequest >(Arena* arena) {
-  return Arena::CreateMessageInternal< ::protobuf::CancelOrderRequest >(arena);
-}
-PROTOBUF_NAMESPACE_CLOSE
-
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
 // @@protoc_insertion_point(global_scope)
-#include <google/protobuf/port_undef.inc>
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY2 static ::std::false_type
+    _static_init2_ PROTOBUF_UNUSED =
+        (::_pbi::AddDescriptors(&descriptor_table_CancelOrderRequest_2eproto),
+         ::std::false_type{});
+#include "google/protobuf/port_undef.inc"
