@@ -1,183 +1,218 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: ComboLeg.proto
+// Protobuf C++ Version: 5.29.3
 
-#ifndef GOOGLE_PROTOBUF_INCLUDED_ComboLeg_2eproto
-#define GOOGLE_PROTOBUF_INCLUDED_ComboLeg_2eproto
+#ifndef ComboLeg_2eproto_2epb_2eh
+#define ComboLeg_2eproto_2epb_2eh
 
 #include <limits>
 #include <string>
+#include <type_traits>
+#include <utility>
 
-#include <google/protobuf/port_def.inc>
-#if PROTOBUF_VERSION < 3012000
-#error This file was generated by a newer version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please update
-#error your headers.
+#include "google/protobuf/runtime_version.h"
+#if PROTOBUF_VERSION != 5029003
+#error "Protobuf C++ gencode is built with an incompatible version of"
+#error "Protobuf C++ headers/runtime. See"
+#error "https://protobuf.dev/support/cross-version-runtime-guarantee/#cpp"
 #endif
-#if 3012004 < PROTOBUF_MIN_PROTOC_VERSION
-#error This file was generated by an older version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please
-#error regenerate this file with a newer version of protoc.
-#endif
-
-#include <google/protobuf/port_undef.inc>
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/arena.h>
-#include <google/protobuf/arenastring.h>
-#include <google/protobuf/generated_message_table_driven.h>
-#include <google/protobuf/generated_message_util.h>
-#include <google/protobuf/inlined_string_field.h>
-#include <google/protobuf/metadata_lite.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/message.h>
-#include <google/protobuf/repeated_field.h>  // IWYU pragma: export
-#include <google/protobuf/extension_set.h>  // IWYU pragma: export
-#include <google/protobuf/unknown_field_set.h>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/arena.h"
+#include "google/protobuf/arenastring.h"
+#include "google/protobuf/generated_message_tctable_decl.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/metadata_lite.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/message.h"
+#include "google/protobuf/message_lite.h"
+#include "google/protobuf/repeated_field.h"  // IWYU pragma: export
+#include "google/protobuf/extension_set.h"  // IWYU pragma: export
+#include "google/protobuf/unknown_field_set.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
+
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+
 #define PROTOBUF_INTERNAL_EXPORT_ComboLeg_2eproto
-PROTOBUF_NAMESPACE_OPEN
+
+namespace google {
+namespace protobuf {
 namespace internal {
-class AnyMetadata;
+template <typename T>
+::absl::string_view GetAnyMessageName();
 }  // namespace internal
-PROTOBUF_NAMESPACE_CLOSE
+}  // namespace protobuf
+}  // namespace google
 
 // Internal implementation detail -- do not use these members.
 struct TableStruct_ComboLeg_2eproto {
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTableField entries[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxillaryParseTableField aux[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTable schema[1]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::FieldMetadata field_metadata[];
-  static const ::PROTOBUF_NAMESPACE_ID::internal::SerializationTable serialization_table[];
-  static const ::PROTOBUF_NAMESPACE_ID::uint32 offsets[];
+  static const ::uint32_t offsets[];
 };
-extern const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_ComboLeg_2eproto;
+extern const ::google::protobuf::internal::DescriptorTable
+    descriptor_table_ComboLeg_2eproto;
 namespace protobuf {
 class ComboLeg;
-class ComboLegDefaultTypeInternal;
+struct ComboLegDefaultTypeInternal;
 extern ComboLegDefaultTypeInternal _ComboLeg_default_instance_;
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> ::protobuf::ComboLeg* Arena::CreateMaybeMessage<::protobuf::ComboLeg>(Arena*);
-PROTOBUF_NAMESPACE_CLOSE
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
+
 namespace protobuf {
 
 // ===================================================================
 
-class ComboLeg PROTOBUF_FINAL :
-    public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:protobuf.ComboLeg) */ {
+
+// -------------------------------------------------------------------
+
+class ComboLeg final : public ::google::protobuf::Message
+/* @@protoc_insertion_point(class_definition:protobuf.ComboLeg) */ {
  public:
-  inline ComboLeg() : ComboLeg(nullptr) {};
-  virtual ~ComboLeg();
+  inline ComboLeg() : ComboLeg(nullptr) {}
+  ~ComboLeg() PROTOBUF_FINAL;
 
-  ComboLeg(const ComboLeg& from);
-  ComboLeg(ComboLeg&& from) noexcept
-    : ComboLeg() {
-    *this = ::std::move(from);
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+  void operator delete(ComboLeg* msg, std::destroying_delete_t) {
+    SharedDtor(*msg);
+    ::google::protobuf::internal::SizedDelete(msg, sizeof(ComboLeg));
   }
+#endif
 
+  template <typename = void>
+  explicit PROTOBUF_CONSTEXPR ComboLeg(
+      ::google::protobuf::internal::ConstantInitialized);
+
+  inline ComboLeg(const ComboLeg& from) : ComboLeg(nullptr, from) {}
+  inline ComboLeg(ComboLeg&& from) noexcept
+      : ComboLeg(nullptr, std::move(from)) {}
   inline ComboLeg& operator=(const ComboLeg& from) {
     CopyFrom(from);
     return *this;
   }
   inline ComboLeg& operator=(ComboLeg&& from) noexcept {
-    if (GetArena() == from.GetArena()) {
-      if (this != &from) InternalSwap(&from);
+    if (this == &from) return *this;
+    if (::google::protobuf::internal::CanMoveWithInternalSwap(GetArena(), from.GetArena())) {
+      InternalSwap(&from);
     } else {
       CopyFrom(from);
     }
     return *this;
   }
 
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* descriptor() {
+  inline const ::google::protobuf::UnknownFieldSet& unknown_fields() const
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance);
+  }
+  inline ::google::protobuf::UnknownFieldSet* mutable_unknown_fields()
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.mutable_unknown_fields<::google::protobuf::UnknownFieldSet>();
+  }
+
+  static const ::google::protobuf::Descriptor* descriptor() {
     return GetDescriptor();
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* GetDescriptor() {
-    return GetMetadataStatic().descriptor;
+  static const ::google::protobuf::Descriptor* GetDescriptor() {
+    return default_instance().GetMetadata().descriptor;
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Reflection* GetReflection() {
-    return GetMetadataStatic().reflection;
+  static const ::google::protobuf::Reflection* GetReflection() {
+    return default_instance().GetMetadata().reflection;
   }
-  static const ComboLeg& default_instance();
-
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
+  static const ComboLeg& default_instance() {
+    return *internal_default_instance();
+  }
   static inline const ComboLeg* internal_default_instance() {
     return reinterpret_cast<const ComboLeg*>(
-               &_ComboLeg_default_instance_);
+        &_ComboLeg_default_instance_);
   }
-  static constexpr int kIndexInFileMessages =
-    0;
-
-  friend void swap(ComboLeg& a, ComboLeg& b) {
-    a.Swap(&b);
-  }
+  static constexpr int kIndexInFileMessages = 0;
+  friend void swap(ComboLeg& a, ComboLeg& b) { a.Swap(&b); }
   inline void Swap(ComboLeg* other) {
     if (other == this) return;
-    if (GetArena() == other->GetArena()) {
+    if (::google::protobuf::internal::CanUseInternalSwap(GetArena(), other->GetArena())) {
       InternalSwap(other);
     } else {
-      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+      ::google::protobuf::internal::GenericSwap(this, other);
     }
   }
   void UnsafeArenaSwap(ComboLeg* other) {
     if (other == this) return;
-    GOOGLE_DCHECK(GetArena() == other->GetArena());
+    ABSL_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
   // implements Message ----------------------------------------------
 
-  inline ComboLeg* New() const final {
-    return CreateMaybeMessage<ComboLeg>(nullptr);
+  ComboLeg* New(::google::protobuf::Arena* arena = nullptr) const {
+    return ::google::protobuf::Message::DefaultConstruct<ComboLeg>(arena);
   }
-
-  ComboLeg* New(::PROTOBUF_NAMESPACE_ID::Arena* arena) const final {
-    return CreateMaybeMessage<ComboLeg>(arena);
-  }
-  void CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
-  void MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
+  using ::google::protobuf::Message::CopyFrom;
   void CopyFrom(const ComboLeg& from);
-  void MergeFrom(const ComboLeg& from);
-  PROTOBUF_ATTRIBUTE_REINITIALIZES void Clear() final;
-  bool IsInitialized() const final;
-
-  size_t ByteSizeLong() const final;
-  const char* _InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) final;
-  ::PROTOBUF_NAMESPACE_ID::uint8* _InternalSerialize(
-      ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const final;
-  int GetCachedSize() const final { return _cached_size_.Get(); }
+  using ::google::protobuf::Message::MergeFrom;
+  void MergeFrom(const ComboLeg& from) { ComboLeg::MergeImpl(*this, from); }
 
   private:
-  inline void SharedCtor();
-  inline void SharedDtor();
-  void SetCachedSize(int size) const final;
+  static void MergeImpl(
+      ::google::protobuf::MessageLite& to_msg,
+      const ::google::protobuf::MessageLite& from_msg);
+
+  public:
+  bool IsInitialized() const {
+    return true;
+  }
+  ABSL_ATTRIBUTE_REINITIALIZES void Clear() PROTOBUF_FINAL;
+  #if defined(PROTOBUF_CUSTOM_VTABLE)
+  private:
+  static ::size_t ByteSizeLong(const ::google::protobuf::MessageLite& msg);
+  static ::uint8_t* _InternalSerialize(
+      const MessageLite& msg, ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream);
+
+  public:
+  ::size_t ByteSizeLong() const { return ByteSizeLong(*this); }
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+    return _InternalSerialize(*this, target, stream);
+  }
+  #else   // PROTOBUF_CUSTOM_VTABLE
+  ::size_t ByteSizeLong() const final;
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const final;
+  #endif  // PROTOBUF_CUSTOM_VTABLE
+  int GetCachedSize() const { return _impl_._cached_size_.Get(); }
+
+  private:
+  void SharedCtor(::google::protobuf::Arena* arena);
+  static void SharedDtor(MessageLite& self);
   void InternalSwap(ComboLeg* other);
-  friend class ::PROTOBUF_NAMESPACE_ID::internal::AnyMetadata;
-  static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
-    return "protobuf.ComboLeg";
+ private:
+  template <typename T>
+  friend ::absl::string_view(
+      ::google::protobuf::internal::GetAnyMessageName)();
+  static ::absl::string_view FullMessageName() { return "protobuf.ComboLeg"; }
+
+ protected:
+  explicit ComboLeg(::google::protobuf::Arena* arena);
+  ComboLeg(::google::protobuf::Arena* arena, const ComboLeg& from);
+  ComboLeg(::google::protobuf::Arena* arena, ComboLeg&& from) noexcept
+      : ComboLeg(arena) {
+    *this = ::std::move(from);
   }
-  protected:
-  explicit ComboLeg(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  private:
-  static void ArenaDtor(void* object);
-  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  public:
+  const ::google::protobuf::internal::ClassData* GetClassData() const PROTOBUF_FINAL;
+  static void* PlacementNew_(const void*, void* mem,
+                             ::google::protobuf::Arena* arena);
+  static constexpr auto InternalNewImpl_();
+  static const ::google::protobuf::internal::ClassDataFull _class_data_;
 
-  ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
-  private:
-  static ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadataStatic() {
-    ::PROTOBUF_NAMESPACE_ID::internal::AssignDescriptors(&::descriptor_table_ComboLeg_2eproto);
-    return ::descriptor_table_ComboLeg_2eproto.file_level_metadata[kIndexInFileMessages];
-  }
-
-  public:
-
+ public:
+  ::google::protobuf::Metadata GetMetadata() const;
   // nested types ----------------------------------------------------
 
   // accessors -------------------------------------------------------
-
   enum : int {
     kActionFieldNumber = 3,
     kExchangeFieldNumber = 4,
@@ -189,658 +224,564 @@ class ComboLeg PROTOBUF_FINAL :
     kPerLegPriceFieldNumber = 9,
     kExemptCodeFieldNumber = 8,
   };
-  // string action = 3;
+  // optional string action = 3;
   bool has_action() const;
-  private:
-  bool _internal_has_action() const;
-  public:
-  void clear_action();
+  void clear_action() ;
   const std::string& action() const;
-  void set_action(const std::string& value);
-  void set_action(std::string&& value);
-  void set_action(const char* value);
-  void set_action(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_action(Arg_&& arg, Args_... args);
   std::string* mutable_action();
-  std::string* release_action();
-  void set_allocated_action(std::string* action);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_action();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_action(
-      std::string* action);
+  PROTOBUF_NODISCARD std::string* release_action();
+  void set_allocated_action(std::string* value);
+
   private:
   const std::string& _internal_action() const;
-  void _internal_set_action(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_action(
+      const std::string& value);
   std::string* _internal_mutable_action();
-  public:
 
-  // string exchange = 4;
-  bool has_exchange() const;
-  private:
-  bool _internal_has_exchange() const;
   public:
-  void clear_exchange();
+  // optional string exchange = 4;
+  bool has_exchange() const;
+  void clear_exchange() ;
   const std::string& exchange() const;
-  void set_exchange(const std::string& value);
-  void set_exchange(std::string&& value);
-  void set_exchange(const char* value);
-  void set_exchange(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_exchange(Arg_&& arg, Args_... args);
   std::string* mutable_exchange();
-  std::string* release_exchange();
-  void set_allocated_exchange(std::string* exchange);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_exchange();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_exchange(
-      std::string* exchange);
+  PROTOBUF_NODISCARD std::string* release_exchange();
+  void set_allocated_exchange(std::string* value);
+
   private:
   const std::string& _internal_exchange() const;
-  void _internal_set_exchange(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_exchange(
+      const std::string& value);
   std::string* _internal_mutable_exchange();
-  public:
 
-  // string designatedLocation = 7;
-  bool has_designatedlocation() const;
-  private:
-  bool _internal_has_designatedlocation() const;
   public:
-  void clear_designatedlocation();
+  // optional string designatedLocation = 7;
+  bool has_designatedlocation() const;
+  void clear_designatedlocation() ;
   const std::string& designatedlocation() const;
-  void set_designatedlocation(const std::string& value);
-  void set_designatedlocation(std::string&& value);
-  void set_designatedlocation(const char* value);
-  void set_designatedlocation(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_designatedlocation(Arg_&& arg, Args_... args);
   std::string* mutable_designatedlocation();
-  std::string* release_designatedlocation();
-  void set_allocated_designatedlocation(std::string* designatedlocation);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_designatedlocation();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_designatedlocation(
-      std::string* designatedlocation);
+  PROTOBUF_NODISCARD std::string* release_designatedlocation();
+  void set_allocated_designatedlocation(std::string* value);
+
   private:
   const std::string& _internal_designatedlocation() const;
-  void _internal_set_designatedlocation(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_designatedlocation(
+      const std::string& value);
   std::string* _internal_mutable_designatedlocation();
-  public:
 
-  // int32 conId = 1;
+  public:
+  // optional int32 conId = 1;
   bool has_conid() const;
-  private:
-  bool _internal_has_conid() const;
-  public:
-  void clear_conid();
-  ::PROTOBUF_NAMESPACE_ID::int32 conid() const;
-  void set_conid(::PROTOBUF_NAMESPACE_ID::int32 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int32 _internal_conid() const;
-  void _internal_set_conid(::PROTOBUF_NAMESPACE_ID::int32 value);
-  public:
+  void clear_conid() ;
+  ::int32_t conid() const;
+  void set_conid(::int32_t value);
 
-  // int32 ratio = 2;
+  private:
+  ::int32_t _internal_conid() const;
+  void _internal_set_conid(::int32_t value);
+
+  public:
+  // optional int32 ratio = 2;
   bool has_ratio() const;
-  private:
-  bool _internal_has_ratio() const;
-  public:
-  void clear_ratio();
-  ::PROTOBUF_NAMESPACE_ID::int32 ratio() const;
-  void set_ratio(::PROTOBUF_NAMESPACE_ID::int32 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int32 _internal_ratio() const;
-  void _internal_set_ratio(::PROTOBUF_NAMESPACE_ID::int32 value);
-  public:
+  void clear_ratio() ;
+  ::int32_t ratio() const;
+  void set_ratio(::int32_t value);
 
-  // int32 openClose = 5;
+  private:
+  ::int32_t _internal_ratio() const;
+  void _internal_set_ratio(::int32_t value);
+
+  public:
+  // optional int32 openClose = 5;
   bool has_openclose() const;
-  private:
-  bool _internal_has_openclose() const;
-  public:
-  void clear_openclose();
-  ::PROTOBUF_NAMESPACE_ID::int32 openclose() const;
-  void set_openclose(::PROTOBUF_NAMESPACE_ID::int32 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int32 _internal_openclose() const;
-  void _internal_set_openclose(::PROTOBUF_NAMESPACE_ID::int32 value);
-  public:
+  void clear_openclose() ;
+  ::int32_t openclose() const;
+  void set_openclose(::int32_t value);
 
-  // int32 shortSalesSlot = 6;
+  private:
+  ::int32_t _internal_openclose() const;
+  void _internal_set_openclose(::int32_t value);
+
+  public:
+  // optional int32 shortSalesSlot = 6;
   bool has_shortsalesslot() const;
-  private:
-  bool _internal_has_shortsalesslot() const;
-  public:
-  void clear_shortsalesslot();
-  ::PROTOBUF_NAMESPACE_ID::int32 shortsalesslot() const;
-  void set_shortsalesslot(::PROTOBUF_NAMESPACE_ID::int32 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int32 _internal_shortsalesslot() const;
-  void _internal_set_shortsalesslot(::PROTOBUF_NAMESPACE_ID::int32 value);
-  public:
+  void clear_shortsalesslot() ;
+  ::int32_t shortsalesslot() const;
+  void set_shortsalesslot(::int32_t value);
 
-  // double perLegPrice = 9;
-  bool has_perlegprice() const;
   private:
-  bool _internal_has_perlegprice() const;
+  ::int32_t _internal_shortsalesslot() const;
+  void _internal_set_shortsalesslot(::int32_t value);
+
   public:
-  void clear_perlegprice();
+  // optional double perLegPrice = 9;
+  bool has_perlegprice() const;
+  void clear_perlegprice() ;
   double perlegprice() const;
   void set_perlegprice(double value);
+
   private:
   double _internal_perlegprice() const;
   void _internal_set_perlegprice(double value);
-  public:
 
-  // int32 exemptCode = 8;
+  public:
+  // optional int32 exemptCode = 8;
   bool has_exemptcode() const;
-  private:
-  bool _internal_has_exemptcode() const;
-  public:
-  void clear_exemptcode();
-  ::PROTOBUF_NAMESPACE_ID::int32 exemptcode() const;
-  void set_exemptcode(::PROTOBUF_NAMESPACE_ID::int32 value);
-  private:
-  ::PROTOBUF_NAMESPACE_ID::int32 _internal_exemptcode() const;
-  void _internal_set_exemptcode(::PROTOBUF_NAMESPACE_ID::int32 value);
-  public:
+  void clear_exemptcode() ;
+  ::int32_t exemptcode() const;
+  void set_exemptcode(::int32_t value);
 
+  private:
+  ::int32_t _internal_exemptcode() const;
+  void _internal_set_exemptcode(::int32_t value);
+
+  public:
   // @@protoc_insertion_point(class_scope:protobuf.ComboLeg)
  private:
   class _Internal;
+  friend class ::google::protobuf::internal::TcParser;
+  static const ::google::protobuf::internal::TcParseTable<
+      4, 9, 0,
+      66, 2>
+      _table_;
 
-  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
-  typedef void InternalArenaConstructable_;
-  typedef void DestructorSkippable_;
-  ::PROTOBUF_NAMESPACE_ID::internal::HasBits<1> _has_bits_;
-  mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr action_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr exchange_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr designatedlocation_;
-  ::PROTOBUF_NAMESPACE_ID::int32 conid_;
-  ::PROTOBUF_NAMESPACE_ID::int32 ratio_;
-  ::PROTOBUF_NAMESPACE_ID::int32 openclose_;
-  ::PROTOBUF_NAMESPACE_ID::int32 shortsalesslot_;
-  double perlegprice_;
-  ::PROTOBUF_NAMESPACE_ID::int32 exemptcode_;
+  friend class ::google::protobuf::MessageLite;
+  friend class ::google::protobuf::Arena;
+  template <typename T>
+  friend class ::google::protobuf::Arena::InternalHelper;
+  using InternalArenaConstructable_ = void;
+  using DestructorSkippable_ = void;
+  struct Impl_ {
+    inline explicit constexpr Impl_(
+        ::google::protobuf::internal::ConstantInitialized) noexcept;
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena);
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena, const Impl_& from,
+                          const ComboLeg& from_msg);
+    ::google::protobuf::internal::HasBits<1> _has_bits_;
+    ::google::protobuf::internal::CachedSize _cached_size_;
+    ::google::protobuf::internal::ArenaStringPtr action_;
+    ::google::protobuf::internal::ArenaStringPtr exchange_;
+    ::google::protobuf::internal::ArenaStringPtr designatedlocation_;
+    ::int32_t conid_;
+    ::int32_t ratio_;
+    ::int32_t openclose_;
+    ::int32_t shortsalesslot_;
+    double perlegprice_;
+    ::int32_t exemptcode_;
+    PROTOBUF_TSAN_DECLARE_MEMBER
+  };
+  union { Impl_ _impl_; };
   friend struct ::TableStruct_ComboLeg_2eproto;
 };
+
 // ===================================================================
 
 
+
+
 // ===================================================================
+
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic push
-  #pragma GCC diagnostic ignored "-Wstrict-aliasing"
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wstrict-aliasing"
 #endif  // __GNUC__
+// -------------------------------------------------------------------
+
 // ComboLeg
 
-// int32 conId = 1;
-inline bool ComboLeg::_internal_has_conid() const {
-  bool value = (_has_bits_[0] & 0x00000008u) != 0;
+// optional int32 conId = 1;
+inline bool ComboLeg::has_conid() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000008u) != 0;
   return value;
 }
-inline bool ComboLeg::has_conid() const {
-  return _internal_has_conid();
-}
 inline void ComboLeg::clear_conid() {
-  conid_ = 0;
-  _has_bits_[0] &= ~0x00000008u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.conid_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000008u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int32 ComboLeg::_internal_conid() const {
-  return conid_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int32 ComboLeg::conid() const {
+inline ::int32_t ComboLeg::conid() const {
   // @@protoc_insertion_point(field_get:protobuf.ComboLeg.conId)
   return _internal_conid();
 }
-inline void ComboLeg::_internal_set_conid(::PROTOBUF_NAMESPACE_ID::int32 value) {
-  _has_bits_[0] |= 0x00000008u;
-  conid_ = value;
-}
-inline void ComboLeg::set_conid(::PROTOBUF_NAMESPACE_ID::int32 value) {
+inline void ComboLeg::set_conid(::int32_t value) {
   _internal_set_conid(value);
+  _impl_._has_bits_[0] |= 0x00000008u;
   // @@protoc_insertion_point(field_set:protobuf.ComboLeg.conId)
 }
+inline ::int32_t ComboLeg::_internal_conid() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.conid_;
+}
+inline void ComboLeg::_internal_set_conid(::int32_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.conid_ = value;
+}
 
-// int32 ratio = 2;
-inline bool ComboLeg::_internal_has_ratio() const {
-  bool value = (_has_bits_[0] & 0x00000010u) != 0;
+// optional int32 ratio = 2;
+inline bool ComboLeg::has_ratio() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000010u) != 0;
   return value;
 }
-inline bool ComboLeg::has_ratio() const {
-  return _internal_has_ratio();
-}
 inline void ComboLeg::clear_ratio() {
-  ratio_ = 0;
-  _has_bits_[0] &= ~0x00000010u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.ratio_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000010u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int32 ComboLeg::_internal_ratio() const {
-  return ratio_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int32 ComboLeg::ratio() const {
+inline ::int32_t ComboLeg::ratio() const {
   // @@protoc_insertion_point(field_get:protobuf.ComboLeg.ratio)
   return _internal_ratio();
 }
-inline void ComboLeg::_internal_set_ratio(::PROTOBUF_NAMESPACE_ID::int32 value) {
-  _has_bits_[0] |= 0x00000010u;
-  ratio_ = value;
-}
-inline void ComboLeg::set_ratio(::PROTOBUF_NAMESPACE_ID::int32 value) {
+inline void ComboLeg::set_ratio(::int32_t value) {
   _internal_set_ratio(value);
+  _impl_._has_bits_[0] |= 0x00000010u;
   // @@protoc_insertion_point(field_set:protobuf.ComboLeg.ratio)
 }
+inline ::int32_t ComboLeg::_internal_ratio() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.ratio_;
+}
+inline void ComboLeg::_internal_set_ratio(::int32_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.ratio_ = value;
+}
 
-// string action = 3;
-inline bool ComboLeg::_internal_has_action() const {
-  bool value = (_has_bits_[0] & 0x00000001u) != 0;
+// optional string action = 3;
+inline bool ComboLeg::has_action() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000001u) != 0;
   return value;
 }
-inline bool ComboLeg::has_action() const {
-  return _internal_has_action();
-}
 inline void ComboLeg::clear_action() {
-  action_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000001u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.action_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000001u;
 }
-inline const std::string& ComboLeg::action() const {
+inline const std::string& ComboLeg::action() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.ComboLeg.action)
   return _internal_action();
 }
-inline void ComboLeg::set_action(const std::string& value) {
-  _internal_set_action(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void ComboLeg::set_action(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  _impl_.action_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.ComboLeg.action)
 }
-inline std::string* ComboLeg::mutable_action() {
+inline std::string* ComboLeg::mutable_action() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_action();
   // @@protoc_insertion_point(field_mutable:protobuf.ComboLeg.action)
-  return _internal_mutable_action();
+  return _s;
 }
 inline const std::string& ComboLeg::_internal_action() const {
-  return action_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.action_.Get();
 }
 inline void ComboLeg::_internal_set_action(const std::string& value) {
-  _has_bits_[0] |= 0x00000001u;
-  action_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void ComboLeg::set_action(std::string&& value) {
-  _has_bits_[0] |= 0x00000001u;
-  action_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.ComboLeg.action)
-}
-inline void ComboLeg::set_action(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000001u;
-  action_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.ComboLeg.action)
-}
-inline void ComboLeg::set_action(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000001u;
-  action_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.ComboLeg.action)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  _impl_.action_.Set(value, GetArena());
 }
 inline std::string* ComboLeg::_internal_mutable_action() {
-  _has_bits_[0] |= 0x00000001u;
-  return action_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  return _impl_.action_.Mutable( GetArena());
 }
 inline std::string* ComboLeg::release_action() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.ComboLeg.action)
-  if (!_internal_has_action()) {
+  if ((_impl_._has_bits_[0] & 0x00000001u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000001u;
-  return action_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void ComboLeg::set_allocated_action(std::string* action) {
-  if (action != nullptr) {
-    _has_bits_[0] |= 0x00000001u;
-  } else {
-    _has_bits_[0] &= ~0x00000001u;
+  _impl_._has_bits_[0] &= ~0x00000001u;
+  auto* released = _impl_.action_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.action_.Set("", GetArena());
   }
-  action_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), action,
-      GetArena());
+  return released;
+}
+inline void ComboLeg::set_allocated_action(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000001u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000001u;
+  }
+  _impl_.action_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.action_.IsDefault()) {
+    _impl_.action_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.ComboLeg.action)
 }
-inline std::string* ComboLeg::unsafe_arena_release_action() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.ComboLeg.action)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000001u;
-  return action_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void ComboLeg::unsafe_arena_set_allocated_action(
-    std::string* action) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (action != nullptr) {
-    _has_bits_[0] |= 0x00000001u;
-  } else {
-    _has_bits_[0] &= ~0x00000001u;
-  }
-  action_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      action, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.ComboLeg.action)
-}
 
-// string exchange = 4;
-inline bool ComboLeg::_internal_has_exchange() const {
-  bool value = (_has_bits_[0] & 0x00000002u) != 0;
+// optional string exchange = 4;
+inline bool ComboLeg::has_exchange() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000002u) != 0;
   return value;
 }
-inline bool ComboLeg::has_exchange() const {
-  return _internal_has_exchange();
-}
 inline void ComboLeg::clear_exchange() {
-  exchange_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000002u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.exchange_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000002u;
 }
-inline const std::string& ComboLeg::exchange() const {
+inline const std::string& ComboLeg::exchange() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.ComboLeg.exchange)
   return _internal_exchange();
 }
-inline void ComboLeg::set_exchange(const std::string& value) {
-  _internal_set_exchange(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void ComboLeg::set_exchange(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  _impl_.exchange_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.ComboLeg.exchange)
 }
-inline std::string* ComboLeg::mutable_exchange() {
+inline std::string* ComboLeg::mutable_exchange() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_exchange();
   // @@protoc_insertion_point(field_mutable:protobuf.ComboLeg.exchange)
-  return _internal_mutable_exchange();
+  return _s;
 }
 inline const std::string& ComboLeg::_internal_exchange() const {
-  return exchange_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.exchange_.Get();
 }
 inline void ComboLeg::_internal_set_exchange(const std::string& value) {
-  _has_bits_[0] |= 0x00000002u;
-  exchange_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void ComboLeg::set_exchange(std::string&& value) {
-  _has_bits_[0] |= 0x00000002u;
-  exchange_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.ComboLeg.exchange)
-}
-inline void ComboLeg::set_exchange(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000002u;
-  exchange_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.ComboLeg.exchange)
-}
-inline void ComboLeg::set_exchange(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000002u;
-  exchange_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.ComboLeg.exchange)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  _impl_.exchange_.Set(value, GetArena());
 }
 inline std::string* ComboLeg::_internal_mutable_exchange() {
-  _has_bits_[0] |= 0x00000002u;
-  return exchange_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  return _impl_.exchange_.Mutable( GetArena());
 }
 inline std::string* ComboLeg::release_exchange() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.ComboLeg.exchange)
-  if (!_internal_has_exchange()) {
+  if ((_impl_._has_bits_[0] & 0x00000002u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000002u;
-  return exchange_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void ComboLeg::set_allocated_exchange(std::string* exchange) {
-  if (exchange != nullptr) {
-    _has_bits_[0] |= 0x00000002u;
-  } else {
-    _has_bits_[0] &= ~0x00000002u;
+  _impl_._has_bits_[0] &= ~0x00000002u;
+  auto* released = _impl_.exchange_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.exchange_.Set("", GetArena());
   }
-  exchange_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), exchange,
-      GetArena());
+  return released;
+}
+inline void ComboLeg::set_allocated_exchange(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000002u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000002u;
+  }
+  _impl_.exchange_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.exchange_.IsDefault()) {
+    _impl_.exchange_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.ComboLeg.exchange)
 }
-inline std::string* ComboLeg::unsafe_arena_release_exchange() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.ComboLeg.exchange)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000002u;
-  return exchange_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void ComboLeg::unsafe_arena_set_allocated_exchange(
-    std::string* exchange) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (exchange != nullptr) {
-    _has_bits_[0] |= 0x00000002u;
-  } else {
-    _has_bits_[0] &= ~0x00000002u;
-  }
-  exchange_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      exchange, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.ComboLeg.exchange)
-}
 
-// int32 openClose = 5;
-inline bool ComboLeg::_internal_has_openclose() const {
-  bool value = (_has_bits_[0] & 0x00000020u) != 0;
+// optional int32 openClose = 5;
+inline bool ComboLeg::has_openclose() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000020u) != 0;
   return value;
 }
-inline bool ComboLeg::has_openclose() const {
-  return _internal_has_openclose();
-}
 inline void ComboLeg::clear_openclose() {
-  openclose_ = 0;
-  _has_bits_[0] &= ~0x00000020u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.openclose_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000020u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int32 ComboLeg::_internal_openclose() const {
-  return openclose_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int32 ComboLeg::openclose() const {
+inline ::int32_t ComboLeg::openclose() const {
   // @@protoc_insertion_point(field_get:protobuf.ComboLeg.openClose)
   return _internal_openclose();
 }
-inline void ComboLeg::_internal_set_openclose(::PROTOBUF_NAMESPACE_ID::int32 value) {
-  _has_bits_[0] |= 0x00000020u;
-  openclose_ = value;
-}
-inline void ComboLeg::set_openclose(::PROTOBUF_NAMESPACE_ID::int32 value) {
+inline void ComboLeg::set_openclose(::int32_t value) {
   _internal_set_openclose(value);
+  _impl_._has_bits_[0] |= 0x00000020u;
   // @@protoc_insertion_point(field_set:protobuf.ComboLeg.openClose)
 }
+inline ::int32_t ComboLeg::_internal_openclose() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.openclose_;
+}
+inline void ComboLeg::_internal_set_openclose(::int32_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.openclose_ = value;
+}
 
-// int32 shortSalesSlot = 6;
-inline bool ComboLeg::_internal_has_shortsalesslot() const {
-  bool value = (_has_bits_[0] & 0x00000040u) != 0;
+// optional int32 shortSalesSlot = 6;
+inline bool ComboLeg::has_shortsalesslot() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000040u) != 0;
   return value;
 }
-inline bool ComboLeg::has_shortsalesslot() const {
-  return _internal_has_shortsalesslot();
-}
 inline void ComboLeg::clear_shortsalesslot() {
-  shortsalesslot_ = 0;
-  _has_bits_[0] &= ~0x00000040u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.shortsalesslot_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000040u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int32 ComboLeg::_internal_shortsalesslot() const {
-  return shortsalesslot_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int32 ComboLeg::shortsalesslot() const {
+inline ::int32_t ComboLeg::shortsalesslot() const {
   // @@protoc_insertion_point(field_get:protobuf.ComboLeg.shortSalesSlot)
   return _internal_shortsalesslot();
 }
-inline void ComboLeg::_internal_set_shortsalesslot(::PROTOBUF_NAMESPACE_ID::int32 value) {
-  _has_bits_[0] |= 0x00000040u;
-  shortsalesslot_ = value;
-}
-inline void ComboLeg::set_shortsalesslot(::PROTOBUF_NAMESPACE_ID::int32 value) {
+inline void ComboLeg::set_shortsalesslot(::int32_t value) {
   _internal_set_shortsalesslot(value);
+  _impl_._has_bits_[0] |= 0x00000040u;
   // @@protoc_insertion_point(field_set:protobuf.ComboLeg.shortSalesSlot)
 }
+inline ::int32_t ComboLeg::_internal_shortsalesslot() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.shortsalesslot_;
+}
+inline void ComboLeg::_internal_set_shortsalesslot(::int32_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.shortsalesslot_ = value;
+}
 
-// string designatedLocation = 7;
-inline bool ComboLeg::_internal_has_designatedlocation() const {
-  bool value = (_has_bits_[0] & 0x00000004u) != 0;
+// optional string designatedLocation = 7;
+inline bool ComboLeg::has_designatedlocation() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000004u) != 0;
   return value;
 }
-inline bool ComboLeg::has_designatedlocation() const {
-  return _internal_has_designatedlocation();
-}
 inline void ComboLeg::clear_designatedlocation() {
-  designatedlocation_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000004u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.designatedlocation_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000004u;
 }
-inline const std::string& ComboLeg::designatedlocation() const {
+inline const std::string& ComboLeg::designatedlocation() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.ComboLeg.designatedLocation)
   return _internal_designatedlocation();
 }
-inline void ComboLeg::set_designatedlocation(const std::string& value) {
-  _internal_set_designatedlocation(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void ComboLeg::set_designatedlocation(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000004u;
+  _impl_.designatedlocation_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.ComboLeg.designatedLocation)
 }
-inline std::string* ComboLeg::mutable_designatedlocation() {
+inline std::string* ComboLeg::mutable_designatedlocation() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_designatedlocation();
   // @@protoc_insertion_point(field_mutable:protobuf.ComboLeg.designatedLocation)
-  return _internal_mutable_designatedlocation();
+  return _s;
 }
 inline const std::string& ComboLeg::_internal_designatedlocation() const {
-  return designatedlocation_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.designatedlocation_.Get();
 }
 inline void ComboLeg::_internal_set_designatedlocation(const std::string& value) {
-  _has_bits_[0] |= 0x00000004u;
-  designatedlocation_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void ComboLeg::set_designatedlocation(std::string&& value) {
-  _has_bits_[0] |= 0x00000004u;
-  designatedlocation_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.ComboLeg.designatedLocation)
-}
-inline void ComboLeg::set_designatedlocation(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000004u;
-  designatedlocation_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.ComboLeg.designatedLocation)
-}
-inline void ComboLeg::set_designatedlocation(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000004u;
-  designatedlocation_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.ComboLeg.designatedLocation)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000004u;
+  _impl_.designatedlocation_.Set(value, GetArena());
 }
 inline std::string* ComboLeg::_internal_mutable_designatedlocation() {
-  _has_bits_[0] |= 0x00000004u;
-  return designatedlocation_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000004u;
+  return _impl_.designatedlocation_.Mutable( GetArena());
 }
 inline std::string* ComboLeg::release_designatedlocation() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.ComboLeg.designatedLocation)
-  if (!_internal_has_designatedlocation()) {
+  if ((_impl_._has_bits_[0] & 0x00000004u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000004u;
-  return designatedlocation_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void ComboLeg::set_allocated_designatedlocation(std::string* designatedlocation) {
-  if (designatedlocation != nullptr) {
-    _has_bits_[0] |= 0x00000004u;
-  } else {
-    _has_bits_[0] &= ~0x00000004u;
+  _impl_._has_bits_[0] &= ~0x00000004u;
+  auto* released = _impl_.designatedlocation_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.designatedlocation_.Set("", GetArena());
   }
-  designatedlocation_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), designatedlocation,
-      GetArena());
+  return released;
+}
+inline void ComboLeg::set_allocated_designatedlocation(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000004u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000004u;
+  }
+  _impl_.designatedlocation_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.designatedlocation_.IsDefault()) {
+    _impl_.designatedlocation_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.ComboLeg.designatedLocation)
 }
-inline std::string* ComboLeg::unsafe_arena_release_designatedlocation() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.ComboLeg.designatedLocation)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000004u;
-  return designatedlocation_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void ComboLeg::unsafe_arena_set_allocated_designatedlocation(
-    std::string* designatedlocation) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (designatedlocation != nullptr) {
-    _has_bits_[0] |= 0x00000004u;
-  } else {
-    _has_bits_[0] &= ~0x00000004u;
-  }
-  designatedlocation_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      designatedlocation, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.ComboLeg.designatedLocation)
-}
 
-// int32 exemptCode = 8;
-inline bool ComboLeg::_internal_has_exemptcode() const {
-  bool value = (_has_bits_[0] & 0x00000100u) != 0;
+// optional int32 exemptCode = 8;
+inline bool ComboLeg::has_exemptcode() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000100u) != 0;
   return value;
 }
-inline bool ComboLeg::has_exemptcode() const {
-  return _internal_has_exemptcode();
-}
 inline void ComboLeg::clear_exemptcode() {
-  exemptcode_ = 0;
-  _has_bits_[0] &= ~0x00000100u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.exemptcode_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000100u;
 }
-inline ::PROTOBUF_NAMESPACE_ID::int32 ComboLeg::_internal_exemptcode() const {
-  return exemptcode_;
-}
-inline ::PROTOBUF_NAMESPACE_ID::int32 ComboLeg::exemptcode() const {
+inline ::int32_t ComboLeg::exemptcode() const {
   // @@protoc_insertion_point(field_get:protobuf.ComboLeg.exemptCode)
   return _internal_exemptcode();
 }
-inline void ComboLeg::_internal_set_exemptcode(::PROTOBUF_NAMESPACE_ID::int32 value) {
-  _has_bits_[0] |= 0x00000100u;
-  exemptcode_ = value;
-}
-inline void ComboLeg::set_exemptcode(::PROTOBUF_NAMESPACE_ID::int32 value) {
+inline void ComboLeg::set_exemptcode(::int32_t value) {
   _internal_set_exemptcode(value);
+  _impl_._has_bits_[0] |= 0x00000100u;
   // @@protoc_insertion_point(field_set:protobuf.ComboLeg.exemptCode)
 }
+inline ::int32_t ComboLeg::_internal_exemptcode() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.exemptcode_;
+}
+inline void ComboLeg::_internal_set_exemptcode(::int32_t value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.exemptcode_ = value;
+}
 
-// double perLegPrice = 9;
-inline bool ComboLeg::_internal_has_perlegprice() const {
-  bool value = (_has_bits_[0] & 0x00000080u) != 0;
+// optional double perLegPrice = 9;
+inline bool ComboLeg::has_perlegprice() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000080u) != 0;
   return value;
 }
-inline bool ComboLeg::has_perlegprice() const {
-  return _internal_has_perlegprice();
-}
 inline void ComboLeg::clear_perlegprice() {
-  perlegprice_ = 0;
-  _has_bits_[0] &= ~0x00000080u;
-}
-inline double ComboLeg::_internal_perlegprice() const {
-  return perlegprice_;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.perlegprice_ = 0;
+  _impl_._has_bits_[0] &= ~0x00000080u;
 }
 inline double ComboLeg::perlegprice() const {
   // @@protoc_insertion_point(field_get:protobuf.ComboLeg.perLegPrice)
   return _internal_perlegprice();
 }
-inline void ComboLeg::_internal_set_perlegprice(double value) {
-  _has_bits_[0] |= 0x00000080u;
-  perlegprice_ = value;
-}
 inline void ComboLeg::set_perlegprice(double value) {
   _internal_set_perlegprice(value);
+  _impl_._has_bits_[0] |= 0x00000080u;
   // @@protoc_insertion_point(field_set:protobuf.ComboLeg.perLegPrice)
+}
+inline double ComboLeg::_internal_perlegprice() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.perlegprice_;
+}
+inline void ComboLeg::_internal_set_perlegprice(double value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.perlegprice_ = value;
 }
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic pop
+#pragma GCC diagnostic pop
 #endif  // __GNUC__
 
 // @@protoc_insertion_point(namespace_scope)
-
 }  // namespace protobuf
+
 
 // @@protoc_insertion_point(global_scope)
 
-#include <google/protobuf/port_undef.inc>
-#endif  // GOOGLE_PROTOBUF_INCLUDED_GOOGLE_PROTOBUF_INCLUDED_ComboLeg_2eproto
+#include "google/protobuf/port_undef.inc"
+
+#endif  // ComboLeg_2eproto_2epb_2eh
