@@ -1,434 +1,468 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: ErrorMessage.proto
+// Protobuf C++ Version: 5.29.3
 
 #include "ErrorMessage.pb.h"
 
 #include <algorithm>
-
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/extension_set.h>
-#include <google/protobuf/wire_format_lite.h>
-#include <google/protobuf/descriptor.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/reflection_ops.h>
-#include <google/protobuf/wire_format.h>
+#include <type_traits>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/generated_message_tctable_impl.h"
+#include "google/protobuf/extension_set.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/wire_format_lite.h"
+#include "google/protobuf/descriptor.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/reflection_ops.h"
+#include "google/protobuf/wire_format.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
-namespace protobuf {
-class ErrorMessageDefaultTypeInternal {
- public:
-  ::PROTOBUF_NAMESPACE_ID::internal::ExplicitlyConstructed<ErrorMessage> _instance;
-} _ErrorMessage_default_instance_;
-}  // namespace protobuf
-static void InitDefaultsscc_info_ErrorMessage_ErrorMessage_2eproto() {
-  GOOGLE_PROTOBUF_VERIFY_VERSION;
 
-  {
-    void* ptr = &::protobuf::_ErrorMessage_default_instance_;
-    new (ptr) ::protobuf::ErrorMessage();
-    ::PROTOBUF_NAMESPACE_ID::internal::OnShutdownDestroyMessage(ptr);
-  }
-  ::protobuf::ErrorMessage::InitAsDefaultInstance();
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+PROTOBUF_PRAGMA_INIT_SEG
+namespace _pb = ::google::protobuf;
+namespace _pbi = ::google::protobuf::internal;
+namespace _fl = ::google::protobuf::internal::field_layout;
+namespace protobuf {
+
+inline constexpr ErrorMessage::Impl_::Impl_(
+    ::_pbi::ConstantInitialized) noexcept
+      : _cached_size_{0},
+        errormsg_(
+            &::google::protobuf::internal::fixed_address_empty_string,
+            ::_pbi::ConstantInitialized()),
+        advancedorderrejectjson_(
+            &::google::protobuf::internal::fixed_address_empty_string,
+            ::_pbi::ConstantInitialized()),
+        errortime_{::int64_t{0}},
+        id_{0},
+        errorcode_{0} {}
+
+template <typename>
+PROTOBUF_CONSTEXPR ErrorMessage::ErrorMessage(::_pbi::ConstantInitialized)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(_class_data_.base()),
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(),
+#endif  // PROTOBUF_CUSTOM_VTABLE
+      _impl_(::_pbi::ConstantInitialized()) {
 }
-
-::PROTOBUF_NAMESPACE_ID::internal::SCCInfo<0> scc_info_ErrorMessage_ErrorMessage_2eproto =
-    {{ATOMIC_VAR_INIT(::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase::kUninitialized), 0, 0, InitDefaultsscc_info_ErrorMessage_ErrorMessage_2eproto}, {}};
-
-static ::PROTOBUF_NAMESPACE_ID::Metadata file_level_metadata_ErrorMessage_2eproto[1];
-static constexpr ::PROTOBUF_NAMESPACE_ID::EnumDescriptor const** file_level_enum_descriptors_ErrorMessage_2eproto = nullptr;
-static constexpr ::PROTOBUF_NAMESPACE_ID::ServiceDescriptor const** file_level_service_descriptors_ErrorMessage_2eproto = nullptr;
-
-const ::PROTOBUF_NAMESPACE_ID::uint32 TableStruct_ErrorMessage_2eproto::offsets[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, _has_bits_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, _internal_metadata_),
-  ~0u,  // no _extensions_
-  ~0u,  // no _oneof_case_
-  ~0u,  // no _weak_field_map_
-  PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, id_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, errortime_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, errorcode_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, errormsg_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, advancedorderrejectjson_),
-  3,
-  2,
-  4,
-  0,
-  1,
-};
-static const ::PROTOBUF_NAMESPACE_ID::internal::MigrationSchema schemas[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  { 0, 10, sizeof(::protobuf::ErrorMessage)},
+struct ErrorMessageDefaultTypeInternal {
+  PROTOBUF_CONSTEXPR ErrorMessageDefaultTypeInternal() : _instance(::_pbi::ConstantInitialized{}) {}
+  ~ErrorMessageDefaultTypeInternal() {}
+  union {
+    ErrorMessage _instance;
+  };
 };
 
-static ::PROTOBUF_NAMESPACE_ID::Message const * const file_default_instances[] = {
-  reinterpret_cast<const ::PROTOBUF_NAMESPACE_ID::Message*>(&::protobuf::_ErrorMessage_default_instance_),
+PROTOBUF_ATTRIBUTE_NO_DESTROY PROTOBUF_CONSTINIT
+    PROTOBUF_ATTRIBUTE_INIT_PRIORITY1 ErrorMessageDefaultTypeInternal _ErrorMessage_default_instance_;
+}  // namespace protobuf
+static constexpr const ::_pb::EnumDescriptor**
+    file_level_enum_descriptors_ErrorMessage_2eproto = nullptr;
+static constexpr const ::_pb::ServiceDescriptor**
+    file_level_service_descriptors_ErrorMessage_2eproto = nullptr;
+const ::uint32_t
+    TableStruct_ErrorMessage_2eproto::offsets[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+        protodesc_cold) = {
+        PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, _impl_._has_bits_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, _internal_metadata_),
+        ~0u,  // no _extensions_
+        ~0u,  // no _oneof_case_
+        ~0u,  // no _weak_field_map_
+        ~0u,  // no _inlined_string_donated_
+        ~0u,  // no _split_
+        ~0u,  // no sizeof(Split)
+        PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, _impl_.id_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, _impl_.errortime_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, _impl_.errorcode_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, _impl_.errormsg_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ErrorMessage, _impl_.advancedorderrejectjson_),
+        3,
+        2,
+        4,
+        0,
+        1,
 };
 
-const char descriptor_table_protodef_ErrorMessage_2eproto[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) =
-  "\n\022ErrorMessage.proto\022\010protobuf\"\330\001\n\014Error"
-  "Message\022\017\n\002id\030\001 \001(\005H\000\210\001\001\022\026\n\terrorTime\030\002 "
-  "\001(\003H\001\210\001\001\022\026\n\terrorCode\030\003 \001(\005H\002\210\001\001\022\025\n\010erro"
-  "rMsg\030\004 \001(\tH\003\210\001\001\022$\n\027advancedOrderRejectJs"
-  "on\030\005 \001(\tH\004\210\001\001B\005\n\003_idB\014\n\n_errorTimeB\014\n\n_e"
-  "rrorCodeB\013\n\t_errorMsgB\032\n\030_advancedOrderR"
-  "ejectJsonB<\n\026com.ib.client.protobufB\021Err"
-  "orMessageProto\252\002\016IBApi.protobufb\006proto3"
-  ;
-static const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable*const descriptor_table_ErrorMessage_2eproto_deps[1] = {
+static const ::_pbi::MigrationSchema
+    schemas[] ABSL_ATTRIBUTE_SECTION_VARIABLE(protodesc_cold) = {
+        {0, 13, -1, sizeof(::protobuf::ErrorMessage)},
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase*const descriptor_table_ErrorMessage_2eproto_sccs[1] = {
-  &scc_info_ErrorMessage_ErrorMessage_2eproto.base,
+static const ::_pb::Message* const file_default_instances[] = {
+    &::protobuf::_ErrorMessage_default_instance_._instance,
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::once_flag descriptor_table_ErrorMessage_2eproto_once;
-const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_ErrorMessage_2eproto = {
-  false, false, descriptor_table_protodef_ErrorMessage_2eproto, "ErrorMessage.proto", 319,
-  &descriptor_table_ErrorMessage_2eproto_once, descriptor_table_ErrorMessage_2eproto_sccs, descriptor_table_ErrorMessage_2eproto_deps, 1, 0,
-  schemas, file_default_instances, TableStruct_ErrorMessage_2eproto::offsets,
-  file_level_metadata_ErrorMessage_2eproto, 1, file_level_enum_descriptors_ErrorMessage_2eproto, file_level_service_descriptors_ErrorMessage_2eproto,
+const char descriptor_table_protodef_ErrorMessage_2eproto[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+    protodesc_cold) = {
+    "\n\022ErrorMessage.proto\022\010protobuf\"\330\001\n\014Error"
+    "Message\022\017\n\002id\030\001 \001(\005H\000\210\001\001\022\026\n\terrorTime\030\002 "
+    "\001(\003H\001\210\001\001\022\026\n\terrorCode\030\003 \001(\005H\002\210\001\001\022\025\n\010erro"
+    "rMsg\030\004 \001(\tH\003\210\001\001\022$\n\027advancedOrderRejectJs"
+    "on\030\005 \001(\tH\004\210\001\001B\005\n\003_idB\014\n\n_errorTimeB\014\n\n_e"
+    "rrorCodeB\013\n\t_errorMsgB\032\n\030_advancedOrderR"
+    "ejectJsonB<\n\026com.ib.client.protobufB\021Err"
+    "orMessageProto\252\002\016IBApi.protobufb\006proto3"
 };
-
-// Force running AddDescriptors() at dynamic initialization time.
-static bool dynamic_init_dummy_ErrorMessage_2eproto = (static_cast<void>(::PROTOBUF_NAMESPACE_ID::internal::AddDescriptors(&descriptor_table_ErrorMessage_2eproto)), true);
+static ::absl::once_flag descriptor_table_ErrorMessage_2eproto_once;
+PROTOBUF_CONSTINIT const ::_pbi::DescriptorTable descriptor_table_ErrorMessage_2eproto = {
+    false,
+    false,
+    319,
+    descriptor_table_protodef_ErrorMessage_2eproto,
+    "ErrorMessage.proto",
+    &descriptor_table_ErrorMessage_2eproto_once,
+    nullptr,
+    0,
+    1,
+    schemas,
+    file_default_instances,
+    TableStruct_ErrorMessage_2eproto::offsets,
+    file_level_enum_descriptors_ErrorMessage_2eproto,
+    file_level_service_descriptors_ErrorMessage_2eproto,
+};
 namespace protobuf {
-
 // ===================================================================
 
-void ErrorMessage::InitAsDefaultInstance() {
-}
 class ErrorMessage::_Internal {
  public:
-  using HasBits = decltype(std::declval<ErrorMessage>()._has_bits_);
-  static void set_has_id(HasBits* has_bits) {
-    (*has_bits)[0] |= 8u;
-  }
-  static void set_has_errortime(HasBits* has_bits) {
-    (*has_bits)[0] |= 4u;
-  }
-  static void set_has_errorcode(HasBits* has_bits) {
-    (*has_bits)[0] |= 16u;
-  }
-  static void set_has_errormsg(HasBits* has_bits) {
-    (*has_bits)[0] |= 1u;
-  }
-  static void set_has_advancedorderrejectjson(HasBits* has_bits) {
-    (*has_bits)[0] |= 2u;
-  }
+  using HasBits =
+      decltype(std::declval<ErrorMessage>()._impl_._has_bits_);
+  static constexpr ::int32_t kHasBitsOffset =
+      8 * PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_._has_bits_);
 };
 
-ErrorMessage::ErrorMessage(::PROTOBUF_NAMESPACE_ID::Arena* arena)
-  : ::PROTOBUF_NAMESPACE_ID::Message(arena) {
-  SharedCtor();
-  RegisterArenaDtor(arena);
+ErrorMessage::ErrorMessage(::google::protobuf::Arena* arena)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  SharedCtor(arena);
   // @@protoc_insertion_point(arena_constructor:protobuf.ErrorMessage)
 }
-ErrorMessage::ErrorMessage(const ErrorMessage& from)
-  : ::PROTOBUF_NAMESPACE_ID::Message(),
-      _has_bits_(from._has_bits_) {
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  errormsg_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  if (from._internal_has_errormsg()) {
-    errormsg_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), from._internal_errormsg(),
-      GetArena());
-  }
-  advancedorderrejectjson_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  if (from._internal_has_advancedorderrejectjson()) {
-    advancedorderrejectjson_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), from._internal_advancedorderrejectjson(),
-      GetArena());
-  }
-  ::memcpy(&errortime_, &from.errortime_,
-    static_cast<size_t>(reinterpret_cast<char*>(&errorcode_) -
-    reinterpret_cast<char*>(&errortime_)) + sizeof(errorcode_));
+inline PROTOBUF_NDEBUG_INLINE ErrorMessage::Impl_::Impl_(
+    ::google::protobuf::internal::InternalVisibility visibility, ::google::protobuf::Arena* arena,
+    const Impl_& from, const ::protobuf::ErrorMessage& from_msg)
+      : _has_bits_{from._has_bits_},
+        _cached_size_{0},
+        errormsg_(arena, from.errormsg_),
+        advancedorderrejectjson_(arena, from.advancedorderrejectjson_) {}
+
+ErrorMessage::ErrorMessage(
+    ::google::protobuf::Arena* arena,
+    const ErrorMessage& from)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  ErrorMessage* const _this = this;
+  (void)_this;
+  _internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(
+      from._internal_metadata_);
+  new (&_impl_) Impl_(internal_visibility(), arena, from._impl_, from);
+  ::memcpy(reinterpret_cast<char *>(&_impl_) +
+               offsetof(Impl_, errortime_),
+           reinterpret_cast<const char *>(&from._impl_) +
+               offsetof(Impl_, errortime_),
+           offsetof(Impl_, errorcode_) -
+               offsetof(Impl_, errortime_) +
+               sizeof(Impl_::errorcode_));
+
   // @@protoc_insertion_point(copy_constructor:protobuf.ErrorMessage)
 }
+inline PROTOBUF_NDEBUG_INLINE ErrorMessage::Impl_::Impl_(
+    ::google::protobuf::internal::InternalVisibility visibility,
+    ::google::protobuf::Arena* arena)
+      : _cached_size_{0},
+        errormsg_(arena),
+        advancedorderrejectjson_(arena) {}
 
-void ErrorMessage::SharedCtor() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&scc_info_ErrorMessage_ErrorMessage_2eproto.base);
-  errormsg_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  advancedorderrejectjson_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  ::memset(&errortime_, 0, static_cast<size_t>(
-      reinterpret_cast<char*>(&errorcode_) -
-      reinterpret_cast<char*>(&errortime_)) + sizeof(errorcode_));
+inline void ErrorMessage::SharedCtor(::_pb::Arena* arena) {
+  new (&_impl_) Impl_(internal_visibility(), arena);
+  ::memset(reinterpret_cast<char *>(&_impl_) +
+               offsetof(Impl_, errortime_),
+           0,
+           offsetof(Impl_, errorcode_) -
+               offsetof(Impl_, errortime_) +
+               sizeof(Impl_::errorcode_));
 }
-
 ErrorMessage::~ErrorMessage() {
   // @@protoc_insertion_point(destructor:protobuf.ErrorMessage)
-  SharedDtor();
-  _internal_metadata_.Delete<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  SharedDtor(*this);
+}
+inline void ErrorMessage::SharedDtor(MessageLite& self) {
+  ErrorMessage& this_ = static_cast<ErrorMessage&>(self);
+  this_._internal_metadata_.Delete<::google::protobuf::UnknownFieldSet>();
+  ABSL_DCHECK(this_.GetArena() == nullptr);
+  this_._impl_.errormsg_.Destroy();
+  this_._impl_.advancedorderrejectjson_.Destroy();
+  this_._impl_.~Impl_();
 }
 
-void ErrorMessage::SharedDtor() {
-  GOOGLE_DCHECK(GetArena() == nullptr);
-  errormsg_.DestroyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  advancedorderrejectjson_.DestroyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+inline void* ErrorMessage::PlacementNew_(const void*, void* mem,
+                                        ::google::protobuf::Arena* arena) {
+  return ::new (mem) ErrorMessage(arena);
 }
+constexpr auto ErrorMessage::InternalNewImpl_() {
+  return ::google::protobuf::internal::MessageCreator::CopyInit(sizeof(ErrorMessage),
+                                            alignof(ErrorMessage));
+}
+PROTOBUF_CONSTINIT
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::google::protobuf::internal::ClassDataFull ErrorMessage::_class_data_ = {
+    ::google::protobuf::internal::ClassData{
+        &_ErrorMessage_default_instance_._instance,
+        &_table_.header,
+        nullptr,  // OnDemandRegisterArenaDtor
+        nullptr,  // IsInitialized
+        &ErrorMessage::MergeImpl,
+        ::google::protobuf::Message::GetNewImpl<ErrorMessage>(),
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        &ErrorMessage::SharedDtor,
+        ::google::protobuf::Message::GetClearImpl<ErrorMessage>(), &ErrorMessage::ByteSizeLong,
+            &ErrorMessage::_InternalSerialize,
+#endif  // PROTOBUF_CUSTOM_VTABLE
+        PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_._cached_size_),
+        false,
+    },
+    &ErrorMessage::kDescriptorMethods,
+    &descriptor_table_ErrorMessage_2eproto,
+    nullptr,  // tracker
+};
+const ::google::protobuf::internal::ClassData* ErrorMessage::GetClassData() const {
+  ::google::protobuf::internal::PrefetchToLocalCache(&_class_data_);
+  ::google::protobuf::internal::PrefetchToLocalCache(_class_data_.tc_table);
+  return _class_data_.base();
+}
+PROTOBUF_CONSTINIT PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::_pbi::TcParseTable<3, 5, 0, 61, 2> ErrorMessage::_table_ = {
+  {
+    PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_._has_bits_),
+    0, // no _extensions_
+    5, 56,  // max_field_number, fast_idx_mask
+    offsetof(decltype(_table_), field_lookup_table),
+    4294967264,  // skipmap
+    offsetof(decltype(_table_), field_entries),
+    5,  // num_field_entries
+    0,  // num_aux_entries
+    offsetof(decltype(_table_), field_names),  // no aux_entries
+    _class_data_.base(),
+    nullptr,  // post_loop_handler
+    ::_pbi::TcParser::GenericFallback,  // fallback
+    #ifdef PROTOBUF_PREFETCH_PARSE_TABLE
+    ::_pbi::TcParser::GetTable<::protobuf::ErrorMessage>(),  // to_prefetch
+    #endif  // PROTOBUF_PREFETCH_PARSE_TABLE
+  }, {{
+    {::_pbi::TcParser::MiniParse, {}},
+    // optional int32 id = 1;
+    {::_pbi::TcParser::SingularVarintNoZag1<::uint32_t, offsetof(ErrorMessage, _impl_.id_), 3>(),
+     {8, 3, 0, PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.id_)}},
+    // optional int64 errorTime = 2;
+    {::_pbi::TcParser::SingularVarintNoZag1<::uint64_t, offsetof(ErrorMessage, _impl_.errortime_), 2>(),
+     {16, 2, 0, PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.errortime_)}},
+    // optional int32 errorCode = 3;
+    {::_pbi::TcParser::SingularVarintNoZag1<::uint32_t, offsetof(ErrorMessage, _impl_.errorcode_), 4>(),
+     {24, 4, 0, PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.errorcode_)}},
+    // optional string errorMsg = 4;
+    {::_pbi::TcParser::FastUS1,
+     {34, 0, 0, PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.errormsg_)}},
+    // optional string advancedOrderRejectJson = 5;
+    {::_pbi::TcParser::FastUS1,
+     {42, 1, 0, PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.advancedorderrejectjson_)}},
+    {::_pbi::TcParser::MiniParse, {}},
+    {::_pbi::TcParser::MiniParse, {}},
+  }}, {{
+    65535, 65535
+  }}, {{
+    // optional int32 id = 1;
+    {PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.id_), _Internal::kHasBitsOffset + 3, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kInt32)},
+    // optional int64 errorTime = 2;
+    {PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.errortime_), _Internal::kHasBitsOffset + 2, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kInt64)},
+    // optional int32 errorCode = 3;
+    {PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.errorcode_), _Internal::kHasBitsOffset + 4, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kInt32)},
+    // optional string errorMsg = 4;
+    {PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.errormsg_), _Internal::kHasBitsOffset + 0, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kUtf8String | ::_fl::kRepAString)},
+    // optional string advancedOrderRejectJson = 5;
+    {PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.advancedorderrejectjson_), _Internal::kHasBitsOffset + 1, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kUtf8String | ::_fl::kRepAString)},
+  }},
+  // no aux_entries
+  {{
+    "\25\0\0\0\10\27\0\0"
+    "protobuf.ErrorMessage"
+    "errorMsg"
+    "advancedOrderRejectJson"
+  }},
+};
 
-void ErrorMessage::ArenaDtor(void* object) {
-  ErrorMessage* _this = reinterpret_cast< ErrorMessage* >(object);
-  (void)_this;
-}
-void ErrorMessage::RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena*) {
-}
-void ErrorMessage::SetCachedSize(int size) const {
-  _cached_size_.Set(size);
-}
-const ErrorMessage& ErrorMessage::default_instance() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&::scc_info_ErrorMessage_ErrorMessage_2eproto.base);
-  return *internal_default_instance();
-}
-
-
-void ErrorMessage::Clear() {
+PROTOBUF_NOINLINE void ErrorMessage::Clear() {
 // @@protoc_insertion_point(message_clear_start:protobuf.ErrorMessage)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  ::uint32_t cached_has_bits = 0;
   // Prevent compiler warnings about cached_has_bits being unused
   (void) cached_has_bits;
 
-  cached_has_bits = _has_bits_[0];
+  cached_has_bits = _impl_._has_bits_[0];
   if (cached_has_bits & 0x00000003u) {
     if (cached_has_bits & 0x00000001u) {
-      errormsg_.ClearNonDefaultToEmpty();
+      _impl_.errormsg_.ClearNonDefaultToEmpty();
     }
     if (cached_has_bits & 0x00000002u) {
-      advancedorderrejectjson_.ClearNonDefaultToEmpty();
+      _impl_.advancedorderrejectjson_.ClearNonDefaultToEmpty();
     }
   }
   if (cached_has_bits & 0x0000001cu) {
-    ::memset(&errortime_, 0, static_cast<size_t>(
-        reinterpret_cast<char*>(&errorcode_) -
-        reinterpret_cast<char*>(&errortime_)) + sizeof(errorcode_));
+    ::memset(&_impl_.errortime_, 0, static_cast<::size_t>(
+        reinterpret_cast<char*>(&_impl_.errorcode_) -
+        reinterpret_cast<char*>(&_impl_.errortime_)) + sizeof(_impl_.errorcode_));
   }
-  _has_bits_.Clear();
-  _internal_metadata_.Clear<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  _impl_._has_bits_.Clear();
+  _internal_metadata_.Clear<::google::protobuf::UnknownFieldSet>();
 }
 
-const char* ErrorMessage::_InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) {
-#define CHK_(x) if (PROTOBUF_PREDICT_FALSE(!(x))) goto failure
-  _Internal::HasBits has_bits{};
-  ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena(); (void)arena;
-  while (!ctx->Done(&ptr)) {
-    ::PROTOBUF_NAMESPACE_ID::uint32 tag;
-    ptr = ::PROTOBUF_NAMESPACE_ID::internal::ReadTag(ptr, &tag);
-    CHK_(ptr);
-    switch (tag >> 3) {
-      // int32 id = 1;
-      case 1:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 8)) {
-          _Internal::set_has_id(&has_bits);
-          id_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // int64 errorTime = 2;
-      case 2:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 16)) {
-          _Internal::set_has_errortime(&has_bits);
-          errortime_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // int32 errorCode = 3;
-      case 3:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 24)) {
-          _Internal::set_has_errorcode(&has_bits);
-          errorcode_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // string errorMsg = 4;
-      case 4:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 34)) {
-          auto str = _internal_mutable_errormsg();
-          ptr = ::PROTOBUF_NAMESPACE_ID::internal::InlineGreedyStringParser(str, ptr, ctx);
-          CHK_(::PROTOBUF_NAMESPACE_ID::internal::VerifyUTF8(str, "protobuf.ErrorMessage.errorMsg"));
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // string advancedOrderRejectJson = 5;
-      case 5:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 42)) {
-          auto str = _internal_mutable_advancedorderrejectjson();
-          ptr = ::PROTOBUF_NAMESPACE_ID::internal::InlineGreedyStringParser(str, ptr, ctx);
-          CHK_(::PROTOBUF_NAMESPACE_ID::internal::VerifyUTF8(str, "protobuf.ErrorMessage.advancedOrderRejectJson"));
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      default: {
-      handle_unusual:
-        if ((tag & 7) == 4 || tag == 0) {
-          ctx->SetLastTag(tag);
-          goto success;
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::uint8_t* ErrorMessage::_InternalSerialize(
+            const MessageLite& base, ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) {
+          const ErrorMessage& this_ = static_cast<const ErrorMessage&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::uint8_t* ErrorMessage::_InternalSerialize(
+            ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+          const ErrorMessage& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(serialize_to_array_start:protobuf.ErrorMessage)
+          ::uint32_t cached_has_bits = 0;
+          (void)cached_has_bits;
+
+          cached_has_bits = this_._impl_._has_bits_[0];
+          // optional int32 id = 1;
+          if (cached_has_bits & 0x00000008u) {
+            target = ::google::protobuf::internal::WireFormatLite::
+                WriteInt32ToArrayWithField<1>(
+                    stream, this_._internal_id(), target);
+          }
+
+          // optional int64 errorTime = 2;
+          if (cached_has_bits & 0x00000004u) {
+            target = ::google::protobuf::internal::WireFormatLite::
+                WriteInt64ToArrayWithField<2>(
+                    stream, this_._internal_errortime(), target);
+          }
+
+          // optional int32 errorCode = 3;
+          if (cached_has_bits & 0x00000010u) {
+            target = ::google::protobuf::internal::WireFormatLite::
+                WriteInt32ToArrayWithField<3>(
+                    stream, this_._internal_errorcode(), target);
+          }
+
+          // optional string errorMsg = 4;
+          if (cached_has_bits & 0x00000001u) {
+            const std::string& _s = this_._internal_errormsg();
+            ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
+                _s.data(), static_cast<int>(_s.length()), ::google::protobuf::internal::WireFormatLite::SERIALIZE, "protobuf.ErrorMessage.errorMsg");
+            target = stream->WriteStringMaybeAliased(4, _s, target);
+          }
+
+          // optional string advancedOrderRejectJson = 5;
+          if (cached_has_bits & 0x00000002u) {
+            const std::string& _s = this_._internal_advancedorderrejectjson();
+            ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
+                _s.data(), static_cast<int>(_s.length()), ::google::protobuf::internal::WireFormatLite::SERIALIZE, "protobuf.ErrorMessage.advancedOrderRejectJson");
+            target = stream->WriteStringMaybeAliased(5, _s, target);
+          }
+
+          if (PROTOBUF_PREDICT_FALSE(this_._internal_metadata_.have_unknown_fields())) {
+            target =
+                ::_pbi::WireFormat::InternalSerializeUnknownFieldsToArray(
+                    this_._internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance), target, stream);
+          }
+          // @@protoc_insertion_point(serialize_to_array_end:protobuf.ErrorMessage)
+          return target;
         }
-        ptr = UnknownFieldParse(tag,
-            _internal_metadata_.mutable_unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(),
-            ptr, ctx);
-        CHK_(ptr != nullptr);
-        continue;
-      }
-    }  // switch
-  }  // while
-success:
-  _has_bits_.Or(has_bits);
-  return ptr;
-failure:
-  ptr = nullptr;
-  goto success;
-#undef CHK_
-}
 
-::PROTOBUF_NAMESPACE_ID::uint8* ErrorMessage::_InternalSerialize(
-    ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const {
-  // @@protoc_insertion_point(serialize_to_array_start:protobuf.ErrorMessage)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::size_t ErrorMessage::ByteSizeLong(const MessageLite& base) {
+          const ErrorMessage& this_ = static_cast<const ErrorMessage&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::size_t ErrorMessage::ByteSizeLong() const {
+          const ErrorMessage& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(message_byte_size_start:protobuf.ErrorMessage)
+          ::size_t total_size = 0;
+
+          ::uint32_t cached_has_bits = 0;
+          // Prevent compiler warnings about cached_has_bits being unused
+          (void)cached_has_bits;
+
+          ::_pbi::Prefetch5LinesFrom7Lines(&this_);
+          cached_has_bits = this_._impl_._has_bits_[0];
+          if (cached_has_bits & 0x0000001fu) {
+            // optional string errorMsg = 4;
+            if (cached_has_bits & 0x00000001u) {
+              total_size += 1 + ::google::protobuf::internal::WireFormatLite::StringSize(
+                                              this_._internal_errormsg());
+            }
+            // optional string advancedOrderRejectJson = 5;
+            if (cached_has_bits & 0x00000002u) {
+              total_size += 1 + ::google::protobuf::internal::WireFormatLite::StringSize(
+                                              this_._internal_advancedorderrejectjson());
+            }
+            // optional int64 errorTime = 2;
+            if (cached_has_bits & 0x00000004u) {
+              total_size += ::_pbi::WireFormatLite::Int64SizePlusOne(
+                  this_._internal_errortime());
+            }
+            // optional int32 id = 1;
+            if (cached_has_bits & 0x00000008u) {
+              total_size += ::_pbi::WireFormatLite::Int32SizePlusOne(
+                  this_._internal_id());
+            }
+            // optional int32 errorCode = 3;
+            if (cached_has_bits & 0x00000010u) {
+              total_size += ::_pbi::WireFormatLite::Int32SizePlusOne(
+                  this_._internal_errorcode());
+            }
+          }
+          return this_.MaybeComputeUnknownFieldsSize(total_size,
+                                                     &this_._impl_._cached_size_);
+        }
+
+void ErrorMessage::MergeImpl(::google::protobuf::MessageLite& to_msg, const ::google::protobuf::MessageLite& from_msg) {
+  auto* const _this = static_cast<ErrorMessage*>(&to_msg);
+  auto& from = static_cast<const ErrorMessage&>(from_msg);
+  // @@protoc_insertion_point(class_specific_merge_from_start:protobuf.ErrorMessage)
+  ABSL_DCHECK_NE(&from, _this);
+  ::uint32_t cached_has_bits = 0;
   (void) cached_has_bits;
 
-  // int32 id = 1;
-  if (_internal_has_id()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteInt32ToArray(1, this->_internal_id(), target);
-  }
-
-  // int64 errorTime = 2;
-  if (_internal_has_errortime()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteInt64ToArray(2, this->_internal_errortime(), target);
-  }
-
-  // int32 errorCode = 3;
-  if (_internal_has_errorcode()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteInt32ToArray(3, this->_internal_errorcode(), target);
-  }
-
-  // string errorMsg = 4;
-  if (_internal_has_errormsg()) {
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
-      this->_internal_errormsg().data(), static_cast<int>(this->_internal_errormsg().length()),
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
-      "protobuf.ErrorMessage.errorMsg");
-    target = stream->WriteStringMaybeAliased(
-        4, this->_internal_errormsg(), target);
-  }
-
-  // string advancedOrderRejectJson = 5;
-  if (_internal_has_advancedorderrejectjson()) {
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
-      this->_internal_advancedorderrejectjson().data(), static_cast<int>(this->_internal_advancedorderrejectjson().length()),
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
-      "protobuf.ErrorMessage.advancedOrderRejectJson");
-    target = stream->WriteStringMaybeAliased(
-        5, this->_internal_advancedorderrejectjson(), target);
-  }
-
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormat::InternalSerializeUnknownFieldsToArray(
-        _internal_metadata_.unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(::PROTOBUF_NAMESPACE_ID::UnknownFieldSet::default_instance), target, stream);
-  }
-  // @@protoc_insertion_point(serialize_to_array_end:protobuf.ErrorMessage)
-  return target;
-}
-
-size_t ErrorMessage::ByteSizeLong() const {
-// @@protoc_insertion_point(message_byte_size_start:protobuf.ErrorMessage)
-  size_t total_size = 0;
-
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  // Prevent compiler warnings about cached_has_bits being unused
-  (void) cached_has_bits;
-
-  cached_has_bits = _has_bits_[0];
-  if (cached_has_bits & 0x0000001fu) {
-    // string errorMsg = 4;
-    if (cached_has_bits & 0x00000001u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::StringSize(
-          this->_internal_errormsg());
-    }
-
-    // string advancedOrderRejectJson = 5;
-    if (cached_has_bits & 0x00000002u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::StringSize(
-          this->_internal_advancedorderrejectjson());
-    }
-
-    // int64 errorTime = 2;
-    if (cached_has_bits & 0x00000004u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::Int64Size(
-          this->_internal_errortime());
-    }
-
-    // int32 id = 1;
-    if (cached_has_bits & 0x00000008u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::Int32Size(
-          this->_internal_id());
-    }
-
-    // int32 errorCode = 3;
-    if (cached_has_bits & 0x00000010u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::Int32Size(
-          this->_internal_errorcode());
-    }
-
-  }
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    return ::PROTOBUF_NAMESPACE_ID::internal::ComputeUnknownFieldsSize(
-        _internal_metadata_, total_size, &_cached_size_);
-  }
-  int cached_size = ::PROTOBUF_NAMESPACE_ID::internal::ToCachedSize(total_size);
-  SetCachedSize(cached_size);
-  return total_size;
-}
-
-void ErrorMessage::MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_merge_from_start:protobuf.ErrorMessage)
-  GOOGLE_DCHECK_NE(&from, this);
-  const ErrorMessage* source =
-      ::PROTOBUF_NAMESPACE_ID::DynamicCastToGenerated<ErrorMessage>(
-          &from);
-  if (source == nullptr) {
-  // @@protoc_insertion_point(generalized_merge_from_cast_fail:protobuf.ErrorMessage)
-    ::PROTOBUF_NAMESPACE_ID::internal::ReflectionOps::Merge(from, this);
-  } else {
-  // @@protoc_insertion_point(generalized_merge_from_cast_success:protobuf.ErrorMessage)
-    MergeFrom(*source);
-  }
-}
-
-void ErrorMessage::MergeFrom(const ErrorMessage& from) {
-// @@protoc_insertion_point(class_specific_merge_from_start:protobuf.ErrorMessage)
-  GOOGLE_DCHECK_NE(&from, this);
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  (void) cached_has_bits;
-
-  cached_has_bits = from._has_bits_[0];
+  cached_has_bits = from._impl_._has_bits_[0];
   if (cached_has_bits & 0x0000001fu) {
     if (cached_has_bits & 0x00000001u) {
-      _internal_set_errormsg(from._internal_errormsg());
+      _this->_internal_set_errormsg(from._internal_errormsg());
     }
     if (cached_has_bits & 0x00000002u) {
-      _internal_set_advancedorderrejectjson(from._internal_advancedorderrejectjson());
+      _this->_internal_set_advancedorderrejectjson(from._internal_advancedorderrejectjson());
     }
     if (cached_has_bits & 0x00000004u) {
-      errortime_ = from.errortime_;
+      _this->_impl_.errortime_ = from._impl_.errortime_;
     }
     if (cached_has_bits & 0x00000008u) {
-      id_ = from.id_;
+      _this->_impl_.id_ = from._impl_.id_;
     }
     if (cached_has_bits & 0x00000010u) {
-      errorcode_ = from.errorcode_;
+      _this->_impl_.errorcode_ = from._impl_.errorcode_;
     }
-    _has_bits_[0] |= cached_has_bits;
   }
-}
-
-void ErrorMessage::CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_copy_from_start:protobuf.ErrorMessage)
-  if (&from == this) return;
-  Clear();
-  MergeFrom(from);
+  _this->_impl_._has_bits_[0] |= cached_has_bits;
+  _this->_internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(from._internal_metadata_);
 }
 
 void ErrorMessage::CopyFrom(const ErrorMessage& from) {
@@ -438,36 +472,35 @@ void ErrorMessage::CopyFrom(const ErrorMessage& from) {
   MergeFrom(from);
 }
 
-bool ErrorMessage::IsInitialized() const {
-  return true;
-}
 
-void ErrorMessage::InternalSwap(ErrorMessage* other) {
+void ErrorMessage::InternalSwap(ErrorMessage* PROTOBUF_RESTRICT other) {
   using std::swap;
-  _internal_metadata_.Swap<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(&other->_internal_metadata_);
-  swap(_has_bits_[0], other->_has_bits_[0]);
-  errormsg_.Swap(&other->errormsg_, &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  advancedorderrejectjson_.Swap(&other->advancedorderrejectjson_, &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  ::PROTOBUF_NAMESPACE_ID::internal::memswap<
-      PROTOBUF_FIELD_OFFSET(ErrorMessage, errorcode_)
-      + sizeof(ErrorMessage::errorcode_)
-      - PROTOBUF_FIELD_OFFSET(ErrorMessage, errortime_)>(
-          reinterpret_cast<char*>(&errortime_),
-          reinterpret_cast<char*>(&other->errortime_));
+  auto* arena = GetArena();
+  ABSL_DCHECK_EQ(arena, other->GetArena());
+  _internal_metadata_.InternalSwap(&other->_internal_metadata_);
+  swap(_impl_._has_bits_[0], other->_impl_._has_bits_[0]);
+  ::_pbi::ArenaStringPtr::InternalSwap(&_impl_.errormsg_, &other->_impl_.errormsg_, arena);
+  ::_pbi::ArenaStringPtr::InternalSwap(&_impl_.advancedorderrejectjson_, &other->_impl_.advancedorderrejectjson_, arena);
+  ::google::protobuf::internal::memswap<
+      PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.errorcode_)
+      + sizeof(ErrorMessage::_impl_.errorcode_)
+      - PROTOBUF_FIELD_OFFSET(ErrorMessage, _impl_.errortime_)>(
+          reinterpret_cast<char*>(&_impl_.errortime_),
+          reinterpret_cast<char*>(&other->_impl_.errortime_));
 }
 
-::PROTOBUF_NAMESPACE_ID::Metadata ErrorMessage::GetMetadata() const {
-  return GetMetadataStatic();
+::google::protobuf::Metadata ErrorMessage::GetMetadata() const {
+  return ::google::protobuf::Message::GetMetadataImpl(GetClassData()->full());
 }
-
-
 // @@protoc_insertion_point(namespace_scope)
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> PROTOBUF_NOINLINE ::protobuf::ErrorMessage* Arena::CreateMaybeMessage< ::protobuf::ErrorMessage >(Arena* arena) {
-  return Arena::CreateMessageInternal< ::protobuf::ErrorMessage >(arena);
-}
-PROTOBUF_NAMESPACE_CLOSE
-
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
 // @@protoc_insertion_point(global_scope)
-#include <google/protobuf/port_undef.inc>
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY2 static ::std::false_type
+    _static_init2_ PROTOBUF_UNUSED =
+        (::_pbi::AddDescriptors(&descriptor_table_ErrorMessage_2eproto),
+         ::std::false_type{});
+#include "google/protobuf/port_undef.inc"
