@@ -1,266 +1,305 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: ExecutionDetailsEnd.proto
+// Protobuf C++ Version: 5.29.3
 
 #include "ExecutionDetailsEnd.pb.h"
 
 #include <algorithm>
-
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/extension_set.h>
-#include <google/protobuf/wire_format_lite.h>
-#include <google/protobuf/descriptor.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/reflection_ops.h>
-#include <google/protobuf/wire_format.h>
+#include <type_traits>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/generated_message_tctable_impl.h"
+#include "google/protobuf/extension_set.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/wire_format_lite.h"
+#include "google/protobuf/descriptor.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/reflection_ops.h"
+#include "google/protobuf/wire_format.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
-namespace protobuf {
-class ExecutionDetailsEndDefaultTypeInternal {
- public:
-  ::PROTOBUF_NAMESPACE_ID::internal::ExplicitlyConstructed<ExecutionDetailsEnd> _instance;
-} _ExecutionDetailsEnd_default_instance_;
-}  // namespace protobuf
-static void InitDefaultsscc_info_ExecutionDetailsEnd_ExecutionDetailsEnd_2eproto() {
-  GOOGLE_PROTOBUF_VERIFY_VERSION;
 
-  {
-    void* ptr = &::protobuf::_ExecutionDetailsEnd_default_instance_;
-    new (ptr) ::protobuf::ExecutionDetailsEnd();
-    ::PROTOBUF_NAMESPACE_ID::internal::OnShutdownDestroyMessage(ptr);
-  }
-  ::protobuf::ExecutionDetailsEnd::InitAsDefaultInstance();
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+PROTOBUF_PRAGMA_INIT_SEG
+namespace _pb = ::google::protobuf;
+namespace _pbi = ::google::protobuf::internal;
+namespace _fl = ::google::protobuf::internal::field_layout;
+namespace protobuf {
+
+inline constexpr ExecutionDetailsEnd::Impl_::Impl_(
+    ::_pbi::ConstantInitialized) noexcept
+      : _cached_size_{0},
+        reqid_{0} {}
+
+template <typename>
+PROTOBUF_CONSTEXPR ExecutionDetailsEnd::ExecutionDetailsEnd(::_pbi::ConstantInitialized)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(_class_data_.base()),
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(),
+#endif  // PROTOBUF_CUSTOM_VTABLE
+      _impl_(::_pbi::ConstantInitialized()) {
 }
-
-::PROTOBUF_NAMESPACE_ID::internal::SCCInfo<0> scc_info_ExecutionDetailsEnd_ExecutionDetailsEnd_2eproto =
-    {{ATOMIC_VAR_INIT(::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase::kUninitialized), 0, 0, InitDefaultsscc_info_ExecutionDetailsEnd_ExecutionDetailsEnd_2eproto}, {}};
-
-static ::PROTOBUF_NAMESPACE_ID::Metadata file_level_metadata_ExecutionDetailsEnd_2eproto[1];
-static constexpr ::PROTOBUF_NAMESPACE_ID::EnumDescriptor const** file_level_enum_descriptors_ExecutionDetailsEnd_2eproto = nullptr;
-static constexpr ::PROTOBUF_NAMESPACE_ID::ServiceDescriptor const** file_level_service_descriptors_ExecutionDetailsEnd_2eproto = nullptr;
-
-const ::PROTOBUF_NAMESPACE_ID::uint32 TableStruct_ExecutionDetailsEnd_2eproto::offsets[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionDetailsEnd, _has_bits_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionDetailsEnd, _internal_metadata_),
-  ~0u,  // no _extensions_
-  ~0u,  // no _oneof_case_
-  ~0u,  // no _weak_field_map_
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionDetailsEnd, reqid_),
-  0,
-};
-static const ::PROTOBUF_NAMESPACE_ID::internal::MigrationSchema schemas[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  { 0, 6, sizeof(::protobuf::ExecutionDetailsEnd)},
+struct ExecutionDetailsEndDefaultTypeInternal {
+  PROTOBUF_CONSTEXPR ExecutionDetailsEndDefaultTypeInternal() : _instance(::_pbi::ConstantInitialized{}) {}
+  ~ExecutionDetailsEndDefaultTypeInternal() {}
+  union {
+    ExecutionDetailsEnd _instance;
+  };
 };
 
-static ::PROTOBUF_NAMESPACE_ID::Message const * const file_default_instances[] = {
-  reinterpret_cast<const ::PROTOBUF_NAMESPACE_ID::Message*>(&::protobuf::_ExecutionDetailsEnd_default_instance_),
+PROTOBUF_ATTRIBUTE_NO_DESTROY PROTOBUF_CONSTINIT
+    PROTOBUF_ATTRIBUTE_INIT_PRIORITY1 ExecutionDetailsEndDefaultTypeInternal _ExecutionDetailsEnd_default_instance_;
+}  // namespace protobuf
+static constexpr const ::_pb::EnumDescriptor**
+    file_level_enum_descriptors_ExecutionDetailsEnd_2eproto = nullptr;
+static constexpr const ::_pb::ServiceDescriptor**
+    file_level_service_descriptors_ExecutionDetailsEnd_2eproto = nullptr;
+const ::uint32_t
+    TableStruct_ExecutionDetailsEnd_2eproto::offsets[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+        protodesc_cold) = {
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionDetailsEnd, _impl_._has_bits_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionDetailsEnd, _internal_metadata_),
+        ~0u,  // no _extensions_
+        ~0u,  // no _oneof_case_
+        ~0u,  // no _weak_field_map_
+        ~0u,  // no _inlined_string_donated_
+        ~0u,  // no _split_
+        ~0u,  // no sizeof(Split)
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionDetailsEnd, _impl_.reqid_),
+        0,
 };
 
-const char descriptor_table_protodef_ExecutionDetailsEnd_2eproto[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) =
-  "\n\031ExecutionDetailsEnd.proto\022\010protobuf\"3\n"
-  "\023ExecutionDetailsEnd\022\022\n\005reqId\030\001 \001(\005H\000\210\001\001"
-  "B\010\n\006_reqIdB2\n\026com.ib.client.protobufB\030Ex"
-  "ecutionDetailsEndProtob\006proto3"
-  ;
-static const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable*const descriptor_table_ExecutionDetailsEnd_2eproto_deps[1] = {
+static const ::_pbi::MigrationSchema
+    schemas[] ABSL_ATTRIBUTE_SECTION_VARIABLE(protodesc_cold) = {
+        {0, 9, -1, sizeof(::protobuf::ExecutionDetailsEnd)},
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase*const descriptor_table_ExecutionDetailsEnd_2eproto_sccs[1] = {
-  &scc_info_ExecutionDetailsEnd_ExecutionDetailsEnd_2eproto.base,
+static const ::_pb::Message* const file_default_instances[] = {
+    &::protobuf::_ExecutionDetailsEnd_default_instance_._instance,
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::once_flag descriptor_table_ExecutionDetailsEnd_2eproto_once;
-const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_ExecutionDetailsEnd_2eproto = {
-  false, false, descriptor_table_protodef_ExecutionDetailsEnd_2eproto, "ExecutionDetailsEnd.proto", 150,
-  &descriptor_table_ExecutionDetailsEnd_2eproto_once, descriptor_table_ExecutionDetailsEnd_2eproto_sccs, descriptor_table_ExecutionDetailsEnd_2eproto_deps, 1, 0,
-  schemas, file_default_instances, TableStruct_ExecutionDetailsEnd_2eproto::offsets,
-  file_level_metadata_ExecutionDetailsEnd_2eproto, 1, file_level_enum_descriptors_ExecutionDetailsEnd_2eproto, file_level_service_descriptors_ExecutionDetailsEnd_2eproto,
+const char descriptor_table_protodef_ExecutionDetailsEnd_2eproto[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+    protodesc_cold) = {
+    "\n\031ExecutionDetailsEnd.proto\022\010protobuf\"3\n"
+    "\023ExecutionDetailsEnd\022\022\n\005reqId\030\001 \001(\005H\000\210\001\001"
+    "B\010\n\006_reqIdBC\n\026com.ib.client.protobufB\030Ex"
+    "ecutionDetailsEndProto\252\002\016IBApi.protobufb"
+    "\006proto3"
 };
-
-// Force running AddDescriptors() at dynamic initialization time.
-static bool dynamic_init_dummy_ExecutionDetailsEnd_2eproto = (static_cast<void>(::PROTOBUF_NAMESPACE_ID::internal::AddDescriptors(&descriptor_table_ExecutionDetailsEnd_2eproto)), true);
+static ::absl::once_flag descriptor_table_ExecutionDetailsEnd_2eproto_once;
+PROTOBUF_CONSTINIT const ::_pbi::DescriptorTable descriptor_table_ExecutionDetailsEnd_2eproto = {
+    false,
+    false,
+    167,
+    descriptor_table_protodef_ExecutionDetailsEnd_2eproto,
+    "ExecutionDetailsEnd.proto",
+    &descriptor_table_ExecutionDetailsEnd_2eproto_once,
+    nullptr,
+    0,
+    1,
+    schemas,
+    file_default_instances,
+    TableStruct_ExecutionDetailsEnd_2eproto::offsets,
+    file_level_enum_descriptors_ExecutionDetailsEnd_2eproto,
+    file_level_service_descriptors_ExecutionDetailsEnd_2eproto,
+};
 namespace protobuf {
-
 // ===================================================================
 
-void ExecutionDetailsEnd::InitAsDefaultInstance() {
-}
 class ExecutionDetailsEnd::_Internal {
  public:
-  using HasBits = decltype(std::declval<ExecutionDetailsEnd>()._has_bits_);
-  static void set_has_reqid(HasBits* has_bits) {
-    (*has_bits)[0] |= 1u;
-  }
+  using HasBits =
+      decltype(std::declval<ExecutionDetailsEnd>()._impl_._has_bits_);
+  static constexpr ::int32_t kHasBitsOffset =
+      8 * PROTOBUF_FIELD_OFFSET(ExecutionDetailsEnd, _impl_._has_bits_);
 };
 
-ExecutionDetailsEnd::ExecutionDetailsEnd(::PROTOBUF_NAMESPACE_ID::Arena* arena)
-  : ::PROTOBUF_NAMESPACE_ID::Message(arena) {
-  SharedCtor();
-  RegisterArenaDtor(arena);
+ExecutionDetailsEnd::ExecutionDetailsEnd(::google::protobuf::Arena* arena)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  SharedCtor(arena);
   // @@protoc_insertion_point(arena_constructor:protobuf.ExecutionDetailsEnd)
 }
-ExecutionDetailsEnd::ExecutionDetailsEnd(const ExecutionDetailsEnd& from)
-  : ::PROTOBUF_NAMESPACE_ID::Message(),
-      _has_bits_(from._has_bits_) {
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  reqid_ = from.reqid_;
-  // @@protoc_insertion_point(copy_constructor:protobuf.ExecutionDetailsEnd)
+ExecutionDetailsEnd::ExecutionDetailsEnd(
+    ::google::protobuf::Arena* arena, const ExecutionDetailsEnd& from)
+    : ExecutionDetailsEnd(arena) {
+  MergeFrom(from);
 }
+inline PROTOBUF_NDEBUG_INLINE ExecutionDetailsEnd::Impl_::Impl_(
+    ::google::protobuf::internal::InternalVisibility visibility,
+    ::google::protobuf::Arena* arena)
+      : _cached_size_{0} {}
 
-void ExecutionDetailsEnd::SharedCtor() {
-  reqid_ = 0;
+inline void ExecutionDetailsEnd::SharedCtor(::_pb::Arena* arena) {
+  new (&_impl_) Impl_(internal_visibility(), arena);
+  _impl_.reqid_ = {};
 }
-
 ExecutionDetailsEnd::~ExecutionDetailsEnd() {
   // @@protoc_insertion_point(destructor:protobuf.ExecutionDetailsEnd)
-  SharedDtor();
-  _internal_metadata_.Delete<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  SharedDtor(*this);
+}
+inline void ExecutionDetailsEnd::SharedDtor(MessageLite& self) {
+  ExecutionDetailsEnd& this_ = static_cast<ExecutionDetailsEnd&>(self);
+  this_._internal_metadata_.Delete<::google::protobuf::UnknownFieldSet>();
+  ABSL_DCHECK(this_.GetArena() == nullptr);
+  this_._impl_.~Impl_();
 }
 
-void ExecutionDetailsEnd::SharedDtor() {
-  GOOGLE_DCHECK(GetArena() == nullptr);
+inline void* ExecutionDetailsEnd::PlacementNew_(const void*, void* mem,
+                                        ::google::protobuf::Arena* arena) {
+  return ::new (mem) ExecutionDetailsEnd(arena);
 }
+constexpr auto ExecutionDetailsEnd::InternalNewImpl_() {
+  return ::google::protobuf::internal::MessageCreator::ZeroInit(sizeof(ExecutionDetailsEnd),
+                                            alignof(ExecutionDetailsEnd));
+}
+PROTOBUF_CONSTINIT
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::google::protobuf::internal::ClassDataFull ExecutionDetailsEnd::_class_data_ = {
+    ::google::protobuf::internal::ClassData{
+        &_ExecutionDetailsEnd_default_instance_._instance,
+        &_table_.header,
+        nullptr,  // OnDemandRegisterArenaDtor
+        nullptr,  // IsInitialized
+        &ExecutionDetailsEnd::MergeImpl,
+        ::google::protobuf::Message::GetNewImpl<ExecutionDetailsEnd>(),
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        &ExecutionDetailsEnd::SharedDtor,
+        ::google::protobuf::Message::GetClearImpl<ExecutionDetailsEnd>(), &ExecutionDetailsEnd::ByteSizeLong,
+            &ExecutionDetailsEnd::_InternalSerialize,
+#endif  // PROTOBUF_CUSTOM_VTABLE
+        PROTOBUF_FIELD_OFFSET(ExecutionDetailsEnd, _impl_._cached_size_),
+        false,
+    },
+    &ExecutionDetailsEnd::kDescriptorMethods,
+    &descriptor_table_ExecutionDetailsEnd_2eproto,
+    nullptr,  // tracker
+};
+const ::google::protobuf::internal::ClassData* ExecutionDetailsEnd::GetClassData() const {
+  ::google::protobuf::internal::PrefetchToLocalCache(&_class_data_);
+  ::google::protobuf::internal::PrefetchToLocalCache(_class_data_.tc_table);
+  return _class_data_.base();
+}
+PROTOBUF_CONSTINIT PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::_pbi::TcParseTable<0, 1, 0, 0, 2> ExecutionDetailsEnd::_table_ = {
+  {
+    PROTOBUF_FIELD_OFFSET(ExecutionDetailsEnd, _impl_._has_bits_),
+    0, // no _extensions_
+    1, 0,  // max_field_number, fast_idx_mask
+    offsetof(decltype(_table_), field_lookup_table),
+    4294967294,  // skipmap
+    offsetof(decltype(_table_), field_entries),
+    1,  // num_field_entries
+    0,  // num_aux_entries
+    offsetof(decltype(_table_), field_names),  // no aux_entries
+    _class_data_.base(),
+    nullptr,  // post_loop_handler
+    ::_pbi::TcParser::GenericFallback,  // fallback
+    #ifdef PROTOBUF_PREFETCH_PARSE_TABLE
+    ::_pbi::TcParser::GetTable<::protobuf::ExecutionDetailsEnd>(),  // to_prefetch
+    #endif  // PROTOBUF_PREFETCH_PARSE_TABLE
+  }, {{
+    // optional int32 reqId = 1;
+    {::_pbi::TcParser::SingularVarintNoZag1<::uint32_t, offsetof(ExecutionDetailsEnd, _impl_.reqid_), 0>(),
+     {8, 0, 0, PROTOBUF_FIELD_OFFSET(ExecutionDetailsEnd, _impl_.reqid_)}},
+  }}, {{
+    65535, 65535
+  }}, {{
+    // optional int32 reqId = 1;
+    {PROTOBUF_FIELD_OFFSET(ExecutionDetailsEnd, _impl_.reqid_), _Internal::kHasBitsOffset + 0, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kInt32)},
+  }},
+  // no aux_entries
+  {{
+  }},
+};
 
-void ExecutionDetailsEnd::ArenaDtor(void* object) {
-  ExecutionDetailsEnd* _this = reinterpret_cast< ExecutionDetailsEnd* >(object);
-  (void)_this;
-}
-void ExecutionDetailsEnd::RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena*) {
-}
-void ExecutionDetailsEnd::SetCachedSize(int size) const {
-  _cached_size_.Set(size);
-}
-const ExecutionDetailsEnd& ExecutionDetailsEnd::default_instance() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&::scc_info_ExecutionDetailsEnd_ExecutionDetailsEnd_2eproto.base);
-  return *internal_default_instance();
-}
-
-
-void ExecutionDetailsEnd::Clear() {
+PROTOBUF_NOINLINE void ExecutionDetailsEnd::Clear() {
 // @@protoc_insertion_point(message_clear_start:protobuf.ExecutionDetailsEnd)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  ::uint32_t cached_has_bits = 0;
   // Prevent compiler warnings about cached_has_bits being unused
   (void) cached_has_bits;
 
-  reqid_ = 0;
-  _has_bits_.Clear();
-  _internal_metadata_.Clear<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  _impl_.reqid_ = 0;
+  _impl_._has_bits_.Clear();
+  _internal_metadata_.Clear<::google::protobuf::UnknownFieldSet>();
 }
 
-const char* ExecutionDetailsEnd::_InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) {
-#define CHK_(x) if (PROTOBUF_PREDICT_FALSE(!(x))) goto failure
-  _Internal::HasBits has_bits{};
-  ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena(); (void)arena;
-  while (!ctx->Done(&ptr)) {
-    ::PROTOBUF_NAMESPACE_ID::uint32 tag;
-    ptr = ::PROTOBUF_NAMESPACE_ID::internal::ReadTag(ptr, &tag);
-    CHK_(ptr);
-    switch (tag >> 3) {
-      // int32 reqId = 1;
-      case 1:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 8)) {
-          _Internal::set_has_reqid(&has_bits);
-          reqid_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      default: {
-      handle_unusual:
-        if ((tag & 7) == 4 || tag == 0) {
-          ctx->SetLastTag(tag);
-          goto success;
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::uint8_t* ExecutionDetailsEnd::_InternalSerialize(
+            const MessageLite& base, ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) {
+          const ExecutionDetailsEnd& this_ = static_cast<const ExecutionDetailsEnd&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::uint8_t* ExecutionDetailsEnd::_InternalSerialize(
+            ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+          const ExecutionDetailsEnd& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(serialize_to_array_start:protobuf.ExecutionDetailsEnd)
+          ::uint32_t cached_has_bits = 0;
+          (void)cached_has_bits;
+
+          cached_has_bits = this_._impl_._has_bits_[0];
+          // optional int32 reqId = 1;
+          if (cached_has_bits & 0x00000001u) {
+            target = ::google::protobuf::internal::WireFormatLite::
+                WriteInt32ToArrayWithField<1>(
+                    stream, this_._internal_reqid(), target);
+          }
+
+          if (PROTOBUF_PREDICT_FALSE(this_._internal_metadata_.have_unknown_fields())) {
+            target =
+                ::_pbi::WireFormat::InternalSerializeUnknownFieldsToArray(
+                    this_._internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance), target, stream);
+          }
+          // @@protoc_insertion_point(serialize_to_array_end:protobuf.ExecutionDetailsEnd)
+          return target;
         }
-        ptr = UnknownFieldParse(tag,
-            _internal_metadata_.mutable_unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(),
-            ptr, ctx);
-        CHK_(ptr != nullptr);
-        continue;
-      }
-    }  // switch
-  }  // while
-success:
-  _has_bits_.Or(has_bits);
-  return ptr;
-failure:
-  ptr = nullptr;
-  goto success;
-#undef CHK_
-}
 
-::PROTOBUF_NAMESPACE_ID::uint8* ExecutionDetailsEnd::_InternalSerialize(
-    ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const {
-  // @@protoc_insertion_point(serialize_to_array_start:protobuf.ExecutionDetailsEnd)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::size_t ExecutionDetailsEnd::ByteSizeLong(const MessageLite& base) {
+          const ExecutionDetailsEnd& this_ = static_cast<const ExecutionDetailsEnd&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::size_t ExecutionDetailsEnd::ByteSizeLong() const {
+          const ExecutionDetailsEnd& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(message_byte_size_start:protobuf.ExecutionDetailsEnd)
+          ::size_t total_size = 0;
+
+          ::uint32_t cached_has_bits = 0;
+          // Prevent compiler warnings about cached_has_bits being unused
+          (void)cached_has_bits;
+
+           {
+            // optional int32 reqId = 1;
+            cached_has_bits = this_._impl_._has_bits_[0];
+            if (cached_has_bits & 0x00000001u) {
+              total_size += ::_pbi::WireFormatLite::Int32SizePlusOne(
+                  this_._internal_reqid());
+            }
+          }
+          return this_.MaybeComputeUnknownFieldsSize(total_size,
+                                                     &this_._impl_._cached_size_);
+        }
+
+void ExecutionDetailsEnd::MergeImpl(::google::protobuf::MessageLite& to_msg, const ::google::protobuf::MessageLite& from_msg) {
+  auto* const _this = static_cast<ExecutionDetailsEnd*>(&to_msg);
+  auto& from = static_cast<const ExecutionDetailsEnd&>(from_msg);
+  // @@protoc_insertion_point(class_specific_merge_from_start:protobuf.ExecutionDetailsEnd)
+  ABSL_DCHECK_NE(&from, _this);
+  ::uint32_t cached_has_bits = 0;
   (void) cached_has_bits;
 
-  // int32 reqId = 1;
-  if (_internal_has_reqid()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteInt32ToArray(1, this->_internal_reqid(), target);
-  }
-
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormat::InternalSerializeUnknownFieldsToArray(
-        _internal_metadata_.unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(::PROTOBUF_NAMESPACE_ID::UnknownFieldSet::default_instance), target, stream);
-  }
-  // @@protoc_insertion_point(serialize_to_array_end:protobuf.ExecutionDetailsEnd)
-  return target;
-}
-
-size_t ExecutionDetailsEnd::ByteSizeLong() const {
-// @@protoc_insertion_point(message_byte_size_start:protobuf.ExecutionDetailsEnd)
-  size_t total_size = 0;
-
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  // Prevent compiler warnings about cached_has_bits being unused
-  (void) cached_has_bits;
-
-  // int32 reqId = 1;
-  cached_has_bits = _has_bits_[0];
+  cached_has_bits = from._impl_._has_bits_[0];
   if (cached_has_bits & 0x00000001u) {
-    total_size += 1 +
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::Int32Size(
-        this->_internal_reqid());
+    _this->_impl_.reqid_ = from._impl_.reqid_;
   }
-
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    return ::PROTOBUF_NAMESPACE_ID::internal::ComputeUnknownFieldsSize(
-        _internal_metadata_, total_size, &_cached_size_);
-  }
-  int cached_size = ::PROTOBUF_NAMESPACE_ID::internal::ToCachedSize(total_size);
-  SetCachedSize(cached_size);
-  return total_size;
-}
-
-void ExecutionDetailsEnd::MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_merge_from_start:protobuf.ExecutionDetailsEnd)
-  GOOGLE_DCHECK_NE(&from, this);
-  const ExecutionDetailsEnd* source =
-      ::PROTOBUF_NAMESPACE_ID::DynamicCastToGenerated<ExecutionDetailsEnd>(
-          &from);
-  if (source == nullptr) {
-  // @@protoc_insertion_point(generalized_merge_from_cast_fail:protobuf.ExecutionDetailsEnd)
-    ::PROTOBUF_NAMESPACE_ID::internal::ReflectionOps::Merge(from, this);
-  } else {
-  // @@protoc_insertion_point(generalized_merge_from_cast_success:protobuf.ExecutionDetailsEnd)
-    MergeFrom(*source);
-  }
-}
-
-void ExecutionDetailsEnd::MergeFrom(const ExecutionDetailsEnd& from) {
-// @@protoc_insertion_point(class_specific_merge_from_start:protobuf.ExecutionDetailsEnd)
-  GOOGLE_DCHECK_NE(&from, this);
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  (void) cached_has_bits;
-
-  if (from._internal_has_reqid()) {
-    _internal_set_reqid(from._internal_reqid());
-  }
-}
-
-void ExecutionDetailsEnd::CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_copy_from_start:protobuf.ExecutionDetailsEnd)
-  if (&from == this) return;
-  Clear();
-  MergeFrom(from);
+  _this->_impl_._has_bits_[0] |= cached_has_bits;
+  _this->_internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(from._internal_metadata_);
 }
 
 void ExecutionDetailsEnd::CopyFrom(const ExecutionDetailsEnd& from) {
@@ -270,29 +309,26 @@ void ExecutionDetailsEnd::CopyFrom(const ExecutionDetailsEnd& from) {
   MergeFrom(from);
 }
 
-bool ExecutionDetailsEnd::IsInitialized() const {
-  return true;
-}
 
-void ExecutionDetailsEnd::InternalSwap(ExecutionDetailsEnd* other) {
+void ExecutionDetailsEnd::InternalSwap(ExecutionDetailsEnd* PROTOBUF_RESTRICT other) {
   using std::swap;
-  _internal_metadata_.Swap<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(&other->_internal_metadata_);
-  swap(_has_bits_[0], other->_has_bits_[0]);
-  swap(reqid_, other->reqid_);
+  _internal_metadata_.InternalSwap(&other->_internal_metadata_);
+  swap(_impl_._has_bits_[0], other->_impl_._has_bits_[0]);
+        swap(_impl_.reqid_, other->_impl_.reqid_);
 }
 
-::PROTOBUF_NAMESPACE_ID::Metadata ExecutionDetailsEnd::GetMetadata() const {
-  return GetMetadataStatic();
+::google::protobuf::Metadata ExecutionDetailsEnd::GetMetadata() const {
+  return ::google::protobuf::Message::GetMetadataImpl(GetClassData()->full());
 }
-
-
 // @@protoc_insertion_point(namespace_scope)
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> PROTOBUF_NOINLINE ::protobuf::ExecutionDetailsEnd* Arena::CreateMaybeMessage< ::protobuf::ExecutionDetailsEnd >(Arena* arena) {
-  return Arena::CreateMessageInternal< ::protobuf::ExecutionDetailsEnd >(arena);
-}
-PROTOBUF_NAMESPACE_CLOSE
-
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
 // @@protoc_insertion_point(global_scope)
-#include <google/protobuf/port_undef.inc>
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY2 static ::std::false_type
+    _static_init2_ PROTOBUF_UNUSED =
+        (::_pbi::AddDescriptors(&descriptor_table_ExecutionDetailsEnd_2eproto),
+         ::std::false_type{});
+#include "google/protobuf/port_undef.inc"
