@@ -1,623 +1,633 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: ExecutionFilter.proto
+// Protobuf C++ Version: 5.29.3
 
 #include "ExecutionFilter.pb.h"
 
 #include <algorithm>
-
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/extension_set.h>
-#include <google/protobuf/wire_format_lite.h>
-#include <google/protobuf/descriptor.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/reflection_ops.h>
-#include <google/protobuf/wire_format.h>
+#include <type_traits>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/generated_message_tctable_impl.h"
+#include "google/protobuf/extension_set.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/wire_format_lite.h"
+#include "google/protobuf/descriptor.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/reflection_ops.h"
+#include "google/protobuf/wire_format.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
-namespace protobuf {
-class ExecutionFilterDefaultTypeInternal {
- public:
-  ::PROTOBUF_NAMESPACE_ID::internal::ExplicitlyConstructed<ExecutionFilter> _instance;
-} _ExecutionFilter_default_instance_;
-}  // namespace protobuf
-static void InitDefaultsscc_info_ExecutionFilter_ExecutionFilter_2eproto() {
-  GOOGLE_PROTOBUF_VERIFY_VERSION;
 
-  {
-    void* ptr = &::protobuf::_ExecutionFilter_default_instance_;
-    new (ptr) ::protobuf::ExecutionFilter();
-    ::PROTOBUF_NAMESPACE_ID::internal::OnShutdownDestroyMessage(ptr);
-  }
-  ::protobuf::ExecutionFilter::InitAsDefaultInstance();
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+PROTOBUF_PRAGMA_INIT_SEG
+namespace _pb = ::google::protobuf;
+namespace _pbi = ::google::protobuf::internal;
+namespace _fl = ::google::protobuf::internal::field_layout;
+namespace protobuf {
+
+inline constexpr ExecutionFilter::Impl_::Impl_(
+    ::_pbi::ConstantInitialized) noexcept
+      : _cached_size_{0},
+        specificdates_{},
+        _specificdates_cached_byte_size_{0},
+        acctcode_(
+            &::google::protobuf::internal::fixed_address_empty_string,
+            ::_pbi::ConstantInitialized()),
+        time_(
+            &::google::protobuf::internal::fixed_address_empty_string,
+            ::_pbi::ConstantInitialized()),
+        symbol_(
+            &::google::protobuf::internal::fixed_address_empty_string,
+            ::_pbi::ConstantInitialized()),
+        sectype_(
+            &::google::protobuf::internal::fixed_address_empty_string,
+            ::_pbi::ConstantInitialized()),
+        exchange_(
+            &::google::protobuf::internal::fixed_address_empty_string,
+            ::_pbi::ConstantInitialized()),
+        side_(
+            &::google::protobuf::internal::fixed_address_empty_string,
+            ::_pbi::ConstantInitialized()),
+        clientid_{0},
+        lastndays_{0} {}
+
+template <typename>
+PROTOBUF_CONSTEXPR ExecutionFilter::ExecutionFilter(::_pbi::ConstantInitialized)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(_class_data_.base()),
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(),
+#endif  // PROTOBUF_CUSTOM_VTABLE
+      _impl_(::_pbi::ConstantInitialized()) {
 }
-
-::PROTOBUF_NAMESPACE_ID::internal::SCCInfo<0> scc_info_ExecutionFilter_ExecutionFilter_2eproto =
-    {{ATOMIC_VAR_INIT(::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase::kUninitialized), 0, 0, InitDefaultsscc_info_ExecutionFilter_ExecutionFilter_2eproto}, {}};
-
-static ::PROTOBUF_NAMESPACE_ID::Metadata file_level_metadata_ExecutionFilter_2eproto[1];
-static constexpr ::PROTOBUF_NAMESPACE_ID::EnumDescriptor const** file_level_enum_descriptors_ExecutionFilter_2eproto = nullptr;
-static constexpr ::PROTOBUF_NAMESPACE_ID::ServiceDescriptor const** file_level_service_descriptors_ExecutionFilter_2eproto = nullptr;
-
-const ::PROTOBUF_NAMESPACE_ID::uint32 TableStruct_ExecutionFilter_2eproto::offsets[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _has_bits_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _internal_metadata_),
-  ~0u,  // no _extensions_
-  ~0u,  // no _oneof_case_
-  ~0u,  // no _weak_field_map_
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, clientid_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, acctcode_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, time_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, symbol_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, sectype_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, exchange_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, side_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, lastndays_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, specificdates_),
-  6,
-  0,
-  1,
-  2,
-  3,
-  4,
-  5,
-  7,
-  ~0u,
-};
-static const ::PROTOBUF_NAMESPACE_ID::internal::MigrationSchema schemas[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  { 0, 14, sizeof(::protobuf::ExecutionFilter)},
+struct ExecutionFilterDefaultTypeInternal {
+  PROTOBUF_CONSTEXPR ExecutionFilterDefaultTypeInternal() : _instance(::_pbi::ConstantInitialized{}) {}
+  ~ExecutionFilterDefaultTypeInternal() {}
+  union {
+    ExecutionFilter _instance;
+  };
 };
 
-static ::PROTOBUF_NAMESPACE_ID::Message const * const file_default_instances[] = {
-  reinterpret_cast<const ::PROTOBUF_NAMESPACE_ID::Message*>(&::protobuf::_ExecutionFilter_default_instance_),
+PROTOBUF_ATTRIBUTE_NO_DESTROY PROTOBUF_CONSTINIT
+    PROTOBUF_ATTRIBUTE_INIT_PRIORITY1 ExecutionFilterDefaultTypeInternal _ExecutionFilter_default_instance_;
+}  // namespace protobuf
+static constexpr const ::_pb::EnumDescriptor**
+    file_level_enum_descriptors_ExecutionFilter_2eproto = nullptr;
+static constexpr const ::_pb::ServiceDescriptor**
+    file_level_service_descriptors_ExecutionFilter_2eproto = nullptr;
+const ::uint32_t
+    TableStruct_ExecutionFilter_2eproto::offsets[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+        protodesc_cold) = {
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _impl_._has_bits_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _internal_metadata_),
+        ~0u,  // no _extensions_
+        ~0u,  // no _oneof_case_
+        ~0u,  // no _weak_field_map_
+        ~0u,  // no _inlined_string_donated_
+        ~0u,  // no _split_
+        ~0u,  // no sizeof(Split)
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _impl_.clientid_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _impl_.acctcode_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _impl_.time_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _impl_.symbol_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _impl_.sectype_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _impl_.exchange_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _impl_.side_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _impl_.lastndays_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::ExecutionFilter, _impl_.specificdates_),
+        6,
+        0,
+        1,
+        2,
+        3,
+        4,
+        5,
+        7,
+        ~0u,
 };
 
-const char descriptor_table_protodef_ExecutionFilter_2eproto[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) =
-  "\n\025ExecutionFilter.proto\022\010protobuf\"\264\002\n\017Ex"
-  "ecutionFilter\022\025\n\010clientId\030\001 \001(\005H\000\210\001\001\022\025\n\010"
-  "acctCode\030\002 \001(\tH\001\210\001\001\022\021\n\004time\030\003 \001(\tH\002\210\001\001\022\023"
-  "\n\006symbol\030\004 \001(\tH\003\210\001\001\022\024\n\007secType\030\005 \001(\tH\004\210\001"
-  "\001\022\025\n\010exchange\030\006 \001(\tH\005\210\001\001\022\021\n\004side\030\007 \001(\tH\006"
-  "\210\001\001\022\026\n\tlastNDays\030\010 \001(\005H\007\210\001\001\022\025\n\rspecificD"
-  "ates\030\t \003(\005B\013\n\t_clientIdB\013\n\t_acctCodeB\007\n\005"
-  "_timeB\t\n\007_symbolB\n\n\010_secTypeB\013\n\t_exchang"
-  "eB\007\n\005_sideB\014\n\n_lastNDaysB.\n\026com.ib.clien"
-  "t.protobufB\024ExecutionFilterProtob\006proto3"
-  ;
-static const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable*const descriptor_table_ExecutionFilter_2eproto_deps[1] = {
+static const ::_pbi::MigrationSchema
+    schemas[] ABSL_ATTRIBUTE_SECTION_VARIABLE(protodesc_cold) = {
+        {0, 17, -1, sizeof(::protobuf::ExecutionFilter)},
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase*const descriptor_table_ExecutionFilter_2eproto_sccs[1] = {
-  &scc_info_ExecutionFilter_ExecutionFilter_2eproto.base,
+static const ::_pb::Message* const file_default_instances[] = {
+    &::protobuf::_ExecutionFilter_default_instance_._instance,
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::once_flag descriptor_table_ExecutionFilter_2eproto_once;
-const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_ExecutionFilter_2eproto = {
-  false, false, descriptor_table_protodef_ExecutionFilter_2eproto, "ExecutionFilter.proto", 400,
-  &descriptor_table_ExecutionFilter_2eproto_once, descriptor_table_ExecutionFilter_2eproto_sccs, descriptor_table_ExecutionFilter_2eproto_deps, 1, 0,
-  schemas, file_default_instances, TableStruct_ExecutionFilter_2eproto::offsets,
-  file_level_metadata_ExecutionFilter_2eproto, 1, file_level_enum_descriptors_ExecutionFilter_2eproto, file_level_service_descriptors_ExecutionFilter_2eproto,
+const char descriptor_table_protodef_ExecutionFilter_2eproto[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+    protodesc_cold) = {
+    "\n\025ExecutionFilter.proto\022\010protobuf\"\264\002\n\017Ex"
+    "ecutionFilter\022\025\n\010clientId\030\001 \001(\005H\000\210\001\001\022\025\n\010"
+    "acctCode\030\002 \001(\tH\001\210\001\001\022\021\n\004time\030\003 \001(\tH\002\210\001\001\022\023"
+    "\n\006symbol\030\004 \001(\tH\003\210\001\001\022\024\n\007secType\030\005 \001(\tH\004\210\001"
+    "\001\022\025\n\010exchange\030\006 \001(\tH\005\210\001\001\022\021\n\004side\030\007 \001(\tH\006"
+    "\210\001\001\022\026\n\tlastNDays\030\010 \001(\005H\007\210\001\001\022\025\n\rspecificD"
+    "ates\030\t \003(\005B\013\n\t_clientIdB\013\n\t_acctCodeB\007\n\005"
+    "_timeB\t\n\007_symbolB\n\n\010_secTypeB\013\n\t_exchang"
+    "eB\007\n\005_sideB\014\n\n_lastNDaysB\?\n\026com.ib.clien"
+    "t.protobufB\024ExecutionFilterProto\252\002\016IBApi"
+    ".protobufb\006proto3"
 };
-
-// Force running AddDescriptors() at dynamic initialization time.
-static bool dynamic_init_dummy_ExecutionFilter_2eproto = (static_cast<void>(::PROTOBUF_NAMESPACE_ID::internal::AddDescriptors(&descriptor_table_ExecutionFilter_2eproto)), true);
+static ::absl::once_flag descriptor_table_ExecutionFilter_2eproto_once;
+PROTOBUF_CONSTINIT const ::_pbi::DescriptorTable descriptor_table_ExecutionFilter_2eproto = {
+    false,
+    false,
+    417,
+    descriptor_table_protodef_ExecutionFilter_2eproto,
+    "ExecutionFilter.proto",
+    &descriptor_table_ExecutionFilter_2eproto_once,
+    nullptr,
+    0,
+    1,
+    schemas,
+    file_default_instances,
+    TableStruct_ExecutionFilter_2eproto::offsets,
+    file_level_enum_descriptors_ExecutionFilter_2eproto,
+    file_level_service_descriptors_ExecutionFilter_2eproto,
+};
 namespace protobuf {
-
 // ===================================================================
 
-void ExecutionFilter::InitAsDefaultInstance() {
-}
 class ExecutionFilter::_Internal {
  public:
-  using HasBits = decltype(std::declval<ExecutionFilter>()._has_bits_);
-  static void set_has_clientid(HasBits* has_bits) {
-    (*has_bits)[0] |= 64u;
-  }
-  static void set_has_acctcode(HasBits* has_bits) {
-    (*has_bits)[0] |= 1u;
-  }
-  static void set_has_time(HasBits* has_bits) {
-    (*has_bits)[0] |= 2u;
-  }
-  static void set_has_symbol(HasBits* has_bits) {
-    (*has_bits)[0] |= 4u;
-  }
-  static void set_has_sectype(HasBits* has_bits) {
-    (*has_bits)[0] |= 8u;
-  }
-  static void set_has_exchange(HasBits* has_bits) {
-    (*has_bits)[0] |= 16u;
-  }
-  static void set_has_side(HasBits* has_bits) {
-    (*has_bits)[0] |= 32u;
-  }
-  static void set_has_lastndays(HasBits* has_bits) {
-    (*has_bits)[0] |= 128u;
-  }
+  using HasBits =
+      decltype(std::declval<ExecutionFilter>()._impl_._has_bits_);
+  static constexpr ::int32_t kHasBitsOffset =
+      8 * PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_._has_bits_);
 };
 
-ExecutionFilter::ExecutionFilter(::PROTOBUF_NAMESPACE_ID::Arena* arena)
-  : ::PROTOBUF_NAMESPACE_ID::Message(arena),
-  specificdates_(arena) {
-  SharedCtor();
-  RegisterArenaDtor(arena);
+ExecutionFilter::ExecutionFilter(::google::protobuf::Arena* arena)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  SharedCtor(arena);
   // @@protoc_insertion_point(arena_constructor:protobuf.ExecutionFilter)
 }
-ExecutionFilter::ExecutionFilter(const ExecutionFilter& from)
-  : ::PROTOBUF_NAMESPACE_ID::Message(),
-      _has_bits_(from._has_bits_),
-      specificdates_(from.specificdates_) {
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  acctcode_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  if (from._internal_has_acctcode()) {
-    acctcode_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), from._internal_acctcode(),
-      GetArena());
-  }
-  time_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  if (from._internal_has_time()) {
-    time_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), from._internal_time(),
-      GetArena());
-  }
-  symbol_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  if (from._internal_has_symbol()) {
-    symbol_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), from._internal_symbol(),
-      GetArena());
-  }
-  sectype_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  if (from._internal_has_sectype()) {
-    sectype_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), from._internal_sectype(),
-      GetArena());
-  }
-  exchange_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  if (from._internal_has_exchange()) {
-    exchange_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), from._internal_exchange(),
-      GetArena());
-  }
-  side_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  if (from._internal_has_side()) {
-    side_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), from._internal_side(),
-      GetArena());
-  }
-  ::memcpy(&clientid_, &from.clientid_,
-    static_cast<size_t>(reinterpret_cast<char*>(&lastndays_) -
-    reinterpret_cast<char*>(&clientid_)) + sizeof(lastndays_));
+inline PROTOBUF_NDEBUG_INLINE ExecutionFilter::Impl_::Impl_(
+    ::google::protobuf::internal::InternalVisibility visibility, ::google::protobuf::Arena* arena,
+    const Impl_& from, const ::protobuf::ExecutionFilter& from_msg)
+      : _has_bits_{from._has_bits_},
+        _cached_size_{0},
+        specificdates_{visibility, arena, from.specificdates_},
+        _specificdates_cached_byte_size_{0},
+        acctcode_(arena, from.acctcode_),
+        time_(arena, from.time_),
+        symbol_(arena, from.symbol_),
+        sectype_(arena, from.sectype_),
+        exchange_(arena, from.exchange_),
+        side_(arena, from.side_) {}
+
+ExecutionFilter::ExecutionFilter(
+    ::google::protobuf::Arena* arena,
+    const ExecutionFilter& from)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  ExecutionFilter* const _this = this;
+  (void)_this;
+  _internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(
+      from._internal_metadata_);
+  new (&_impl_) Impl_(internal_visibility(), arena, from._impl_, from);
+  ::memcpy(reinterpret_cast<char *>(&_impl_) +
+               offsetof(Impl_, clientid_),
+           reinterpret_cast<const char *>(&from._impl_) +
+               offsetof(Impl_, clientid_),
+           offsetof(Impl_, lastndays_) -
+               offsetof(Impl_, clientid_) +
+               sizeof(Impl_::lastndays_));
+
   // @@protoc_insertion_point(copy_constructor:protobuf.ExecutionFilter)
 }
+inline PROTOBUF_NDEBUG_INLINE ExecutionFilter::Impl_::Impl_(
+    ::google::protobuf::internal::InternalVisibility visibility,
+    ::google::protobuf::Arena* arena)
+      : _cached_size_{0},
+        specificdates_{visibility, arena},
+        _specificdates_cached_byte_size_{0},
+        acctcode_(arena),
+        time_(arena),
+        symbol_(arena),
+        sectype_(arena),
+        exchange_(arena),
+        side_(arena) {}
 
-void ExecutionFilter::SharedCtor() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&scc_info_ExecutionFilter_ExecutionFilter_2eproto.base);
-  acctcode_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  time_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  symbol_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  sectype_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  exchange_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  side_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  ::memset(&clientid_, 0, static_cast<size_t>(
-      reinterpret_cast<char*>(&lastndays_) -
-      reinterpret_cast<char*>(&clientid_)) + sizeof(lastndays_));
+inline void ExecutionFilter::SharedCtor(::_pb::Arena* arena) {
+  new (&_impl_) Impl_(internal_visibility(), arena);
+  ::memset(reinterpret_cast<char *>(&_impl_) +
+               offsetof(Impl_, clientid_),
+           0,
+           offsetof(Impl_, lastndays_) -
+               offsetof(Impl_, clientid_) +
+               sizeof(Impl_::lastndays_));
 }
-
 ExecutionFilter::~ExecutionFilter() {
   // @@protoc_insertion_point(destructor:protobuf.ExecutionFilter)
-  SharedDtor();
-  _internal_metadata_.Delete<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  SharedDtor(*this);
+}
+inline void ExecutionFilter::SharedDtor(MessageLite& self) {
+  ExecutionFilter& this_ = static_cast<ExecutionFilter&>(self);
+  this_._internal_metadata_.Delete<::google::protobuf::UnknownFieldSet>();
+  ABSL_DCHECK(this_.GetArena() == nullptr);
+  this_._impl_.acctcode_.Destroy();
+  this_._impl_.time_.Destroy();
+  this_._impl_.symbol_.Destroy();
+  this_._impl_.sectype_.Destroy();
+  this_._impl_.exchange_.Destroy();
+  this_._impl_.side_.Destroy();
+  this_._impl_.~Impl_();
 }
 
-void ExecutionFilter::SharedDtor() {
-  GOOGLE_DCHECK(GetArena() == nullptr);
-  acctcode_.DestroyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  time_.DestroyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  symbol_.DestroyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  sectype_.DestroyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  exchange_.DestroyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  side_.DestroyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+inline void* ExecutionFilter::PlacementNew_(const void*, void* mem,
+                                        ::google::protobuf::Arena* arena) {
+  return ::new (mem) ExecutionFilter(arena);
 }
+constexpr auto ExecutionFilter::InternalNewImpl_() {
+  constexpr auto arena_bits = ::google::protobuf::internal::EncodePlacementArenaOffsets({
+      PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.specificdates_) +
+          decltype(ExecutionFilter::_impl_.specificdates_)::
+              InternalGetArenaOffset(
+                  ::google::protobuf::Message::internal_visibility()),
+  });
+  if (arena_bits.has_value()) {
+    return ::google::protobuf::internal::MessageCreator::CopyInit(
+        sizeof(ExecutionFilter), alignof(ExecutionFilter), *arena_bits);
+  } else {
+    return ::google::protobuf::internal::MessageCreator(&ExecutionFilter::PlacementNew_,
+                                 sizeof(ExecutionFilter),
+                                 alignof(ExecutionFilter));
+  }
+}
+PROTOBUF_CONSTINIT
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::google::protobuf::internal::ClassDataFull ExecutionFilter::_class_data_ = {
+    ::google::protobuf::internal::ClassData{
+        &_ExecutionFilter_default_instance_._instance,
+        &_table_.header,
+        nullptr,  // OnDemandRegisterArenaDtor
+        nullptr,  // IsInitialized
+        &ExecutionFilter::MergeImpl,
+        ::google::protobuf::Message::GetNewImpl<ExecutionFilter>(),
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        &ExecutionFilter::SharedDtor,
+        ::google::protobuf::Message::GetClearImpl<ExecutionFilter>(), &ExecutionFilter::ByteSizeLong,
+            &ExecutionFilter::_InternalSerialize,
+#endif  // PROTOBUF_CUSTOM_VTABLE
+        PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_._cached_size_),
+        false,
+    },
+    &ExecutionFilter::kDescriptorMethods,
+    &descriptor_table_ExecutionFilter_2eproto,
+    nullptr,  // tracker
+};
+const ::google::protobuf::internal::ClassData* ExecutionFilter::GetClassData() const {
+  ::google::protobuf::internal::PrefetchToLocalCache(&_class_data_);
+  ::google::protobuf::internal::PrefetchToLocalCache(_class_data_.tc_table);
+  return _class_data_.base();
+}
+PROTOBUF_CONSTINIT PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::_pbi::TcParseTable<4, 9, 0, 78, 2> ExecutionFilter::_table_ = {
+  {
+    PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_._has_bits_),
+    0, // no _extensions_
+    9, 120,  // max_field_number, fast_idx_mask
+    offsetof(decltype(_table_), field_lookup_table),
+    4294966784,  // skipmap
+    offsetof(decltype(_table_), field_entries),
+    9,  // num_field_entries
+    0,  // num_aux_entries
+    offsetof(decltype(_table_), field_names),  // no aux_entries
+    _class_data_.base(),
+    nullptr,  // post_loop_handler
+    ::_pbi::TcParser::GenericFallback,  // fallback
+    #ifdef PROTOBUF_PREFETCH_PARSE_TABLE
+    ::_pbi::TcParser::GetTable<::protobuf::ExecutionFilter>(),  // to_prefetch
+    #endif  // PROTOBUF_PREFETCH_PARSE_TABLE
+  }, {{
+    {::_pbi::TcParser::MiniParse, {}},
+    // optional int32 clientId = 1;
+    {::_pbi::TcParser::SingularVarintNoZag1<::uint32_t, offsetof(ExecutionFilter, _impl_.clientid_), 6>(),
+     {8, 6, 0, PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.clientid_)}},
+    // optional string acctCode = 2;
+    {::_pbi::TcParser::FastUS1,
+     {18, 0, 0, PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.acctcode_)}},
+    // optional string time = 3;
+    {::_pbi::TcParser::FastUS1,
+     {26, 1, 0, PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.time_)}},
+    // optional string symbol = 4;
+    {::_pbi::TcParser::FastUS1,
+     {34, 2, 0, PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.symbol_)}},
+    // optional string secType = 5;
+    {::_pbi::TcParser::FastUS1,
+     {42, 3, 0, PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.sectype_)}},
+    // optional string exchange = 6;
+    {::_pbi::TcParser::FastUS1,
+     {50, 4, 0, PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.exchange_)}},
+    // optional string side = 7;
+    {::_pbi::TcParser::FastUS1,
+     {58, 5, 0, PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.side_)}},
+    // optional int32 lastNDays = 8;
+    {::_pbi::TcParser::SingularVarintNoZag1<::uint32_t, offsetof(ExecutionFilter, _impl_.lastndays_), 7>(),
+     {64, 7, 0, PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.lastndays_)}},
+    // repeated int32 specificDates = 9;
+    {::_pbi::TcParser::FastV32P1,
+     {74, 63, 0, PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.specificdates_)}},
+    {::_pbi::TcParser::MiniParse, {}},
+    {::_pbi::TcParser::MiniParse, {}},
+    {::_pbi::TcParser::MiniParse, {}},
+    {::_pbi::TcParser::MiniParse, {}},
+    {::_pbi::TcParser::MiniParse, {}},
+    {::_pbi::TcParser::MiniParse, {}},
+  }}, {{
+    65535, 65535
+  }}, {{
+    // optional int32 clientId = 1;
+    {PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.clientid_), _Internal::kHasBitsOffset + 6, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kInt32)},
+    // optional string acctCode = 2;
+    {PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.acctcode_), _Internal::kHasBitsOffset + 0, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kUtf8String | ::_fl::kRepAString)},
+    // optional string time = 3;
+    {PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.time_), _Internal::kHasBitsOffset + 1, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kUtf8String | ::_fl::kRepAString)},
+    // optional string symbol = 4;
+    {PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.symbol_), _Internal::kHasBitsOffset + 2, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kUtf8String | ::_fl::kRepAString)},
+    // optional string secType = 5;
+    {PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.sectype_), _Internal::kHasBitsOffset + 3, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kUtf8String | ::_fl::kRepAString)},
+    // optional string exchange = 6;
+    {PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.exchange_), _Internal::kHasBitsOffset + 4, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kUtf8String | ::_fl::kRepAString)},
+    // optional string side = 7;
+    {PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.side_), _Internal::kHasBitsOffset + 5, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kUtf8String | ::_fl::kRepAString)},
+    // optional int32 lastNDays = 8;
+    {PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.lastndays_), _Internal::kHasBitsOffset + 7, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kInt32)},
+    // repeated int32 specificDates = 9;
+    {PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.specificdates_), -1, 0,
+    (0 | ::_fl::kFcRepeated | ::_fl::kPackedInt32)},
+  }},
+  // no aux_entries
+  {{
+    "\30\0\10\4\6\7\10\4\0\0\0\0\0\0\0\0"
+    "protobuf.ExecutionFilter"
+    "acctCode"
+    "time"
+    "symbol"
+    "secType"
+    "exchange"
+    "side"
+  }},
+};
 
-void ExecutionFilter::ArenaDtor(void* object) {
-  ExecutionFilter* _this = reinterpret_cast< ExecutionFilter* >(object);
-  (void)_this;
-}
-void ExecutionFilter::RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena*) {
-}
-void ExecutionFilter::SetCachedSize(int size) const {
-  _cached_size_.Set(size);
-}
-const ExecutionFilter& ExecutionFilter::default_instance() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&::scc_info_ExecutionFilter_ExecutionFilter_2eproto.base);
-  return *internal_default_instance();
-}
-
-
-void ExecutionFilter::Clear() {
+PROTOBUF_NOINLINE void ExecutionFilter::Clear() {
 // @@protoc_insertion_point(message_clear_start:protobuf.ExecutionFilter)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  ::uint32_t cached_has_bits = 0;
   // Prevent compiler warnings about cached_has_bits being unused
   (void) cached_has_bits;
 
-  specificdates_.Clear();
-  cached_has_bits = _has_bits_[0];
+  _impl_.specificdates_.Clear();
+  cached_has_bits = _impl_._has_bits_[0];
   if (cached_has_bits & 0x0000003fu) {
     if (cached_has_bits & 0x00000001u) {
-      acctcode_.ClearNonDefaultToEmpty();
+      _impl_.acctcode_.ClearNonDefaultToEmpty();
     }
     if (cached_has_bits & 0x00000002u) {
-      time_.ClearNonDefaultToEmpty();
+      _impl_.time_.ClearNonDefaultToEmpty();
     }
     if (cached_has_bits & 0x00000004u) {
-      symbol_.ClearNonDefaultToEmpty();
+      _impl_.symbol_.ClearNonDefaultToEmpty();
     }
     if (cached_has_bits & 0x00000008u) {
-      sectype_.ClearNonDefaultToEmpty();
+      _impl_.sectype_.ClearNonDefaultToEmpty();
     }
     if (cached_has_bits & 0x00000010u) {
-      exchange_.ClearNonDefaultToEmpty();
+      _impl_.exchange_.ClearNonDefaultToEmpty();
     }
     if (cached_has_bits & 0x00000020u) {
-      side_.ClearNonDefaultToEmpty();
+      _impl_.side_.ClearNonDefaultToEmpty();
     }
   }
   if (cached_has_bits & 0x000000c0u) {
-    ::memset(&clientid_, 0, static_cast<size_t>(
-        reinterpret_cast<char*>(&lastndays_) -
-        reinterpret_cast<char*>(&clientid_)) + sizeof(lastndays_));
+    ::memset(&_impl_.clientid_, 0, static_cast<::size_t>(
+        reinterpret_cast<char*>(&_impl_.lastndays_) -
+        reinterpret_cast<char*>(&_impl_.clientid_)) + sizeof(_impl_.lastndays_));
   }
-  _has_bits_.Clear();
-  _internal_metadata_.Clear<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  _impl_._has_bits_.Clear();
+  _internal_metadata_.Clear<::google::protobuf::UnknownFieldSet>();
 }
 
-const char* ExecutionFilter::_InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) {
-#define CHK_(x) if (PROTOBUF_PREDICT_FALSE(!(x))) goto failure
-  _Internal::HasBits has_bits{};
-  ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena(); (void)arena;
-  while (!ctx->Done(&ptr)) {
-    ::PROTOBUF_NAMESPACE_ID::uint32 tag;
-    ptr = ::PROTOBUF_NAMESPACE_ID::internal::ReadTag(ptr, &tag);
-    CHK_(ptr);
-    switch (tag >> 3) {
-      // int32 clientId = 1;
-      case 1:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 8)) {
-          _Internal::set_has_clientid(&has_bits);
-          clientid_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // string acctCode = 2;
-      case 2:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 18)) {
-          auto str = _internal_mutable_acctcode();
-          ptr = ::PROTOBUF_NAMESPACE_ID::internal::InlineGreedyStringParser(str, ptr, ctx);
-          CHK_(::PROTOBUF_NAMESPACE_ID::internal::VerifyUTF8(str, "protobuf.ExecutionFilter.acctCode"));
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // string time = 3;
-      case 3:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 26)) {
-          auto str = _internal_mutable_time();
-          ptr = ::PROTOBUF_NAMESPACE_ID::internal::InlineGreedyStringParser(str, ptr, ctx);
-          CHK_(::PROTOBUF_NAMESPACE_ID::internal::VerifyUTF8(str, "protobuf.ExecutionFilter.time"));
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // string symbol = 4;
-      case 4:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 34)) {
-          auto str = _internal_mutable_symbol();
-          ptr = ::PROTOBUF_NAMESPACE_ID::internal::InlineGreedyStringParser(str, ptr, ctx);
-          CHK_(::PROTOBUF_NAMESPACE_ID::internal::VerifyUTF8(str, "protobuf.ExecutionFilter.symbol"));
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // string secType = 5;
-      case 5:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 42)) {
-          auto str = _internal_mutable_sectype();
-          ptr = ::PROTOBUF_NAMESPACE_ID::internal::InlineGreedyStringParser(str, ptr, ctx);
-          CHK_(::PROTOBUF_NAMESPACE_ID::internal::VerifyUTF8(str, "protobuf.ExecutionFilter.secType"));
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // string exchange = 6;
-      case 6:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 50)) {
-          auto str = _internal_mutable_exchange();
-          ptr = ::PROTOBUF_NAMESPACE_ID::internal::InlineGreedyStringParser(str, ptr, ctx);
-          CHK_(::PROTOBUF_NAMESPACE_ID::internal::VerifyUTF8(str, "protobuf.ExecutionFilter.exchange"));
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // string side = 7;
-      case 7:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 58)) {
-          auto str = _internal_mutable_side();
-          ptr = ::PROTOBUF_NAMESPACE_ID::internal::InlineGreedyStringParser(str, ptr, ctx);
-          CHK_(::PROTOBUF_NAMESPACE_ID::internal::VerifyUTF8(str, "protobuf.ExecutionFilter.side"));
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // int32 lastNDays = 8;
-      case 8:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 64)) {
-          _Internal::set_has_lastndays(&has_bits);
-          lastndays_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // repeated int32 specificDates = 9;
-      case 9:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 74)) {
-          ptr = ::PROTOBUF_NAMESPACE_ID::internal::PackedInt32Parser(_internal_mutable_specificdates(), ptr, ctx);
-          CHK_(ptr);
-        } else if (static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 72) {
-          _internal_add_specificdates(::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr));
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      default: {
-      handle_unusual:
-        if ((tag & 7) == 4 || tag == 0) {
-          ctx->SetLastTag(tag);
-          goto success;
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::uint8_t* ExecutionFilter::_InternalSerialize(
+            const MessageLite& base, ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) {
+          const ExecutionFilter& this_ = static_cast<const ExecutionFilter&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::uint8_t* ExecutionFilter::_InternalSerialize(
+            ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+          const ExecutionFilter& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(serialize_to_array_start:protobuf.ExecutionFilter)
+          ::uint32_t cached_has_bits = 0;
+          (void)cached_has_bits;
+
+          cached_has_bits = this_._impl_._has_bits_[0];
+          // optional int32 clientId = 1;
+          if (cached_has_bits & 0x00000040u) {
+            target = ::google::protobuf::internal::WireFormatLite::
+                WriteInt32ToArrayWithField<1>(
+                    stream, this_._internal_clientid(), target);
+          }
+
+          // optional string acctCode = 2;
+          if (cached_has_bits & 0x00000001u) {
+            const std::string& _s = this_._internal_acctcode();
+            ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
+                _s.data(), static_cast<int>(_s.length()), ::google::protobuf::internal::WireFormatLite::SERIALIZE, "protobuf.ExecutionFilter.acctCode");
+            target = stream->WriteStringMaybeAliased(2, _s, target);
+          }
+
+          // optional string time = 3;
+          if (cached_has_bits & 0x00000002u) {
+            const std::string& _s = this_._internal_time();
+            ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
+                _s.data(), static_cast<int>(_s.length()), ::google::protobuf::internal::WireFormatLite::SERIALIZE, "protobuf.ExecutionFilter.time");
+            target = stream->WriteStringMaybeAliased(3, _s, target);
+          }
+
+          // optional string symbol = 4;
+          if (cached_has_bits & 0x00000004u) {
+            const std::string& _s = this_._internal_symbol();
+            ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
+                _s.data(), static_cast<int>(_s.length()), ::google::protobuf::internal::WireFormatLite::SERIALIZE, "protobuf.ExecutionFilter.symbol");
+            target = stream->WriteStringMaybeAliased(4, _s, target);
+          }
+
+          // optional string secType = 5;
+          if (cached_has_bits & 0x00000008u) {
+            const std::string& _s = this_._internal_sectype();
+            ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
+                _s.data(), static_cast<int>(_s.length()), ::google::protobuf::internal::WireFormatLite::SERIALIZE, "protobuf.ExecutionFilter.secType");
+            target = stream->WriteStringMaybeAliased(5, _s, target);
+          }
+
+          // optional string exchange = 6;
+          if (cached_has_bits & 0x00000010u) {
+            const std::string& _s = this_._internal_exchange();
+            ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
+                _s.data(), static_cast<int>(_s.length()), ::google::protobuf::internal::WireFormatLite::SERIALIZE, "protobuf.ExecutionFilter.exchange");
+            target = stream->WriteStringMaybeAliased(6, _s, target);
+          }
+
+          // optional string side = 7;
+          if (cached_has_bits & 0x00000020u) {
+            const std::string& _s = this_._internal_side();
+            ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
+                _s.data(), static_cast<int>(_s.length()), ::google::protobuf::internal::WireFormatLite::SERIALIZE, "protobuf.ExecutionFilter.side");
+            target = stream->WriteStringMaybeAliased(7, _s, target);
+          }
+
+          // optional int32 lastNDays = 8;
+          if (cached_has_bits & 0x00000080u) {
+            target = ::google::protobuf::internal::WireFormatLite::
+                WriteInt32ToArrayWithField<8>(
+                    stream, this_._internal_lastndays(), target);
+          }
+
+          // repeated int32 specificDates = 9;
+          {
+            int byte_size = this_._impl_._specificdates_cached_byte_size_.Get();
+            if (byte_size > 0) {
+              target = stream->WriteInt32Packed(
+                  9, this_._internal_specificdates(), byte_size, target);
+            }
+          }
+
+          if (PROTOBUF_PREDICT_FALSE(this_._internal_metadata_.have_unknown_fields())) {
+            target =
+                ::_pbi::WireFormat::InternalSerializeUnknownFieldsToArray(
+                    this_._internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance), target, stream);
+          }
+          // @@protoc_insertion_point(serialize_to_array_end:protobuf.ExecutionFilter)
+          return target;
         }
-        ptr = UnknownFieldParse(tag,
-            _internal_metadata_.mutable_unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(),
-            ptr, ctx);
-        CHK_(ptr != nullptr);
-        continue;
-      }
-    }  // switch
-  }  // while
-success:
-  _has_bits_.Or(has_bits);
-  return ptr;
-failure:
-  ptr = nullptr;
-  goto success;
-#undef CHK_
-}
 
-::PROTOBUF_NAMESPACE_ID::uint8* ExecutionFilter::_InternalSerialize(
-    ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const {
-  // @@protoc_insertion_point(serialize_to_array_start:protobuf.ExecutionFilter)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::size_t ExecutionFilter::ByteSizeLong(const MessageLite& base) {
+          const ExecutionFilter& this_ = static_cast<const ExecutionFilter&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::size_t ExecutionFilter::ByteSizeLong() const {
+          const ExecutionFilter& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(message_byte_size_start:protobuf.ExecutionFilter)
+          ::size_t total_size = 0;
+
+          ::uint32_t cached_has_bits = 0;
+          // Prevent compiler warnings about cached_has_bits being unused
+          (void)cached_has_bits;
+
+          ::_pbi::Prefetch5LinesFrom7Lines(&this_);
+           {
+            // repeated int32 specificDates = 9;
+            {
+              total_size +=
+                  ::_pbi::WireFormatLite::Int32SizeWithPackedTagSize(
+                      this_._internal_specificdates(), 1,
+                      this_._impl_._specificdates_cached_byte_size_);
+            }
+          }
+          cached_has_bits = this_._impl_._has_bits_[0];
+          if (cached_has_bits & 0x000000ffu) {
+            // optional string acctCode = 2;
+            if (cached_has_bits & 0x00000001u) {
+              total_size += 1 + ::google::protobuf::internal::WireFormatLite::StringSize(
+                                              this_._internal_acctcode());
+            }
+            // optional string time = 3;
+            if (cached_has_bits & 0x00000002u) {
+              total_size += 1 + ::google::protobuf::internal::WireFormatLite::StringSize(
+                                              this_._internal_time());
+            }
+            // optional string symbol = 4;
+            if (cached_has_bits & 0x00000004u) {
+              total_size += 1 + ::google::protobuf::internal::WireFormatLite::StringSize(
+                                              this_._internal_symbol());
+            }
+            // optional string secType = 5;
+            if (cached_has_bits & 0x00000008u) {
+              total_size += 1 + ::google::protobuf::internal::WireFormatLite::StringSize(
+                                              this_._internal_sectype());
+            }
+            // optional string exchange = 6;
+            if (cached_has_bits & 0x00000010u) {
+              total_size += 1 + ::google::protobuf::internal::WireFormatLite::StringSize(
+                                              this_._internal_exchange());
+            }
+            // optional string side = 7;
+            if (cached_has_bits & 0x00000020u) {
+              total_size += 1 + ::google::protobuf::internal::WireFormatLite::StringSize(
+                                              this_._internal_side());
+            }
+            // optional int32 clientId = 1;
+            if (cached_has_bits & 0x00000040u) {
+              total_size += ::_pbi::WireFormatLite::Int32SizePlusOne(
+                  this_._internal_clientid());
+            }
+            // optional int32 lastNDays = 8;
+            if (cached_has_bits & 0x00000080u) {
+              total_size += ::_pbi::WireFormatLite::Int32SizePlusOne(
+                  this_._internal_lastndays());
+            }
+          }
+          return this_.MaybeComputeUnknownFieldsSize(total_size,
+                                                     &this_._impl_._cached_size_);
+        }
+
+void ExecutionFilter::MergeImpl(::google::protobuf::MessageLite& to_msg, const ::google::protobuf::MessageLite& from_msg) {
+  auto* const _this = static_cast<ExecutionFilter*>(&to_msg);
+  auto& from = static_cast<const ExecutionFilter&>(from_msg);
+  // @@protoc_insertion_point(class_specific_merge_from_start:protobuf.ExecutionFilter)
+  ABSL_DCHECK_NE(&from, _this);
+  ::uint32_t cached_has_bits = 0;
   (void) cached_has_bits;
 
-  // int32 clientId = 1;
-  if (_internal_has_clientid()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteInt32ToArray(1, this->_internal_clientid(), target);
-  }
-
-  // string acctCode = 2;
-  if (_internal_has_acctcode()) {
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
-      this->_internal_acctcode().data(), static_cast<int>(this->_internal_acctcode().length()),
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
-      "protobuf.ExecutionFilter.acctCode");
-    target = stream->WriteStringMaybeAliased(
-        2, this->_internal_acctcode(), target);
-  }
-
-  // string time = 3;
-  if (_internal_has_time()) {
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
-      this->_internal_time().data(), static_cast<int>(this->_internal_time().length()),
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
-      "protobuf.ExecutionFilter.time");
-    target = stream->WriteStringMaybeAliased(
-        3, this->_internal_time(), target);
-  }
-
-  // string symbol = 4;
-  if (_internal_has_symbol()) {
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
-      this->_internal_symbol().data(), static_cast<int>(this->_internal_symbol().length()),
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
-      "protobuf.ExecutionFilter.symbol");
-    target = stream->WriteStringMaybeAliased(
-        4, this->_internal_symbol(), target);
-  }
-
-  // string secType = 5;
-  if (_internal_has_sectype()) {
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
-      this->_internal_sectype().data(), static_cast<int>(this->_internal_sectype().length()),
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
-      "protobuf.ExecutionFilter.secType");
-    target = stream->WriteStringMaybeAliased(
-        5, this->_internal_sectype(), target);
-  }
-
-  // string exchange = 6;
-  if (_internal_has_exchange()) {
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
-      this->_internal_exchange().data(), static_cast<int>(this->_internal_exchange().length()),
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
-      "protobuf.ExecutionFilter.exchange");
-    target = stream->WriteStringMaybeAliased(
-        6, this->_internal_exchange(), target);
-  }
-
-  // string side = 7;
-  if (_internal_has_side()) {
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
-      this->_internal_side().data(), static_cast<int>(this->_internal_side().length()),
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
-      "protobuf.ExecutionFilter.side");
-    target = stream->WriteStringMaybeAliased(
-        7, this->_internal_side(), target);
-  }
-
-  // int32 lastNDays = 8;
-  if (_internal_has_lastndays()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteInt32ToArray(8, this->_internal_lastndays(), target);
-  }
-
-  // repeated int32 specificDates = 9;
-  {
-    int byte_size = _specificdates_cached_byte_size_.load(std::memory_order_relaxed);
-    if (byte_size > 0) {
-      target = stream->WriteInt32Packed(
-          9, _internal_specificdates(), byte_size, target);
-    }
-  }
-
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormat::InternalSerializeUnknownFieldsToArray(
-        _internal_metadata_.unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(::PROTOBUF_NAMESPACE_ID::UnknownFieldSet::default_instance), target, stream);
-  }
-  // @@protoc_insertion_point(serialize_to_array_end:protobuf.ExecutionFilter)
-  return target;
-}
-
-size_t ExecutionFilter::ByteSizeLong() const {
-// @@protoc_insertion_point(message_byte_size_start:protobuf.ExecutionFilter)
-  size_t total_size = 0;
-
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  // Prevent compiler warnings about cached_has_bits being unused
-  (void) cached_has_bits;
-
-  // repeated int32 specificDates = 9;
-  {
-    size_t data_size = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::
-      Int32Size(this->specificdates_);
-    if (data_size > 0) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::Int32Size(
-            static_cast<::PROTOBUF_NAMESPACE_ID::int32>(data_size));
-    }
-    int cached_size = ::PROTOBUF_NAMESPACE_ID::internal::ToCachedSize(data_size);
-    _specificdates_cached_byte_size_.store(cached_size,
-                                    std::memory_order_relaxed);
-    total_size += data_size;
-  }
-
-  cached_has_bits = _has_bits_[0];
-  if (cached_has_bits & 0x000000ffu) {
-    // string acctCode = 2;
-    if (cached_has_bits & 0x00000001u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::StringSize(
-          this->_internal_acctcode());
-    }
-
-    // string time = 3;
-    if (cached_has_bits & 0x00000002u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::StringSize(
-          this->_internal_time());
-    }
-
-    // string symbol = 4;
-    if (cached_has_bits & 0x00000004u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::StringSize(
-          this->_internal_symbol());
-    }
-
-    // string secType = 5;
-    if (cached_has_bits & 0x00000008u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::StringSize(
-          this->_internal_sectype());
-    }
-
-    // string exchange = 6;
-    if (cached_has_bits & 0x00000010u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::StringSize(
-          this->_internal_exchange());
-    }
-
-    // string side = 7;
-    if (cached_has_bits & 0x00000020u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::StringSize(
-          this->_internal_side());
-    }
-
-    // int32 clientId = 1;
-    if (cached_has_bits & 0x00000040u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::Int32Size(
-          this->_internal_clientid());
-    }
-
-    // int32 lastNDays = 8;
-    if (cached_has_bits & 0x00000080u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::Int32Size(
-          this->_internal_lastndays());
-    }
-
-  }
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    return ::PROTOBUF_NAMESPACE_ID::internal::ComputeUnknownFieldsSize(
-        _internal_metadata_, total_size, &_cached_size_);
-  }
-  int cached_size = ::PROTOBUF_NAMESPACE_ID::internal::ToCachedSize(total_size);
-  SetCachedSize(cached_size);
-  return total_size;
-}
-
-void ExecutionFilter::MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_merge_from_start:protobuf.ExecutionFilter)
-  GOOGLE_DCHECK_NE(&from, this);
-  const ExecutionFilter* source =
-      ::PROTOBUF_NAMESPACE_ID::DynamicCastToGenerated<ExecutionFilter>(
-          &from);
-  if (source == nullptr) {
-  // @@protoc_insertion_point(generalized_merge_from_cast_fail:protobuf.ExecutionFilter)
-    ::PROTOBUF_NAMESPACE_ID::internal::ReflectionOps::Merge(from, this);
-  } else {
-  // @@protoc_insertion_point(generalized_merge_from_cast_success:protobuf.ExecutionFilter)
-    MergeFrom(*source);
-  }
-}
-
-void ExecutionFilter::MergeFrom(const ExecutionFilter& from) {
-// @@protoc_insertion_point(class_specific_merge_from_start:protobuf.ExecutionFilter)
-  GOOGLE_DCHECK_NE(&from, this);
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  (void) cached_has_bits;
-
-  specificdates_.MergeFrom(from.specificdates_);
-  cached_has_bits = from._has_bits_[0];
+  _this->_internal_mutable_specificdates()->MergeFrom(from._internal_specificdates());
+  cached_has_bits = from._impl_._has_bits_[0];
   if (cached_has_bits & 0x000000ffu) {
     if (cached_has_bits & 0x00000001u) {
-      _internal_set_acctcode(from._internal_acctcode());
+      _this->_internal_set_acctcode(from._internal_acctcode());
     }
     if (cached_has_bits & 0x00000002u) {
-      _internal_set_time(from._internal_time());
+      _this->_internal_set_time(from._internal_time());
     }
     if (cached_has_bits & 0x00000004u) {
-      _internal_set_symbol(from._internal_symbol());
+      _this->_internal_set_symbol(from._internal_symbol());
     }
     if (cached_has_bits & 0x00000008u) {
-      _internal_set_sectype(from._internal_sectype());
+      _this->_internal_set_sectype(from._internal_sectype());
     }
     if (cached_has_bits & 0x00000010u) {
-      _internal_set_exchange(from._internal_exchange());
+      _this->_internal_set_exchange(from._internal_exchange());
     }
     if (cached_has_bits & 0x00000020u) {
-      _internal_set_side(from._internal_side());
+      _this->_internal_set_side(from._internal_side());
     }
     if (cached_has_bits & 0x00000040u) {
-      clientid_ = from.clientid_;
+      _this->_impl_.clientid_ = from._impl_.clientid_;
     }
     if (cached_has_bits & 0x00000080u) {
-      lastndays_ = from.lastndays_;
+      _this->_impl_.lastndays_ = from._impl_.lastndays_;
     }
-    _has_bits_[0] |= cached_has_bits;
   }
-}
-
-void ExecutionFilter::CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_copy_from_start:protobuf.ExecutionFilter)
-  if (&from == this) return;
-  Clear();
-  MergeFrom(from);
+  _this->_impl_._has_bits_[0] |= cached_has_bits;
+  _this->_internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(from._internal_metadata_);
 }
 
 void ExecutionFilter::CopyFrom(const ExecutionFilter& from) {
@@ -627,41 +637,40 @@ void ExecutionFilter::CopyFrom(const ExecutionFilter& from) {
   MergeFrom(from);
 }
 
-bool ExecutionFilter::IsInitialized() const {
-  return true;
-}
 
-void ExecutionFilter::InternalSwap(ExecutionFilter* other) {
+void ExecutionFilter::InternalSwap(ExecutionFilter* PROTOBUF_RESTRICT other) {
   using std::swap;
-  _internal_metadata_.Swap<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(&other->_internal_metadata_);
-  swap(_has_bits_[0], other->_has_bits_[0]);
-  specificdates_.InternalSwap(&other->specificdates_);
-  acctcode_.Swap(&other->acctcode_, &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  time_.Swap(&other->time_, &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  symbol_.Swap(&other->symbol_, &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  sectype_.Swap(&other->sectype_, &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  exchange_.Swap(&other->exchange_, &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  side_.Swap(&other->side_, &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  ::PROTOBUF_NAMESPACE_ID::internal::memswap<
-      PROTOBUF_FIELD_OFFSET(ExecutionFilter, lastndays_)
-      + sizeof(ExecutionFilter::lastndays_)
-      - PROTOBUF_FIELD_OFFSET(ExecutionFilter, clientid_)>(
-          reinterpret_cast<char*>(&clientid_),
-          reinterpret_cast<char*>(&other->clientid_));
+  auto* arena = GetArena();
+  ABSL_DCHECK_EQ(arena, other->GetArena());
+  _internal_metadata_.InternalSwap(&other->_internal_metadata_);
+  swap(_impl_._has_bits_[0], other->_impl_._has_bits_[0]);
+  _impl_.specificdates_.InternalSwap(&other->_impl_.specificdates_);
+  ::_pbi::ArenaStringPtr::InternalSwap(&_impl_.acctcode_, &other->_impl_.acctcode_, arena);
+  ::_pbi::ArenaStringPtr::InternalSwap(&_impl_.time_, &other->_impl_.time_, arena);
+  ::_pbi::ArenaStringPtr::InternalSwap(&_impl_.symbol_, &other->_impl_.symbol_, arena);
+  ::_pbi::ArenaStringPtr::InternalSwap(&_impl_.sectype_, &other->_impl_.sectype_, arena);
+  ::_pbi::ArenaStringPtr::InternalSwap(&_impl_.exchange_, &other->_impl_.exchange_, arena);
+  ::_pbi::ArenaStringPtr::InternalSwap(&_impl_.side_, &other->_impl_.side_, arena);
+  ::google::protobuf::internal::memswap<
+      PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.lastndays_)
+      + sizeof(ExecutionFilter::_impl_.lastndays_)
+      - PROTOBUF_FIELD_OFFSET(ExecutionFilter, _impl_.clientid_)>(
+          reinterpret_cast<char*>(&_impl_.clientid_),
+          reinterpret_cast<char*>(&other->_impl_.clientid_));
 }
 
-::PROTOBUF_NAMESPACE_ID::Metadata ExecutionFilter::GetMetadata() const {
-  return GetMetadataStatic();
+::google::protobuf::Metadata ExecutionFilter::GetMetadata() const {
+  return ::google::protobuf::Message::GetMetadataImpl(GetClassData()->full());
 }
-
-
 // @@protoc_insertion_point(namespace_scope)
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> PROTOBUF_NOINLINE ::protobuf::ExecutionFilter* Arena::CreateMaybeMessage< ::protobuf::ExecutionFilter >(Arena* arena) {
-  return Arena::CreateMessageInternal< ::protobuf::ExecutionFilter >(arena);
-}
-PROTOBUF_NAMESPACE_CLOSE
-
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
 // @@protoc_insertion_point(global_scope)
-#include <google/protobuf/port_undef.inc>
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY2 static ::std::false_type
+    _static_init2_ PROTOBUF_UNUSED =
+        (::_pbi::AddDescriptors(&descriptor_table_ExecutionFilter_2eproto),
+         ::std::false_type{});
+#include "google/protobuf/port_undef.inc"
