@@ -1,460 +1,477 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: OpenOrder.proto
+// Protobuf C++ Version: 5.29.3
 
 #include "OpenOrder.pb.h"
 
 #include <algorithm>
-
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/extension_set.h>
-#include <google/protobuf/wire_format_lite.h>
-#include <google/protobuf/descriptor.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/reflection_ops.h>
-#include <google/protobuf/wire_format.h>
+#include <type_traits>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/generated_message_tctable_impl.h"
+#include "google/protobuf/extension_set.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/wire_format_lite.h"
+#include "google/protobuf/descriptor.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/reflection_ops.h"
+#include "google/protobuf/wire_format.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
-extern PROTOBUF_INTERNAL_EXPORT_Contract_2eproto ::PROTOBUF_NAMESPACE_ID::internal::SCCInfo<2> scc_info_Contract_Contract_2eproto;
-extern PROTOBUF_INTERNAL_EXPORT_Order_2eproto ::PROTOBUF_NAMESPACE_ID::internal::SCCInfo<5> scc_info_Order_Order_2eproto;
-extern PROTOBUF_INTERNAL_EXPORT_OrderState_2eproto ::PROTOBUF_NAMESPACE_ID::internal::SCCInfo<1> scc_info_OrderState_OrderState_2eproto;
-namespace protobuf {
-class OpenOrderDefaultTypeInternal {
- public:
-  ::PROTOBUF_NAMESPACE_ID::internal::ExplicitlyConstructed<OpenOrder> _instance;
-} _OpenOrder_default_instance_;
-}  // namespace protobuf
-static void InitDefaultsscc_info_OpenOrder_OpenOrder_2eproto() {
-  GOOGLE_PROTOBUF_VERIFY_VERSION;
 
-  {
-    void* ptr = &::protobuf::_OpenOrder_default_instance_;
-    new (ptr) ::protobuf::OpenOrder();
-    ::PROTOBUF_NAMESPACE_ID::internal::OnShutdownDestroyMessage(ptr);
-  }
-  ::protobuf::OpenOrder::InitAsDefaultInstance();
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+PROTOBUF_PRAGMA_INIT_SEG
+namespace _pb = ::google::protobuf;
+namespace _pbi = ::google::protobuf::internal;
+namespace _fl = ::google::protobuf::internal::field_layout;
+namespace protobuf {
+
+inline constexpr OpenOrder::Impl_::Impl_(
+    ::_pbi::ConstantInitialized) noexcept
+      : _cached_size_{0},
+        contract_{nullptr},
+        order_{nullptr},
+        orderstate_{nullptr},
+        orderid_{0} {}
+
+template <typename>
+PROTOBUF_CONSTEXPR OpenOrder::OpenOrder(::_pbi::ConstantInitialized)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(_class_data_.base()),
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(),
+#endif  // PROTOBUF_CUSTOM_VTABLE
+      _impl_(::_pbi::ConstantInitialized()) {
 }
-
-::PROTOBUF_NAMESPACE_ID::internal::SCCInfo<3> scc_info_OpenOrder_OpenOrder_2eproto =
-    {{ATOMIC_VAR_INIT(::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase::kUninitialized), 3, 0, InitDefaultsscc_info_OpenOrder_OpenOrder_2eproto}, {
-      &scc_info_Contract_Contract_2eproto.base,
-      &scc_info_Order_Order_2eproto.base,
-      &scc_info_OrderState_OrderState_2eproto.base,}};
-
-static ::PROTOBUF_NAMESPACE_ID::Metadata file_level_metadata_OpenOrder_2eproto[1];
-static constexpr ::PROTOBUF_NAMESPACE_ID::EnumDescriptor const** file_level_enum_descriptors_OpenOrder_2eproto = nullptr;
-static constexpr ::PROTOBUF_NAMESPACE_ID::ServiceDescriptor const** file_level_service_descriptors_OpenOrder_2eproto = nullptr;
-
-const ::PROTOBUF_NAMESPACE_ID::uint32 TableStruct_OpenOrder_2eproto::offsets[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, _has_bits_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, _internal_metadata_),
-  ~0u,  // no _extensions_
-  ~0u,  // no _oneof_case_
-  ~0u,  // no _weak_field_map_
-  PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, orderid_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, contract_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, order_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, orderstate_),
-  3,
-  0,
-  1,
-  2,
-};
-static const ::PROTOBUF_NAMESPACE_ID::internal::MigrationSchema schemas[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  { 0, 9, sizeof(::protobuf::OpenOrder)},
+struct OpenOrderDefaultTypeInternal {
+  PROTOBUF_CONSTEXPR OpenOrderDefaultTypeInternal() : _instance(::_pbi::ConstantInitialized{}) {}
+  ~OpenOrderDefaultTypeInternal() {}
+  union {
+    OpenOrder _instance;
+  };
 };
 
-static ::PROTOBUF_NAMESPACE_ID::Message const * const file_default_instances[] = {
-  reinterpret_cast<const ::PROTOBUF_NAMESPACE_ID::Message*>(&::protobuf::_OpenOrder_default_instance_),
+PROTOBUF_ATTRIBUTE_NO_DESTROY PROTOBUF_CONSTINIT
+    PROTOBUF_ATTRIBUTE_INIT_PRIORITY1 OpenOrderDefaultTypeInternal _OpenOrder_default_instance_;
+}  // namespace protobuf
+static constexpr const ::_pb::EnumDescriptor**
+    file_level_enum_descriptors_OpenOrder_2eproto = nullptr;
+static constexpr const ::_pb::ServiceDescriptor**
+    file_level_service_descriptors_OpenOrder_2eproto = nullptr;
+const ::uint32_t
+    TableStruct_OpenOrder_2eproto::offsets[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+        protodesc_cold) = {
+        PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, _impl_._has_bits_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, _internal_metadata_),
+        ~0u,  // no _extensions_
+        ~0u,  // no _oneof_case_
+        ~0u,  // no _weak_field_map_
+        ~0u,  // no _inlined_string_donated_
+        ~0u,  // no _split_
+        ~0u,  // no sizeof(Split)
+        PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, _impl_.orderid_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, _impl_.contract_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, _impl_.order_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrder, _impl_.orderstate_),
+        3,
+        0,
+        1,
+        2,
 };
 
-const char descriptor_table_protodef_OpenOrder_2eproto[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) =
-  "\n\017OpenOrder.proto\022\010protobuf\032\016Contract.pr"
-  "oto\032\013Order.proto\032\020OrderState.proto\"\322\001\n\tO"
-  "penOrder\022\024\n\007orderId\030\001 \001(\005H\000\210\001\001\022)\n\010contra"
-  "ct\030\002 \001(\0132\022.protobuf.ContractH\001\210\001\001\022#\n\005ord"
-  "er\030\003 \001(\0132\017.protobuf.OrderH\002\210\001\001\022-\n\norderS"
-  "tate\030\004 \001(\0132\024.protobuf.OrderStateH\003\210\001\001B\n\n"
-  "\010_orderIdB\013\n\t_contractB\010\n\006_orderB\r\n\013_ord"
-  "erStateB9\n\026com.ib.client.protobufB\016OpenO"
-  "rderProto\252\002\016IBApi.protobufb\006proto3"
-  ;
-static const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable*const descriptor_table_OpenOrder_2eproto_deps[3] = {
-  &::descriptor_table_Contract_2eproto,
-  &::descriptor_table_Order_2eproto,
-  &::descriptor_table_OrderState_2eproto,
+static const ::_pbi::MigrationSchema
+    schemas[] ABSL_ATTRIBUTE_SECTION_VARIABLE(protodesc_cold) = {
+        {0, 12, -1, sizeof(::protobuf::OpenOrder)},
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase*const descriptor_table_OpenOrder_2eproto_sccs[1] = {
-  &scc_info_OpenOrder_OpenOrder_2eproto.base,
+static const ::_pb::Message* const file_default_instances[] = {
+    &::protobuf::_OpenOrder_default_instance_._instance,
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::once_flag descriptor_table_OpenOrder_2eproto_once;
-const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_OpenOrder_2eproto = {
-  false, false, descriptor_table_protodef_OpenOrder_2eproto, "OpenOrder.proto", 354,
-  &descriptor_table_OpenOrder_2eproto_once, descriptor_table_OpenOrder_2eproto_sccs, descriptor_table_OpenOrder_2eproto_deps, 1, 3,
-  schemas, file_default_instances, TableStruct_OpenOrder_2eproto::offsets,
-  file_level_metadata_OpenOrder_2eproto, 1, file_level_enum_descriptors_OpenOrder_2eproto, file_level_service_descriptors_OpenOrder_2eproto,
+const char descriptor_table_protodef_OpenOrder_2eproto[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+    protodesc_cold) = {
+    "\n\017OpenOrder.proto\022\010protobuf\032\016Contract.pr"
+    "oto\032\013Order.proto\032\020OrderState.proto\"\322\001\n\tO"
+    "penOrder\022\024\n\007orderId\030\001 \001(\005H\000\210\001\001\022)\n\010contra"
+    "ct\030\002 \001(\0132\022.protobuf.ContractH\001\210\001\001\022#\n\005ord"
+    "er\030\003 \001(\0132\017.protobuf.OrderH\002\210\001\001\022-\n\norderS"
+    "tate\030\004 \001(\0132\024.protobuf.OrderStateH\003\210\001\001B\n\n"
+    "\010_orderIdB\013\n\t_contractB\010\n\006_orderB\r\n\013_ord"
+    "erStateB9\n\026com.ib.client.protobufB\016OpenO"
+    "rderProto\252\002\016IBApi.protobufb\006proto3"
 };
-
-// Force running AddDescriptors() at dynamic initialization time.
-static bool dynamic_init_dummy_OpenOrder_2eproto = (static_cast<void>(::PROTOBUF_NAMESPACE_ID::internal::AddDescriptors(&descriptor_table_OpenOrder_2eproto)), true);
+static const ::_pbi::DescriptorTable* const descriptor_table_OpenOrder_2eproto_deps[3] =
+    {
+        &::descriptor_table_Contract_2eproto,
+        &::descriptor_table_Order_2eproto,
+        &::descriptor_table_OrderState_2eproto,
+};
+static ::absl::once_flag descriptor_table_OpenOrder_2eproto_once;
+PROTOBUF_CONSTINIT const ::_pbi::DescriptorTable descriptor_table_OpenOrder_2eproto = {
+    false,
+    false,
+    354,
+    descriptor_table_protodef_OpenOrder_2eproto,
+    "OpenOrder.proto",
+    &descriptor_table_OpenOrder_2eproto_once,
+    descriptor_table_OpenOrder_2eproto_deps,
+    3,
+    1,
+    schemas,
+    file_default_instances,
+    TableStruct_OpenOrder_2eproto::offsets,
+    file_level_enum_descriptors_OpenOrder_2eproto,
+    file_level_service_descriptors_OpenOrder_2eproto,
+};
 namespace protobuf {
-
 // ===================================================================
 
-void OpenOrder::InitAsDefaultInstance() {
-  ::protobuf::_OpenOrder_default_instance_._instance.get_mutable()->contract_ = const_cast< ::protobuf::Contract*>(
-      ::protobuf::Contract::internal_default_instance());
-  ::protobuf::_OpenOrder_default_instance_._instance.get_mutable()->order_ = const_cast< ::protobuf::Order*>(
-      ::protobuf::Order::internal_default_instance());
-  ::protobuf::_OpenOrder_default_instance_._instance.get_mutable()->orderstate_ = const_cast< ::protobuf::OrderState*>(
-      ::protobuf::OrderState::internal_default_instance());
-}
 class OpenOrder::_Internal {
  public:
-  using HasBits = decltype(std::declval<OpenOrder>()._has_bits_);
-  static void set_has_orderid(HasBits* has_bits) {
-    (*has_bits)[0] |= 8u;
-  }
-  static const ::protobuf::Contract& contract(const OpenOrder* msg);
-  static void set_has_contract(HasBits* has_bits) {
-    (*has_bits)[0] |= 1u;
-  }
-  static const ::protobuf::Order& order(const OpenOrder* msg);
-  static void set_has_order(HasBits* has_bits) {
-    (*has_bits)[0] |= 2u;
-  }
-  static const ::protobuf::OrderState& orderstate(const OpenOrder* msg);
-  static void set_has_orderstate(HasBits* has_bits) {
-    (*has_bits)[0] |= 4u;
-  }
+  using HasBits =
+      decltype(std::declval<OpenOrder>()._impl_._has_bits_);
+  static constexpr ::int32_t kHasBitsOffset =
+      8 * PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_._has_bits_);
 };
 
-const ::protobuf::Contract&
-OpenOrder::_Internal::contract(const OpenOrder* msg) {
-  return *msg->contract_;
-}
-const ::protobuf::Order&
-OpenOrder::_Internal::order(const OpenOrder* msg) {
-  return *msg->order_;
-}
-const ::protobuf::OrderState&
-OpenOrder::_Internal::orderstate(const OpenOrder* msg) {
-  return *msg->orderstate_;
-}
 void OpenOrder::clear_contract() {
-  if (GetArena() == nullptr && contract_ != nullptr) {
-    delete contract_;
-  }
-  contract_ = nullptr;
-  _has_bits_[0] &= ~0x00000001u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (_impl_.contract_ != nullptr) _impl_.contract_->Clear();
+  _impl_._has_bits_[0] &= ~0x00000001u;
 }
 void OpenOrder::clear_order() {
-  if (GetArena() == nullptr && order_ != nullptr) {
-    delete order_;
-  }
-  order_ = nullptr;
-  _has_bits_[0] &= ~0x00000002u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (_impl_.order_ != nullptr) _impl_.order_->Clear();
+  _impl_._has_bits_[0] &= ~0x00000002u;
 }
 void OpenOrder::clear_orderstate() {
-  if (GetArena() == nullptr && orderstate_ != nullptr) {
-    delete orderstate_;
-  }
-  orderstate_ = nullptr;
-  _has_bits_[0] &= ~0x00000004u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (_impl_.orderstate_ != nullptr) _impl_.orderstate_->Clear();
+  _impl_._has_bits_[0] &= ~0x00000004u;
 }
-OpenOrder::OpenOrder(::PROTOBUF_NAMESPACE_ID::Arena* arena)
-  : ::PROTOBUF_NAMESPACE_ID::Message(arena) {
-  SharedCtor();
-  RegisterArenaDtor(arena);
+OpenOrder::OpenOrder(::google::protobuf::Arena* arena)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  SharedCtor(arena);
   // @@protoc_insertion_point(arena_constructor:protobuf.OpenOrder)
 }
-OpenOrder::OpenOrder(const OpenOrder& from)
-  : ::PROTOBUF_NAMESPACE_ID::Message(),
-      _has_bits_(from._has_bits_) {
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  if (from._internal_has_contract()) {
-    contract_ = new ::protobuf::Contract(*from.contract_);
-  } else {
-    contract_ = nullptr;
-  }
-  if (from._internal_has_order()) {
-    order_ = new ::protobuf::Order(*from.order_);
-  } else {
-    order_ = nullptr;
-  }
-  if (from._internal_has_orderstate()) {
-    orderstate_ = new ::protobuf::OrderState(*from.orderstate_);
-  } else {
-    orderstate_ = nullptr;
-  }
-  orderid_ = from.orderid_;
+inline PROTOBUF_NDEBUG_INLINE OpenOrder::Impl_::Impl_(
+    ::google::protobuf::internal::InternalVisibility visibility, ::google::protobuf::Arena* arena,
+    const Impl_& from, const ::protobuf::OpenOrder& from_msg)
+      : _has_bits_{from._has_bits_},
+        _cached_size_{0} {}
+
+OpenOrder::OpenOrder(
+    ::google::protobuf::Arena* arena,
+    const OpenOrder& from)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  OpenOrder* const _this = this;
+  (void)_this;
+  _internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(
+      from._internal_metadata_);
+  new (&_impl_) Impl_(internal_visibility(), arena, from._impl_, from);
+  ::uint32_t cached_has_bits = _impl_._has_bits_[0];
+  _impl_.contract_ = (cached_has_bits & 0x00000001u) ? ::google::protobuf::Message::CopyConstruct<::protobuf::Contract>(
+                              arena, *from._impl_.contract_)
+                        : nullptr;
+  _impl_.order_ = (cached_has_bits & 0x00000002u) ? ::google::protobuf::Message::CopyConstruct<::protobuf::Order>(
+                              arena, *from._impl_.order_)
+                        : nullptr;
+  _impl_.orderstate_ = (cached_has_bits & 0x00000004u) ? ::google::protobuf::Message::CopyConstruct<::protobuf::OrderState>(
+                              arena, *from._impl_.orderstate_)
+                        : nullptr;
+  _impl_.orderid_ = from._impl_.orderid_;
+
   // @@protoc_insertion_point(copy_constructor:protobuf.OpenOrder)
 }
+inline PROTOBUF_NDEBUG_INLINE OpenOrder::Impl_::Impl_(
+    ::google::protobuf::internal::InternalVisibility visibility,
+    ::google::protobuf::Arena* arena)
+      : _cached_size_{0} {}
 
-void OpenOrder::SharedCtor() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&scc_info_OpenOrder_OpenOrder_2eproto.base);
-  ::memset(&contract_, 0, static_cast<size_t>(
-      reinterpret_cast<char*>(&orderid_) -
-      reinterpret_cast<char*>(&contract_)) + sizeof(orderid_));
+inline void OpenOrder::SharedCtor(::_pb::Arena* arena) {
+  new (&_impl_) Impl_(internal_visibility(), arena);
+  ::memset(reinterpret_cast<char *>(&_impl_) +
+               offsetof(Impl_, contract_),
+           0,
+           offsetof(Impl_, orderid_) -
+               offsetof(Impl_, contract_) +
+               sizeof(Impl_::orderid_));
 }
-
 OpenOrder::~OpenOrder() {
   // @@protoc_insertion_point(destructor:protobuf.OpenOrder)
-  SharedDtor();
-  _internal_metadata_.Delete<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  SharedDtor(*this);
+}
+inline void OpenOrder::SharedDtor(MessageLite& self) {
+  OpenOrder& this_ = static_cast<OpenOrder&>(self);
+  this_._internal_metadata_.Delete<::google::protobuf::UnknownFieldSet>();
+  ABSL_DCHECK(this_.GetArena() == nullptr);
+  delete this_._impl_.contract_;
+  delete this_._impl_.order_;
+  delete this_._impl_.orderstate_;
+  this_._impl_.~Impl_();
 }
 
-void OpenOrder::SharedDtor() {
-  GOOGLE_DCHECK(GetArena() == nullptr);
-  if (this != internal_default_instance()) delete contract_;
-  if (this != internal_default_instance()) delete order_;
-  if (this != internal_default_instance()) delete orderstate_;
+inline void* OpenOrder::PlacementNew_(const void*, void* mem,
+                                        ::google::protobuf::Arena* arena) {
+  return ::new (mem) OpenOrder(arena);
 }
+constexpr auto OpenOrder::InternalNewImpl_() {
+  return ::google::protobuf::internal::MessageCreator::ZeroInit(sizeof(OpenOrder),
+                                            alignof(OpenOrder));
+}
+PROTOBUF_CONSTINIT
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::google::protobuf::internal::ClassDataFull OpenOrder::_class_data_ = {
+    ::google::protobuf::internal::ClassData{
+        &_OpenOrder_default_instance_._instance,
+        &_table_.header,
+        nullptr,  // OnDemandRegisterArenaDtor
+        nullptr,  // IsInitialized
+        &OpenOrder::MergeImpl,
+        ::google::protobuf::Message::GetNewImpl<OpenOrder>(),
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        &OpenOrder::SharedDtor,
+        ::google::protobuf::Message::GetClearImpl<OpenOrder>(), &OpenOrder::ByteSizeLong,
+            &OpenOrder::_InternalSerialize,
+#endif  // PROTOBUF_CUSTOM_VTABLE
+        PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_._cached_size_),
+        false,
+    },
+    &OpenOrder::kDescriptorMethods,
+    &descriptor_table_OpenOrder_2eproto,
+    nullptr,  // tracker
+};
+const ::google::protobuf::internal::ClassData* OpenOrder::GetClassData() const {
+  ::google::protobuf::internal::PrefetchToLocalCache(&_class_data_);
+  ::google::protobuf::internal::PrefetchToLocalCache(_class_data_.tc_table);
+  return _class_data_.base();
+}
+PROTOBUF_CONSTINIT PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::_pbi::TcParseTable<2, 4, 3, 0, 2> OpenOrder::_table_ = {
+  {
+    PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_._has_bits_),
+    0, // no _extensions_
+    4, 24,  // max_field_number, fast_idx_mask
+    offsetof(decltype(_table_), field_lookup_table),
+    4294967280,  // skipmap
+    offsetof(decltype(_table_), field_entries),
+    4,  // num_field_entries
+    3,  // num_aux_entries
+    offsetof(decltype(_table_), aux_entries),
+    _class_data_.base(),
+    nullptr,  // post_loop_handler
+    ::_pbi::TcParser::GenericFallback,  // fallback
+    #ifdef PROTOBUF_PREFETCH_PARSE_TABLE
+    ::_pbi::TcParser::GetTable<::protobuf::OpenOrder>(),  // to_prefetch
+    #endif  // PROTOBUF_PREFETCH_PARSE_TABLE
+  }, {{
+    // optional .protobuf.OrderState orderState = 4;
+    {::_pbi::TcParser::FastMtS1,
+     {34, 2, 2, PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_.orderstate_)}},
+    // optional int32 orderId = 1;
+    {::_pbi::TcParser::SingularVarintNoZag1<::uint32_t, offsetof(OpenOrder, _impl_.orderid_), 3>(),
+     {8, 3, 0, PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_.orderid_)}},
+    // optional .protobuf.Contract contract = 2;
+    {::_pbi::TcParser::FastMtS1,
+     {18, 0, 0, PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_.contract_)}},
+    // optional .protobuf.Order order = 3;
+    {::_pbi::TcParser::FastMtS1,
+     {26, 1, 1, PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_.order_)}},
+  }}, {{
+    65535, 65535
+  }}, {{
+    // optional int32 orderId = 1;
+    {PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_.orderid_), _Internal::kHasBitsOffset + 3, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kInt32)},
+    // optional .protobuf.Contract contract = 2;
+    {PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_.contract_), _Internal::kHasBitsOffset + 0, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kMessage | ::_fl::kTvTable)},
+    // optional .protobuf.Order order = 3;
+    {PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_.order_), _Internal::kHasBitsOffset + 1, 1,
+    (0 | ::_fl::kFcOptional | ::_fl::kMessage | ::_fl::kTvTable)},
+    // optional .protobuf.OrderState orderState = 4;
+    {PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_.orderstate_), _Internal::kHasBitsOffset + 2, 2,
+    (0 | ::_fl::kFcOptional | ::_fl::kMessage | ::_fl::kTvTable)},
+  }}, {{
+    {::_pbi::TcParser::GetTable<::protobuf::Contract>()},
+    {::_pbi::TcParser::GetTable<::protobuf::Order>()},
+    {::_pbi::TcParser::GetTable<::protobuf::OrderState>()},
+  }}, {{
+  }},
+};
 
-void OpenOrder::ArenaDtor(void* object) {
-  OpenOrder* _this = reinterpret_cast< OpenOrder* >(object);
-  (void)_this;
-}
-void OpenOrder::RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena*) {
-}
-void OpenOrder::SetCachedSize(int size) const {
-  _cached_size_.Set(size);
-}
-const OpenOrder& OpenOrder::default_instance() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&::scc_info_OpenOrder_OpenOrder_2eproto.base);
-  return *internal_default_instance();
-}
-
-
-void OpenOrder::Clear() {
+PROTOBUF_NOINLINE void OpenOrder::Clear() {
 // @@protoc_insertion_point(message_clear_start:protobuf.OpenOrder)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  ::uint32_t cached_has_bits = 0;
   // Prevent compiler warnings about cached_has_bits being unused
   (void) cached_has_bits;
 
-  cached_has_bits = _has_bits_[0];
+  cached_has_bits = _impl_._has_bits_[0];
   if (cached_has_bits & 0x00000007u) {
     if (cached_has_bits & 0x00000001u) {
-      if (GetArena() == nullptr && contract_ != nullptr) {
-        delete contract_;
-      }
-      contract_ = nullptr;
+      ABSL_DCHECK(_impl_.contract_ != nullptr);
+      _impl_.contract_->Clear();
     }
     if (cached_has_bits & 0x00000002u) {
-      if (GetArena() == nullptr && order_ != nullptr) {
-        delete order_;
-      }
-      order_ = nullptr;
+      ABSL_DCHECK(_impl_.order_ != nullptr);
+      _impl_.order_->Clear();
     }
     if (cached_has_bits & 0x00000004u) {
-      if (GetArena() == nullptr && orderstate_ != nullptr) {
-        delete orderstate_;
-      }
-      orderstate_ = nullptr;
+      ABSL_DCHECK(_impl_.orderstate_ != nullptr);
+      _impl_.orderstate_->Clear();
     }
   }
-  orderid_ = 0;
-  _has_bits_.Clear();
-  _internal_metadata_.Clear<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  _impl_.orderid_ = 0;
+  _impl_._has_bits_.Clear();
+  _internal_metadata_.Clear<::google::protobuf::UnknownFieldSet>();
 }
 
-const char* OpenOrder::_InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) {
-#define CHK_(x) if (PROTOBUF_PREDICT_FALSE(!(x))) goto failure
-  _Internal::HasBits has_bits{};
-  ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena(); (void)arena;
-  while (!ctx->Done(&ptr)) {
-    ::PROTOBUF_NAMESPACE_ID::uint32 tag;
-    ptr = ::PROTOBUF_NAMESPACE_ID::internal::ReadTag(ptr, &tag);
-    CHK_(ptr);
-    switch (tag >> 3) {
-      // int32 orderId = 1;
-      case 1:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 8)) {
-          _Internal::set_has_orderid(&has_bits);
-          orderid_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // .protobuf.Contract contract = 2;
-      case 2:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 18)) {
-          ptr = ctx->ParseMessage(_internal_mutable_contract(), ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // .protobuf.Order order = 3;
-      case 3:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 26)) {
-          ptr = ctx->ParseMessage(_internal_mutable_order(), ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // .protobuf.OrderState orderState = 4;
-      case 4:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 34)) {
-          ptr = ctx->ParseMessage(_internal_mutable_orderstate(), ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      default: {
-      handle_unusual:
-        if ((tag & 7) == 4 || tag == 0) {
-          ctx->SetLastTag(tag);
-          goto success;
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::uint8_t* OpenOrder::_InternalSerialize(
+            const MessageLite& base, ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) {
+          const OpenOrder& this_ = static_cast<const OpenOrder&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::uint8_t* OpenOrder::_InternalSerialize(
+            ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+          const OpenOrder& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(serialize_to_array_start:protobuf.OpenOrder)
+          ::uint32_t cached_has_bits = 0;
+          (void)cached_has_bits;
+
+          cached_has_bits = this_._impl_._has_bits_[0];
+          // optional int32 orderId = 1;
+          if (cached_has_bits & 0x00000008u) {
+            target = ::google::protobuf::internal::WireFormatLite::
+                WriteInt32ToArrayWithField<1>(
+                    stream, this_._internal_orderid(), target);
+          }
+
+          // optional .protobuf.Contract contract = 2;
+          if (cached_has_bits & 0x00000001u) {
+            target = ::google::protobuf::internal::WireFormatLite::InternalWriteMessage(
+                2, *this_._impl_.contract_, this_._impl_.contract_->GetCachedSize(), target,
+                stream);
+          }
+
+          // optional .protobuf.Order order = 3;
+          if (cached_has_bits & 0x00000002u) {
+            target = ::google::protobuf::internal::WireFormatLite::InternalWriteMessage(
+                3, *this_._impl_.order_, this_._impl_.order_->GetCachedSize(), target,
+                stream);
+          }
+
+          // optional .protobuf.OrderState orderState = 4;
+          if (cached_has_bits & 0x00000004u) {
+            target = ::google::protobuf::internal::WireFormatLite::InternalWriteMessage(
+                4, *this_._impl_.orderstate_, this_._impl_.orderstate_->GetCachedSize(), target,
+                stream);
+          }
+
+          if (PROTOBUF_PREDICT_FALSE(this_._internal_metadata_.have_unknown_fields())) {
+            target =
+                ::_pbi::WireFormat::InternalSerializeUnknownFieldsToArray(
+                    this_._internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance), target, stream);
+          }
+          // @@protoc_insertion_point(serialize_to_array_end:protobuf.OpenOrder)
+          return target;
         }
-        ptr = UnknownFieldParse(tag,
-            _internal_metadata_.mutable_unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(),
-            ptr, ctx);
-        CHK_(ptr != nullptr);
-        continue;
+
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::size_t OpenOrder::ByteSizeLong(const MessageLite& base) {
+          const OpenOrder& this_ = static_cast<const OpenOrder&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::size_t OpenOrder::ByteSizeLong() const {
+          const OpenOrder& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(message_byte_size_start:protobuf.OpenOrder)
+          ::size_t total_size = 0;
+
+          ::uint32_t cached_has_bits = 0;
+          // Prevent compiler warnings about cached_has_bits being unused
+          (void)cached_has_bits;
+
+          ::_pbi::Prefetch5LinesFrom7Lines(&this_);
+          cached_has_bits = this_._impl_._has_bits_[0];
+          if (cached_has_bits & 0x0000000fu) {
+            // optional .protobuf.Contract contract = 2;
+            if (cached_has_bits & 0x00000001u) {
+              total_size += 1 +
+                            ::google::protobuf::internal::WireFormatLite::MessageSize(*this_._impl_.contract_);
+            }
+            // optional .protobuf.Order order = 3;
+            if (cached_has_bits & 0x00000002u) {
+              total_size += 1 +
+                            ::google::protobuf::internal::WireFormatLite::MessageSize(*this_._impl_.order_);
+            }
+            // optional .protobuf.OrderState orderState = 4;
+            if (cached_has_bits & 0x00000004u) {
+              total_size += 1 +
+                            ::google::protobuf::internal::WireFormatLite::MessageSize(*this_._impl_.orderstate_);
+            }
+            // optional int32 orderId = 1;
+            if (cached_has_bits & 0x00000008u) {
+              total_size += ::_pbi::WireFormatLite::Int32SizePlusOne(
+                  this_._internal_orderid());
+            }
+          }
+          return this_.MaybeComputeUnknownFieldsSize(total_size,
+                                                     &this_._impl_._cached_size_);
+        }
+
+void OpenOrder::MergeImpl(::google::protobuf::MessageLite& to_msg, const ::google::protobuf::MessageLite& from_msg) {
+  auto* const _this = static_cast<OpenOrder*>(&to_msg);
+  auto& from = static_cast<const OpenOrder&>(from_msg);
+  ::google::protobuf::Arena* arena = _this->GetArena();
+  // @@protoc_insertion_point(class_specific_merge_from_start:protobuf.OpenOrder)
+  ABSL_DCHECK_NE(&from, _this);
+  ::uint32_t cached_has_bits = 0;
+  (void) cached_has_bits;
+
+  cached_has_bits = from._impl_._has_bits_[0];
+  if (cached_has_bits & 0x0000000fu) {
+    if (cached_has_bits & 0x00000001u) {
+      ABSL_DCHECK(from._impl_.contract_ != nullptr);
+      if (_this->_impl_.contract_ == nullptr) {
+        _this->_impl_.contract_ =
+            ::google::protobuf::Message::CopyConstruct<::protobuf::Contract>(arena, *from._impl_.contract_);
+      } else {
+        _this->_impl_.contract_->MergeFrom(*from._impl_.contract_);
       }
-    }  // switch
-  }  // while
-success:
-  _has_bits_.Or(has_bits);
-  return ptr;
-failure:
-  ptr = nullptr;
-  goto success;
-#undef CHK_
-}
-
-::PROTOBUF_NAMESPACE_ID::uint8* OpenOrder::_InternalSerialize(
-    ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const {
-  // @@protoc_insertion_point(serialize_to_array_start:protobuf.OpenOrder)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  (void) cached_has_bits;
-
-  // int32 orderId = 1;
-  if (_internal_has_orderid()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteInt32ToArray(1, this->_internal_orderid(), target);
-  }
-
-  // .protobuf.Contract contract = 2;
-  if (_internal_has_contract()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::
-      InternalWriteMessage(
-        2, _Internal::contract(this), target, stream);
-  }
-
-  // .protobuf.Order order = 3;
-  if (_internal_has_order()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::
-      InternalWriteMessage(
-        3, _Internal::order(this), target, stream);
-  }
-
-  // .protobuf.OrderState orderState = 4;
-  if (_internal_has_orderstate()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::
-      InternalWriteMessage(
-        4, _Internal::orderstate(this), target, stream);
-  }
-
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormat::InternalSerializeUnknownFieldsToArray(
-        _internal_metadata_.unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(::PROTOBUF_NAMESPACE_ID::UnknownFieldSet::default_instance), target, stream);
-  }
-  // @@protoc_insertion_point(serialize_to_array_end:protobuf.OpenOrder)
-  return target;
-}
-
-size_t OpenOrder::ByteSizeLong() const {
-// @@protoc_insertion_point(message_byte_size_start:protobuf.OpenOrder)
-  size_t total_size = 0;
-
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  // Prevent compiler warnings about cached_has_bits being unused
-  (void) cached_has_bits;
-
-  cached_has_bits = _has_bits_[0];
-  if (cached_has_bits & 0x0000000fu) {
-    // .protobuf.Contract contract = 2;
-    if (cached_has_bits & 0x00000001u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::MessageSize(
-          *contract_);
-    }
-
-    // .protobuf.Order order = 3;
-    if (cached_has_bits & 0x00000002u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::MessageSize(
-          *order_);
-    }
-
-    // .protobuf.OrderState orderState = 4;
-    if (cached_has_bits & 0x00000004u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::MessageSize(
-          *orderstate_);
-    }
-
-    // int32 orderId = 1;
-    if (cached_has_bits & 0x00000008u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::Int32Size(
-          this->_internal_orderid());
-    }
-
-  }
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    return ::PROTOBUF_NAMESPACE_ID::internal::ComputeUnknownFieldsSize(
-        _internal_metadata_, total_size, &_cached_size_);
-  }
-  int cached_size = ::PROTOBUF_NAMESPACE_ID::internal::ToCachedSize(total_size);
-  SetCachedSize(cached_size);
-  return total_size;
-}
-
-void OpenOrder::MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_merge_from_start:protobuf.OpenOrder)
-  GOOGLE_DCHECK_NE(&from, this);
-  const OpenOrder* source =
-      ::PROTOBUF_NAMESPACE_ID::DynamicCastToGenerated<OpenOrder>(
-          &from);
-  if (source == nullptr) {
-  // @@protoc_insertion_point(generalized_merge_from_cast_fail:protobuf.OpenOrder)
-    ::PROTOBUF_NAMESPACE_ID::internal::ReflectionOps::Merge(from, this);
-  } else {
-  // @@protoc_insertion_point(generalized_merge_from_cast_success:protobuf.OpenOrder)
-    MergeFrom(*source);
-  }
-}
-
-void OpenOrder::MergeFrom(const OpenOrder& from) {
-// @@protoc_insertion_point(class_specific_merge_from_start:protobuf.OpenOrder)
-  GOOGLE_DCHECK_NE(&from, this);
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  (void) cached_has_bits;
-
-  cached_has_bits = from._has_bits_[0];
-  if (cached_has_bits & 0x0000000fu) {
-    if (cached_has_bits & 0x00000001u) {
-      _internal_mutable_contract()->::protobuf::Contract::MergeFrom(from._internal_contract());
     }
     if (cached_has_bits & 0x00000002u) {
-      _internal_mutable_order()->::protobuf::Order::MergeFrom(from._internal_order());
+      ABSL_DCHECK(from._impl_.order_ != nullptr);
+      if (_this->_impl_.order_ == nullptr) {
+        _this->_impl_.order_ =
+            ::google::protobuf::Message::CopyConstruct<::protobuf::Order>(arena, *from._impl_.order_);
+      } else {
+        _this->_impl_.order_->MergeFrom(*from._impl_.order_);
+      }
     }
     if (cached_has_bits & 0x00000004u) {
-      _internal_mutable_orderstate()->::protobuf::OrderState::MergeFrom(from._internal_orderstate());
+      ABSL_DCHECK(from._impl_.orderstate_ != nullptr);
+      if (_this->_impl_.orderstate_ == nullptr) {
+        _this->_impl_.orderstate_ =
+            ::google::protobuf::Message::CopyConstruct<::protobuf::OrderState>(arena, *from._impl_.orderstate_);
+      } else {
+        _this->_impl_.orderstate_->MergeFrom(*from._impl_.orderstate_);
+      }
     }
     if (cached_has_bits & 0x00000008u) {
-      orderid_ = from.orderid_;
+      _this->_impl_.orderid_ = from._impl_.orderid_;
     }
-    _has_bits_[0] |= cached_has_bits;
   }
-}
-
-void OpenOrder::CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_copy_from_start:protobuf.OpenOrder)
-  if (&from == this) return;
-  Clear();
-  MergeFrom(from);
+  _this->_impl_._has_bits_[0] |= cached_has_bits;
+  _this->_internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(from._internal_metadata_);
 }
 
 void OpenOrder::CopyFrom(const OpenOrder& from) {
@@ -464,34 +481,31 @@ void OpenOrder::CopyFrom(const OpenOrder& from) {
   MergeFrom(from);
 }
 
-bool OpenOrder::IsInitialized() const {
-  return true;
-}
 
-void OpenOrder::InternalSwap(OpenOrder* other) {
+void OpenOrder::InternalSwap(OpenOrder* PROTOBUF_RESTRICT other) {
   using std::swap;
-  _internal_metadata_.Swap<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(&other->_internal_metadata_);
-  swap(_has_bits_[0], other->_has_bits_[0]);
-  ::PROTOBUF_NAMESPACE_ID::internal::memswap<
-      PROTOBUF_FIELD_OFFSET(OpenOrder, orderid_)
-      + sizeof(OpenOrder::orderid_)
-      - PROTOBUF_FIELD_OFFSET(OpenOrder, contract_)>(
-          reinterpret_cast<char*>(&contract_),
-          reinterpret_cast<char*>(&other->contract_));
+  _internal_metadata_.InternalSwap(&other->_internal_metadata_);
+  swap(_impl_._has_bits_[0], other->_impl_._has_bits_[0]);
+  ::google::protobuf::internal::memswap<
+      PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_.orderid_)
+      + sizeof(OpenOrder::_impl_.orderid_)
+      - PROTOBUF_FIELD_OFFSET(OpenOrder, _impl_.contract_)>(
+          reinterpret_cast<char*>(&_impl_.contract_),
+          reinterpret_cast<char*>(&other->_impl_.contract_));
 }
 
-::PROTOBUF_NAMESPACE_ID::Metadata OpenOrder::GetMetadata() const {
-  return GetMetadataStatic();
+::google::protobuf::Metadata OpenOrder::GetMetadata() const {
+  return ::google::protobuf::Message::GetMetadataImpl(GetClassData()->full());
 }
-
-
 // @@protoc_insertion_point(namespace_scope)
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> PROTOBUF_NOINLINE ::protobuf::OpenOrder* Arena::CreateMaybeMessage< ::protobuf::OpenOrder >(Arena* arena) {
-  return Arena::CreateMessageInternal< ::protobuf::OpenOrder >(arena);
-}
-PROTOBUF_NAMESPACE_CLOSE
-
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
 // @@protoc_insertion_point(global_scope)
-#include <google/protobuf/port_undef.inc>
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY2 static ::std::false_type
+    _static_init2_ PROTOBUF_UNUSED =
+        (::_pbi::AddDescriptors(&descriptor_table_OpenOrder_2eproto),
+         ::std::false_type{});
+#include "google/protobuf/port_undef.inc"
