@@ -1,252 +1,209 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: OpenOrdersEnd.proto
+// Protobuf C++ Version: 5.29.3
 
 #include "OpenOrdersEnd.pb.h"
 
 #include <algorithm>
-
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/extension_set.h>
-#include <google/protobuf/wire_format_lite.h>
-#include <google/protobuf/descriptor.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/reflection_ops.h>
-#include <google/protobuf/wire_format.h>
+#include <type_traits>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/generated_message_tctable_impl.h"
+#include "google/protobuf/extension_set.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/wire_format_lite.h"
+#include "google/protobuf/descriptor.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/reflection_ops.h"
+#include "google/protobuf/wire_format.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
-namespace protobuf {
-class OpenOrdersEndDefaultTypeInternal {
- public:
-  ::PROTOBUF_NAMESPACE_ID::internal::ExplicitlyConstructed<OpenOrdersEnd> _instance;
-} _OpenOrdersEnd_default_instance_;
-}  // namespace protobuf
-static void InitDefaultsscc_info_OpenOrdersEnd_OpenOrdersEnd_2eproto() {
-  GOOGLE_PROTOBUF_VERIFY_VERSION;
 
-  {
-    void* ptr = &::protobuf::_OpenOrdersEnd_default_instance_;
-    new (ptr) ::protobuf::OpenOrdersEnd();
-    ::PROTOBUF_NAMESPACE_ID::internal::OnShutdownDestroyMessage(ptr);
-  }
-  ::protobuf::OpenOrdersEnd::InitAsDefaultInstance();
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+PROTOBUF_PRAGMA_INIT_SEG
+namespace _pb = ::google::protobuf;
+namespace _pbi = ::google::protobuf::internal;
+namespace _fl = ::google::protobuf::internal::field_layout;
+namespace protobuf {
+              template <typename>
+PROTOBUF_CONSTEXPR OpenOrdersEnd::OpenOrdersEnd(::_pbi::ConstantInitialized)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::internal::ZeroFieldsBase(_class_data_.base()){}
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::internal::ZeroFieldsBase() {
 }
-
-::PROTOBUF_NAMESPACE_ID::internal::SCCInfo<0> scc_info_OpenOrdersEnd_OpenOrdersEnd_2eproto =
-    {{ATOMIC_VAR_INIT(::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase::kUninitialized), 0, 0, InitDefaultsscc_info_OpenOrdersEnd_OpenOrdersEnd_2eproto}, {}};
-
-static ::PROTOBUF_NAMESPACE_ID::Metadata file_level_metadata_OpenOrdersEnd_2eproto[1];
-static constexpr ::PROTOBUF_NAMESPACE_ID::EnumDescriptor const** file_level_enum_descriptors_OpenOrdersEnd_2eproto = nullptr;
-static constexpr ::PROTOBUF_NAMESPACE_ID::ServiceDescriptor const** file_level_service_descriptors_OpenOrdersEnd_2eproto = nullptr;
-
-const ::PROTOBUF_NAMESPACE_ID::uint32 TableStruct_OpenOrdersEnd_2eproto::offsets[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  ~0u,  // no _has_bits_
-  PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrdersEnd, _internal_metadata_),
-  ~0u,  // no _extensions_
-  ~0u,  // no _oneof_case_
-  ~0u,  // no _weak_field_map_
-};
-static const ::PROTOBUF_NAMESPACE_ID::internal::MigrationSchema schemas[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  { 0, -1, sizeof(::protobuf::OpenOrdersEnd)},
+#endif  // PROTOBUF_CUSTOM_VTABLE
+struct OpenOrdersEndDefaultTypeInternal {
+  PROTOBUF_CONSTEXPR OpenOrdersEndDefaultTypeInternal() : _instance(::_pbi::ConstantInitialized{}) {}
+  ~OpenOrdersEndDefaultTypeInternal() {}
+  union {
+    OpenOrdersEnd _instance;
+  };
 };
 
-static ::PROTOBUF_NAMESPACE_ID::Message const * const file_default_instances[] = {
-  reinterpret_cast<const ::PROTOBUF_NAMESPACE_ID::Message*>(&::protobuf::_OpenOrdersEnd_default_instance_),
+PROTOBUF_ATTRIBUTE_NO_DESTROY PROTOBUF_CONSTINIT
+    PROTOBUF_ATTRIBUTE_INIT_PRIORITY1 OpenOrdersEndDefaultTypeInternal _OpenOrdersEnd_default_instance_;
+}  // namespace protobuf
+static constexpr const ::_pb::EnumDescriptor**
+    file_level_enum_descriptors_OpenOrdersEnd_2eproto = nullptr;
+static constexpr const ::_pb::ServiceDescriptor**
+    file_level_service_descriptors_OpenOrdersEnd_2eproto = nullptr;
+const ::uint32_t
+    TableStruct_OpenOrdersEnd_2eproto::offsets[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+        protodesc_cold) = {
+        ~0u,  // no _has_bits_
+        PROTOBUF_FIELD_OFFSET(::protobuf::OpenOrdersEnd, _internal_metadata_),
+        ~0u,  // no _extensions_
+        ~0u,  // no _oneof_case_
+        ~0u,  // no _weak_field_map_
+        ~0u,  // no _inlined_string_donated_
+        ~0u,  // no _split_
+        ~0u,  // no sizeof(Split)
 };
 
-const char descriptor_table_protodef_OpenOrdersEnd_2eproto[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) =
-  "\n\023OpenOrdersEnd.proto\022\010protobuf\"\017\n\rOpenO"
-  "rdersEndB=\n\026com.ib.client.protobufB\022Open"
-  "OrdersEndProto\252\002\016IBApi.protobufb\006proto3"
-  ;
-static const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable*const descriptor_table_OpenOrdersEnd_2eproto_deps[1] = {
+static const ::_pbi::MigrationSchema
+    schemas[] ABSL_ATTRIBUTE_SECTION_VARIABLE(protodesc_cold) = {
+        {0, -1, -1, sizeof(::protobuf::OpenOrdersEnd)},
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase*const descriptor_table_OpenOrdersEnd_2eproto_sccs[1] = {
-  &scc_info_OpenOrdersEnd_OpenOrdersEnd_2eproto.base,
+static const ::_pb::Message* const file_default_instances[] = {
+    &::protobuf::_OpenOrdersEnd_default_instance_._instance,
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::once_flag descriptor_table_OpenOrdersEnd_2eproto_once;
-const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_OpenOrdersEnd_2eproto = {
-  false, false, descriptor_table_protodef_OpenOrdersEnd_2eproto, "OpenOrdersEnd.proto", 119,
-  &descriptor_table_OpenOrdersEnd_2eproto_once, descriptor_table_OpenOrdersEnd_2eproto_sccs, descriptor_table_OpenOrdersEnd_2eproto_deps, 1, 0,
-  schemas, file_default_instances, TableStruct_OpenOrdersEnd_2eproto::offsets,
-  file_level_metadata_OpenOrdersEnd_2eproto, 1, file_level_enum_descriptors_OpenOrdersEnd_2eproto, file_level_service_descriptors_OpenOrdersEnd_2eproto,
+const char descriptor_table_protodef_OpenOrdersEnd_2eproto[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+    protodesc_cold) = {
+    "\n\023OpenOrdersEnd.proto\022\010protobuf\"\017\n\rOpenO"
+    "rdersEndB=\n\026com.ib.client.protobufB\022Open"
+    "OrdersEndProto\252\002\016IBApi.protobufb\006proto3"
 };
-
-// Force running AddDescriptors() at dynamic initialization time.
-static bool dynamic_init_dummy_OpenOrdersEnd_2eproto = (static_cast<void>(::PROTOBUF_NAMESPACE_ID::internal::AddDescriptors(&descriptor_table_OpenOrdersEnd_2eproto)), true);
+static ::absl::once_flag descriptor_table_OpenOrdersEnd_2eproto_once;
+PROTOBUF_CONSTINIT const ::_pbi::DescriptorTable descriptor_table_OpenOrdersEnd_2eproto = {
+    false,
+    false,
+    119,
+    descriptor_table_protodef_OpenOrdersEnd_2eproto,
+    "OpenOrdersEnd.proto",
+    &descriptor_table_OpenOrdersEnd_2eproto_once,
+    nullptr,
+    0,
+    1,
+    schemas,
+    file_default_instances,
+    TableStruct_OpenOrdersEnd_2eproto::offsets,
+    file_level_enum_descriptors_OpenOrdersEnd_2eproto,
+    file_level_service_descriptors_OpenOrdersEnd_2eproto,
+};
 namespace protobuf {
-
 // ===================================================================
 
-void OpenOrdersEnd::InitAsDefaultInstance() {
-}
 class OpenOrdersEnd::_Internal {
  public:
 };
 
-OpenOrdersEnd::OpenOrdersEnd(::PROTOBUF_NAMESPACE_ID::Arena* arena)
-  : ::PROTOBUF_NAMESPACE_ID::Message(arena) {
-  SharedCtor();
-  RegisterArenaDtor(arena);
+OpenOrdersEnd::OpenOrdersEnd(::google::protobuf::Arena* arena)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::internal::ZeroFieldsBase(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::internal::ZeroFieldsBase(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
   // @@protoc_insertion_point(arena_constructor:protobuf.OpenOrdersEnd)
 }
-OpenOrdersEnd::OpenOrdersEnd(const OpenOrdersEnd& from)
-  : ::PROTOBUF_NAMESPACE_ID::Message() {
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
+OpenOrdersEnd::OpenOrdersEnd(
+    ::google::protobuf::Arena* arena,
+    const OpenOrdersEnd& from)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::internal::ZeroFieldsBase(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::internal::ZeroFieldsBase(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  OpenOrdersEnd* const _this = this;
+  (void)_this;
+  _internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(
+      from._internal_metadata_);
+
   // @@protoc_insertion_point(copy_constructor:protobuf.OpenOrdersEnd)
 }
 
-void OpenOrdersEnd::SharedCtor() {
+inline void* OpenOrdersEnd::PlacementNew_(const void*, void* mem,
+                                        ::google::protobuf::Arena* arena) {
+  return ::new (mem) OpenOrdersEnd(arena);
 }
-
-OpenOrdersEnd::~OpenOrdersEnd() {
-  // @@protoc_insertion_point(destructor:protobuf.OpenOrdersEnd)
-  SharedDtor();
-  _internal_metadata_.Delete<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+constexpr auto OpenOrdersEnd::InternalNewImpl_() {
+  return ::google::protobuf::internal::MessageCreator::ZeroInit(sizeof(OpenOrdersEnd),
+                                            alignof(OpenOrdersEnd));
 }
-
-void OpenOrdersEnd::SharedDtor() {
-  GOOGLE_DCHECK(GetArena() == nullptr);
+PROTOBUF_CONSTINIT
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::google::protobuf::internal::ClassDataFull OpenOrdersEnd::_class_data_ = {
+    ::google::protobuf::internal::ClassData{
+        &_OpenOrdersEnd_default_instance_._instance,
+        &_table_.header,
+        nullptr,  // OnDemandRegisterArenaDtor
+        nullptr,  // IsInitialized
+        &OpenOrdersEnd::MergeImpl,
+        ::google::protobuf::internal::ZeroFieldsBase::GetNewImpl<OpenOrdersEnd>(),
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        &OpenOrdersEnd::SharedDtor,
+        ::google::protobuf::internal::ZeroFieldsBase::GetClearImpl<OpenOrdersEnd>(), &OpenOrdersEnd::ByteSizeLong,
+            &OpenOrdersEnd::_InternalSerialize,
+#endif  // PROTOBUF_CUSTOM_VTABLE
+        PROTOBUF_FIELD_OFFSET(OpenOrdersEnd, _impl_._cached_size_),
+        false,
+    },
+    &OpenOrdersEnd::kDescriptorMethods,
+    &descriptor_table_OpenOrdersEnd_2eproto,
+    nullptr,  // tracker
+};
+const ::google::protobuf::internal::ClassData* OpenOrdersEnd::GetClassData() const {
+  ::google::protobuf::internal::PrefetchToLocalCache(&_class_data_);
+  ::google::protobuf::internal::PrefetchToLocalCache(_class_data_.tc_table);
+  return _class_data_.base();
 }
+PROTOBUF_CONSTINIT PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::_pbi::TcParseTable<0, 0, 0, 0, 2> OpenOrdersEnd::_table_ = {
+  {
+    0,  // no _has_bits_
+    0, // no _extensions_
+    0, 0,  // max_field_number, fast_idx_mask
+    offsetof(decltype(_table_), field_lookup_table),
+    4294967295,  // skipmap
+    offsetof(decltype(_table_), field_names),  // no field_entries
+    0,  // num_field_entries
+    0,  // num_aux_entries
+    offsetof(decltype(_table_), field_names),  // no aux_entries
+    _class_data_.base(),
+    nullptr,  // post_loop_handler
+    ::_pbi::TcParser::GenericFallback,  // fallback
+    #ifdef PROTOBUF_PREFETCH_PARSE_TABLE
+    ::_pbi::TcParser::GetTable<::protobuf::OpenOrdersEnd>(),  // to_prefetch
+    #endif  // PROTOBUF_PREFETCH_PARSE_TABLE
+  }, {{
+    {::_pbi::TcParser::MiniParse, {}},
+  }}, {{
+    65535, 65535
+  }},
+  // no field_entries, or aux_entries
+  {{
+  }},
+};
 
-void OpenOrdersEnd::ArenaDtor(void* object) {
-  OpenOrdersEnd* _this = reinterpret_cast< OpenOrdersEnd* >(object);
-  (void)_this;
+
+
+
+
+
+
+
+::google::protobuf::Metadata OpenOrdersEnd::GetMetadata() const {
+  return ::google::protobuf::internal::ZeroFieldsBase::GetMetadataImpl(GetClassData()->full());
 }
-void OpenOrdersEnd::RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena*) {
-}
-void OpenOrdersEnd::SetCachedSize(int size) const {
-  _cached_size_.Set(size);
-}
-const OpenOrdersEnd& OpenOrdersEnd::default_instance() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&::scc_info_OpenOrdersEnd_OpenOrdersEnd_2eproto.base);
-  return *internal_default_instance();
-}
-
-
-void OpenOrdersEnd::Clear() {
-// @@protoc_insertion_point(message_clear_start:protobuf.OpenOrdersEnd)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  // Prevent compiler warnings about cached_has_bits being unused
-  (void) cached_has_bits;
-
-  _internal_metadata_.Clear<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
-}
-
-const char* OpenOrdersEnd::_InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) {
-#define CHK_(x) if (PROTOBUF_PREDICT_FALSE(!(x))) goto failure
-  ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena(); (void)arena;
-  while (!ctx->Done(&ptr)) {
-    ::PROTOBUF_NAMESPACE_ID::uint32 tag;
-    ptr = ::PROTOBUF_NAMESPACE_ID::internal::ReadTag(ptr, &tag);
-    CHK_(ptr);
-        if ((tag & 7) == 4 || tag == 0) {
-          ctx->SetLastTag(tag);
-          goto success;
-        }
-        ptr = UnknownFieldParse(tag,
-            _internal_metadata_.mutable_unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(),
-            ptr, ctx);
-        CHK_(ptr != nullptr);
-        continue;
-  }  // while
-success:
-  return ptr;
-failure:
-  ptr = nullptr;
-  goto success;
-#undef CHK_
-}
-
-::PROTOBUF_NAMESPACE_ID::uint8* OpenOrdersEnd::_InternalSerialize(
-    ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const {
-  // @@protoc_insertion_point(serialize_to_array_start:protobuf.OpenOrdersEnd)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  (void) cached_has_bits;
-
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormat::InternalSerializeUnknownFieldsToArray(
-        _internal_metadata_.unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(::PROTOBUF_NAMESPACE_ID::UnknownFieldSet::default_instance), target, stream);
-  }
-  // @@protoc_insertion_point(serialize_to_array_end:protobuf.OpenOrdersEnd)
-  return target;
-}
-
-size_t OpenOrdersEnd::ByteSizeLong() const {
-// @@protoc_insertion_point(message_byte_size_start:protobuf.OpenOrdersEnd)
-  size_t total_size = 0;
-
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  // Prevent compiler warnings about cached_has_bits being unused
-  (void) cached_has_bits;
-
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    return ::PROTOBUF_NAMESPACE_ID::internal::ComputeUnknownFieldsSize(
-        _internal_metadata_, total_size, &_cached_size_);
-  }
-  int cached_size = ::PROTOBUF_NAMESPACE_ID::internal::ToCachedSize(total_size);
-  SetCachedSize(cached_size);
-  return total_size;
-}
-
-void OpenOrdersEnd::MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_merge_from_start:protobuf.OpenOrdersEnd)
-  GOOGLE_DCHECK_NE(&from, this);
-  const OpenOrdersEnd* source =
-      ::PROTOBUF_NAMESPACE_ID::DynamicCastToGenerated<OpenOrdersEnd>(
-          &from);
-  if (source == nullptr) {
-  // @@protoc_insertion_point(generalized_merge_from_cast_fail:protobuf.OpenOrdersEnd)
-    ::PROTOBUF_NAMESPACE_ID::internal::ReflectionOps::Merge(from, this);
-  } else {
-  // @@protoc_insertion_point(generalized_merge_from_cast_success:protobuf.OpenOrdersEnd)
-    MergeFrom(*source);
-  }
-}
-
-void OpenOrdersEnd::MergeFrom(const OpenOrdersEnd& from) {
-// @@protoc_insertion_point(class_specific_merge_from_start:protobuf.OpenOrdersEnd)
-  GOOGLE_DCHECK_NE(&from, this);
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  (void) cached_has_bits;
-
-}
-
-void OpenOrdersEnd::CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_copy_from_start:protobuf.OpenOrdersEnd)
-  if (&from == this) return;
-  Clear();
-  MergeFrom(from);
-}
-
-void OpenOrdersEnd::CopyFrom(const OpenOrdersEnd& from) {
-// @@protoc_insertion_point(class_specific_copy_from_start:protobuf.OpenOrdersEnd)
-  if (&from == this) return;
-  Clear();
-  MergeFrom(from);
-}
-
-bool OpenOrdersEnd::IsInitialized() const {
-  return true;
-}
-
-void OpenOrdersEnd::InternalSwap(OpenOrdersEnd* other) {
-  using std::swap;
-  _internal_metadata_.Swap<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(&other->_internal_metadata_);
-}
-
-::PROTOBUF_NAMESPACE_ID::Metadata OpenOrdersEnd::GetMetadata() const {
-  return GetMetadataStatic();
-}
-
-
 // @@protoc_insertion_point(namespace_scope)
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> PROTOBUF_NOINLINE ::protobuf::OpenOrdersEnd* Arena::CreateMaybeMessage< ::protobuf::OpenOrdersEnd >(Arena* arena) {
-  return Arena::CreateMessageInternal< ::protobuf::OpenOrdersEnd >(arena);
-}
-PROTOBUF_NAMESPACE_CLOSE
-
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
 // @@protoc_insertion_point(global_scope)
-#include <google/protobuf/port_undef.inc>
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY2 static ::std::false_type
+    _static_init2_ PROTOBUF_UNUSED =
+        (::_pbi::AddDescriptors(&descriptor_table_OpenOrdersEnd_2eproto),
+         ::std::false_type{});
+#include "google/protobuf/port_undef.inc"
