@@ -1,183 +1,218 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: OrderAllocation.proto
+// Protobuf C++ Version: 5.29.3
 
-#ifndef GOOGLE_PROTOBUF_INCLUDED_OrderAllocation_2eproto
-#define GOOGLE_PROTOBUF_INCLUDED_OrderAllocation_2eproto
+#ifndef OrderAllocation_2eproto_2epb_2eh
+#define OrderAllocation_2eproto_2epb_2eh
 
 #include <limits>
 #include <string>
+#include <type_traits>
+#include <utility>
 
-#include <google/protobuf/port_def.inc>
-#if PROTOBUF_VERSION < 3012000
-#error This file was generated by a newer version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please update
-#error your headers.
+#include "google/protobuf/runtime_version.h"
+#if PROTOBUF_VERSION != 5029003
+#error "Protobuf C++ gencode is built with an incompatible version of"
+#error "Protobuf C++ headers/runtime. See"
+#error "https://protobuf.dev/support/cross-version-runtime-guarantee/#cpp"
 #endif
-#if 3012004 < PROTOBUF_MIN_PROTOC_VERSION
-#error This file was generated by an older version of protoc which is
-#error incompatible with your Protocol Buffer headers. Please
-#error regenerate this file with a newer version of protoc.
-#endif
-
-#include <google/protobuf/port_undef.inc>
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/arena.h>
-#include <google/protobuf/arenastring.h>
-#include <google/protobuf/generated_message_table_driven.h>
-#include <google/protobuf/generated_message_util.h>
-#include <google/protobuf/inlined_string_field.h>
-#include <google/protobuf/metadata_lite.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/message.h>
-#include <google/protobuf/repeated_field.h>  // IWYU pragma: export
-#include <google/protobuf/extension_set.h>  // IWYU pragma: export
-#include <google/protobuf/unknown_field_set.h>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/arena.h"
+#include "google/protobuf/arenastring.h"
+#include "google/protobuf/generated_message_tctable_decl.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/metadata_lite.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/message.h"
+#include "google/protobuf/message_lite.h"
+#include "google/protobuf/repeated_field.h"  // IWYU pragma: export
+#include "google/protobuf/extension_set.h"  // IWYU pragma: export
+#include "google/protobuf/unknown_field_set.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
+
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+
 #define PROTOBUF_INTERNAL_EXPORT_OrderAllocation_2eproto
-PROTOBUF_NAMESPACE_OPEN
+
+namespace google {
+namespace protobuf {
 namespace internal {
-class AnyMetadata;
+template <typename T>
+::absl::string_view GetAnyMessageName();
 }  // namespace internal
-PROTOBUF_NAMESPACE_CLOSE
+}  // namespace protobuf
+}  // namespace google
 
 // Internal implementation detail -- do not use these members.
 struct TableStruct_OrderAllocation_2eproto {
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTableField entries[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxillaryParseTableField aux[]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTable schema[1]
-    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::FieldMetadata field_metadata[];
-  static const ::PROTOBUF_NAMESPACE_ID::internal::SerializationTable serialization_table[];
-  static const ::PROTOBUF_NAMESPACE_ID::uint32 offsets[];
+  static const ::uint32_t offsets[];
 };
-extern const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_OrderAllocation_2eproto;
+extern const ::google::protobuf::internal::DescriptorTable
+    descriptor_table_OrderAllocation_2eproto;
 namespace protobuf {
 class OrderAllocation;
-class OrderAllocationDefaultTypeInternal;
+struct OrderAllocationDefaultTypeInternal;
 extern OrderAllocationDefaultTypeInternal _OrderAllocation_default_instance_;
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> ::protobuf::OrderAllocation* Arena::CreateMaybeMessage<::protobuf::OrderAllocation>(Arena*);
-PROTOBUF_NAMESPACE_CLOSE
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
+
 namespace protobuf {
 
 // ===================================================================
 
-class OrderAllocation PROTOBUF_FINAL :
-    public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:protobuf.OrderAllocation) */ {
+
+// -------------------------------------------------------------------
+
+class OrderAllocation final : public ::google::protobuf::Message
+/* @@protoc_insertion_point(class_definition:protobuf.OrderAllocation) */ {
  public:
-  inline OrderAllocation() : OrderAllocation(nullptr) {};
-  virtual ~OrderAllocation();
+  inline OrderAllocation() : OrderAllocation(nullptr) {}
+  ~OrderAllocation() PROTOBUF_FINAL;
 
-  OrderAllocation(const OrderAllocation& from);
-  OrderAllocation(OrderAllocation&& from) noexcept
-    : OrderAllocation() {
-    *this = ::std::move(from);
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+  void operator delete(OrderAllocation* msg, std::destroying_delete_t) {
+    SharedDtor(*msg);
+    ::google::protobuf::internal::SizedDelete(msg, sizeof(OrderAllocation));
   }
+#endif
 
+  template <typename = void>
+  explicit PROTOBUF_CONSTEXPR OrderAllocation(
+      ::google::protobuf::internal::ConstantInitialized);
+
+  inline OrderAllocation(const OrderAllocation& from) : OrderAllocation(nullptr, from) {}
+  inline OrderAllocation(OrderAllocation&& from) noexcept
+      : OrderAllocation(nullptr, std::move(from)) {}
   inline OrderAllocation& operator=(const OrderAllocation& from) {
     CopyFrom(from);
     return *this;
   }
   inline OrderAllocation& operator=(OrderAllocation&& from) noexcept {
-    if (GetArena() == from.GetArena()) {
-      if (this != &from) InternalSwap(&from);
+    if (this == &from) return *this;
+    if (::google::protobuf::internal::CanMoveWithInternalSwap(GetArena(), from.GetArena())) {
+      InternalSwap(&from);
     } else {
       CopyFrom(from);
     }
     return *this;
   }
 
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* descriptor() {
+  inline const ::google::protobuf::UnknownFieldSet& unknown_fields() const
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance);
+  }
+  inline ::google::protobuf::UnknownFieldSet* mutable_unknown_fields()
+      ABSL_ATTRIBUTE_LIFETIME_BOUND {
+    return _internal_metadata_.mutable_unknown_fields<::google::protobuf::UnknownFieldSet>();
+  }
+
+  static const ::google::protobuf::Descriptor* descriptor() {
     return GetDescriptor();
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Descriptor* GetDescriptor() {
-    return GetMetadataStatic().descriptor;
+  static const ::google::protobuf::Descriptor* GetDescriptor() {
+    return default_instance().GetMetadata().descriptor;
   }
-  static const ::PROTOBUF_NAMESPACE_ID::Reflection* GetReflection() {
-    return GetMetadataStatic().reflection;
+  static const ::google::protobuf::Reflection* GetReflection() {
+    return default_instance().GetMetadata().reflection;
   }
-  static const OrderAllocation& default_instance();
-
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
+  static const OrderAllocation& default_instance() {
+    return *internal_default_instance();
+  }
   static inline const OrderAllocation* internal_default_instance() {
     return reinterpret_cast<const OrderAllocation*>(
-               &_OrderAllocation_default_instance_);
+        &_OrderAllocation_default_instance_);
   }
-  static constexpr int kIndexInFileMessages =
-    0;
-
-  friend void swap(OrderAllocation& a, OrderAllocation& b) {
-    a.Swap(&b);
-  }
+  static constexpr int kIndexInFileMessages = 0;
+  friend void swap(OrderAllocation& a, OrderAllocation& b) { a.Swap(&b); }
   inline void Swap(OrderAllocation* other) {
     if (other == this) return;
-    if (GetArena() == other->GetArena()) {
+    if (::google::protobuf::internal::CanUseInternalSwap(GetArena(), other->GetArena())) {
       InternalSwap(other);
     } else {
-      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+      ::google::protobuf::internal::GenericSwap(this, other);
     }
   }
   void UnsafeArenaSwap(OrderAllocation* other) {
     if (other == this) return;
-    GOOGLE_DCHECK(GetArena() == other->GetArena());
+    ABSL_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
   // implements Message ----------------------------------------------
 
-  inline OrderAllocation* New() const final {
-    return CreateMaybeMessage<OrderAllocation>(nullptr);
+  OrderAllocation* New(::google::protobuf::Arena* arena = nullptr) const {
+    return ::google::protobuf::Message::DefaultConstruct<OrderAllocation>(arena);
   }
-
-  OrderAllocation* New(::PROTOBUF_NAMESPACE_ID::Arena* arena) const final {
-    return CreateMaybeMessage<OrderAllocation>(arena);
-  }
-  void CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
-  void MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) final;
+  using ::google::protobuf::Message::CopyFrom;
   void CopyFrom(const OrderAllocation& from);
-  void MergeFrom(const OrderAllocation& from);
-  PROTOBUF_ATTRIBUTE_REINITIALIZES void Clear() final;
-  bool IsInitialized() const final;
-
-  size_t ByteSizeLong() const final;
-  const char* _InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) final;
-  ::PROTOBUF_NAMESPACE_ID::uint8* _InternalSerialize(
-      ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const final;
-  int GetCachedSize() const final { return _cached_size_.Get(); }
+  using ::google::protobuf::Message::MergeFrom;
+  void MergeFrom(const OrderAllocation& from) { OrderAllocation::MergeImpl(*this, from); }
 
   private:
-  inline void SharedCtor();
-  inline void SharedDtor();
-  void SetCachedSize(int size) const final;
+  static void MergeImpl(
+      ::google::protobuf::MessageLite& to_msg,
+      const ::google::protobuf::MessageLite& from_msg);
+
+  public:
+  bool IsInitialized() const {
+    return true;
+  }
+  ABSL_ATTRIBUTE_REINITIALIZES void Clear() PROTOBUF_FINAL;
+  #if defined(PROTOBUF_CUSTOM_VTABLE)
+  private:
+  static ::size_t ByteSizeLong(const ::google::protobuf::MessageLite& msg);
+  static ::uint8_t* _InternalSerialize(
+      const MessageLite& msg, ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream);
+
+  public:
+  ::size_t ByteSizeLong() const { return ByteSizeLong(*this); }
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+    return _InternalSerialize(*this, target, stream);
+  }
+  #else   // PROTOBUF_CUSTOM_VTABLE
+  ::size_t ByteSizeLong() const final;
+  ::uint8_t* _InternalSerialize(
+      ::uint8_t* target,
+      ::google::protobuf::io::EpsCopyOutputStream* stream) const final;
+  #endif  // PROTOBUF_CUSTOM_VTABLE
+  int GetCachedSize() const { return _impl_._cached_size_.Get(); }
+
+  private:
+  void SharedCtor(::google::protobuf::Arena* arena);
+  static void SharedDtor(MessageLite& self);
   void InternalSwap(OrderAllocation* other);
-  friend class ::PROTOBUF_NAMESPACE_ID::internal::AnyMetadata;
-  static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
-    return "protobuf.OrderAllocation";
+ private:
+  template <typename T>
+  friend ::absl::string_view(
+      ::google::protobuf::internal::GetAnyMessageName)();
+  static ::absl::string_view FullMessageName() { return "protobuf.OrderAllocation"; }
+
+ protected:
+  explicit OrderAllocation(::google::protobuf::Arena* arena);
+  OrderAllocation(::google::protobuf::Arena* arena, const OrderAllocation& from);
+  OrderAllocation(::google::protobuf::Arena* arena, OrderAllocation&& from) noexcept
+      : OrderAllocation(arena) {
+    *this = ::std::move(from);
   }
-  protected:
-  explicit OrderAllocation(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  private:
-  static void ArenaDtor(void* object);
-  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  public:
+  const ::google::protobuf::internal::ClassData* GetClassData() const PROTOBUF_FINAL;
+  static void* PlacementNew_(const void*, void* mem,
+                             ::google::protobuf::Arena* arena);
+  static constexpr auto InternalNewImpl_();
+  static const ::google::protobuf::internal::ClassDataFull _class_data_;
 
-  ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
-  private:
-  static ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadataStatic() {
-    ::PROTOBUF_NAMESPACE_ID::internal::AssignDescriptors(&::descriptor_table_OrderAllocation_2eproto);
-    return ::descriptor_table_OrderAllocation_2eproto.file_level_metadata[kIndexInFileMessages];
-  }
-
-  public:
-
+ public:
+  ::google::protobuf::Metadata GetMetadata() const;
   // nested types ----------------------------------------------------
 
   // accessors -------------------------------------------------------
-
   enum : int {
     kAccountFieldNumber = 1,
     kPositionFieldNumber = 2,
@@ -187,817 +222,625 @@ class OrderAllocation PROTOBUF_FINAL :
     kAllowedAllocQtyFieldNumber = 6,
     kIsMonetaryFieldNumber = 7,
   };
-  // string account = 1;
+  // optional string account = 1;
   bool has_account() const;
-  private:
-  bool _internal_has_account() const;
-  public:
-  void clear_account();
+  void clear_account() ;
   const std::string& account() const;
-  void set_account(const std::string& value);
-  void set_account(std::string&& value);
-  void set_account(const char* value);
-  void set_account(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_account(Arg_&& arg, Args_... args);
   std::string* mutable_account();
-  std::string* release_account();
-  void set_allocated_account(std::string* account);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_account();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_account(
-      std::string* account);
+  PROTOBUF_NODISCARD std::string* release_account();
+  void set_allocated_account(std::string* value);
+
   private:
   const std::string& _internal_account() const;
-  void _internal_set_account(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_account(
+      const std::string& value);
   std::string* _internal_mutable_account();
-  public:
 
-  // string position = 2;
-  bool has_position() const;
-  private:
-  bool _internal_has_position() const;
   public:
-  void clear_position();
+  // optional string position = 2;
+  bool has_position() const;
+  void clear_position() ;
   const std::string& position() const;
-  void set_position(const std::string& value);
-  void set_position(std::string&& value);
-  void set_position(const char* value);
-  void set_position(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_position(Arg_&& arg, Args_... args);
   std::string* mutable_position();
-  std::string* release_position();
-  void set_allocated_position(std::string* position);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_position();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_position(
-      std::string* position);
+  PROTOBUF_NODISCARD std::string* release_position();
+  void set_allocated_position(std::string* value);
+
   private:
   const std::string& _internal_position() const;
-  void _internal_set_position(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_position(
+      const std::string& value);
   std::string* _internal_mutable_position();
-  public:
 
-  // string positionDesired = 3;
-  bool has_positiondesired() const;
-  private:
-  bool _internal_has_positiondesired() const;
   public:
-  void clear_positiondesired();
+  // optional string positionDesired = 3;
+  bool has_positiondesired() const;
+  void clear_positiondesired() ;
   const std::string& positiondesired() const;
-  void set_positiondesired(const std::string& value);
-  void set_positiondesired(std::string&& value);
-  void set_positiondesired(const char* value);
-  void set_positiondesired(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_positiondesired(Arg_&& arg, Args_... args);
   std::string* mutable_positiondesired();
-  std::string* release_positiondesired();
-  void set_allocated_positiondesired(std::string* positiondesired);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_positiondesired();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_positiondesired(
-      std::string* positiondesired);
+  PROTOBUF_NODISCARD std::string* release_positiondesired();
+  void set_allocated_positiondesired(std::string* value);
+
   private:
   const std::string& _internal_positiondesired() const;
-  void _internal_set_positiondesired(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_positiondesired(
+      const std::string& value);
   std::string* _internal_mutable_positiondesired();
-  public:
 
-  // string positionAfter = 4;
-  bool has_positionafter() const;
-  private:
-  bool _internal_has_positionafter() const;
   public:
-  void clear_positionafter();
+  // optional string positionAfter = 4;
+  bool has_positionafter() const;
+  void clear_positionafter() ;
   const std::string& positionafter() const;
-  void set_positionafter(const std::string& value);
-  void set_positionafter(std::string&& value);
-  void set_positionafter(const char* value);
-  void set_positionafter(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_positionafter(Arg_&& arg, Args_... args);
   std::string* mutable_positionafter();
-  std::string* release_positionafter();
-  void set_allocated_positionafter(std::string* positionafter);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_positionafter();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_positionafter(
-      std::string* positionafter);
+  PROTOBUF_NODISCARD std::string* release_positionafter();
+  void set_allocated_positionafter(std::string* value);
+
   private:
   const std::string& _internal_positionafter() const;
-  void _internal_set_positionafter(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_positionafter(
+      const std::string& value);
   std::string* _internal_mutable_positionafter();
-  public:
 
-  // string desiredAllocQty = 5;
-  bool has_desiredallocqty() const;
-  private:
-  bool _internal_has_desiredallocqty() const;
   public:
-  void clear_desiredallocqty();
+  // optional string desiredAllocQty = 5;
+  bool has_desiredallocqty() const;
+  void clear_desiredallocqty() ;
   const std::string& desiredallocqty() const;
-  void set_desiredallocqty(const std::string& value);
-  void set_desiredallocqty(std::string&& value);
-  void set_desiredallocqty(const char* value);
-  void set_desiredallocqty(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_desiredallocqty(Arg_&& arg, Args_... args);
   std::string* mutable_desiredallocqty();
-  std::string* release_desiredallocqty();
-  void set_allocated_desiredallocqty(std::string* desiredallocqty);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_desiredallocqty();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_desiredallocqty(
-      std::string* desiredallocqty);
+  PROTOBUF_NODISCARD std::string* release_desiredallocqty();
+  void set_allocated_desiredallocqty(std::string* value);
+
   private:
   const std::string& _internal_desiredallocqty() const;
-  void _internal_set_desiredallocqty(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_desiredallocqty(
+      const std::string& value);
   std::string* _internal_mutable_desiredallocqty();
-  public:
 
-  // string allowedAllocQty = 6;
-  bool has_allowedallocqty() const;
-  private:
-  bool _internal_has_allowedallocqty() const;
   public:
-  void clear_allowedallocqty();
+  // optional string allowedAllocQty = 6;
+  bool has_allowedallocqty() const;
+  void clear_allowedallocqty() ;
   const std::string& allowedallocqty() const;
-  void set_allowedallocqty(const std::string& value);
-  void set_allowedallocqty(std::string&& value);
-  void set_allowedallocqty(const char* value);
-  void set_allowedallocqty(const char* value, size_t size);
+  template <typename Arg_ = const std::string&, typename... Args_>
+  void set_allowedallocqty(Arg_&& arg, Args_... args);
   std::string* mutable_allowedallocqty();
-  std::string* release_allowedallocqty();
-  void set_allocated_allowedallocqty(std::string* allowedallocqty);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_allowedallocqty();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_allowedallocqty(
-      std::string* allowedallocqty);
+  PROTOBUF_NODISCARD std::string* release_allowedallocqty();
+  void set_allocated_allowedallocqty(std::string* value);
+
   private:
   const std::string& _internal_allowedallocqty() const;
-  void _internal_set_allowedallocqty(const std::string& value);
+  inline PROTOBUF_ALWAYS_INLINE void _internal_set_allowedallocqty(
+      const std::string& value);
   std::string* _internal_mutable_allowedallocqty();
-  public:
 
-  // bool isMonetary = 7;
-  bool has_ismonetary() const;
-  private:
-  bool _internal_has_ismonetary() const;
   public:
-  void clear_ismonetary();
+  // optional bool isMonetary = 7;
+  bool has_ismonetary() const;
+  void clear_ismonetary() ;
   bool ismonetary() const;
   void set_ismonetary(bool value);
+
   private:
   bool _internal_ismonetary() const;
   void _internal_set_ismonetary(bool value);
-  public:
 
+  public:
   // @@protoc_insertion_point(class_scope:protobuf.OrderAllocation)
  private:
   class _Internal;
+  friend class ::google::protobuf::internal::TcParser;
+  static const ::google::protobuf::internal::TcParseTable<
+      3, 7, 0,
+      106, 2>
+      _table_;
 
-  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
-  typedef void InternalArenaConstructable_;
-  typedef void DestructorSkippable_;
-  ::PROTOBUF_NAMESPACE_ID::internal::HasBits<1> _has_bits_;
-  mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr account_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr position_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr positiondesired_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr positionafter_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr desiredallocqty_;
-  ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr allowedallocqty_;
-  bool ismonetary_;
+  friend class ::google::protobuf::MessageLite;
+  friend class ::google::protobuf::Arena;
+  template <typename T>
+  friend class ::google::protobuf::Arena::InternalHelper;
+  using InternalArenaConstructable_ = void;
+  using DestructorSkippable_ = void;
+  struct Impl_ {
+    inline explicit constexpr Impl_(
+        ::google::protobuf::internal::ConstantInitialized) noexcept;
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena);
+    inline explicit Impl_(::google::protobuf::internal::InternalVisibility visibility,
+                          ::google::protobuf::Arena* arena, const Impl_& from,
+                          const OrderAllocation& from_msg);
+    ::google::protobuf::internal::HasBits<1> _has_bits_;
+    ::google::protobuf::internal::CachedSize _cached_size_;
+    ::google::protobuf::internal::ArenaStringPtr account_;
+    ::google::protobuf::internal::ArenaStringPtr position_;
+    ::google::protobuf::internal::ArenaStringPtr positiondesired_;
+    ::google::protobuf::internal::ArenaStringPtr positionafter_;
+    ::google::protobuf::internal::ArenaStringPtr desiredallocqty_;
+    ::google::protobuf::internal::ArenaStringPtr allowedallocqty_;
+    bool ismonetary_;
+    PROTOBUF_TSAN_DECLARE_MEMBER
+  };
+  union { Impl_ _impl_; };
   friend struct ::TableStruct_OrderAllocation_2eproto;
 };
+
 // ===================================================================
 
 
+
+
 // ===================================================================
+
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic push
-  #pragma GCC diagnostic ignored "-Wstrict-aliasing"
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wstrict-aliasing"
 #endif  // __GNUC__
+// -------------------------------------------------------------------
+
 // OrderAllocation
 
-// string account = 1;
-inline bool OrderAllocation::_internal_has_account() const {
-  bool value = (_has_bits_[0] & 0x00000001u) != 0;
+// optional string account = 1;
+inline bool OrderAllocation::has_account() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000001u) != 0;
   return value;
 }
-inline bool OrderAllocation::has_account() const {
-  return _internal_has_account();
-}
 inline void OrderAllocation::clear_account() {
-  account_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000001u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.account_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000001u;
 }
-inline const std::string& OrderAllocation::account() const {
+inline const std::string& OrderAllocation::account() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.OrderAllocation.account)
   return _internal_account();
 }
-inline void OrderAllocation::set_account(const std::string& value) {
-  _internal_set_account(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void OrderAllocation::set_account(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  _impl_.account_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.OrderAllocation.account)
 }
-inline std::string* OrderAllocation::mutable_account() {
+inline std::string* OrderAllocation::mutable_account() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_account();
   // @@protoc_insertion_point(field_mutable:protobuf.OrderAllocation.account)
-  return _internal_mutable_account();
+  return _s;
 }
 inline const std::string& OrderAllocation::_internal_account() const {
-  return account_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.account_.Get();
 }
 inline void OrderAllocation::_internal_set_account(const std::string& value) {
-  _has_bits_[0] |= 0x00000001u;
-  account_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void OrderAllocation::set_account(std::string&& value) {
-  _has_bits_[0] |= 0x00000001u;
-  account_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.OrderAllocation.account)
-}
-inline void OrderAllocation::set_account(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000001u;
-  account_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.OrderAllocation.account)
-}
-inline void OrderAllocation::set_account(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000001u;
-  account_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.OrderAllocation.account)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  _impl_.account_.Set(value, GetArena());
 }
 inline std::string* OrderAllocation::_internal_mutable_account() {
-  _has_bits_[0] |= 0x00000001u;
-  return account_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000001u;
+  return _impl_.account_.Mutable( GetArena());
 }
 inline std::string* OrderAllocation::release_account() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.OrderAllocation.account)
-  if (!_internal_has_account()) {
+  if ((_impl_._has_bits_[0] & 0x00000001u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000001u;
-  return account_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void OrderAllocation::set_allocated_account(std::string* account) {
-  if (account != nullptr) {
-    _has_bits_[0] |= 0x00000001u;
-  } else {
-    _has_bits_[0] &= ~0x00000001u;
+  _impl_._has_bits_[0] &= ~0x00000001u;
+  auto* released = _impl_.account_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.account_.Set("", GetArena());
   }
-  account_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), account,
-      GetArena());
+  return released;
+}
+inline void OrderAllocation::set_allocated_account(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000001u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000001u;
+  }
+  _impl_.account_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.account_.IsDefault()) {
+    _impl_.account_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.OrderAllocation.account)
 }
-inline std::string* OrderAllocation::unsafe_arena_release_account() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.OrderAllocation.account)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000001u;
-  return account_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void OrderAllocation::unsafe_arena_set_allocated_account(
-    std::string* account) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (account != nullptr) {
-    _has_bits_[0] |= 0x00000001u;
-  } else {
-    _has_bits_[0] &= ~0x00000001u;
-  }
-  account_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      account, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.OrderAllocation.account)
-}
 
-// string position = 2;
-inline bool OrderAllocation::_internal_has_position() const {
-  bool value = (_has_bits_[0] & 0x00000002u) != 0;
+// optional string position = 2;
+inline bool OrderAllocation::has_position() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000002u) != 0;
   return value;
 }
-inline bool OrderAllocation::has_position() const {
-  return _internal_has_position();
-}
 inline void OrderAllocation::clear_position() {
-  position_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000002u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.position_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000002u;
 }
-inline const std::string& OrderAllocation::position() const {
+inline const std::string& OrderAllocation::position() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.OrderAllocation.position)
   return _internal_position();
 }
-inline void OrderAllocation::set_position(const std::string& value) {
-  _internal_set_position(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void OrderAllocation::set_position(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  _impl_.position_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.OrderAllocation.position)
 }
-inline std::string* OrderAllocation::mutable_position() {
+inline std::string* OrderAllocation::mutable_position() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_position();
   // @@protoc_insertion_point(field_mutable:protobuf.OrderAllocation.position)
-  return _internal_mutable_position();
+  return _s;
 }
 inline const std::string& OrderAllocation::_internal_position() const {
-  return position_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.position_.Get();
 }
 inline void OrderAllocation::_internal_set_position(const std::string& value) {
-  _has_bits_[0] |= 0x00000002u;
-  position_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void OrderAllocation::set_position(std::string&& value) {
-  _has_bits_[0] |= 0x00000002u;
-  position_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.OrderAllocation.position)
-}
-inline void OrderAllocation::set_position(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000002u;
-  position_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.OrderAllocation.position)
-}
-inline void OrderAllocation::set_position(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000002u;
-  position_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.OrderAllocation.position)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  _impl_.position_.Set(value, GetArena());
 }
 inline std::string* OrderAllocation::_internal_mutable_position() {
-  _has_bits_[0] |= 0x00000002u;
-  return position_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000002u;
+  return _impl_.position_.Mutable( GetArena());
 }
 inline std::string* OrderAllocation::release_position() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.OrderAllocation.position)
-  if (!_internal_has_position()) {
+  if ((_impl_._has_bits_[0] & 0x00000002u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000002u;
-  return position_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void OrderAllocation::set_allocated_position(std::string* position) {
-  if (position != nullptr) {
-    _has_bits_[0] |= 0x00000002u;
-  } else {
-    _has_bits_[0] &= ~0x00000002u;
+  _impl_._has_bits_[0] &= ~0x00000002u;
+  auto* released = _impl_.position_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.position_.Set("", GetArena());
   }
-  position_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), position,
-      GetArena());
+  return released;
+}
+inline void OrderAllocation::set_allocated_position(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000002u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000002u;
+  }
+  _impl_.position_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.position_.IsDefault()) {
+    _impl_.position_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.OrderAllocation.position)
 }
-inline std::string* OrderAllocation::unsafe_arena_release_position() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.OrderAllocation.position)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000002u;
-  return position_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void OrderAllocation::unsafe_arena_set_allocated_position(
-    std::string* position) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (position != nullptr) {
-    _has_bits_[0] |= 0x00000002u;
-  } else {
-    _has_bits_[0] &= ~0x00000002u;
-  }
-  position_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      position, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.OrderAllocation.position)
-}
 
-// string positionDesired = 3;
-inline bool OrderAllocation::_internal_has_positiondesired() const {
-  bool value = (_has_bits_[0] & 0x00000004u) != 0;
+// optional string positionDesired = 3;
+inline bool OrderAllocation::has_positiondesired() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000004u) != 0;
   return value;
 }
-inline bool OrderAllocation::has_positiondesired() const {
-  return _internal_has_positiondesired();
-}
 inline void OrderAllocation::clear_positiondesired() {
-  positiondesired_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000004u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.positiondesired_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000004u;
 }
-inline const std::string& OrderAllocation::positiondesired() const {
+inline const std::string& OrderAllocation::positiondesired() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.OrderAllocation.positionDesired)
   return _internal_positiondesired();
 }
-inline void OrderAllocation::set_positiondesired(const std::string& value) {
-  _internal_set_positiondesired(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void OrderAllocation::set_positiondesired(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000004u;
+  _impl_.positiondesired_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.OrderAllocation.positionDesired)
 }
-inline std::string* OrderAllocation::mutable_positiondesired() {
+inline std::string* OrderAllocation::mutable_positiondesired() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_positiondesired();
   // @@protoc_insertion_point(field_mutable:protobuf.OrderAllocation.positionDesired)
-  return _internal_mutable_positiondesired();
+  return _s;
 }
 inline const std::string& OrderAllocation::_internal_positiondesired() const {
-  return positiondesired_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.positiondesired_.Get();
 }
 inline void OrderAllocation::_internal_set_positiondesired(const std::string& value) {
-  _has_bits_[0] |= 0x00000004u;
-  positiondesired_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void OrderAllocation::set_positiondesired(std::string&& value) {
-  _has_bits_[0] |= 0x00000004u;
-  positiondesired_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.OrderAllocation.positionDesired)
-}
-inline void OrderAllocation::set_positiondesired(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000004u;
-  positiondesired_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.OrderAllocation.positionDesired)
-}
-inline void OrderAllocation::set_positiondesired(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000004u;
-  positiondesired_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.OrderAllocation.positionDesired)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000004u;
+  _impl_.positiondesired_.Set(value, GetArena());
 }
 inline std::string* OrderAllocation::_internal_mutable_positiondesired() {
-  _has_bits_[0] |= 0x00000004u;
-  return positiondesired_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000004u;
+  return _impl_.positiondesired_.Mutable( GetArena());
 }
 inline std::string* OrderAllocation::release_positiondesired() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.OrderAllocation.positionDesired)
-  if (!_internal_has_positiondesired()) {
+  if ((_impl_._has_bits_[0] & 0x00000004u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000004u;
-  return positiondesired_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void OrderAllocation::set_allocated_positiondesired(std::string* positiondesired) {
-  if (positiondesired != nullptr) {
-    _has_bits_[0] |= 0x00000004u;
-  } else {
-    _has_bits_[0] &= ~0x00000004u;
+  _impl_._has_bits_[0] &= ~0x00000004u;
+  auto* released = _impl_.positiondesired_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.positiondesired_.Set("", GetArena());
   }
-  positiondesired_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), positiondesired,
-      GetArena());
+  return released;
+}
+inline void OrderAllocation::set_allocated_positiondesired(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000004u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000004u;
+  }
+  _impl_.positiondesired_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.positiondesired_.IsDefault()) {
+    _impl_.positiondesired_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.OrderAllocation.positionDesired)
 }
-inline std::string* OrderAllocation::unsafe_arena_release_positiondesired() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.OrderAllocation.positionDesired)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000004u;
-  return positiondesired_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void OrderAllocation::unsafe_arena_set_allocated_positiondesired(
-    std::string* positiondesired) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (positiondesired != nullptr) {
-    _has_bits_[0] |= 0x00000004u;
-  } else {
-    _has_bits_[0] &= ~0x00000004u;
-  }
-  positiondesired_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      positiondesired, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.OrderAllocation.positionDesired)
-}
 
-// string positionAfter = 4;
-inline bool OrderAllocation::_internal_has_positionafter() const {
-  bool value = (_has_bits_[0] & 0x00000008u) != 0;
+// optional string positionAfter = 4;
+inline bool OrderAllocation::has_positionafter() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000008u) != 0;
   return value;
 }
-inline bool OrderAllocation::has_positionafter() const {
-  return _internal_has_positionafter();
-}
 inline void OrderAllocation::clear_positionafter() {
-  positionafter_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000008u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.positionafter_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000008u;
 }
-inline const std::string& OrderAllocation::positionafter() const {
+inline const std::string& OrderAllocation::positionafter() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.OrderAllocation.positionAfter)
   return _internal_positionafter();
 }
-inline void OrderAllocation::set_positionafter(const std::string& value) {
-  _internal_set_positionafter(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void OrderAllocation::set_positionafter(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000008u;
+  _impl_.positionafter_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.OrderAllocation.positionAfter)
 }
-inline std::string* OrderAllocation::mutable_positionafter() {
+inline std::string* OrderAllocation::mutable_positionafter() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_positionafter();
   // @@protoc_insertion_point(field_mutable:protobuf.OrderAllocation.positionAfter)
-  return _internal_mutable_positionafter();
+  return _s;
 }
 inline const std::string& OrderAllocation::_internal_positionafter() const {
-  return positionafter_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.positionafter_.Get();
 }
 inline void OrderAllocation::_internal_set_positionafter(const std::string& value) {
-  _has_bits_[0] |= 0x00000008u;
-  positionafter_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void OrderAllocation::set_positionafter(std::string&& value) {
-  _has_bits_[0] |= 0x00000008u;
-  positionafter_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.OrderAllocation.positionAfter)
-}
-inline void OrderAllocation::set_positionafter(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000008u;
-  positionafter_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.OrderAllocation.positionAfter)
-}
-inline void OrderAllocation::set_positionafter(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000008u;
-  positionafter_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.OrderAllocation.positionAfter)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000008u;
+  _impl_.positionafter_.Set(value, GetArena());
 }
 inline std::string* OrderAllocation::_internal_mutable_positionafter() {
-  _has_bits_[0] |= 0x00000008u;
-  return positionafter_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000008u;
+  return _impl_.positionafter_.Mutable( GetArena());
 }
 inline std::string* OrderAllocation::release_positionafter() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.OrderAllocation.positionAfter)
-  if (!_internal_has_positionafter()) {
+  if ((_impl_._has_bits_[0] & 0x00000008u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000008u;
-  return positionafter_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void OrderAllocation::set_allocated_positionafter(std::string* positionafter) {
-  if (positionafter != nullptr) {
-    _has_bits_[0] |= 0x00000008u;
-  } else {
-    _has_bits_[0] &= ~0x00000008u;
+  _impl_._has_bits_[0] &= ~0x00000008u;
+  auto* released = _impl_.positionafter_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.positionafter_.Set("", GetArena());
   }
-  positionafter_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), positionafter,
-      GetArena());
+  return released;
+}
+inline void OrderAllocation::set_allocated_positionafter(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000008u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000008u;
+  }
+  _impl_.positionafter_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.positionafter_.IsDefault()) {
+    _impl_.positionafter_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.OrderAllocation.positionAfter)
 }
-inline std::string* OrderAllocation::unsafe_arena_release_positionafter() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.OrderAllocation.positionAfter)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000008u;
-  return positionafter_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void OrderAllocation::unsafe_arena_set_allocated_positionafter(
-    std::string* positionafter) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (positionafter != nullptr) {
-    _has_bits_[0] |= 0x00000008u;
-  } else {
-    _has_bits_[0] &= ~0x00000008u;
-  }
-  positionafter_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      positionafter, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.OrderAllocation.positionAfter)
-}
 
-// string desiredAllocQty = 5;
-inline bool OrderAllocation::_internal_has_desiredallocqty() const {
-  bool value = (_has_bits_[0] & 0x00000010u) != 0;
+// optional string desiredAllocQty = 5;
+inline bool OrderAllocation::has_desiredallocqty() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000010u) != 0;
   return value;
 }
-inline bool OrderAllocation::has_desiredallocqty() const {
-  return _internal_has_desiredallocqty();
-}
 inline void OrderAllocation::clear_desiredallocqty() {
-  desiredallocqty_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000010u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.desiredallocqty_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000010u;
 }
-inline const std::string& OrderAllocation::desiredallocqty() const {
+inline const std::string& OrderAllocation::desiredallocqty() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.OrderAllocation.desiredAllocQty)
   return _internal_desiredallocqty();
 }
-inline void OrderAllocation::set_desiredallocqty(const std::string& value) {
-  _internal_set_desiredallocqty(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void OrderAllocation::set_desiredallocqty(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000010u;
+  _impl_.desiredallocqty_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.OrderAllocation.desiredAllocQty)
 }
-inline std::string* OrderAllocation::mutable_desiredallocqty() {
+inline std::string* OrderAllocation::mutable_desiredallocqty() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_desiredallocqty();
   // @@protoc_insertion_point(field_mutable:protobuf.OrderAllocation.desiredAllocQty)
-  return _internal_mutable_desiredallocqty();
+  return _s;
 }
 inline const std::string& OrderAllocation::_internal_desiredallocqty() const {
-  return desiredallocqty_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.desiredallocqty_.Get();
 }
 inline void OrderAllocation::_internal_set_desiredallocqty(const std::string& value) {
-  _has_bits_[0] |= 0x00000010u;
-  desiredallocqty_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void OrderAllocation::set_desiredallocqty(std::string&& value) {
-  _has_bits_[0] |= 0x00000010u;
-  desiredallocqty_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.OrderAllocation.desiredAllocQty)
-}
-inline void OrderAllocation::set_desiredallocqty(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000010u;
-  desiredallocqty_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.OrderAllocation.desiredAllocQty)
-}
-inline void OrderAllocation::set_desiredallocqty(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000010u;
-  desiredallocqty_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.OrderAllocation.desiredAllocQty)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000010u;
+  _impl_.desiredallocqty_.Set(value, GetArena());
 }
 inline std::string* OrderAllocation::_internal_mutable_desiredallocqty() {
-  _has_bits_[0] |= 0x00000010u;
-  return desiredallocqty_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000010u;
+  return _impl_.desiredallocqty_.Mutable( GetArena());
 }
 inline std::string* OrderAllocation::release_desiredallocqty() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.OrderAllocation.desiredAllocQty)
-  if (!_internal_has_desiredallocqty()) {
+  if ((_impl_._has_bits_[0] & 0x00000010u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000010u;
-  return desiredallocqty_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void OrderAllocation::set_allocated_desiredallocqty(std::string* desiredallocqty) {
-  if (desiredallocqty != nullptr) {
-    _has_bits_[0] |= 0x00000010u;
-  } else {
-    _has_bits_[0] &= ~0x00000010u;
+  _impl_._has_bits_[0] &= ~0x00000010u;
+  auto* released = _impl_.desiredallocqty_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.desiredallocqty_.Set("", GetArena());
   }
-  desiredallocqty_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), desiredallocqty,
-      GetArena());
+  return released;
+}
+inline void OrderAllocation::set_allocated_desiredallocqty(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000010u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000010u;
+  }
+  _impl_.desiredallocqty_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.desiredallocqty_.IsDefault()) {
+    _impl_.desiredallocqty_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.OrderAllocation.desiredAllocQty)
 }
-inline std::string* OrderAllocation::unsafe_arena_release_desiredallocqty() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.OrderAllocation.desiredAllocQty)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000010u;
-  return desiredallocqty_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void OrderAllocation::unsafe_arena_set_allocated_desiredallocqty(
-    std::string* desiredallocqty) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (desiredallocqty != nullptr) {
-    _has_bits_[0] |= 0x00000010u;
-  } else {
-    _has_bits_[0] &= ~0x00000010u;
-  }
-  desiredallocqty_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      desiredallocqty, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.OrderAllocation.desiredAllocQty)
-}
 
-// string allowedAllocQty = 6;
-inline bool OrderAllocation::_internal_has_allowedallocqty() const {
-  bool value = (_has_bits_[0] & 0x00000020u) != 0;
+// optional string allowedAllocQty = 6;
+inline bool OrderAllocation::has_allowedallocqty() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000020u) != 0;
   return value;
 }
-inline bool OrderAllocation::has_allowedallocqty() const {
-  return _internal_has_allowedallocqty();
-}
 inline void OrderAllocation::clear_allowedallocqty() {
-  allowedallocqty_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  _has_bits_[0] &= ~0x00000020u;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.allowedallocqty_.ClearToEmpty();
+  _impl_._has_bits_[0] &= ~0x00000020u;
 }
-inline const std::string& OrderAllocation::allowedallocqty() const {
+inline const std::string& OrderAllocation::allowedallocqty() const
+    ABSL_ATTRIBUTE_LIFETIME_BOUND {
   // @@protoc_insertion_point(field_get:protobuf.OrderAllocation.allowedAllocQty)
   return _internal_allowedallocqty();
 }
-inline void OrderAllocation::set_allowedallocqty(const std::string& value) {
-  _internal_set_allowedallocqty(value);
+template <typename Arg_, typename... Args_>
+inline PROTOBUF_ALWAYS_INLINE void OrderAllocation::set_allowedallocqty(Arg_&& arg,
+                                                     Args_... args) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000020u;
+  _impl_.allowedallocqty_.Set(static_cast<Arg_&&>(arg), args..., GetArena());
   // @@protoc_insertion_point(field_set:protobuf.OrderAllocation.allowedAllocQty)
 }
-inline std::string* OrderAllocation::mutable_allowedallocqty() {
+inline std::string* OrderAllocation::mutable_allowedallocqty() ABSL_ATTRIBUTE_LIFETIME_BOUND {
+  std::string* _s = _internal_mutable_allowedallocqty();
   // @@protoc_insertion_point(field_mutable:protobuf.OrderAllocation.allowedAllocQty)
-  return _internal_mutable_allowedallocqty();
+  return _s;
 }
 inline const std::string& OrderAllocation::_internal_allowedallocqty() const {
-  return allowedallocqty_.Get();
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.allowedallocqty_.Get();
 }
 inline void OrderAllocation::_internal_set_allowedallocqty(const std::string& value) {
-  _has_bits_[0] |= 0x00000020u;
-  allowedallocqty_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArena());
-}
-inline void OrderAllocation::set_allowedallocqty(std::string&& value) {
-  _has_bits_[0] |= 0x00000020u;
-  allowedallocqty_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArena());
-  // @@protoc_insertion_point(field_set_rvalue:protobuf.OrderAllocation.allowedAllocQty)
-}
-inline void OrderAllocation::set_allowedallocqty(const char* value) {
-  GOOGLE_DCHECK(value != nullptr);
-  _has_bits_[0] |= 0x00000020u;
-  allowedallocqty_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArena());
-  // @@protoc_insertion_point(field_set_char:protobuf.OrderAllocation.allowedAllocQty)
-}
-inline void OrderAllocation::set_allowedallocqty(const char* value,
-    size_t size) {
-  _has_bits_[0] |= 0x00000020u;
-  allowedallocqty_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArena());
-  // @@protoc_insertion_point(field_set_pointer:protobuf.OrderAllocation.allowedAllocQty)
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000020u;
+  _impl_.allowedallocqty_.Set(value, GetArena());
 }
 inline std::string* OrderAllocation::_internal_mutable_allowedallocqty() {
-  _has_bits_[0] |= 0x00000020u;
-  return allowedallocqty_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_._has_bits_[0] |= 0x00000020u;
+  return _impl_.allowedallocqty_.Mutable( GetArena());
 }
 inline std::string* OrderAllocation::release_allowedallocqty() {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
   // @@protoc_insertion_point(field_release:protobuf.OrderAllocation.allowedAllocQty)
-  if (!_internal_has_allowedallocqty()) {
+  if ((_impl_._has_bits_[0] & 0x00000020u) == 0) {
     return nullptr;
   }
-  _has_bits_[0] &= ~0x00000020u;
-  return allowedallocqty_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-}
-inline void OrderAllocation::set_allocated_allowedallocqty(std::string* allowedallocqty) {
-  if (allowedallocqty != nullptr) {
-    _has_bits_[0] |= 0x00000020u;
-  } else {
-    _has_bits_[0] &= ~0x00000020u;
+  _impl_._has_bits_[0] &= ~0x00000020u;
+  auto* released = _impl_.allowedallocqty_.Release();
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString()) {
+    _impl_.allowedallocqty_.Set("", GetArena());
   }
-  allowedallocqty_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), allowedallocqty,
-      GetArena());
+  return released;
+}
+inline void OrderAllocation::set_allocated_allowedallocqty(std::string* value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  if (value != nullptr) {
+    _impl_._has_bits_[0] |= 0x00000020u;
+  } else {
+    _impl_._has_bits_[0] &= ~0x00000020u;
+  }
+  _impl_.allowedallocqty_.SetAllocated(value, GetArena());
+  if (::google::protobuf::internal::DebugHardenForceCopyDefaultString() && _impl_.allowedallocqty_.IsDefault()) {
+    _impl_.allowedallocqty_.Set("", GetArena());
+  }
   // @@protoc_insertion_point(field_set_allocated:protobuf.OrderAllocation.allowedAllocQty)
 }
-inline std::string* OrderAllocation::unsafe_arena_release_allowedallocqty() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:protobuf.OrderAllocation.allowedAllocQty)
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  _has_bits_[0] &= ~0x00000020u;
-  return allowedallocqty_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArena());
-}
-inline void OrderAllocation::unsafe_arena_set_allocated_allowedallocqty(
-    std::string* allowedallocqty) {
-  GOOGLE_DCHECK(GetArena() != nullptr);
-  if (allowedallocqty != nullptr) {
-    _has_bits_[0] |= 0x00000020u;
-  } else {
-    _has_bits_[0] &= ~0x00000020u;
-  }
-  allowedallocqty_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      allowedallocqty, GetArena());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:protobuf.OrderAllocation.allowedAllocQty)
-}
 
-// bool isMonetary = 7;
-inline bool OrderAllocation::_internal_has_ismonetary() const {
-  bool value = (_has_bits_[0] & 0x00000040u) != 0;
+// optional bool isMonetary = 7;
+inline bool OrderAllocation::has_ismonetary() const {
+  bool value = (_impl_._has_bits_[0] & 0x00000040u) != 0;
   return value;
 }
-inline bool OrderAllocation::has_ismonetary() const {
-  return _internal_has_ismonetary();
-}
 inline void OrderAllocation::clear_ismonetary() {
-  ismonetary_ = false;
-  _has_bits_[0] &= ~0x00000040u;
-}
-inline bool OrderAllocation::_internal_ismonetary() const {
-  return ismonetary_;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.ismonetary_ = false;
+  _impl_._has_bits_[0] &= ~0x00000040u;
 }
 inline bool OrderAllocation::ismonetary() const {
   // @@protoc_insertion_point(field_get:protobuf.OrderAllocation.isMonetary)
   return _internal_ismonetary();
 }
-inline void OrderAllocation::_internal_set_ismonetary(bool value) {
-  _has_bits_[0] |= 0x00000040u;
-  ismonetary_ = value;
-}
 inline void OrderAllocation::set_ismonetary(bool value) {
   _internal_set_ismonetary(value);
+  _impl_._has_bits_[0] |= 0x00000040u;
   // @@protoc_insertion_point(field_set:protobuf.OrderAllocation.isMonetary)
+}
+inline bool OrderAllocation::_internal_ismonetary() const {
+  ::google::protobuf::internal::TSanRead(&_impl_);
+  return _impl_.ismonetary_;
+}
+inline void OrderAllocation::_internal_set_ismonetary(bool value) {
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  _impl_.ismonetary_ = value;
 }
 
 #ifdef __GNUC__
-  #pragma GCC diagnostic pop
+#pragma GCC diagnostic pop
 #endif  // __GNUC__
 
 // @@protoc_insertion_point(namespace_scope)
-
 }  // namespace protobuf
+
 
 // @@protoc_insertion_point(global_scope)
 
-#include <google/protobuf/port_undef.inc>
-#endif  // GOOGLE_PROTOBUF_INCLUDED_GOOGLE_PROTOBUF_INCLUDED_OrderAllocation_2eproto
+#include "google/protobuf/port_undef.inc"
+
+#endif  // OrderAllocation_2eproto_2epb_2eh
