@@ -1,367 +1,402 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
+// NO CHECKED-IN PROTOBUF GENCODE
 // source: OrderCancel.proto
+// Protobuf C++ Version: 5.29.3
 
 #include "OrderCancel.pb.h"
 
 #include <algorithm>
-
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/extension_set.h>
-#include <google/protobuf/wire_format_lite.h>
-#include <google/protobuf/descriptor.h>
-#include <google/protobuf/generated_message_reflection.h>
-#include <google/protobuf/reflection_ops.h>
-#include <google/protobuf/wire_format.h>
+#include <type_traits>
+#include "google/protobuf/io/coded_stream.h"
+#include "google/protobuf/generated_message_tctable_impl.h"
+#include "google/protobuf/extension_set.h"
+#include "google/protobuf/generated_message_util.h"
+#include "google/protobuf/wire_format_lite.h"
+#include "google/protobuf/descriptor.h"
+#include "google/protobuf/generated_message_reflection.h"
+#include "google/protobuf/reflection_ops.h"
+#include "google/protobuf/wire_format.h"
 // @@protoc_insertion_point(includes)
-#include <google/protobuf/port_def.inc>
-namespace protobuf {
-class OrderCancelDefaultTypeInternal {
- public:
-  ::PROTOBUF_NAMESPACE_ID::internal::ExplicitlyConstructed<OrderCancel> _instance;
-} _OrderCancel_default_instance_;
-}  // namespace protobuf
-static void InitDefaultsscc_info_OrderCancel_OrderCancel_2eproto() {
-  GOOGLE_PROTOBUF_VERIFY_VERSION;
 
-  {
-    void* ptr = &::protobuf::_OrderCancel_default_instance_;
-    new (ptr) ::protobuf::OrderCancel();
-    ::PROTOBUF_NAMESPACE_ID::internal::OnShutdownDestroyMessage(ptr);
-  }
-  ::protobuf::OrderCancel::InitAsDefaultInstance();
+// Must be included last.
+#include "google/protobuf/port_def.inc"
+PROTOBUF_PRAGMA_INIT_SEG
+namespace _pb = ::google::protobuf;
+namespace _pbi = ::google::protobuf::internal;
+namespace _fl = ::google::protobuf::internal::field_layout;
+namespace protobuf {
+
+inline constexpr OrderCancel::Impl_::Impl_(
+    ::_pbi::ConstantInitialized) noexcept
+      : _cached_size_{0},
+        manualordercanceltime_(
+            &::google::protobuf::internal::fixed_address_empty_string,
+            ::_pbi::ConstantInitialized()),
+        extoperator_(
+            &::google::protobuf::internal::fixed_address_empty_string,
+            ::_pbi::ConstantInitialized()),
+        manualorderindicator_{0} {}
+
+template <typename>
+PROTOBUF_CONSTEXPR OrderCancel::OrderCancel(::_pbi::ConstantInitialized)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(_class_data_.base()),
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(),
+#endif  // PROTOBUF_CUSTOM_VTABLE
+      _impl_(::_pbi::ConstantInitialized()) {
 }
-
-::PROTOBUF_NAMESPACE_ID::internal::SCCInfo<0> scc_info_OrderCancel_OrderCancel_2eproto =
-    {{ATOMIC_VAR_INIT(::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase::kUninitialized), 0, 0, InitDefaultsscc_info_OrderCancel_OrderCancel_2eproto}, {}};
-
-static ::PROTOBUF_NAMESPACE_ID::Metadata file_level_metadata_OrderCancel_2eproto[1];
-static constexpr ::PROTOBUF_NAMESPACE_ID::EnumDescriptor const** file_level_enum_descriptors_OrderCancel_2eproto = nullptr;
-static constexpr ::PROTOBUF_NAMESPACE_ID::ServiceDescriptor const** file_level_service_descriptors_OrderCancel_2eproto = nullptr;
-
-const ::PROTOBUF_NAMESPACE_ID::uint32 TableStruct_OrderCancel_2eproto::offsets[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  PROTOBUF_FIELD_OFFSET(::protobuf::OrderCancel, _has_bits_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::OrderCancel, _internal_metadata_),
-  ~0u,  // no _extensions_
-  ~0u,  // no _oneof_case_
-  ~0u,  // no _weak_field_map_
-  PROTOBUF_FIELD_OFFSET(::protobuf::OrderCancel, manualordercanceltime_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::OrderCancel, extoperator_),
-  PROTOBUF_FIELD_OFFSET(::protobuf::OrderCancel, manualorderindicator_),
-  0,
-  1,
-  2,
-};
-static const ::PROTOBUF_NAMESPACE_ID::internal::MigrationSchema schemas[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
-  { 0, 8, sizeof(::protobuf::OrderCancel)},
+struct OrderCancelDefaultTypeInternal {
+  PROTOBUF_CONSTEXPR OrderCancelDefaultTypeInternal() : _instance(::_pbi::ConstantInitialized{}) {}
+  ~OrderCancelDefaultTypeInternal() {}
+  union {
+    OrderCancel _instance;
+  };
 };
 
-static ::PROTOBUF_NAMESPACE_ID::Message const * const file_default_instances[] = {
-  reinterpret_cast<const ::PROTOBUF_NAMESPACE_ID::Message*>(&::protobuf::_OrderCancel_default_instance_),
+PROTOBUF_ATTRIBUTE_NO_DESTROY PROTOBUF_CONSTINIT
+    PROTOBUF_ATTRIBUTE_INIT_PRIORITY1 OrderCancelDefaultTypeInternal _OrderCancel_default_instance_;
+}  // namespace protobuf
+static constexpr const ::_pb::EnumDescriptor**
+    file_level_enum_descriptors_OrderCancel_2eproto = nullptr;
+static constexpr const ::_pb::ServiceDescriptor**
+    file_level_service_descriptors_OrderCancel_2eproto = nullptr;
+const ::uint32_t
+    TableStruct_OrderCancel_2eproto::offsets[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+        protodesc_cold) = {
+        PROTOBUF_FIELD_OFFSET(::protobuf::OrderCancel, _impl_._has_bits_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::OrderCancel, _internal_metadata_),
+        ~0u,  // no _extensions_
+        ~0u,  // no _oneof_case_
+        ~0u,  // no _weak_field_map_
+        ~0u,  // no _inlined_string_donated_
+        ~0u,  // no _split_
+        ~0u,  // no sizeof(Split)
+        PROTOBUF_FIELD_OFFSET(::protobuf::OrderCancel, _impl_.manualordercanceltime_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::OrderCancel, _impl_.extoperator_),
+        PROTOBUF_FIELD_OFFSET(::protobuf::OrderCancel, _impl_.manualorderindicator_),
+        0,
+        1,
+        2,
 };
 
-const char descriptor_table_protodef_OrderCancel_2eproto[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) =
-  "\n\021OrderCancel.proto\022\010protobuf\"\261\001\n\013OrderC"
-  "ancel\022\"\n\025manualOrderCancelTime\030\001 \001(\tH\000\210\001"
-  "\001\022\030\n\013extOperator\030\002 \001(\tH\001\210\001\001\022!\n\024manualOrd"
-  "erIndicator\030\003 \001(\005H\002\210\001\001B\030\n\026_manualOrderCa"
-  "ncelTimeB\016\n\014_extOperatorB\027\n\025_manualOrder"
-  "IndicatorB;\n\026com.ib.client.protobufB\020Ord"
-  "erCancelProto\252\002\016IBApi.protobufb\006proto3"
-  ;
-static const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable*const descriptor_table_OrderCancel_2eproto_deps[1] = {
+static const ::_pbi::MigrationSchema
+    schemas[] ABSL_ATTRIBUTE_SECTION_VARIABLE(protodesc_cold) = {
+        {0, 11, -1, sizeof(::protobuf::OrderCancel)},
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase*const descriptor_table_OrderCancel_2eproto_sccs[1] = {
-  &scc_info_OrderCancel_OrderCancel_2eproto.base,
+static const ::_pb::Message* const file_default_instances[] = {
+    &::protobuf::_OrderCancel_default_instance_._instance,
 };
-static ::PROTOBUF_NAMESPACE_ID::internal::once_flag descriptor_table_OrderCancel_2eproto_once;
-const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_OrderCancel_2eproto = {
-  false, false, descriptor_table_protodef_OrderCancel_2eproto, "OrderCancel.proto", 278,
-  &descriptor_table_OrderCancel_2eproto_once, descriptor_table_OrderCancel_2eproto_sccs, descriptor_table_OrderCancel_2eproto_deps, 1, 0,
-  schemas, file_default_instances, TableStruct_OrderCancel_2eproto::offsets,
-  file_level_metadata_OrderCancel_2eproto, 1, file_level_enum_descriptors_OrderCancel_2eproto, file_level_service_descriptors_OrderCancel_2eproto,
+const char descriptor_table_protodef_OrderCancel_2eproto[] ABSL_ATTRIBUTE_SECTION_VARIABLE(
+    protodesc_cold) = {
+    "\n\021OrderCancel.proto\022\010protobuf\"\261\001\n\013OrderC"
+    "ancel\022\"\n\025manualOrderCancelTime\030\001 \001(\tH\000\210\001"
+    "\001\022\030\n\013extOperator\030\002 \001(\tH\001\210\001\001\022!\n\024manualOrd"
+    "erIndicator\030\003 \001(\005H\002\210\001\001B\030\n\026_manualOrderCa"
+    "ncelTimeB\016\n\014_extOperatorB\027\n\025_manualOrder"
+    "IndicatorB;\n\026com.ib.client.protobufB\020Ord"
+    "erCancelProto\252\002\016IBApi.protobufb\006proto3"
 };
-
-// Force running AddDescriptors() at dynamic initialization time.
-static bool dynamic_init_dummy_OrderCancel_2eproto = (static_cast<void>(::PROTOBUF_NAMESPACE_ID::internal::AddDescriptors(&descriptor_table_OrderCancel_2eproto)), true);
+static ::absl::once_flag descriptor_table_OrderCancel_2eproto_once;
+PROTOBUF_CONSTINIT const ::_pbi::DescriptorTable descriptor_table_OrderCancel_2eproto = {
+    false,
+    false,
+    278,
+    descriptor_table_protodef_OrderCancel_2eproto,
+    "OrderCancel.proto",
+    &descriptor_table_OrderCancel_2eproto_once,
+    nullptr,
+    0,
+    1,
+    schemas,
+    file_default_instances,
+    TableStruct_OrderCancel_2eproto::offsets,
+    file_level_enum_descriptors_OrderCancel_2eproto,
+    file_level_service_descriptors_OrderCancel_2eproto,
+};
 namespace protobuf {
-
 // ===================================================================
 
-void OrderCancel::InitAsDefaultInstance() {
-}
 class OrderCancel::_Internal {
  public:
-  using HasBits = decltype(std::declval<OrderCancel>()._has_bits_);
-  static void set_has_manualordercanceltime(HasBits* has_bits) {
-    (*has_bits)[0] |= 1u;
-  }
-  static void set_has_extoperator(HasBits* has_bits) {
-    (*has_bits)[0] |= 2u;
-  }
-  static void set_has_manualorderindicator(HasBits* has_bits) {
-    (*has_bits)[0] |= 4u;
-  }
+  using HasBits =
+      decltype(std::declval<OrderCancel>()._impl_._has_bits_);
+  static constexpr ::int32_t kHasBitsOffset =
+      8 * PROTOBUF_FIELD_OFFSET(OrderCancel, _impl_._has_bits_);
 };
 
-OrderCancel::OrderCancel(::PROTOBUF_NAMESPACE_ID::Arena* arena)
-  : ::PROTOBUF_NAMESPACE_ID::Message(arena) {
-  SharedCtor();
-  RegisterArenaDtor(arena);
+OrderCancel::OrderCancel(::google::protobuf::Arena* arena)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  SharedCtor(arena);
   // @@protoc_insertion_point(arena_constructor:protobuf.OrderCancel)
 }
-OrderCancel::OrderCancel(const OrderCancel& from)
-  : ::PROTOBUF_NAMESPACE_ID::Message(),
-      _has_bits_(from._has_bits_) {
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  manualordercanceltime_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  if (from._internal_has_manualordercanceltime()) {
-    manualordercanceltime_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), from._internal_manualordercanceltime(),
-      GetArena());
-  }
-  extoperator_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  if (from._internal_has_extoperator()) {
-    extoperator_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), from._internal_extoperator(),
-      GetArena());
-  }
-  manualorderindicator_ = from.manualorderindicator_;
+inline PROTOBUF_NDEBUG_INLINE OrderCancel::Impl_::Impl_(
+    ::google::protobuf::internal::InternalVisibility visibility, ::google::protobuf::Arena* arena,
+    const Impl_& from, const ::protobuf::OrderCancel& from_msg)
+      : _has_bits_{from._has_bits_},
+        _cached_size_{0},
+        manualordercanceltime_(arena, from.manualordercanceltime_),
+        extoperator_(arena, from.extoperator_) {}
+
+OrderCancel::OrderCancel(
+    ::google::protobuf::Arena* arena,
+    const OrderCancel& from)
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+    : ::google::protobuf::Message(arena, _class_data_.base()) {
+#else   // PROTOBUF_CUSTOM_VTABLE
+    : ::google::protobuf::Message(arena) {
+#endif  // PROTOBUF_CUSTOM_VTABLE
+  OrderCancel* const _this = this;
+  (void)_this;
+  _internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(
+      from._internal_metadata_);
+  new (&_impl_) Impl_(internal_visibility(), arena, from._impl_, from);
+  _impl_.manualorderindicator_ = from._impl_.manualorderindicator_;
+
   // @@protoc_insertion_point(copy_constructor:protobuf.OrderCancel)
 }
+inline PROTOBUF_NDEBUG_INLINE OrderCancel::Impl_::Impl_(
+    ::google::protobuf::internal::InternalVisibility visibility,
+    ::google::protobuf::Arena* arena)
+      : _cached_size_{0},
+        manualordercanceltime_(arena),
+        extoperator_(arena) {}
 
-void OrderCancel::SharedCtor() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&scc_info_OrderCancel_OrderCancel_2eproto.base);
-  manualordercanceltime_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  extoperator_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  manualorderindicator_ = 0;
+inline void OrderCancel::SharedCtor(::_pb::Arena* arena) {
+  new (&_impl_) Impl_(internal_visibility(), arena);
+  _impl_.manualorderindicator_ = {};
 }
-
 OrderCancel::~OrderCancel() {
   // @@protoc_insertion_point(destructor:protobuf.OrderCancel)
-  SharedDtor();
-  _internal_metadata_.Delete<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  SharedDtor(*this);
+}
+inline void OrderCancel::SharedDtor(MessageLite& self) {
+  OrderCancel& this_ = static_cast<OrderCancel&>(self);
+  this_._internal_metadata_.Delete<::google::protobuf::UnknownFieldSet>();
+  ABSL_DCHECK(this_.GetArena() == nullptr);
+  this_._impl_.manualordercanceltime_.Destroy();
+  this_._impl_.extoperator_.Destroy();
+  this_._impl_.~Impl_();
 }
 
-void OrderCancel::SharedDtor() {
-  GOOGLE_DCHECK(GetArena() == nullptr);
-  manualordercanceltime_.DestroyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  extoperator_.DestroyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+inline void* OrderCancel::PlacementNew_(const void*, void* mem,
+                                        ::google::protobuf::Arena* arena) {
+  return ::new (mem) OrderCancel(arena);
 }
+constexpr auto OrderCancel::InternalNewImpl_() {
+  return ::google::protobuf::internal::MessageCreator::CopyInit(sizeof(OrderCancel),
+                                            alignof(OrderCancel));
+}
+PROTOBUF_CONSTINIT
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::google::protobuf::internal::ClassDataFull OrderCancel::_class_data_ = {
+    ::google::protobuf::internal::ClassData{
+        &_OrderCancel_default_instance_._instance,
+        &_table_.header,
+        nullptr,  // OnDemandRegisterArenaDtor
+        nullptr,  // IsInitialized
+        &OrderCancel::MergeImpl,
+        ::google::protobuf::Message::GetNewImpl<OrderCancel>(),
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        &OrderCancel::SharedDtor,
+        ::google::protobuf::Message::GetClearImpl<OrderCancel>(), &OrderCancel::ByteSizeLong,
+            &OrderCancel::_InternalSerialize,
+#endif  // PROTOBUF_CUSTOM_VTABLE
+        PROTOBUF_FIELD_OFFSET(OrderCancel, _impl_._cached_size_),
+        false,
+    },
+    &OrderCancel::kDescriptorMethods,
+    &descriptor_table_OrderCancel_2eproto,
+    nullptr,  // tracker
+};
+const ::google::protobuf::internal::ClassData* OrderCancel::GetClassData() const {
+  ::google::protobuf::internal::PrefetchToLocalCache(&_class_data_);
+  ::google::protobuf::internal::PrefetchToLocalCache(_class_data_.tc_table);
+  return _class_data_.base();
+}
+PROTOBUF_CONSTINIT PROTOBUF_ATTRIBUTE_INIT_PRIORITY1
+const ::_pbi::TcParseTable<2, 3, 0, 61, 2> OrderCancel::_table_ = {
+  {
+    PROTOBUF_FIELD_OFFSET(OrderCancel, _impl_._has_bits_),
+    0, // no _extensions_
+    3, 24,  // max_field_number, fast_idx_mask
+    offsetof(decltype(_table_), field_lookup_table),
+    4294967288,  // skipmap
+    offsetof(decltype(_table_), field_entries),
+    3,  // num_field_entries
+    0,  // num_aux_entries
+    offsetof(decltype(_table_), field_names),  // no aux_entries
+    _class_data_.base(),
+    nullptr,  // post_loop_handler
+    ::_pbi::TcParser::GenericFallback,  // fallback
+    #ifdef PROTOBUF_PREFETCH_PARSE_TABLE
+    ::_pbi::TcParser::GetTable<::protobuf::OrderCancel>(),  // to_prefetch
+    #endif  // PROTOBUF_PREFETCH_PARSE_TABLE
+  }, {{
+    {::_pbi::TcParser::MiniParse, {}},
+    // optional string manualOrderCancelTime = 1;
+    {::_pbi::TcParser::FastUS1,
+     {10, 0, 0, PROTOBUF_FIELD_OFFSET(OrderCancel, _impl_.manualordercanceltime_)}},
+    // optional string extOperator = 2;
+    {::_pbi::TcParser::FastUS1,
+     {18, 1, 0, PROTOBUF_FIELD_OFFSET(OrderCancel, _impl_.extoperator_)}},
+    // optional int32 manualOrderIndicator = 3;
+    {::_pbi::TcParser::SingularVarintNoZag1<::uint32_t, offsetof(OrderCancel, _impl_.manualorderindicator_), 2>(),
+     {24, 2, 0, PROTOBUF_FIELD_OFFSET(OrderCancel, _impl_.manualorderindicator_)}},
+  }}, {{
+    65535, 65535
+  }}, {{
+    // optional string manualOrderCancelTime = 1;
+    {PROTOBUF_FIELD_OFFSET(OrderCancel, _impl_.manualordercanceltime_), _Internal::kHasBitsOffset + 0, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kUtf8String | ::_fl::kRepAString)},
+    // optional string extOperator = 2;
+    {PROTOBUF_FIELD_OFFSET(OrderCancel, _impl_.extoperator_), _Internal::kHasBitsOffset + 1, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kUtf8String | ::_fl::kRepAString)},
+    // optional int32 manualOrderIndicator = 3;
+    {PROTOBUF_FIELD_OFFSET(OrderCancel, _impl_.manualorderindicator_), _Internal::kHasBitsOffset + 2, 0,
+    (0 | ::_fl::kFcOptional | ::_fl::kInt32)},
+  }},
+  // no aux_entries
+  {{
+    "\24\25\13\0\0\0\0\0"
+    "protobuf.OrderCancel"
+    "manualOrderCancelTime"
+    "extOperator"
+  }},
+};
 
-void OrderCancel::ArenaDtor(void* object) {
-  OrderCancel* _this = reinterpret_cast< OrderCancel* >(object);
-  (void)_this;
-}
-void OrderCancel::RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena*) {
-}
-void OrderCancel::SetCachedSize(int size) const {
-  _cached_size_.Set(size);
-}
-const OrderCancel& OrderCancel::default_instance() {
-  ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&::scc_info_OrderCancel_OrderCancel_2eproto.base);
-  return *internal_default_instance();
-}
-
-
-void OrderCancel::Clear() {
+PROTOBUF_NOINLINE void OrderCancel::Clear() {
 // @@protoc_insertion_point(message_clear_start:protobuf.OrderCancel)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
+  ::google::protobuf::internal::TSanWrite(&_impl_);
+  ::uint32_t cached_has_bits = 0;
   // Prevent compiler warnings about cached_has_bits being unused
   (void) cached_has_bits;
 
-  cached_has_bits = _has_bits_[0];
+  cached_has_bits = _impl_._has_bits_[0];
   if (cached_has_bits & 0x00000003u) {
     if (cached_has_bits & 0x00000001u) {
-      manualordercanceltime_.ClearNonDefaultToEmpty();
+      _impl_.manualordercanceltime_.ClearNonDefaultToEmpty();
     }
     if (cached_has_bits & 0x00000002u) {
-      extoperator_.ClearNonDefaultToEmpty();
+      _impl_.extoperator_.ClearNonDefaultToEmpty();
     }
   }
-  manualorderindicator_ = 0;
-  _has_bits_.Clear();
-  _internal_metadata_.Clear<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
+  _impl_.manualorderindicator_ = 0;
+  _impl_._has_bits_.Clear();
+  _internal_metadata_.Clear<::google::protobuf::UnknownFieldSet>();
 }
 
-const char* OrderCancel::_InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::internal::ParseContext* ctx) {
-#define CHK_(x) if (PROTOBUF_PREDICT_FALSE(!(x))) goto failure
-  _Internal::HasBits has_bits{};
-  ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena(); (void)arena;
-  while (!ctx->Done(&ptr)) {
-    ::PROTOBUF_NAMESPACE_ID::uint32 tag;
-    ptr = ::PROTOBUF_NAMESPACE_ID::internal::ReadTag(ptr, &tag);
-    CHK_(ptr);
-    switch (tag >> 3) {
-      // string manualOrderCancelTime = 1;
-      case 1:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 10)) {
-          auto str = _internal_mutable_manualordercanceltime();
-          ptr = ::PROTOBUF_NAMESPACE_ID::internal::InlineGreedyStringParser(str, ptr, ctx);
-          CHK_(::PROTOBUF_NAMESPACE_ID::internal::VerifyUTF8(str, "protobuf.OrderCancel.manualOrderCancelTime"));
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // string extOperator = 2;
-      case 2:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 18)) {
-          auto str = _internal_mutable_extoperator();
-          ptr = ::PROTOBUF_NAMESPACE_ID::internal::InlineGreedyStringParser(str, ptr, ctx);
-          CHK_(::PROTOBUF_NAMESPACE_ID::internal::VerifyUTF8(str, "protobuf.OrderCancel.extOperator"));
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // int32 manualOrderIndicator = 3;
-      case 3:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 24)) {
-          _Internal::set_has_manualorderindicator(&has_bits);
-          manualorderindicator_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      default: {
-      handle_unusual:
-        if ((tag & 7) == 4 || tag == 0) {
-          ctx->SetLastTag(tag);
-          goto success;
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::uint8_t* OrderCancel::_InternalSerialize(
+            const MessageLite& base, ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) {
+          const OrderCancel& this_ = static_cast<const OrderCancel&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::uint8_t* OrderCancel::_InternalSerialize(
+            ::uint8_t* target,
+            ::google::protobuf::io::EpsCopyOutputStream* stream) const {
+          const OrderCancel& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(serialize_to_array_start:protobuf.OrderCancel)
+          ::uint32_t cached_has_bits = 0;
+          (void)cached_has_bits;
+
+          cached_has_bits = this_._impl_._has_bits_[0];
+          // optional string manualOrderCancelTime = 1;
+          if (cached_has_bits & 0x00000001u) {
+            const std::string& _s = this_._internal_manualordercanceltime();
+            ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
+                _s.data(), static_cast<int>(_s.length()), ::google::protobuf::internal::WireFormatLite::SERIALIZE, "protobuf.OrderCancel.manualOrderCancelTime");
+            target = stream->WriteStringMaybeAliased(1, _s, target);
+          }
+
+          // optional string extOperator = 2;
+          if (cached_has_bits & 0x00000002u) {
+            const std::string& _s = this_._internal_extoperator();
+            ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
+                _s.data(), static_cast<int>(_s.length()), ::google::protobuf::internal::WireFormatLite::SERIALIZE, "protobuf.OrderCancel.extOperator");
+            target = stream->WriteStringMaybeAliased(2, _s, target);
+          }
+
+          // optional int32 manualOrderIndicator = 3;
+          if (cached_has_bits & 0x00000004u) {
+            target = ::google::protobuf::internal::WireFormatLite::
+                WriteInt32ToArrayWithField<3>(
+                    stream, this_._internal_manualorderindicator(), target);
+          }
+
+          if (PROTOBUF_PREDICT_FALSE(this_._internal_metadata_.have_unknown_fields())) {
+            target =
+                ::_pbi::WireFormat::InternalSerializeUnknownFieldsToArray(
+                    this_._internal_metadata_.unknown_fields<::google::protobuf::UnknownFieldSet>(::google::protobuf::UnknownFieldSet::default_instance), target, stream);
+          }
+          // @@protoc_insertion_point(serialize_to_array_end:protobuf.OrderCancel)
+          return target;
         }
-        ptr = UnknownFieldParse(tag,
-            _internal_metadata_.mutable_unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(),
-            ptr, ctx);
-        CHK_(ptr != nullptr);
-        continue;
-      }
-    }  // switch
-  }  // while
-success:
-  _has_bits_.Or(has_bits);
-  return ptr;
-failure:
-  ptr = nullptr;
-  goto success;
-#undef CHK_
-}
 
-::PROTOBUF_NAMESPACE_ID::uint8* OrderCancel::_InternalSerialize(
-    ::PROTOBUF_NAMESPACE_ID::uint8* target, ::PROTOBUF_NAMESPACE_ID::io::EpsCopyOutputStream* stream) const {
-  // @@protoc_insertion_point(serialize_to_array_start:protobuf.OrderCancel)
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
+#if defined(PROTOBUF_CUSTOM_VTABLE)
+        ::size_t OrderCancel::ByteSizeLong(const MessageLite& base) {
+          const OrderCancel& this_ = static_cast<const OrderCancel&>(base);
+#else   // PROTOBUF_CUSTOM_VTABLE
+        ::size_t OrderCancel::ByteSizeLong() const {
+          const OrderCancel& this_ = *this;
+#endif  // PROTOBUF_CUSTOM_VTABLE
+          // @@protoc_insertion_point(message_byte_size_start:protobuf.OrderCancel)
+          ::size_t total_size = 0;
+
+          ::uint32_t cached_has_bits = 0;
+          // Prevent compiler warnings about cached_has_bits being unused
+          (void)cached_has_bits;
+
+          ::_pbi::Prefetch5LinesFrom7Lines(&this_);
+          cached_has_bits = this_._impl_._has_bits_[0];
+          if (cached_has_bits & 0x00000007u) {
+            // optional string manualOrderCancelTime = 1;
+            if (cached_has_bits & 0x00000001u) {
+              total_size += 1 + ::google::protobuf::internal::WireFormatLite::StringSize(
+                                              this_._internal_manualordercanceltime());
+            }
+            // optional string extOperator = 2;
+            if (cached_has_bits & 0x00000002u) {
+              total_size += 1 + ::google::protobuf::internal::WireFormatLite::StringSize(
+                                              this_._internal_extoperator());
+            }
+            // optional int32 manualOrderIndicator = 3;
+            if (cached_has_bits & 0x00000004u) {
+              total_size += ::_pbi::WireFormatLite::Int32SizePlusOne(
+                  this_._internal_manualorderindicator());
+            }
+          }
+          return this_.MaybeComputeUnknownFieldsSize(total_size,
+                                                     &this_._impl_._cached_size_);
+        }
+
+void OrderCancel::MergeImpl(::google::protobuf::MessageLite& to_msg, const ::google::protobuf::MessageLite& from_msg) {
+  auto* const _this = static_cast<OrderCancel*>(&to_msg);
+  auto& from = static_cast<const OrderCancel&>(from_msg);
+  // @@protoc_insertion_point(class_specific_merge_from_start:protobuf.OrderCancel)
+  ABSL_DCHECK_NE(&from, _this);
+  ::uint32_t cached_has_bits = 0;
   (void) cached_has_bits;
 
-  // string manualOrderCancelTime = 1;
-  if (_internal_has_manualordercanceltime()) {
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
-      this->_internal_manualordercanceltime().data(), static_cast<int>(this->_internal_manualordercanceltime().length()),
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
-      "protobuf.OrderCancel.manualOrderCancelTime");
-    target = stream->WriteStringMaybeAliased(
-        1, this->_internal_manualordercanceltime(), target);
-  }
-
-  // string extOperator = 2;
-  if (_internal_has_extoperator()) {
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
-      this->_internal_extoperator().data(), static_cast<int>(this->_internal_extoperator().length()),
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
-      "protobuf.OrderCancel.extOperator");
-    target = stream->WriteStringMaybeAliased(
-        2, this->_internal_extoperator(), target);
-  }
-
-  // int32 manualOrderIndicator = 3;
-  if (_internal_has_manualorderindicator()) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteInt32ToArray(3, this->_internal_manualorderindicator(), target);
-  }
-
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormat::InternalSerializeUnknownFieldsToArray(
-        _internal_metadata_.unknown_fields<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(::PROTOBUF_NAMESPACE_ID::UnknownFieldSet::default_instance), target, stream);
-  }
-  // @@protoc_insertion_point(serialize_to_array_end:protobuf.OrderCancel)
-  return target;
-}
-
-size_t OrderCancel::ByteSizeLong() const {
-// @@protoc_insertion_point(message_byte_size_start:protobuf.OrderCancel)
-  size_t total_size = 0;
-
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  // Prevent compiler warnings about cached_has_bits being unused
-  (void) cached_has_bits;
-
-  cached_has_bits = _has_bits_[0];
-  if (cached_has_bits & 0x00000007u) {
-    // string manualOrderCancelTime = 1;
-    if (cached_has_bits & 0x00000001u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::StringSize(
-          this->_internal_manualordercanceltime());
-    }
-
-    // string extOperator = 2;
-    if (cached_has_bits & 0x00000002u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::StringSize(
-          this->_internal_extoperator());
-    }
-
-    // int32 manualOrderIndicator = 3;
-    if (cached_has_bits & 0x00000004u) {
-      total_size += 1 +
-        ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::Int32Size(
-          this->_internal_manualorderindicator());
-    }
-
-  }
-  if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
-    return ::PROTOBUF_NAMESPACE_ID::internal::ComputeUnknownFieldsSize(
-        _internal_metadata_, total_size, &_cached_size_);
-  }
-  int cached_size = ::PROTOBUF_NAMESPACE_ID::internal::ToCachedSize(total_size);
-  SetCachedSize(cached_size);
-  return total_size;
-}
-
-void OrderCancel::MergeFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_merge_from_start:protobuf.OrderCancel)
-  GOOGLE_DCHECK_NE(&from, this);
-  const OrderCancel* source =
-      ::PROTOBUF_NAMESPACE_ID::DynamicCastToGenerated<OrderCancel>(
-          &from);
-  if (source == nullptr) {
-  // @@protoc_insertion_point(generalized_merge_from_cast_fail:protobuf.OrderCancel)
-    ::PROTOBUF_NAMESPACE_ID::internal::ReflectionOps::Merge(from, this);
-  } else {
-  // @@protoc_insertion_point(generalized_merge_from_cast_success:protobuf.OrderCancel)
-    MergeFrom(*source);
-  }
-}
-
-void OrderCancel::MergeFrom(const OrderCancel& from) {
-// @@protoc_insertion_point(class_specific_merge_from_start:protobuf.OrderCancel)
-  GOOGLE_DCHECK_NE(&from, this);
-  _internal_metadata_.MergeFrom<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(from._internal_metadata_);
-  ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
-  (void) cached_has_bits;
-
-  cached_has_bits = from._has_bits_[0];
+  cached_has_bits = from._impl_._has_bits_[0];
   if (cached_has_bits & 0x00000007u) {
     if (cached_has_bits & 0x00000001u) {
-      _internal_set_manualordercanceltime(from._internal_manualordercanceltime());
+      _this->_internal_set_manualordercanceltime(from._internal_manualordercanceltime());
     }
     if (cached_has_bits & 0x00000002u) {
-      _internal_set_extoperator(from._internal_extoperator());
+      _this->_internal_set_extoperator(from._internal_extoperator());
     }
     if (cached_has_bits & 0x00000004u) {
-      manualorderindicator_ = from.manualorderindicator_;
+      _this->_impl_.manualorderindicator_ = from._impl_.manualorderindicator_;
     }
-    _has_bits_[0] |= cached_has_bits;
   }
-}
-
-void OrderCancel::CopyFrom(const ::PROTOBUF_NAMESPACE_ID::Message& from) {
-// @@protoc_insertion_point(generalized_copy_from_start:protobuf.OrderCancel)
-  if (&from == this) return;
-  Clear();
-  MergeFrom(from);
+  _this->_impl_._has_bits_[0] |= cached_has_bits;
+  _this->_internal_metadata_.MergeFrom<::google::protobuf::UnknownFieldSet>(from._internal_metadata_);
 }
 
 void OrderCancel::CopyFrom(const OrderCancel& from) {
@@ -371,31 +406,30 @@ void OrderCancel::CopyFrom(const OrderCancel& from) {
   MergeFrom(from);
 }
 
-bool OrderCancel::IsInitialized() const {
-  return true;
-}
 
-void OrderCancel::InternalSwap(OrderCancel* other) {
+void OrderCancel::InternalSwap(OrderCancel* PROTOBUF_RESTRICT other) {
   using std::swap;
-  _internal_metadata_.Swap<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>(&other->_internal_metadata_);
-  swap(_has_bits_[0], other->_has_bits_[0]);
-  manualordercanceltime_.Swap(&other->manualordercanceltime_, &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  extoperator_.Swap(&other->extoperator_, &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
-  swap(manualorderindicator_, other->manualorderindicator_);
+  auto* arena = GetArena();
+  ABSL_DCHECK_EQ(arena, other->GetArena());
+  _internal_metadata_.InternalSwap(&other->_internal_metadata_);
+  swap(_impl_._has_bits_[0], other->_impl_._has_bits_[0]);
+  ::_pbi::ArenaStringPtr::InternalSwap(&_impl_.manualordercanceltime_, &other->_impl_.manualordercanceltime_, arena);
+  ::_pbi::ArenaStringPtr::InternalSwap(&_impl_.extoperator_, &other->_impl_.extoperator_, arena);
+        swap(_impl_.manualorderindicator_, other->_impl_.manualorderindicator_);
 }
 
-::PROTOBUF_NAMESPACE_ID::Metadata OrderCancel::GetMetadata() const {
-  return GetMetadataStatic();
+::google::protobuf::Metadata OrderCancel::GetMetadata() const {
+  return ::google::protobuf::Message::GetMetadataImpl(GetClassData()->full());
 }
-
-
 // @@protoc_insertion_point(namespace_scope)
 }  // namespace protobuf
-PROTOBUF_NAMESPACE_OPEN
-template<> PROTOBUF_NOINLINE ::protobuf::OrderCancel* Arena::CreateMaybeMessage< ::protobuf::OrderCancel >(Arena* arena) {
-  return Arena::CreateMessageInternal< ::protobuf::OrderCancel >(arena);
-}
-PROTOBUF_NAMESPACE_CLOSE
-
+namespace google {
+namespace protobuf {
+}  // namespace protobuf
+}  // namespace google
 // @@protoc_insertion_point(global_scope)
-#include <google/protobuf/port_undef.inc>
+PROTOBUF_ATTRIBUTE_INIT_PRIORITY2 static ::std::false_type
+    _static_init2_ PROTOBUF_UNUSED =
+        (::_pbi::AddDescriptors(&descriptor_table_OrderCancel_2eproto),
+         ::std::false_type{});
+#include "google/protobuf/port_undef.inc"
